Arithmetic operators for FLCore classes

Overloaded arithmetic operators for FLCore classes

# S4 method for numeric,FLArray
Arith(e1, e2)

# S4 method for FLArray,numeric
Arith(e1, e2)

# S4 method for FLArray,FLArray
Arith(e1, e2)

# S4 method for FLPar,FLPar
Arith(e1, e2)

# S4 method for FLArray,FLPar
Arith(e1, e2)

# S4 method for FLPar,FLArray
Arith(e1, e2)

Details

These methods apply the standard arithmetic operators included in the Arith group ("+", "-", "*", "^", "%%", "%/%", and "/"), so that they return an object of the appropriate class.

When the operation involves objects of two classes (e.g. FLPar and FLQuant), the class is the returned object is that of the more complexs object, in this case FLQuant.

See also

Author

The FLR Team

Examples

flq <- FLQuant(rlnorm(90), dim=c(3,10), units='kg')
flp <- FLPar(a=99)

# FLQuant and numeric
flq * 25
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1        2        3        4        5        6        7        8       
#>     1  42.5319  31.2810  26.7374 126.0505  34.8505  14.8918  14.1749  10.6705
#>     2  15.7280   3.5629 128.3390  19.7844  88.5413  12.9789  18.4162  43.7324
#>     3  32.6496   5.0798  13.1601  78.2297 179.5896  22.4196   8.5348  77.6079
#>      year
#> quant 9        10      
#>     1  19.4437  17.5043
#>     2  22.9333  44.6498
#>     3   6.2182  28.1992
#> 
#> units:  kg 
# Two FLQuant objects
flq + flq
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1        2        3        4        5        6        7        8       
#>     1  3.40255  2.50248  2.13899 10.08404  2.78804  1.19135  1.13400  0.85364
#>     2  1.25824  0.28503 10.26712  1.58276  7.08331  1.03831  1.47330  3.49859
#>     3  2.61197  0.40638  1.05280  6.25838 14.36717  1.79357  0.68279  6.20863
#>      year
#> quant 9        10      
#>     1  1.55550  1.40035
#>     2  1.83466  3.57198
#>     3  0.49745  2.25594
#> 
#> units:  kg 

# FLQuant and FLPar
flq / flp
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1         2         3         4         5         6         7        
#>     1 0.0171846 0.0126388 0.0108030 0.0509295 0.0140810 0.0060169 0.0057272
#>     2 0.0063547 0.0014395 0.0518542 0.0079937 0.0357743 0.0052440 0.0074409
#>     3 0.0131918 0.0020524 0.0053172 0.0316080 0.0725615 0.0090584 0.0034484
#>      year
#> quant 8         9         10       
#>     1 0.0043113 0.0078560 0.0070725
#>     2 0.0176697 0.0092660 0.0180403
#>     3 0.0313567 0.0025124 0.0113936
#> 
#> units:  kg