Class FLArray

A basic 6D array class. No objects of this class are created in FLCore, as it is used only for method inheritance.

Slots

.Data

Internal S4 data representation, of class array.

Validity

Dimensions:

Array must have 6 dimensions

Content:

Array must be of class numeric

See also

Author

The FLR Team