Class FLIndexBiomass

A class for modelling biomass indices.

FLIndexBiomass(object, ...)

# S4 method for FLQuant
FLIndexBiomass(object, plusgroup = dims(object)$max, ...)

# S4 method for missing
FLIndexBiomass(object, ...)

Arguments

object

FLQuant object used for sizing

...

Other objects to be assigned by name to the class slots

Details

The FLIndexBiomass object holds data and parameters related to biomass indices.

Slots

distribution

Statistical distribution of the index values (character).

index

Index values (FLQuant).

index.var

Variance of the index (FLQuant).

catch.n

Catch numbers used to create the index (FLQuant).

catch.wt

Catch weight of the index (FLQuant).

effort

Effort used to create the index (FLQuant).

sel.pattern

Selection pattern for the index (FLQuant).

index.q

Catchability of the index (FLQuant).

name

Name of the stock (character).

desc

General description of the object (character).

range

Range of the object (numeric)

Accessors

All slots in the class have accessor and replacement methods defined that allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot. A numeric vector can also be used when replacing FLQuant slots, and the vector will be used to substitute the values in the slot, but not its other attributes.

Constructor

A construction method exists for this class that can take named arguments for any of its slots. All slots are then created to match the requirements of the class validity. If an unnamed FLQuant object is provided, this is used for sizing but not stored in any slot.

Validity

Dimensions

All FLQuant slots must have iters equal to 1 or 'n'.

Iters

The dimname for iter[1] should be '1'.

Dimnames

The name of the quant dimension must be the same for all FLQuant slots.

See also

computeCatch, dims, iter, plot, propagate, summary, transform, trim, window, FLComp

Examples

idx <- FLIndexBiomass(index=FLQuant(1:10, quant='age'))

data(ple4)
ida <- FLIndexBiomass(index=ssb(ple4),
  catch.n=catch.n(ple4))