Class FLModelSim

A virtual class for statistical simulation models

FLModelSim(object, ...)

# S4 method for missing
FLModelSim(object, ...)

Details

The FLModelSim class provides a virtual class that developers of various statistical models can use to implement classes that allow those models to be tested, fitted and presented.

Slots in this class attempt to map all the usual outputs for a modelling exercise, together with the standard inputs. Input data are stored in slots created by a specified class that is based on FLModelSim. See for example FLSR for a class used for stock-recruitment models.

Various fitting algorithms, similar to those present in the basic R packages, are currently available for FLModelSim, including fmle, nls-FLCore and glm.

Slots

params

Estimated parameter values. FLPar.

distr

character

vcov

array

model

formula

See also