Stock-Recruitment models

A range of stock-recruitment (SR) models commonly used in fisheries science are provided in FLCore.

ricker()

bevholt()

bevholtss3()

segreg()

geomean()

shepherd()

cushing()

rickerSV()

bevholtSV()

shepherdSV()

bevholtAR1()

rickerAR1()

segregAR1()

rickerCa()

Arguments

rho
Autoregression

sigma2
Autoregression

obs
Observed values

hat
estimated values

steepness
Steepness.

vbiomass
Virgin biomass.

spr0
Spawners per recruit at F=0, see spr0.

model
character vector with model name, either 'bevholt' or 'ricker'.

References

Beverton, R.J.H. and Holt, S.J. (1957) On the dynamics of exploited fish populations. MAFF Fish. Invest., Ser: II 19, 533.

Needle, C.L. Recruitment models: diagnosis and prognosis. Reviews in Fish Biology and Fisheries 11: 95-111, 2002.

Ricker, W.E. (1954) Stock and recruitment. J. Fish. Res. Bd Can. 11, 559-623.

Shepherd, J.G. (1982) A versatile new stock-recruitment relationship for fisheries and the construction of sustainable yield curves. J. Cons. Int. Explor. Mer 40, 67-75.

See also

FLSR, FLModel

Examples

# inspect the output of one of the model functions
  bevholt()

#> $logl
#> function (a, b, rec, ssb) 
#> loglAR1(log(rec), log(a * ssb/(b + ssb)))
#> <bytecode: 0x11471398>
#> <environment: 0x939f7d8>
#> 
#> $model
#> rec ~ a * ssb/(b + ssb)
#> <environment: 0x939f7d8>
#> 
#> $initial
#> function (rec, ssb) 
#> {
#>     a <- max(quantile(c(rec), 0.75, na.rm = TRUE))
#>     b <- max(quantile(c(rec)/c(ssb), 0.9, na.rm = TRUE))
#>     return(FLPar(a = a, b = a/b))
#> }
#> <bytecode: 0x11290538>
#> <environment: 0x939f7d8>
#> attr(,"lower")
#> [1] -Inf -Inf
#> attr(,"upper")
#> [1] Inf Inf
#> 

  names(bevholt())

#> [1] "logl"    "model"   "initial"

  bevholt()$logl

#> function(a, b, rec, ssb)
#>       loglAR1(log(rec), log(a*ssb/(b+ssb)))
#> <bytecode: 0x11471398>
#> <environment: 0xdadd130>

# once an FLSR model is in the workspace ...
  data(nsher)

# the three model-definition slots can be modified
# at once by calling 'model<-' with
# (1) a list
  model(nsher) <- bevholt()

# (2) the name of the function returning this list
  model(nsher) <- 'bevholt'

# or (3) the function itself that returns this list
  model(nsher) <- bevholt