Append objects along the year dimension

Method to append objects along the year dimensions, by extending, combining and substituting sections of them.

Usage

# S4 method for FLQuant,FLQuant
append(x, values, after = dims(values)$minyear - 1)

# S4 method for FLStock,FLStock
append(x, values, after = dims(values)$minyear - 1)

Arguments

x

the object to which the values are to be appended to.

values

to be included in the modified object.

after

a year dimname after with the values are to be appended.

Value

An object of the same class as x with values appended.

Details

FLR objects are commonly manipulated along the year dimension, and the append method offers a simple interface for substituting parts of an object with another, or combine them into one, extending them when necessary. The object to be included or added to the first will be placed as defined by the year dimnames, unless the after input argument specifies otherwise.

Attributes like dimnames and units will always be taken from the first argument, unless the necessary chnages to dimnames$year

See also

Author

The FLR Team

Examples

# append(FLQuant, FLQuant)
fq1 <- FLQuant(1, dimnames=list(age=1:3, year=2000:2010))
fq2 <- FLQuant(2, dimnames=list(age=1:3, year=2011:2012))
fq3 <- FLQuant(2, dimnames=list(age=1:3, year=2014:2016))

# Appends by dimnames$year
append(fq1, fq2)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   1 1    1    1    1    1    1    1    1    1    1    1    2    2   
#>   2 1    1    1    1    1    1    1    1    1    1    1    2    2   
#>   3 1    1    1    1    1    1    1    1    1    1    1    2    2   
#> 
#> units:  NA 
# Appends by dimnames$year with gap (2011:2013)
append(fq1, fq3)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012 2013 2014
#>   1  1    1    1    1    1    1    1    1    1    1    1   NA   NA   NA    2  
#>   2  1    1    1    1    1    1    1    1    1    1    1   NA   NA   NA    2  
#>   3  1    1    1    1    1    1    1    1    1    1    1   NA   NA   NA    2  
#>    year
#> age 2015 2016
#>   1  2    2  
#>   2  2    2  
#>   3  2    2  
#> 
#> units:  NA 
# Appends inside x
append(fq1, fq2, after=2009)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011
#>   1 1    1    1    1    1    1    1    1    1    1    2    2   
#>   2 1    1    1    1    1    1    1    1    1    1    2    2   
#>   3 1    1    1    1    1    1    1    1    1    1    2    2   
#> 
#> units:  NA 
# Appends after end of x
append(fq1, fq2, after=2013)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012 2013 2014
#>   1  1    1    1    1    1    1    1    1    1    1    1   NA   NA   NA    2  
#>   2  1    1    1    1    1    1    1    1    1    1    1   NA   NA   NA    2  
#>   3  1    1    1    1    1    1    1    1    1    1    1   NA   NA   NA    2  
#>    year
#> age 2015
#>   1  2  
#>   2  2  
#>   3  2  
#> 
#> units:  NA 

# append(FLStock, FLStock)
data(ple4)
fs1 <- window(ple4, end=2001)
fs2 <- window(ple4, start=2002)
fs3 <- window(ple4, start=2005)

# Appends by dimnames$year
stock.n(append(fs1, fs2))
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957      1958      1959      1960      1961      1962      1963     
#>   1   477074.0  710748.0  874712.0  797702.0  870799.0  615691.0  610017.0
#>   2   268185.0  391105.0  574019.0  696497.0  629331.0  688002.0  494583.0
#>   3   361589.0  204551.0  286216.0  404963.0  485155.0  441333.0  482883.0
#>   4   223134.0  251892.0  135739.0  182546.0  255875.0  308828.0  276136.0
#>   5   147904.0  148258.0  164040.0   87017.8  116368.0  163035.0  195022.0
#>   6    60608.8  103574.0  100146.0  107321.0   55977.1   74519.9  103605.0
#>   7    62627.2   44411.7   73005.1   68029.4   71202.0   36669.7   47899.3
#>   8    49418.3   45231.8   32053.7   52465.8   48326.9   49594.8   24891.5
#>   9    31996.7   35608.8   32755.0   23308.8   38248.3   35143.0   35587.0
#>   10   68585.1   74612.3   79021.3   78690.9   73222.3   82791.3   88328.2
#>     year
#> age  1964      1965      1966      1967      1968      1969      1970     
#>   1  2449900.0  664500.0  579075.0  428110.0  418228.0  666902.0  671454.0
#>   2   502620.0 2057370.0  561599.0  486279.0  351300.0  328477.0  499525.0
#>   3   341643.0  343934.0 1437490.0  402660.0  346712.0  239678.0  214078.0
#>   4   283950.0  189574.0  194255.0  862270.0  249951.0  214175.0  146130.0
#>   5   170312.0  170953.0  113379.0  117592.0  533809.0  158259.0  137189.0
#>   6   121552.0  104036.0  103728.0   69241.0   72675.5  333787.0   99805.0
#>   7    64386.7   73564.1   63413.5   64736.1   43515.6   44971.0  204065.0
#>   8    31671.7   42047.6   48694.7   42848.4   43754.3   28686.1   28864.1
#>   9    17425.6   21773.1   29156.9   34459.9   30486.1   30610.0   19555.5
#>   10   89949.7   74801.9   67158.6   68866.5   75287.6   76834.4   76614.6
#>     year
#> age  1971      1972      1973      1974      1975      1976      1977     
#>   1   433599.0  367450.0 1391430.0 1074920.0  787372.0  674010.0 1033740.0
#>   2   495583.0  322847.0  274502.0 1025520.0  767410.0  534441.0  432591.0
#>   3   324931.0  328032.0  212726.0  175441.0  635424.0  469588.0  324352.0
#>   4   130830.0  198448.0  194134.0  119185.0   95486.7  354654.0  270674.0
#>   5    92885.0   80660.2  115448.0  105314.0   62528.3   51565.8  199395.0
#>   6    86638.5   57736.0   47858.4   64234.3   56842.0   34795.5   29688.2
#>   7    61697.5   54135.3   35370.6   28087.4   36940.6   33525.3   20971.2
#>   8   130295.0   39537.1   34377.0   21998.3   17263.3   22879.5   20930.5
#>   9    19273.9   85654.3   25584.3   21899.6   13875.4   10882.2   14512.6
#>   10   66652.6   57588.9   92986.6   75118.4   60677.9   46683.2   36597.8
#>     year
#> age  1978      1979      1980      1981      1982      1983      1984     
#>   1   879043.0  915553.0 1078660.0  999968.0 1935350.0 1375880.0 1302060.0
#>   2   640577.0  548706.0  603278.0  759629.0  728751.0 1391630.0  948596.0
#>   3   256771.0  368443.0  308663.0  338604.0  431979.0  422976.0  820717.0
#>   4   186122.0  141111.0  188898.0  147121.0  156042.0  206296.0  214367.0
#>   5   153084.0  102391.0   74240.4   94918.5   71554.5   75327.9  100486.0
#>   6   113431.0   83438.8   54395.5   39883.2   51872.3   38991.5   40129.1
#>   7    17442.0   63121.2   45292.9   30397.5   23273.3   31032.9   23210.4
#>   8    12952.3   10501.4   37093.0   26405.0   18034.5   14363.0   19675.9
#>   9    13395.2    8318.7    6668.8   23029.0   16239.1   11328.1    9269.7
#>   10   33351.6   31239.8   26643.8   22198.4   29551.0   29398.7   26037.4
#>     year
#> age  1985      1986      1987      1988      1989      1990      1991     
#>   1  1792220.0 4303680.0 1910200.0 1774940.0 1250510.0 1083810.0  981356.0
#>   2   864871.0 1177750.0 2866780.0 1308470.0 1251000.0  898824.0  785747.0
#>   3   559870.0  503176.0  670566.0 1602170.0  730349.0  716388.0  530410.0
#>   4   429744.0  291704.0  253291.0  320928.0  743636.0  347377.0  357105.0
#>   5   105579.0  212338.0  142492.0  120680.0  150385.0  353208.0  169944.0
#>   6    51597.9   51971.1  101004.0   66591.7   55845.9   68749.9  159986.0
#>   7    22812.3   27066.0   25237.6   47264.9   30819.2   25576.2   30985.4
#>   8    14467.2   13364.5   14671.0   12950.5   23964.9   16156.4   13837.3
#>   9    12765.7    9167.3    8060.4    8304.2    7096.0   13682.2    9708.3
#>   10   22968.3   23691.3   21446.4   18014.1   15057.6   12910.2   16297.5
#>     year
#> age  1992      1993      1994      1995      1996      1997      1998     
#>   1   854841.0  550376.0  566448.0  932162.0  893056.0 2431310.0  778427.0
#>   2   709583.0  614021.0  398693.0  423980.0  722992.0  708162.0 1941390.0
#>   3   465814.0  412376.0  355244.0  239117.0  261727.0  433954.0  398394.0
#>   4   268732.0  230658.0  199926.0  173831.0  115825.0  114808.0  163844.0
#>   5   177919.0  133599.0  113279.0   96739.9   81895.9   51781.9   48217.9
#>   6    77446.0   82752.7   63227.4   53666.9   45201.5   37345.9   22897.6
#>   7    70809.8   33854.3   36415.8   28660.1   24870.8   20578.7   16297.6
#>   8    16300.8   33881.2   15013.9   16696.8   14022.1   12186.0    9697.5
#>   9     8133.4    8552.4   15995.5    7339.9    8872.5    7570.6    6374.5
#>   10   15761.0   13206.8   11093.7   14408.1   12542.4   12623.6   11589.9
#>     year
#> age  1999      2000      2001      2002      2003      2004      2005     
#>   1   683151.0  857525.0  634808.0 1792880.0  557844.0 1235790.0  863893.0
#>   2   617294.0  531403.0  653888.0  481832.0 1362670.0  417612.0  894667.0
#>   3  1091830.0  381345.0  349511.0  413685.0  272903.0  729522.0  238706.0
#>   4   151199.0  510105.0  205661.0  181506.0  185801.0  118238.0  365617.0
#>   5    67219.3   64650.9  232564.0   98489.7   90505.8   97419.1   66113.3
#>   6    20631.8   28114.9   27483.5  106479.0   49763.0   49696.3   57107.9
#>   7     9911.4    9398.3   13484.8   13278.2   51718.8   26043.7   29565.3
#>   8     7927.6    5445.6    5587.2    7752.1    7045.7   28478.9   16720.1
#>   9     5219.9    4778.1    3568.0    3675.4    4964.1    4649.5   20533.5
#>   10   10285.7    9459.5    9463.7    9308.9    9758.3   11423.5   12764.9
#>     year
#> age  2006      2007      2008      2009      2010      2011      2012     
#>   1   875191.0 1379750.0 1135050.0 1088820.0 1444570.0 1608190.0 1278010.0
#>   2   618023.0  651007.0 1067080.0  879336.0  826492.0 1109460.0 1299830.0
#>   3   552652.0  393107.0  420759.0  719298.0  627121.0  613580.0  837139.0
#>   4   136058.0  326369.0  234965.0  266894.0  493740.0  448410.0  436124.0
#>   5   219194.0   87128.7  219954.0  163268.0  188123.0  349991.0  318435.0
#>   6    41505.0  148406.0   62707.2  164048.0  123503.0  142101.0  261180.0
#>   7    37777.6   29245.7  109232.0   48044.5  129629.0   98702.9  112594.0
#>   8    21233.8   28310.9   22299.1   85182.6   38534.8  106302.0   81974.3
#>   9    12941.5   17026.1   23120.2   18460.8   71434.6   32718.3   91209.7
#>   10   27199.2   33875.1   43959.2   58555.1   67452.5  121698.0  135291.0
#>     year
#> age  2013      2014      2015      2016      2017     
#>   1  1455050.0 1640700.0  895620.0 1211320.0 1823000.0
#>   2  1063130.0 1193580.0 1286330.0  689025.0  973599.0
#>   3   980853.0  793858.0  880738.0  948914.0  515997.0
#>   4   582663.0  681616.0  559741.0  626799.0  672732.0
#>   5   309658.0  413410.0  483694.0  398150.0  447979.0
#>   6   234879.0  228655.0  308142.0  362736.0  298156.0
#>   7   204189.0  183635.0  180488.0  244483.0  286575.0
#>   8    93937.9  169885.0  151867.0  148885.0  202450.0
#>   9    70713.1   80901.8  145583.0  130001.0  128177.0
#>   10  198423.0  235842.0  277893.0  372467.0  443492.0
#> 
#> units:  1000 

# Appends by dimnames$year with gap (2011:2013)
stock.n(append(fs1, fs3))
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957      1958      1959      1960      1961      1962      1963     
#>   1   477074.0  710748.0  874712.0  797702.0  870799.0  615691.0  610017.0
#>   2   268185.0  391105.0  574019.0  696497.0  629331.0  688002.0  494583.0
#>   3   361589.0  204551.0  286216.0  404963.0  485155.0  441333.0  482883.0
#>   4   223134.0  251892.0  135739.0  182546.0  255875.0  308828.0  276136.0
#>   5   147904.0  148258.0  164040.0   87017.8  116368.0  163035.0  195022.0
#>   6    60608.8  103574.0  100146.0  107321.0   55977.1   74519.9  103605.0
#>   7    62627.2   44411.7   73005.1   68029.4   71202.0   36669.7   47899.3
#>   8    49418.3   45231.8   32053.7   52465.8   48326.9   49594.8   24891.5
#>   9    31996.7   35608.8   32755.0   23308.8   38248.3   35143.0   35587.0
#>   10   68585.1   74612.3   79021.3   78690.9   73222.3   82791.3   88328.2
#>     year
#> age  1964      1965      1966      1967      1968      1969      1970     
#>   1  2449900.0  664500.0  579075.0  428110.0  418228.0  666902.0  671454.0
#>   2   502620.0 2057370.0  561599.0  486279.0  351300.0  328477.0  499525.0
#>   3   341643.0  343934.0 1437490.0  402660.0  346712.0  239678.0  214078.0
#>   4   283950.0  189574.0  194255.0  862270.0  249951.0  214175.0  146130.0
#>   5   170312.0  170953.0  113379.0  117592.0  533809.0  158259.0  137189.0
#>   6   121552.0  104036.0  103728.0   69241.0   72675.5  333787.0   99805.0
#>   7    64386.7   73564.1   63413.5   64736.1   43515.6   44971.0  204065.0
#>   8    31671.7   42047.6   48694.7   42848.4   43754.3   28686.1   28864.1
#>   9    17425.6   21773.1   29156.9   34459.9   30486.1   30610.0   19555.5
#>   10   89949.7   74801.9   67158.6   68866.5   75287.6   76834.4   76614.6
#>     year
#> age  1971      1972      1973      1974      1975      1976      1977     
#>   1   433599.0  367450.0 1391430.0 1074920.0  787372.0  674010.0 1033740.0
#>   2   495583.0  322847.0  274502.0 1025520.0  767410.0  534441.0  432591.0
#>   3   324931.0  328032.0  212726.0  175441.0  635424.0  469588.0  324352.0
#>   4   130830.0  198448.0  194134.0  119185.0   95486.7  354654.0  270674.0
#>   5    92885.0   80660.2  115448.0  105314.0   62528.3   51565.8  199395.0
#>   6    86638.5   57736.0   47858.4   64234.3   56842.0   34795.5   29688.2
#>   7    61697.5   54135.3   35370.6   28087.4   36940.6   33525.3   20971.2
#>   8   130295.0   39537.1   34377.0   21998.3   17263.3   22879.5   20930.5
#>   9    19273.9   85654.3   25584.3   21899.6   13875.4   10882.2   14512.6
#>   10   66652.6   57588.9   92986.6   75118.4   60677.9   46683.2   36597.8
#>     year
#> age  1978      1979      1980      1981      1982      1983      1984     
#>   1   879043.0  915553.0 1078660.0  999968.0 1935350.0 1375880.0 1302060.0
#>   2   640577.0  548706.0  603278.0  759629.0  728751.0 1391630.0  948596.0
#>   3   256771.0  368443.0  308663.0  338604.0  431979.0  422976.0  820717.0
#>   4   186122.0  141111.0  188898.0  147121.0  156042.0  206296.0  214367.0
#>   5   153084.0  102391.0   74240.4   94918.5   71554.5   75327.9  100486.0
#>   6   113431.0   83438.8   54395.5   39883.2   51872.3   38991.5   40129.1
#>   7    17442.0   63121.2   45292.9   30397.5   23273.3   31032.9   23210.4
#>   8    12952.3   10501.4   37093.0   26405.0   18034.5   14363.0   19675.9
#>   9    13395.2    8318.7    6668.8   23029.0   16239.1   11328.1    9269.7
#>   10   33351.6   31239.8   26643.8   22198.4   29551.0   29398.7   26037.4
#>     year
#> age  1985      1986      1987      1988      1989      1990      1991     
#>   1  1792220.0 4303680.0 1910200.0 1774940.0 1250510.0 1083810.0  981356.0
#>   2   864871.0 1177750.0 2866780.0 1308470.0 1251000.0  898824.0  785747.0
#>   3   559870.0  503176.0  670566.0 1602170.0  730349.0  716388.0  530410.0
#>   4   429744.0  291704.0  253291.0  320928.0  743636.0  347377.0  357105.0
#>   5   105579.0  212338.0  142492.0  120680.0  150385.0  353208.0  169944.0
#>   6    51597.9   51971.1  101004.0   66591.7   55845.9   68749.9  159986.0
#>   7    22812.3   27066.0   25237.6   47264.9   30819.2   25576.2   30985.4
#>   8    14467.2   13364.5   14671.0   12950.5   23964.9   16156.4   13837.3
#>   9    12765.7    9167.3    8060.4    8304.2    7096.0   13682.2    9708.3
#>   10   22968.3   23691.3   21446.4   18014.1   15057.6   12910.2   16297.5
#>     year
#> age  1992      1993      1994      1995      1996      1997      1998     
#>   1   854841.0  550376.0  566448.0  932162.0  893056.0 2431310.0  778427.0
#>   2   709583.0  614021.0  398693.0  423980.0  722992.0  708162.0 1941390.0
#>   3   465814.0  412376.0  355244.0  239117.0  261727.0  433954.0  398394.0
#>   4   268732.0  230658.0  199926.0  173831.0  115825.0  114808.0  163844.0
#>   5   177919.0  133599.0  113279.0   96739.9   81895.9   51781.9   48217.9
#>   6    77446.0   82752.7   63227.4   53666.9   45201.5   37345.9   22897.6
#>   7    70809.8   33854.3   36415.8   28660.1   24870.8   20578.7   16297.6
#>   8    16300.8   33881.2   15013.9   16696.8   14022.1   12186.0    9697.5
#>   9     8133.4    8552.4   15995.5    7339.9    8872.5    7570.6    6374.5
#>   10   15761.0   13206.8   11093.7   14408.1   12542.4   12623.6   11589.9
#>     year
#> age  1999      2000      2001      2002      2003      2004      2005     
#>   1   683151.0  857525.0  634808.0        NA        NA        NA  863893.0
#>   2   617294.0  531403.0  653888.0        NA        NA        NA  894667.0
#>   3  1091830.0  381345.0  349511.0        NA        NA        NA  238706.0
#>   4   151199.0  510105.0  205661.0        NA        NA        NA  365617.0
#>   5    67219.3   64650.9  232564.0        NA        NA        NA   66113.3
#>   6    20631.8   28114.9   27483.5        NA        NA        NA   57107.9
#>   7     9911.4    9398.3   13484.8        NA        NA        NA   29565.3
#>   8     7927.6    5445.6    5587.2        NA        NA        NA   16720.1
#>   9     5219.9    4778.1    3568.0        NA        NA        NA   20533.5
#>   10   10285.7    9459.5    9463.7        NA        NA        NA   12764.9
#>     year
#> age  2006      2007      2008      2009      2010      2011      2012     
#>   1   875191.0 1379750.0 1135050.0 1088820.0 1444570.0 1608190.0 1278010.0
#>   2   618023.0  651007.0 1067080.0  879336.0  826492.0 1109460.0 1299830.0
#>   3   552652.0  393107.0  420759.0  719298.0  627121.0  613580.0  837139.0
#>   4   136058.0  326369.0  234965.0  266894.0  493740.0  448410.0  436124.0
#>   5   219194.0   87128.7  219954.0  163268.0  188123.0  349991.0  318435.0
#>   6    41505.0  148406.0   62707.2  164048.0  123503.0  142101.0  261180.0
#>   7    37777.6   29245.7  109232.0   48044.5  129629.0   98702.9  112594.0
#>   8    21233.8   28310.9   22299.1   85182.6   38534.8  106302.0   81974.3
#>   9    12941.5   17026.1   23120.2   18460.8   71434.6   32718.3   91209.7
#>   10   27199.2   33875.1   43959.2   58555.1   67452.5  121698.0  135291.0
#>     year
#> age  2013      2014      2015      2016      2017     
#>   1  1455050.0 1640700.0  895620.0 1211320.0 1823000.0
#>   2  1063130.0 1193580.0 1286330.0  689025.0  973599.0
#>   3   980853.0  793858.0  880738.0  948914.0  515997.0
#>   4   582663.0  681616.0  559741.0  626799.0  672732.0
#>   5   309658.0  413410.0  483694.0  398150.0  447979.0
#>   6   234879.0  228655.0  308142.0  362736.0  298156.0
#>   7   204189.0  183635.0  180488.0  244483.0  286575.0
#>   8    93937.9  169885.0  151867.0  148885.0  202450.0
#>   9    70713.1   80901.8  145583.0  130001.0  128177.0
#>   10  198423.0  235842.0  277893.0  372467.0  443492.0
#> 
#> units:  1000 

# Appends inside x
stock.n(append(fs1, fs3, after=2000))
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957      1958      1959      1960      1961      1962      1963     
#>   1   477074.0  710748.0  874712.0  797702.0  870799.0  615691.0  610017.0
#>   2   268185.0  391105.0  574019.0  696497.0  629331.0  688002.0  494583.0
#>   3   361589.0  204551.0  286216.0  404963.0  485155.0  441333.0  482883.0
#>   4   223134.0  251892.0  135739.0  182546.0  255875.0  308828.0  276136.0
#>   5   147904.0  148258.0  164040.0   87017.8  116368.0  163035.0  195022.0
#>   6    60608.8  103574.0  100146.0  107321.0   55977.1   74519.9  103605.0
#>   7    62627.2   44411.7   73005.1   68029.4   71202.0   36669.7   47899.3
#>   8    49418.3   45231.8   32053.7   52465.8   48326.9   49594.8   24891.5
#>   9    31996.7   35608.8   32755.0   23308.8   38248.3   35143.0   35587.0
#>   10   68585.1   74612.3   79021.3   78690.9   73222.3   82791.3   88328.2
#>     year
#> age  1964      1965      1966      1967      1968      1969      1970     
#>   1  2449900.0  664500.0  579075.0  428110.0  418228.0  666902.0  671454.0
#>   2   502620.0 2057370.0  561599.0  486279.0  351300.0  328477.0  499525.0
#>   3   341643.0  343934.0 1437490.0  402660.0  346712.0  239678.0  214078.0
#>   4   283950.0  189574.0  194255.0  862270.0  249951.0  214175.0  146130.0
#>   5   170312.0  170953.0  113379.0  117592.0  533809.0  158259.0  137189.0
#>   6   121552.0  104036.0  103728.0   69241.0   72675.5  333787.0   99805.0
#>   7    64386.7   73564.1   63413.5   64736.1   43515.6   44971.0  204065.0
#>   8    31671.7   42047.6   48694.7   42848.4   43754.3   28686.1   28864.1
#>   9    17425.6   21773.1   29156.9   34459.9   30486.1   30610.0   19555.5
#>   10   89949.7   74801.9   67158.6   68866.5   75287.6   76834.4   76614.6
#>     year
#> age  1971      1972      1973      1974      1975      1976      1977     
#>   1   433599.0  367450.0 1391430.0 1074920.0  787372.0  674010.0 1033740.0
#>   2   495583.0  322847.0  274502.0 1025520.0  767410.0  534441.0  432591.0
#>   3   324931.0  328032.0  212726.0  175441.0  635424.0  469588.0  324352.0
#>   4   130830.0  198448.0  194134.0  119185.0   95486.7  354654.0  270674.0
#>   5    92885.0   80660.2  115448.0  105314.0   62528.3   51565.8  199395.0
#>   6    86638.5   57736.0   47858.4   64234.3   56842.0   34795.5   29688.2
#>   7    61697.5   54135.3   35370.6   28087.4   36940.6   33525.3   20971.2
#>   8   130295.0   39537.1   34377.0   21998.3   17263.3   22879.5   20930.5
#>   9    19273.9   85654.3   25584.3   21899.6   13875.4   10882.2   14512.6
#>   10   66652.6   57588.9   92986.6   75118.4   60677.9   46683.2   36597.8
#>     year
#> age  1978      1979      1980      1981      1982      1983      1984     
#>   1   879043.0  915553.0 1078660.0  999968.0 1935350.0 1375880.0 1302060.0
#>   2   640577.0  548706.0  603278.0  759629.0  728751.0 1391630.0  948596.0
#>   3   256771.0  368443.0  308663.0  338604.0  431979.0  422976.0  820717.0
#>   4   186122.0  141111.0  188898.0  147121.0  156042.0  206296.0  214367.0
#>   5   153084.0  102391.0   74240.4   94918.5   71554.5   75327.9  100486.0
#>   6   113431.0   83438.8   54395.5   39883.2   51872.3   38991.5   40129.1
#>   7    17442.0   63121.2   45292.9   30397.5   23273.3   31032.9   23210.4
#>   8    12952.3   10501.4   37093.0   26405.0   18034.5   14363.0   19675.9
#>   9    13395.2    8318.7    6668.8   23029.0   16239.1   11328.1    9269.7
#>   10   33351.6   31239.8   26643.8   22198.4   29551.0   29398.7   26037.4
#>     year
#> age  1985      1986      1987      1988      1989      1990      1991     
#>   1  1792220.0 4303680.0 1910200.0 1774940.0 1250510.0 1083810.0  981356.0
#>   2   864871.0 1177750.0 2866780.0 1308470.0 1251000.0  898824.0  785747.0
#>   3   559870.0  503176.0  670566.0 1602170.0  730349.0  716388.0  530410.0
#>   4   429744.0  291704.0  253291.0  320928.0  743636.0  347377.0  357105.0
#>   5   105579.0  212338.0  142492.0  120680.0  150385.0  353208.0  169944.0
#>   6    51597.9   51971.1  101004.0   66591.7   55845.9   68749.9  159986.0
#>   7    22812.3   27066.0   25237.6   47264.9   30819.2   25576.2   30985.4
#>   8    14467.2   13364.5   14671.0   12950.5   23964.9   16156.4   13837.3
#>   9    12765.7    9167.3    8060.4    8304.2    7096.0   13682.2    9708.3
#>   10   22968.3   23691.3   21446.4   18014.1   15057.6   12910.2   16297.5
#>     year
#> age  1992      1993      1994      1995      1996      1997      1998     
#>   1   854841.0  550376.0  566448.0  932162.0  893056.0 2431310.0  778427.0
#>   2   709583.0  614021.0  398693.0  423980.0  722992.0  708162.0 1941390.0
#>   3   465814.0  412376.0  355244.0  239117.0  261727.0  433954.0  398394.0
#>   4   268732.0  230658.0  199926.0  173831.0  115825.0  114808.0  163844.0
#>   5   177919.0  133599.0  113279.0   96739.9   81895.9   51781.9   48217.9
#>   6    77446.0   82752.7   63227.4   53666.9   45201.5   37345.9   22897.6
#>   7    70809.8   33854.3   36415.8   28660.1   24870.8   20578.7   16297.6
#>   8    16300.8   33881.2   15013.9   16696.8   14022.1   12186.0    9697.5
#>   9     8133.4    8552.4   15995.5    7339.9    8872.5    7570.6    6374.5
#>   10   15761.0   13206.8   11093.7   14408.1   12542.4   12623.6   11589.9
#>     year
#> age  1999      2000      2001      2002      2003      2004      2005     
#>   1   683151.0  857525.0  634808.0        NA        NA        NA  863893.0
#>   2   617294.0  531403.0  653888.0        NA        NA        NA  894667.0
#>   3  1091830.0  381345.0  349511.0        NA        NA        NA  238706.0
#>   4   151199.0  510105.0  205661.0        NA        NA        NA  365617.0
#>   5    67219.3   64650.9  232564.0        NA        NA        NA   66113.3
#>   6    20631.8   28114.9   27483.5        NA        NA        NA   57107.9
#>   7     9911.4    9398.3   13484.8        NA        NA        NA   29565.3
#>   8     7927.6    5445.6    5587.2        NA        NA        NA   16720.1
#>   9     5219.9    4778.1    3568.0        NA        NA        NA   20533.5
#>   10   10285.7    9459.5    9463.7        NA        NA        NA   12764.9
#>     year
#> age  2006      2007      2008      2009      2010      2011      2012     
#>   1   875191.0 1379750.0 1135050.0 1088820.0 1444570.0 1608190.0 1278010.0
#>   2   618023.0  651007.0 1067080.0  879336.0  826492.0 1109460.0 1299830.0
#>   3   552652.0  393107.0  420759.0  719298.0  627121.0  613580.0  837139.0
#>   4   136058.0  326369.0  234965.0  266894.0  493740.0  448410.0  436124.0
#>   5   219194.0   87128.7  219954.0  163268.0  188123.0  349991.0  318435.0
#>   6    41505.0  148406.0   62707.2  164048.0  123503.0  142101.0  261180.0
#>   7    37777.6   29245.7  109232.0   48044.5  129629.0   98702.9  112594.0
#>   8    21233.8   28310.9   22299.1   85182.6   38534.8  106302.0   81974.3
#>   9    12941.5   17026.1   23120.2   18460.8   71434.6   32718.3   91209.7
#>   10   27199.2   33875.1   43959.2   58555.1   67452.5  121698.0  135291.0
#>     year
#> age  2013      2014      2015      2016      2017     
#>   1  1455050.0 1640700.0  895620.0 1211320.0 1823000.0
#>   2  1063130.0 1193580.0 1286330.0  689025.0  973599.0
#>   3   980853.0  793858.0  880738.0  948914.0  515997.0
#>   4   582663.0  681616.0  559741.0  626799.0  672732.0
#>   5   309658.0  413410.0  483694.0  398150.0  447979.0
#>   6   234879.0  228655.0  308142.0  362736.0  298156.0
#>   7   204189.0  183635.0  180488.0  244483.0  286575.0
#>   8    93937.9  169885.0  151867.0  148885.0  202450.0
#>   9    70713.1   80901.8  145583.0  130001.0  128177.0
#>   10  198423.0  235842.0  277893.0  372467.0  443492.0
#> 
#> units:  1000 
# Appends after end of x
stock.n(append(fs1, fs3, after=2005))
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957      1958      1959      1960      1961      1962      1963     
#>   1   477074.0  710748.0  874712.0  797702.0  870799.0  615691.0  610017.0
#>   2   268185.0  391105.0  574019.0  696497.0  629331.0  688002.0  494583.0
#>   3   361589.0  204551.0  286216.0  404963.0  485155.0  441333.0  482883.0
#>   4   223134.0  251892.0  135739.0  182546.0  255875.0  308828.0  276136.0
#>   5   147904.0  148258.0  164040.0   87017.8  116368.0  163035.0  195022.0
#>   6    60608.8  103574.0  100146.0  107321.0   55977.1   74519.9  103605.0
#>   7    62627.2   44411.7   73005.1   68029.4   71202.0   36669.7   47899.3
#>   8    49418.3   45231.8   32053.7   52465.8   48326.9   49594.8   24891.5
#>   9    31996.7   35608.8   32755.0   23308.8   38248.3   35143.0   35587.0
#>   10   68585.1   74612.3   79021.3   78690.9   73222.3   82791.3   88328.2
#>     year
#> age  1964      1965      1966      1967      1968      1969      1970     
#>   1  2449900.0  664500.0  579075.0  428110.0  418228.0  666902.0  671454.0
#>   2   502620.0 2057370.0  561599.0  486279.0  351300.0  328477.0  499525.0
#>   3   341643.0  343934.0 1437490.0  402660.0  346712.0  239678.0  214078.0
#>   4   283950.0  189574.0  194255.0  862270.0  249951.0  214175.0  146130.0
#>   5   170312.0  170953.0  113379.0  117592.0  533809.0  158259.0  137189.0
#>   6   121552.0  104036.0  103728.0   69241.0   72675.5  333787.0   99805.0
#>   7    64386.7   73564.1   63413.5   64736.1   43515.6   44971.0  204065.0
#>   8    31671.7   42047.6   48694.7   42848.4   43754.3   28686.1   28864.1
#>   9    17425.6   21773.1   29156.9   34459.9   30486.1   30610.0   19555.5
#>   10   89949.7   74801.9   67158.6   68866.5   75287.6   76834.4   76614.6
#>     year
#> age  1971      1972      1973      1974      1975      1976      1977     
#>   1   433599.0  367450.0 1391430.0 1074920.0  787372.0  674010.0 1033740.0
#>   2   495583.0  322847.0  274502.0 1025520.0  767410.0  534441.0  432591.0
#>   3   324931.0  328032.0  212726.0  175441.0  635424.0  469588.0  324352.0
#>   4   130830.0  198448.0  194134.0  119185.0   95486.7  354654.0  270674.0
#>   5    92885.0   80660.2  115448.0  105314.0   62528.3   51565.8  199395.0
#>   6    86638.5   57736.0   47858.4   64234.3   56842.0   34795.5   29688.2
#>   7    61697.5   54135.3   35370.6   28087.4   36940.6   33525.3   20971.2
#>   8   130295.0   39537.1   34377.0   21998.3   17263.3   22879.5   20930.5
#>   9    19273.9   85654.3   25584.3   21899.6   13875.4   10882.2   14512.6
#>   10   66652.6   57588.9   92986.6   75118.4   60677.9   46683.2   36597.8
#>     year
#> age  1978      1979      1980      1981      1982      1983      1984     
#>   1   879043.0  915553.0 1078660.0  999968.0 1935350.0 1375880.0 1302060.0
#>   2   640577.0  548706.0  603278.0  759629.0  728751.0 1391630.0  948596.0
#>   3   256771.0  368443.0  308663.0  338604.0  431979.0  422976.0  820717.0
#>   4   186122.0  141111.0  188898.0  147121.0  156042.0  206296.0  214367.0
#>   5   153084.0  102391.0   74240.4   94918.5   71554.5   75327.9  100486.0
#>   6   113431.0   83438.8   54395.5   39883.2   51872.3   38991.5   40129.1
#>   7    17442.0   63121.2   45292.9   30397.5   23273.3   31032.9   23210.4
#>   8    12952.3   10501.4   37093.0   26405.0   18034.5   14363.0   19675.9
#>   9    13395.2    8318.7    6668.8   23029.0   16239.1   11328.1    9269.7
#>   10   33351.6   31239.8   26643.8   22198.4   29551.0   29398.7   26037.4
#>     year
#> age  1985      1986      1987      1988      1989      1990      1991     
#>   1  1792220.0 4303680.0 1910200.0 1774940.0 1250510.0 1083810.0  981356.0
#>   2   864871.0 1177750.0 2866780.0 1308470.0 1251000.0  898824.0  785747.0
#>   3   559870.0  503176.0  670566.0 1602170.0  730349.0  716388.0  530410.0
#>   4   429744.0  291704.0  253291.0  320928.0  743636.0  347377.0  357105.0
#>   5   105579.0  212338.0  142492.0  120680.0  150385.0  353208.0  169944.0
#>   6    51597.9   51971.1  101004.0   66591.7   55845.9   68749.9  159986.0
#>   7    22812.3   27066.0   25237.6   47264.9   30819.2   25576.2   30985.4
#>   8    14467.2   13364.5   14671.0   12950.5   23964.9   16156.4   13837.3
#>   9    12765.7    9167.3    8060.4    8304.2    7096.0   13682.2    9708.3
#>   10   22968.3   23691.3   21446.4   18014.1   15057.6   12910.2   16297.5
#>     year
#> age  1992      1993      1994      1995      1996      1997      1998     
#>   1   854841.0  550376.0  566448.0  932162.0  893056.0 2431310.0  778427.0
#>   2   709583.0  614021.0  398693.0  423980.0  722992.0  708162.0 1941390.0
#>   3   465814.0  412376.0  355244.0  239117.0  261727.0  433954.0  398394.0
#>   4   268732.0  230658.0  199926.0  173831.0  115825.0  114808.0  163844.0
#>   5   177919.0  133599.0  113279.0   96739.9   81895.9   51781.9   48217.9
#>   6    77446.0   82752.7   63227.4   53666.9   45201.5   37345.9   22897.6
#>   7    70809.8   33854.3   36415.8   28660.1   24870.8   20578.7   16297.6
#>   8    16300.8   33881.2   15013.9   16696.8   14022.1   12186.0    9697.5
#>   9     8133.4    8552.4   15995.5    7339.9    8872.5    7570.6    6374.5
#>   10   15761.0   13206.8   11093.7   14408.1   12542.4   12623.6   11589.9
#>     year
#> age  1999      2000      2001      2002      2003      2004      2005     
#>   1   683151.0  857525.0  634808.0        NA        NA        NA  863893.0
#>   2   617294.0  531403.0  653888.0        NA        NA        NA  894667.0
#>   3  1091830.0  381345.0  349511.0        NA        NA        NA  238706.0
#>   4   151199.0  510105.0  205661.0        NA        NA        NA  365617.0
#>   5    67219.3   64650.9  232564.0        NA        NA        NA   66113.3
#>   6    20631.8   28114.9   27483.5        NA        NA        NA   57107.9
#>   7     9911.4    9398.3   13484.8        NA        NA        NA   29565.3
#>   8     7927.6    5445.6    5587.2        NA        NA        NA   16720.1
#>   9     5219.9    4778.1    3568.0        NA        NA        NA   20533.5
#>   10   10285.7    9459.5    9463.7        NA        NA        NA   12764.9
#>     year
#> age  2006      2007      2008      2009      2010      2011      2012     
#>   1   875191.0 1379750.0 1135050.0 1088820.0 1444570.0 1608190.0 1278010.0
#>   2   618023.0  651007.0 1067080.0  879336.0  826492.0 1109460.0 1299830.0
#>   3   552652.0  393107.0  420759.0  719298.0  627121.0  613580.0  837139.0
#>   4   136058.0  326369.0  234965.0  266894.0  493740.0  448410.0  436124.0
#>   5   219194.0   87128.7  219954.0  163268.0  188123.0  349991.0  318435.0
#>   6    41505.0  148406.0   62707.2  164048.0  123503.0  142101.0  261180.0
#>   7    37777.6   29245.7  109232.0   48044.5  129629.0   98702.9  112594.0
#>   8    21233.8   28310.9   22299.1   85182.6   38534.8  106302.0   81974.3
#>   9    12941.5   17026.1   23120.2   18460.8   71434.6   32718.3   91209.7
#>   10   27199.2   33875.1   43959.2   58555.1   67452.5  121698.0  135291.0
#>     year
#> age  2013      2014      2015      2016      2017      2018     
#>   1  1455050.0 1640700.0  895620.0 1211320.0 1823000.0        NA
#>   2  1063130.0 1193580.0 1286330.0  689025.0  973599.0        NA
#>   3   980853.0  793858.0  880738.0  948914.0  515997.0        NA
#>   4   582663.0  681616.0  559741.0  626799.0  672732.0        NA
#>   5   309658.0  413410.0  483694.0  398150.0  447979.0        NA
#>   6   234879.0  228655.0  308142.0  362736.0  298156.0        NA
#>   7   204189.0  183635.0  180488.0  244483.0  286575.0        NA
#>   8    93937.9  169885.0  151867.0  148885.0  202450.0        NA
#>   9    70713.1   80901.8  145583.0  130001.0  128177.0        NA
#>   10  198423.0  235842.0  277893.0  372467.0  443492.0        NA
#> 
#> units:  1000