Computes fishing mortality from abundances, catches and natural mortality

Objects or class 'FLStock' already contain a 'harvest' slot to store estimates of fishing mortality at age, for example those obtained from a stock assessment method. Fishing mortality at age can be recalculated using two methods:

Usage

# S4 method for FLStock,missing
computeHarvest(object, units = NULL)

Arguments

units

Harvest to be computed as 'f' or 'hr', 'character'.

x

An object of class 'FLStock'.

Value

An 'FLQuant' with the calculated fishing mortalities at age.

See also

Author

The FLR Team

Examples

data(ple4)
# Compute 'f' from stock.n and Baranov
computeHarvest(ple4)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957     1958     1959     1960     1961     1962     1963     1964    
#>   1  0.098696 0.113655 0.127831 0.137078 0.135619 0.119030 0.093652 0.074619
#>   2  0.170860 0.212229 0.248867 0.261595 0.254858 0.254017 0.269949 0.279385
#>   3  0.261508 0.310083 0.349745 0.359107 0.351684 0.368906 0.430976 0.488987
#>   4  0.308818 0.328890 0.344621 0.350245 0.350724 0.359672 0.383261 0.407410
#>   5  0.256277 0.292325 0.324286 0.341170 0.345691 0.353379 0.372770 0.392895
#>   6  0.210937 0.249757 0.286689 0.310303 0.322992 0.341965 0.375679 0.402185
#>   7  0.225399 0.226090 0.230368 0.241952 0.261635 0.287424 0.313677 0.326105
#>   8  0.227728 0.222745 0.218582 0.216062 0.218563 0.231906 0.256586 0.274748
#>   9  0.198669 0.232773 0.250974 0.231472 0.197441 0.189074 0.220349 0.261486
#>   10 0.198669 0.232773 0.250974 0.231471 0.197442 0.189073 0.220349 0.261487
#>     year
#> age  1965     1966     1967     1968     1969     1970     1971     1972    
#>   1  0.068247 0.074649 0.097740 0.141560 0.188985 0.203711 0.194942 0.191629
#>   2  0.258530 0.232696 0.238288 0.282344 0.328126 0.330045 0.312624 0.317174
#>   3  0.471278 0.411085 0.376828 0.381701 0.394800 0.392442 0.393086 0.424563
#>   4  0.414043 0.401951 0.379530 0.357032 0.345434 0.353134 0.383654 0.441707
#>   5  0.399617 0.393143 0.381217 0.369535 0.361015 0.359615 0.375481 0.421999
#>   6  0.395060 0.371453 0.364474 0.379986 0.392064 0.380975 0.370258 0.390000
#>   7  0.312587 0.292008 0.291729 0.316707 0.343419 0.348637 0.345004 0.354099
#>   8  0.266111 0.245774 0.240397 0.257263 0.283156 0.303847 0.319482 0.335261
#>   9  0.263264 0.235461 0.216581 0.219652 0.238188 0.266625 0.300163 0.332087
#>   10 0.263264 0.235460 0.216580 0.219653 0.238187 0.266625 0.300162 0.332087
#>     year
#> age  1973     1974     1975     1976     1977     1978     1979     1980    
#>   1  0.205132 0.236980 0.287479 0.343452 0.378569 0.371271 0.317150 0.250645
#>   2  0.347656 0.378663 0.391165 0.399392 0.421608 0.453083 0.475313 0.477547
#>   3  0.479328 0.508316 0.483150 0.450941 0.455427 0.498638 0.568079 0.640995
#>   4  0.511602 0.545058 0.516128 0.475855 0.469928 0.497604 0.542238 0.588189
#>   5  0.486283 0.516671 0.486131 0.452109 0.464093 0.506873 0.532517 0.521353
#>   6  0.432926 0.453226 0.427975 0.406338 0.431869 0.486138 0.510963 0.481921
#>   7  0.374916 0.386738 0.379070 0.371093 0.381877 0.407373 0.431628 0.439597
#>   8  0.350919 0.360848 0.361454 0.355224 0.346311 0.342764 0.354070 0.376674
#>   9  0.356453 0.369329 0.368139 0.352942 0.326880 0.303054 0.295224 0.305917
#>   10 0.356453 0.369329 0.368141 0.352940 0.326881 0.303052 0.295224 0.305917
#>     year
#> age  1981     1982     1983     1984     1985     1986     1987     1988    
#>   1  0.216391 0.229812 0.271866 0.309123 0.319849 0.306281 0.278349 0.249823
#>   2  0.464453 0.444017 0.428053 0.427278 0.441640 0.463239 0.481830 0.483091
#>   3  0.674706 0.639065 0.579626 0.546989 0.551965 0.586401 0.636905 0.667563
#>   4  0.620796 0.628275 0.619294 0.608230 0.605010 0.616454 0.641397 0.658018
#>   5  0.504234 0.507116 0.529749 0.566537 0.608772 0.643019 0.660706 0.670546
#>   6  0.438648 0.413737 0.418743 0.464802 0.545203 0.622353 0.659392 0.670442
#>   7  0.422074 0.382651 0.355653 0.372716 0.434698 0.512406 0.567201 0.579178
#>   8  0.386131 0.365000 0.337902 0.332633 0.356243 0.405640 0.469107 0.501606
#>   9  0.325586 0.343118 0.347356 0.329970 0.310995 0.326655 0.393461 0.458384
#>   10 0.325585 0.343119 0.347357 0.329970 0.310994 0.326654 0.393461 0.458382
#>     year
#> age  1989     1990     1991     1992     1993     1994     1995     1996    
#>   1  0.230220 0.221603 0.224258 0.230886 0.222410 0.189699 0.154108 0.131976
#>   2  0.457477 0.427437 0.422848 0.442742 0.447224 0.411239 0.382384 0.410460
#>   3  0.643112 0.596192 0.579936 0.602850 0.623988 0.614721 0.624873 0.724041
#>   4  0.644495 0.614942 0.596701 0.598872 0.611082 0.625921 0.652635 0.705040
#>   5  0.682723 0.691971 0.685888 0.665472 0.648105 0.647057 0.660896 0.685226
#>   6  0.680934 0.696959 0.715089 0.727515 0.720854 0.691232 0.669102 0.686874
#>   7  0.545822 0.514294 0.542302 0.637137 0.713090 0.679786 0.614871 0.613407
#>   8  0.460495 0.409336 0.431390 0.545003 0.650553 0.615654 0.532264 0.516365
#>   9  0.439977 0.389607 0.400775 0.492909 0.573663 0.531338 0.450409 0.428515
#>   10 0.439977 0.389606 0.400777 0.492910 0.573666 0.531336 0.450410 0.428513
#>     year
#> age  1997     1998     1999     2000     2001     2002     2003     2004    
#>   1  0.125026 0.131930 0.151195 0.171114 0.175727 0.174377 0.189527 0.223014
#>   2  0.475231 0.475549 0.381641 0.318986 0.357831 0.468479 0.524812 0.459320
#>   3  0.874024 0.868845 0.660994 0.517475 0.555246 0.700429 0.736417 0.590803
#>   4  0.767531 0.790954 0.749595 0.685451 0.636277 0.595875 0.545654 0.481330
#>   5  0.716009 0.748897 0.771661 0.755416 0.681218 0.582680 0.499483 0.434080
#>   6  0.729205 0.737348 0.686300 0.634737 0.627462 0.622126 0.547496 0.419329
#>   7  0.652392 0.620671 0.498882 0.420051 0.453605 0.533703 0.496658 0.343164
#>   8  0.547981 0.519383 0.406295 0.322806 0.318828 0.345727 0.315665 0.227106
#>   9  0.455254 0.457637 0.394177 0.308424 0.236415 0.185623 0.153697 0.130443
#>   10 0.455254 0.457640 0.394177 0.308423 0.236415 0.185623 0.153697 0.130443
#>     year
#> age  2005     2006     2007     2008     2009     2010     2011     2012    
#>   1  0.234923 0.195922 0.156976 0.155265 0.175660 0.163938 0.112876 0.084087
#>   2  0.381723 0.352444 0.336460 0.294405 0.238028 0.197880 0.181639 0.181566
#>   3  0.462151 0.426700 0.414645 0.355209 0.276267 0.235432 0.241384 0.262381
#>   4  0.411629 0.345695 0.294610 0.264044 0.249756 0.244102 0.242290 0.242458
#>   5  0.365556 0.290005 0.228910 0.193259 0.179128 0.180558 0.192698 0.204348
#>   6  0.313226 0.250081 0.206478 0.166349 0.135483 0.124151 0.132750 0.146164
#>   7  0.231007 0.188469 0.171186 0.148677 0.120566 0.098392 0.085709 0.081155
#>   8  0.156172 0.120847 0.102540 0.088897 0.076015 0.063627 0.053123 0.047775
#>   9  0.102324 0.069712 0.046624 0.035908 0.032590 0.032114 0.032229 0.032344
#>   10 0.102323 0.069712 0.046624 0.035908 0.032589 0.032114 0.032229 0.032344
#>     year
#> age  2013     2014     2015     2016     2017    
#>   1  0.098083 0.143330 0.162239 0.118466 0.068584
#>   2  0.192068 0.203952 0.204230 0.189177 0.168229
#>   3  0.263956 0.249430 0.240134 0.243971 0.255143
#>   4  0.243169 0.243014 0.240645 0.235880 0.229989
#>   5  0.203254 0.193879 0.187777 0.189212 0.194754
#>   6  0.146120 0.136550 0.131415 0.135675 0.145438
#>   7  0.083924 0.089945 0.092490 0.088653 0.081858
#>   8  0.049398 0.054375 0.055463 0.049762 0.041712
#>   9  0.032053 0.030857 0.028350 0.024854 0.021267
#>   10 0.032053 0.030857 0.028350 0.024854 0.021267
#> 
#> units:  f 
# Recomputes all F at age by solving catch Baranov
recomputeHarvest(ple4)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957     1958     1959     1960     1961     1962     1963     1964    
#>   1  0.098696 0.113654 0.127831 0.137077 0.135620 0.119031 0.093654 0.074617
#>   2  0.170865 0.212230 0.248869 0.261595 0.254859 0.254019 0.269948 0.279385
#>   3  0.261510 0.310081 0.349748 0.359114 0.351689 0.368920 0.430972 0.488987
#>   4  0.308814 0.328892 0.344626 0.350254 0.350732 0.359683 0.383249 0.407404
#>   5  0.256286 0.292328 0.324283 0.341173 0.345694 0.353381 0.372761 0.392885
#>   6  0.210941 0.249751 0.286689 0.310303 0.322991 0.341968 0.375670 0.402172
#>   7  0.225402 0.226093 0.230370 0.241952 0.261636 0.287425 0.313678 0.326106
#>   8  0.227730 0.222746 0.218581 0.216065 0.218564 0.231908 0.256588 0.274751
#>   9  0.198669 0.232773 0.250974 0.231472 0.197441 0.189074 0.220349 0.261486
#>   10 0.198669 0.232773 0.250974 0.231471 0.197442 0.189073 0.220349 0.261487
#>     year
#> age  1965     1966     1967     1968     1969     1970     1971     1972    
#>   1  0.068247 0.074649 0.097739 0.141559 0.188990 0.203713 0.194943 0.191621
#>   2  0.258531 0.232698 0.238291 0.282344 0.328127 0.330046 0.312624 0.317175
#>   3  0.471276 0.411084 0.376814 0.381689 0.394792 0.392436 0.393082 0.424562
#>   4  0.414037 0.401945 0.379520 0.357042 0.345439 0.353141 0.383641 0.441699
#>   5  0.399608 0.393135 0.381204 0.369548 0.361024 0.359625 0.375473 0.421997
#>   6  0.395058 0.371449 0.364484 0.379978 0.392054 0.380965 0.370270 0.389992
#>   7  0.312589 0.292007 0.291729 0.316706 0.343422 0.348644 0.345008 0.354106
#>   8  0.266110 0.245774 0.240398 0.257262 0.283156 0.303847 0.319480 0.335261
#>   9  0.263264 0.235461 0.216581 0.219652 0.238188 0.266625 0.300163 0.332087
#>   10 0.263264 0.235460 0.216580 0.219653 0.238187 0.266625 0.300162 0.332087
#>     year
#> age  1973     1974     1975     1976     1977     1978     1979     1980    
#>   1  0.205136 0.236977 0.287478 0.343453 0.378565 0.371259 0.317151 0.250648
#>   2  0.347660 0.378653 0.391158 0.399388 0.421605 0.453083 0.475313 0.477547
#>   3  0.479325 0.508315 0.483149 0.450938 0.455424 0.498640 0.568083 0.640980
#>   4  0.511604 0.545062 0.516128 0.475859 0.469931 0.497601 0.542233 0.588189
#>   5  0.486286 0.516664 0.486128 0.452108 0.464090 0.506867 0.532523 0.521353
#>   6  0.432925 0.453227 0.427973 0.406337 0.431863 0.486134 0.510963 0.481921
#>   7  0.374905 0.386728 0.379061 0.371085 0.381868 0.407373 0.431629 0.439596
#>   8  0.350924 0.360860 0.361462 0.355232 0.346310 0.342761 0.354073 0.376663
#>   9  0.356453 0.369329 0.368139 0.352942 0.326880 0.303054 0.295224 0.305917
#>   10 0.356453 0.369329 0.368141 0.352940 0.326881 0.303052 0.295224 0.305917
#>     year
#> age  1981     1982     1983     1984     1985     1986     1987     1988    
#>   1  0.216394 0.229811 0.271864 0.309120 0.319854 0.306280 0.278345 0.249823
#>   2  0.464454 0.444014 0.428054 0.427277 0.441640 0.463235 0.481828 0.483094
#>   3  0.674726 0.639052 0.579631 0.546988 0.551966 0.586399 0.636889 0.667582
#>   4  0.620786 0.628256 0.619281 0.608226 0.605009 0.616444 0.641385 0.658038
#>   5  0.504235 0.507115 0.529749 0.566534 0.608759 0.643005 0.660729 0.670558
#>   6  0.438645 0.413733 0.418740 0.464803 0.545202 0.622346 0.659400 0.670456
#>   7  0.422071 0.382644 0.355659 0.372709 0.434695 0.512406 0.567199 0.579179
#>   8  0.386125 0.365012 0.337900 0.332631 0.356248 0.405632 0.469109 0.501606
#>   9  0.325586 0.343118 0.347356 0.329970 0.310995 0.326655 0.393461 0.458384
#>   10 0.325585 0.343119 0.347357 0.329970 0.310994 0.326654 0.393461 0.458382
#>     year
#> age  1989     1990     1991     1992     1993     1994     1995     1996    
#>   1  0.230222 0.221606 0.224259 0.230888 0.222413 0.189693 0.154108 0.131976
#>   2  0.457473 0.427433 0.422846 0.442741 0.447222 0.411234 0.382374 0.410455
#>   3  0.643095 0.596192 0.579939 0.602845 0.623972 0.614710 0.624859 0.724052
#>   4  0.644476 0.614935 0.596703 0.598868 0.611077 0.625912 0.652659 0.705050
#>   5  0.682736 0.691986 0.685901 0.665483 0.648091 0.647042 0.660911 0.685237
#>   6  0.680945 0.696970 0.715092 0.727520 0.720861 0.691244 0.669120 0.686890
#>   7  0.545824 0.514295 0.542299 0.637120 0.713097 0.679797 0.614855 0.613400
#>   8  0.460488 0.409329 0.431385 0.545003 0.650577 0.615649 0.532269 0.516370
#>   9  0.439977 0.389607 0.400775 0.492909 0.573663 0.531338 0.450409 0.428515
#>   10 0.439977 0.389606 0.400777 0.492910 0.573666 0.531336 0.450410 0.428513
#>     year
#> age  1997     1998     1999     2000     2001     2002     2003     2004    
#>   1  0.125023 0.131929 0.151196 0.171117 0.175730 0.174378 0.189520 0.223014
#>   2  0.475230 0.475551 0.381633 0.318986 0.357839 0.468477 0.524814 0.459319
#>   3  0.874023 0.868849 0.661009 0.517475 0.555243 0.700436 0.736424 0.590802
#>   4  0.767544 0.790941 0.749605 0.685465 0.636265 0.595878 0.545656 0.481328
#>   5  0.716016 0.748908 0.771681 0.755433 0.681236 0.582680 0.499483 0.434078
#>   6  0.729204 0.737362 0.686316 0.634718 0.627448 0.622111 0.547498 0.419325
#>   7  0.652410 0.620663 0.498880 0.420048 0.453609 0.533699 0.496658 0.343165
#>   8  0.547980 0.519382 0.406290 0.322807 0.318829 0.345731 0.315665 0.227107
#>   9  0.455254 0.457637 0.394177 0.308424 0.236415 0.185623 0.153697 0.130443
#>   10 0.455254 0.457640 0.394177 0.308423 0.236415 0.185623 0.153697 0.130443
#>     year
#> age  2005     2006     2007     2008     2009     2010     2011     2012    
#>   1  0.234924 0.195926 0.156976 0.155261 0.175660 0.163938 0.112874 0.084084
#>   2  0.381715 0.352452 0.336460 0.294403 0.238031 0.197883 0.181645 0.181574
#>   3  0.462154 0.426697 0.414639 0.355217 0.276266 0.235433 0.241386 0.262383
#>   4  0.411627 0.345698 0.294612 0.264045 0.249755 0.244103 0.242291 0.242460
#>   5  0.365564 0.290004 0.228912 0.193249 0.179127 0.180564 0.192695 0.204349
#>   6  0.313227 0.250085 0.206478 0.166350 0.135489 0.124155 0.132748 0.146160
#>   7  0.231008 0.188474 0.171188 0.148677 0.120566 0.098387 0.085708 0.081156
#>   8  0.156172 0.120844 0.102538 0.088894 0.076016 0.063627 0.053125 0.047775
#>   9  0.102324 0.069712 0.046624 0.035908 0.032590 0.032114 0.032229 0.032344
#>   10 0.102323 0.069712 0.046624 0.035908 0.032589 0.032114 0.032229 0.032344
#>     year
#> age  2013     2014     2015     2016     2017    
#>   1  0.098081 0.143336 0.162238 0.118467 0.068584
#>   2  0.192062 0.203953 0.204229 0.189184 0.168229
#>   3  0.263958 0.249434 0.240136 0.243971 0.255143
#>   4  0.243170 0.243017 0.240647 0.235882 0.229989
#>   5  0.203257 0.193882 0.187781 0.189218 0.194754
#>   6  0.146122 0.136549 0.131416 0.135678 0.145438
#>   7  0.083927 0.089943 0.092493 0.088650 0.081858
#>   8  0.049397 0.054371 0.055460 0.049757 0.041712
#>   9  0.032053 0.030857 0.028350 0.024854 0.021267
#>   10 0.032053 0.030857 0.028350 0.024854 0.021267
#> 
#> units:  f