Compute the ratio of discards to total catch in numbers or weight

A calculation is made of the proportion of discards over total catch at age, either as numbers (value = 'numbers') or weight (value = 'weight'), or for the total discards and catch in biomass (value = 'total').

Usage

discardsRatio(object, value = c("numbers", "weight", "total"))

Arguments

object

An object of class 'FLStock'

value

One of 'numbers' (default), 'weight' or 'total'.

Value

The discards ratio (between 0 and 1), 'FLQuant'

See also

Author

The FLR Team

Examples

data(ple4)
# Discards ratio at age in numbers
discardsRatio(ple4)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957       1958       1959       1960       1961       1962      
#>   1  9.9998e-01 9.9999e-01 1.0000e+00 9.9999e-01 9.9999e-01 9.9999e-01
#>   2  9.0489e-01 9.0656e-01 8.8309e-01 9.6550e-01 9.8435e-01 9.8868e-01
#>   3  1.2189e-01 5.0045e-01 5.9883e-01 4.1916e-01 7.2412e-01 7.1640e-01
#>   4  1.7985e-02 3.7526e-02 3.0462e-01 1.9052e-02 7.7478e-03 2.5174e-01
#>   5  2.6529e-02 5.5647e-02 1.0139e-01 4.7381e-02 4.6307e-02 1.6254e-02
#>   6  2.5776e-03 3.1337e-03 4.5244e-03 1.8534e-02 9.9702e-03 9.2667e-03
#>   7  4.0821e-05 5.7489e-05 3.5276e-05 3.5012e-05 3.3535e-05 5.7089e-05
#>   8  4.9868e-05 5.8768e-05 7.6364e-05 5.5105e-05 5.0500e-05 5.7247e-05
#>   9  9.0523e-05 7.6944e-05 7.4188e-05 9.7626e-05 8.2095e-05 9.8379e-05
#>   10 4.1460e-05 3.5760e-05 3.0246e-05 3.2778e-05 3.3580e-05 4.0406e-05
#>     year
#> age  1963       1964       1965       1966       1967       1968      
#>   1  9.9999e-01 9.9999e-01 9.9999e-01 9.9999e-01 9.9998e-01 9.9999e-01
#>   2  9.6853e-01 9.1184e-01 9.7726e-01 9.7478e-01 9.5631e-01 8.8101e-01
#>   3  8.5569e-01 6.1133e-01 5.8271e-01 8.4070e-01 7.3906e-01 5.2525e-01
#>   4  1.1275e-01 3.4152e-01 5.9842e-02 2.1547e-01 5.6975e-01 4.0331e-01
#>   5  3.6722e-02 4.9491e-02 2.0594e-02 3.6203e-02 1.4558e-02 1.4992e-01
#>   6  5.7818e-03 8.5826e-03 8.7453e-03 4.4828e-03 1.1450e-02 3.3056e-03
#>   7  3.8176e-05 3.1371e-05 2.7078e-05 2.9758e-05 3.0309e-05 4.6979e-05
#>   8  7.1777e-05 7.5241e-05 5.8120e-05 4.5530e-05 6.1246e-05 4.8090e-05
#>   9  6.9598e-05 1.2466e-04 1.1799e-04 7.5918e-05 7.2860e-05 1.2375e-04
#>   10 2.9562e-05 2.9389e-05 2.8270e-05 3.2892e-05 4.3869e-05 3.6350e-05
#>     year
#> age  1969       1970       1971       1972       1973       1974      
#>   1  9.9996e-01 9.9940e-01 9.9971e-01 9.6823e-01 9.9003e-01 9.8920e-01
#>   2  8.0915e-01 9.0049e-01 7.6032e-01 5.9625e-01 5.9905e-01 9.2788e-01
#>   3  2.5938e-01 3.5753e-01 4.5922e-01 3.4642e-01 6.0639e-02 5.9635e-02
#>   4  3.3613e-01 2.2991e-02 7.1814e-02 9.5056e-02 8.6967e-03 6.4941e-03
#>   5  2.0127e-02 1.1322e-01 1.5596e-02 2.3390e-02 2.8225e-02 1.4155e-02
#>   6  1.8408e-02 4.7035e-03 3.4329e-03 4.1874e-03 4.6279e-03 5.3413e-03
#>   7  4.5383e-05 6.5051e-06 2.8617e-05 3.3812e-05 5.3554e-05 6.0234e-05
#>   8  7.7622e-05 8.1300e-05 1.6280e-05 4.4594e-05 4.8317e-05 7.6279e-05
#>   9  5.7670e-05 1.2259e-04 7.5452e-05 2.6162e-05 7.8205e-05 8.6593e-05
#>   10 2.9360e-05 3.4578e-05 2.8622e-05 3.1066e-05 1.9845e-05 2.1939e-05
#>     year
#> age  1975       1976       1977       1978       1979       1980      
#>   1  9.9576e-01 9.8452e-01 9.8759e-01 9.9516e-01 9.9553e-01 9.9562e-01
#>   2  9.0362e-01 8.0334e-01 6.4505e-01 7.2515e-01 7.8780e-01 6.4965e-01
#>   3  7.4428e-01 4.7157e-01 6.4655e-01 3.1197e-01 3.2670e-01 6.7963e-03
#>   4  1.2645e-01 1.5415e-01 3.3663e-01 1.7057e-01 2.7330e-01 8.6237e-03
#>   5  9.3001e-03 1.2199e-02 1.0002e-01 2.5113e-02 1.2181e-02 7.5168e-03
#>   6  5.4566e-03 4.0696e-03 1.3944e-02 1.6388e-02 7.7286e-03 4.6990e-03
#>   7  3.9638e-05 5.3461e-05 8.4422e-05 8.7282e-05 2.0671e-05 3.5666e-05
#>   8  7.9450e-05 7.6286e-05 9.9381e-05 1.5465e-04 1.2034e-04 5.7994e-05
#>   9  1.1460e-04 1.6541e-04 1.2289e-04 2.1427e-04 1.7836e-04 2.6328e-04
#>   10 2.3896e-05 3.8845e-05 5.4307e-05 6.9426e-05 5.3918e-05 9.0942e-05
#>     year
#> age  1981       1982       1983       1984       1985       1986      
#>   1  9.9807e-01 9.9186e-01 9.9538e-01 9.9965e-01 9.9970e-01 9.9851e-01
#>   2  6.5153e-01 8.0882e-01 7.8473e-01 8.3154e-01 7.5893e-01 8.8002e-01
#>   3  1.4482e-02 2.1329e-02 2.1032e-01 1.5805e-01 1.9499e-01 2.2616e-01
#>   4  6.2314e-03 1.5351e-02 2.1765e-02 2.8077e-01 1.0898e-02 1.9913e-01
#>   5  1.1630e-02 7.3621e-03 2.6467e-02 3.4721e-02 5.2569e-03 1.4074e-02
#>   6  4.3338e-03 5.7689e-03 5.5629e-03 4.4878e-03 4.5678e-03 5.2102e-03
#>   7  5.0989e-05 6.4892e-05 6.0685e-05 7.3071e-05 6.7183e-05 5.0813e-05
#>   8  6.0269e-05 9.0055e-05 1.1836e-04 8.9381e-05 1.3224e-04 1.0628e-04
#>   9  9.9881e-05 1.1012e-04 1.6589e-04 1.8314e-04 1.7770e-04 1.7637e-04
#>   10 8.2074e-05 5.6970e-05 6.0174e-05 7.6309e-05 8.5752e-05 7.0586e-05
#>     year
#> age  1987       1988       1989       1990       1991       1992      
#>   1  1.0000e+00 1.0000e+00 9.9415e-01 9.8947e-01 9.9228e-01 9.7617e-01
#>   2  9.4165e-01 9.7561e-01 9.1212e-01 8.9214e-01 8.7504e-01 8.3382e-01
#>   3  6.0447e-01 6.4412e-01 6.4083e-01 6.1679e-01 6.1884e-01 5.2272e-01
#>   4  1.1681e-02 4.1748e-01 2.6469e-01 1.8997e-01 2.4998e-01 2.9657e-01
#>   5  1.6315e-02 1.4507e-02 1.0948e-01 2.8723e-02 6.7748e-02 4.6396e-02
#>   6  5.5421e-03 5.6046e-03 5.3023e-03 5.8924e-03 1.1793e-02 2.2916e-02
#>   7  3.6927e-05 2.5843e-05 4.3082e-05 5.5227e-05 3.2445e-05 1.6423e-05
#>   8  1.0708e-04 8.0653e-05 6.2297e-05 1.0660e-04 9.2742e-05 6.9159e-05
#>   9  2.1302e-04 1.4557e-04 2.1537e-04 1.3143e-04 1.5550e-04 1.5171e-04
#>   10 8.9104e-05 7.8091e-05 8.6105e-05 1.1885e-04 8.8732e-05 7.1668e-05
#>     year
#> age  1993       1994       1995       1996       1997       1998      
#>   1  9.6576e-01 9.7834e-01 9.3857e-01 9.9018e-01 9.9307e-01 9.9812e-01
#>   2  7.4207e-01 6.8234e-01 6.9221e-01 8.8423e-01 9.2217e-01 9.5480e-01
#>   3  3.2516e-01 2.5220e-01 1.6009e-01 2.8828e-01 6.8608e-01 7.2022e-01
#>   4  1.2892e-01 1.1797e-02 1.0048e-02 6.9087e-02 1.1782e-02 4.6117e-01
#>   5  2.5196e-02 1.6994e-02 1.3659e-02 1.1964e-02 8.4785e-03 1.2884e-02
#>   6  6.1402e-03 5.1932e-03 4.8798e-03 6.0436e-03 5.7769e-03 3.6474e-03
#>   7  3.3908e-05 2.8935e-05 3.3905e-05 4.0320e-05 5.2528e-05 8.7658e-05
#>   8  4.3473e-05 7.5053e-05 9.4583e-05 8.4999e-05 9.9243e-05 1.7416e-04
#>   9  1.3494e-04 9.1743e-05 2.0035e-04 1.7857e-04 1.6369e-04 2.5204e-04
#>   10 8.4977e-05 9.1594e-05 1.2832e-04 1.0020e-04 1.1361e-04 1.4877e-04
#>     year
#> age  1999       2000       2001       2002       2003       2004      
#>   1  9.9570e-01 9.7545e-01 8.7028e-01 9.9605e-01 9.8978e-01 9.9767e-01
#>   2  9.5997e-01 9.1622e-01 9.0477e-01 9.1832e-01 9.2114e-01 9.3838e-01
#>   3  5.9450e-01 5.5588e-01 7.6106e-01 5.6531e-01 5.2786e-01 5.1778e-01
#>   4  5.3657e-01 2.7418e-01 5.6650e-01 2.0072e-01 2.6895e-01 2.1428e-01
#>   5  8.6876e-03 4.5296e-02 3.4837e-01 6.0209e-02 1.1162e-01 7.6741e-02
#>   6  4.1384e-03 1.8826e-02 1.3096e-02 1.6410e-02 2.0235e-02 3.2739e-02
#>   7  1.2380e-04 8.1263e-02 9.5098e-03 1.6498e-02 2.4100e-01 7.9902e-03
#>   8  1.9038e-04 1.2854e-01 1.2351e-03 6.6435e-04 8.4458e-04 3.1408e-04
#>   9  3.1286e-04 5.2393e-04 7.6822e-04 6.2855e-04 7.2468e-04 1.0842e-03
#>   10 1.5455e-04 2.3562e-04 2.1144e-04 3.0510e-04 5.6782e-04 8.7654e-04
#>     year
#> age  2005       2006       2007       2008       2009       2010      
#>   1  9.6956e-01 9.9836e-01 9.8185e-01 9.9705e-01 9.8959e-01 9.8752e-01
#>   2  9.4620e-01 9.1821e-01 9.0785e-01 9.5099e-01 9.3959e-01 9.0075e-01
#>   3  6.0973e-01 5.9481e-01 6.3524e-01 4.2033e-01 5.2700e-01 5.5007e-01
#>   4  2.0708e-01 2.2236e-01 1.9949e-01 1.0567e-01 1.8209e-01 2.4629e-01
#>   5  1.8218e-01 4.4350e-02 7.2562e-02 4.9845e-02 7.4644e-02 9.6822e-02
#>   6  3.5531e-01 3.0267e-02 2.4349e-01 4.2579e-02 5.6405e-02 1.2094e-01
#>   7  4.6421e-02 1.1643e-01 5.7974e-02 5.2324e-01 3.9139e-02 1.7520e-01
#>   8  2.3494e-03 7.8296e-02 4.4979e-01 4.7508e-01 4.3050e-01 1.3185e-01
#>   9  2.1699e-04 3.7907e-04 8.2706e-04 6.7963e-04 8.5079e-04 1.9831e-04
#>   10 7.7161e-04 3.5488e-04 3.8333e-04 5.3379e-04 6.0800e-04 3.9426e-04
#>     year
#> age  2011       2012       2013       2014       2015       2016      
#>   1  9.9656e-01 9.9822e-01 9.9573e-01 9.9997e-01 9.9999e-01 9.9999e-01
#>   2  9.2217e-01 9.8069e-01 9.3708e-01 9.7279e-01 9.8687e-01 9.7181e-01
#>   3  5.8026e-01 6.7139e-01 6.6516e-01 7.4080e-01 7.5404e-01 7.7277e-01
#>   4  3.7399e-01 3.4702e-01 2.4346e-01 4.1503e-01 3.6685e-01 3.6275e-01
#>   5  2.0067e-01 1.7350e-01 9.9575e-02 2.6806e-01 1.7118e-01 2.5690e-01
#>   6  1.6485e-01 4.6077e-02 3.3863e-02 1.0637e-01 6.8273e-02 1.5792e-01
#>   7  2.1144e-02 2.9103e-02 1.4099e-02 3.5620e-02 3.9165e-02 9.7503e-02
#>   8  2.4558e-01 2.1995e-03 3.9032e-03 7.9832e-03 2.2991e-02 1.9414e-01
#>   9  3.9604e-04 1.4357e-04 2.0252e-04 2.0219e-04 1.2329e-04 9.4147e-05
#>   10 1.8220e-04 2.5980e-04 1.3172e-04 1.4529e-04 1.0341e-04 9.7543e-05
#>     year
#> age  2017      
#>   1  9.9979e-01
#>   2  9.6470e-01
#>   3  7.5773e-01
#>   4  5.0369e-01
#>   5  2.1738e-01
#>   6  1.0986e-01
#>   7  6.9972e-02
#>   8  3.4123e-02
#>   9  1.6293e-04
#>   10 1.5558e-04
#> 
#> units:   
# Total proportion of discards by year
discardsRatio(ple4, value="total")
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957     1958     1959     1960     1961     1962     1963     1964    
#>   all 0.094872 0.164761 0.256767 0.247596 0.286714 0.284944 0.267392 0.246844
#>      year
#> age   1965     1966     1967     1968     1969     1970     1971     1972    
#>   all 0.303709 0.393999 0.327090 0.192228 0.160876 0.181791 0.169413 0.126828
#>      year
#> age   1973     1974     1975     1976     1977     1978     1979     1980    
#>   all 0.117128 0.270979 0.428883 0.305409 0.345624 0.279514 0.305964 0.184363
#>      year
#> age   1981     1982     1983     1984     1985     1986     1987     1988    
#>   all 0.179898 0.243039 0.324231 0.287314 0.261857 0.403236 0.497518 0.466706
#>      year
#> age   1989     1990     1991     1992     1993     1994     1995     1996    
#>   all 0.357385 0.304026 0.322390 0.300470 0.210351 0.165647 0.173023 0.289076
#>      year
#> age   1997     1998     1999     2000     2001     2002     2003     2004    
#>   all 0.455467 0.566589 0.429636 0.312883 0.480153 0.394062 0.514574 0.410748
#>      year
#> age   2005     2006     2007     2008     2009     2010     2011     2012    
#>   all 0.463110 0.444317 0.430058 0.430545 0.405011 0.375159 0.378231 0.375856
#>      year
#> age   2013     2014     2015     2016     2017    
#>   all 0.311454 0.381583 0.337162 0.341689 0.321577
#> 
#> units: