drop method for FLCore array-based classes

Delete the dimensions of an array which have only one level.

# S4 method for FLArray
drop(x)

Details

This method calls R's base::drop on the @.Data slot of an FLArray. Dimensions of length one are thus dropped, as is the class attribute and the units slot, and an array of equal or less dimensions, a matrix or a vector is returned.

On an FLQuant object with

See also

base::drop

Examples

x <- FLQuant(1:3, dim=c(3,3))
drop(x)

#>      year
#> quant 1 2 3
#>     1 1 1 1
#>     2 2 2 2
#>     3 3 3 3

is(drop(x))

#> [1] "matrix"    "array"     "mMatrix"   "structure" "vector"   

dim(drop(x))

#> [1] 3 3

# Result of drop can be used for matrix algebra
# for example to calculate aging error

data(ple4)
aging.error <- diag(0.8, 10)
diag(aging.error[-1,]) <- c(rep(0.1, 8), 0.2)
diag(aging.error[, -1]) <- c(0.2, rep(0.1, 8))
t(aging.error) %*% drop(catch.n(ple4))

#>        year
#>            1957    1958     1959     1960     1961     1962     1963     1964
#>    [1,] 30875.9 61027.3 107301.5  76536.4  75149.7  59358.8  86782.7  69626.2
#>    [2,] 53303.8 82370.6 146361.7 164551.8 144854.6 169374.9 145725.9 158056.8
#>    [3,] 64784.2 51218.0  79585.5 108046.3 119456.5 125976.3 143740.5 110367.8
#>    [4,] 46678.6 59906.1  41783.3  54536.4  70426.9  88229.4  90284.5  95354.1
#>    [5,] 31639.3 37427.0  42662.6  25897.1  35582.1  51618.1  57068.5  61932.6
#>    [6,] 11504.1 20169.9  24404.5  25935.8  15210.3  22643.1  32522.0  37555.0
#>    [7,] 10617.0  9308.3  14043.4  15034.9  14007.7  10232.9  14160.4  17056.1
#>    [8,]  8822.6  7828.8   7173.3   9137.7   9866.0   8659.7   7455.9   7255.9
#>    [9,]  7053.0  8317.6   9153.6   8001.5   8703.6   7669.3   9639.3   7220.7
#>   [10,]  9184.5 11129.4  13603.6  12631.1  12327.6  10766.5  13978.3  13906.8
#>        year
#>             1965     1966     1967     1968    1969     1970     1971    1972
#>    [1,]  78554.3  47523.7  30111.9  67343.4 76395.5 115893.4  68140.7 66987.8
#>    [2,] 367591.2 182951.1 127445.0  88830.7 89148.4 168440.3 124182.0 97655.2
#>    [3,] 130418.3 354809.2 134762.8  84263.2 63328.5  85858.2  88779.7 91940.2
#>    [4,]  59879.7  97953.2 228314.6  75045.2 54163.6  58418.8  40632.1 58866.8
#>    [5,]  44378.6  39793.4  62377.4 125030.7 45032.0  45865.1  27101.4 26457.1
#>    [6,]  30630.2  27452.9  23668.5  29397.2 87907.2  40651.3  22727.0 17455.8
#>    [7,]  18586.0  17027.9  15971.7  11002.7 19340.4  67263.4  18801.8 14301.4
#>    [8,]   9100.6  10923.9   8693.5   9508.2  6729.0  13355.9  26116.7 12018.9
#>    [9,]   7743.2   9237.6   8417.8   6823.2 10380.6   6947.1  11486.2 19088.5
#>   [10,]  14490.9  12589.1   9612.8  11084.5 13692.8  12290.5  14169.4 14379.3
#>        year
#>             1973     1974     1975     1976     1977     1978     1979     1980
#>    [1,] 115050.8 203842.7 227585.4 166815.6 223944.9 203881.1 262898.7 200626.5
#>    [2,]  98273.1 313817.2 321009.1 191861.3 192762.2 226315.8 295032.9 208909.3
#>    [3,]  66014.2  84754.1 236159.5 147859.9 124107.8  99660.7 175198.7 134305.9
#>    [4,]  69655.5  44016.6  56638.3 107794.4 101214.8  66187.4  76366.8  78552.0
#>    [5,]  43597.4  39567.3  26275.6  23598.0  85301.2  51196.4  49538.6  29761.4
#>    [6,]  16350.1  20789.1  20803.3  10263.3  17310.2  34592.6  39463.9  18347.2
#>    [7,]   9481.2   9013.6  12221.8   8929.6   6158.9   8695.2  23822.7  13661.4
#>    [8,]   9375.3   6414.1   6336.3   6319.5   5014.9   3517.5   6032.3   8328.9
#>    [9,]  10647.6   9482.5   7816.1   5507.1   5589.2   3764.0   4519.3   3412.4
#>   [10,]  19790.8  18140.8  16166.6  10336.3   7758.9   6216.3   7711.1   4488.0
#>        year
#>             1981     1982     1983     1984     1985     1986      1987
#>    [1,] 136932.8 356947.6 265693.8 286300.6 354680.8 951024.2  435147.7
#>    [2,] 274628.0 306169.7 511918.6 396236.2 341297.4 691498.5 1269455.8
#>    [3,] 134531.3 203204.9 182333.7 306970.0 192924.0 237617.7  394735.1
#>    [4,]  62413.8  80772.6  98640.1  99044.4 171199.4 126032.5  126435.4
#>    [5,]  35985.4  31844.5  35559.3  40348.8  46480.6  83190.6   67649.0
#>    [6,]  14549.2  17105.2  14220.3  15526.5  16456.7  28878.9   35736.6
#>    [7,]   9707.3   8300.6   8289.9   7099.4   7424.0  10307.7   13093.3
#>    [8,]   7917.4   5583.1   4477.3   5353.7   3876.9   4795.2    4833.4
#>    [9,]   5889.0   5863.2   4487.1   4043.4   3683.0   4025.4    3252.9
#>   [10,]   5209.8   7370.6   6930.9   5524.0   4908.2   5833.3    4597.8
#>        year
#>             1988     1989     1990     1991     1992     1993     1994     1995
#>    [1,] 341203.1 224901.8 148674.3 182153.7 140089.3 100668.8  64897.9 113216.3
#>    [2,] 639329.4 498902.5 305481.6 335307.2 257261.5 201038.3 139116.0 130438.7
#>    [3,] 643923.7 324261.4 258245.0 230309.5 180802.3 146398.8 132068.5  97568.9
#>    [4,] 184233.8 289661.4 154405.7 155881.1 111144.7  87351.9  82687.6  76815.3
#>    [5,]  54686.6  81759.8 148463.3  86330.2  75720.8  53778.9  41671.8  39459.9
#>    [6,]  26718.0  22559.6  38842.6  72136.7  38084.5  30024.8  21894.3  19151.8
#>    [7,]  16484.1  11145.9   9618.2  20300.9  27763.2  14826.5  14951.5  10061.5
#>    [8,]   6298.4   7317.5   4530.8   5991.5   8850.7  10729.3   7188.3   4703.6
#>    [9,]   4314.3   3568.1   3944.0   4154.3   4642.6   5219.8   6020.9   3056.0
#>   [10,]   5302.6   4676.0   3391.5   4786.9   5862.4   5070.9   4883.2   3236.0
#>        year
#>             1996     1997     1998     1999     2000     2001     2002     2003
#>    [1,] 127239.3 159013.3 151452.3 124005.0 103609.5  66724.0 268200.2 110726.0
#>    [2,] 330518.4 497005.0 588322.7 238020.0 180519.2 342051.6 231010.4 472448.4
#>    [3,] 115975.4 284423.8 287074.2 341373.6 115308.0 243603.1 134590.9 144212.0
#>    [4,]  52688.1  65764.4 115502.5 116993.8 195013.7 137436.3  66886.7  68358.9
#>    [5,]  32551.3  24037.9  25222.8  29673.3  37285.6 128712.7  40982.5  33600.2
#>    [6,]  15979.4  13818.2   7371.6   8226.9   9626.4  20539.9  35254.8  18579.2
#>    [7,]  10394.4   6815.4   4885.9   3139.7   3004.5   4388.8   7853.1  20063.1
#>    [8,]   4945.3   3759.7   2679.5   2227.2   1292.3   1412.5   1981.7   3634.3
#>    [9,]   3530.4   2942.6   2283.2   2060.0   1291.2   1084.2   1089.7    852.7
#>   [10,]   4176.0   3562.7   2671.3   2626.5   1771.6   1910.9   1343.0    752.2
#>        year
#>             2004     2005     2006     2007     2008
#>    [1,] 205831.7 111450.6 201068.8  85312.9 134988.7
#>    [2,] 221726.5 300648.8 257184.8 205857.7 246062.2
#>    [3,] 200512.3 102241.0 170141.4 113510.1  94987.0
#>    [4,]  58944.8  91878.2  51048.5  58533.3  46567.2
#>    [5,]  23249.6  29269.8  40707.5  21334.5  30448.8
#>    [6,]  12023.1  15211.5  10696.6  21341.5   9712.8
#>    [7,]   5484.6   6213.7   5498.0   4910.2  13309.2
#>    [8,]   4169.0   2863.2   2493.2   3124.6   3073.1
#>    [9,]    893.2   2162.5   1476.5   1036.7    897.9
#>   [10,]    473.2    712.7   1178.7   1061.5    784.1