Compute an abundance index from an FLStock

Computes an abundance-at-age index from an FLStock object at a given survey timing, correcting stock numbers for the combined effect of fishing and natural mortality up to that point in the year.

Usage

# S4 method for class 'FLStock'
index(object, sel = catch.sel(object), ages = dimnames(sel)$age, timing = 0.5)

Arguments

object

An FLStock object from which to compute the index.

sel

Selectivity-at-age pattern to apply; defaults to catch.sel(object).

ages

Character vector of ages to include; defaults to all ages in sel.

timing

Fraction of the year at which the index is computed (0–1); defaults to 0.5 (mid-year).

Value

An FLQuant of abundance-at-age corrected for mortality up to timing and filtered to the selected ages.

See also

Author

The FLR Team

Examples

data(ple4)
index(ple4, timing=0.9)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957    1958    1959    1960    1961    1962    1963    1964    1965   
#>   1   127504  202646  260436  245990  271641  163114  111358  319473   82707
#>   2   116282  190548  298391  366430  331378  344478  222059  204106  817352
#>   3   221158  133333  190944  267896  323097  289392  299435  201075  205676
#>   4   154445  171228   89642  118724  170087  199084  158955  149846  104864
#>   5    89073   92574  103822   55580   76589  103846  110227   87815   92461
#>   6    31294   57412   57963   64102   35133   46407   58860   63621   55856
#>   7    34106   22765   35719   33693   38255   20160   24025   29262   33659
#>   8    27134   22911   15039   23752   22547   23126   10751   12701   17078
#>   9    15732   18679   17138   11149   16430   13885   13638    6730    8771
#>   10   33722   39140   41347   37639   31453   32711   33849   34742   30133
#>     year
#> age  1966    1967    1968    1969    1970    1971    1972    1973    1974   
#>   1    89859   91868  124799  246127  265180  164898  122614  423932  345084
#>   2   235635  224179  184201  185708  285275  271874  159261  124677  463083
#>   3   907482  259140  224745  153542  137434  208478  196652  118326   94635
#>   4   120897  557557  154954  125503   87456   82625  121875  111956   66694
#>   5    69566   76258  338685   95570   83126   57836   48174   64742   57308
#>   6    61319   43583   46971  212872   62846   53446   32799   25069   32464
#>   7    31653   34821   24814   26246  121063   36280   28839   16905   12860
#>   8    21327   19890   21381   14573   15538   72597   20283   15714    9619
#>   9    12348   14724   13157   13621    9552   10267   43650   11820    9726
#>   10   28442   29425   32493   34191   37423   35504   29348   42961   33363
#>     year
#> age  1975    1976    1977    1978    1979    1980    1981    1982    1983   
#>   1   309439  326381  541338  421309  351147  307636  241237  517212  432203
#>   2   373810  286173  242702  348074  273552  267261  314631  310309  598039
#>   3   351928  271029  190679  147384  201949  158437  168611  222120  214747
#>   4    54843  211214  162062  106709   75564   93304   70758   79650  107980
#>   5    34751   29808  118522   88660   54320   34519   41181   32877   36558
#>   6    29306   18838   16905   64194   43306   24223   15968   21150   16530
#>   7    17628   17110   11045    8879   29722   19113   11886    9025   11826
#>   8     7981   11338   10322    5880    4350   14193    9757    6778    5284
#>   9     6494    5369    6875    5572    3029    2209    7577    5851    4248
#>   10   28399   23033   17336   13874   11376    8824    7303   10648   11024
#>     year
#> age  1984    1985    1986    1987    1988    1989    1990    1991    1992   
#>   1   457908  645334 1422108  572492  482675  313266  258000  229865  201424
#>   2   414599  385355  511070 1238395  557771  507554  342912  290232  264956
#>   3   412304  282304  247399  333023  799392  352465  327499  233276  205050
#>   4   113327  226442  146750  126168  159197  359201  161058  159176  117936
#>   5    51373   55789  108794   71855   60319   74347  171932   80357   81717
#>   6    18446   25855   26256   50892   33282   27581   33556   76823   36775
#>   7     9295   10067   12429   11885   22154   13778   10858   13182   31942
#>   8     7290    5615    5348    6241    5637    9761    6000    5174    6834
#>   9     3415    4505    3172    3079    3434    2813    4923    3467    3232
#>   10    9593    8106    8198    8192    7450    5969    4645    5820    6263
#>     year
#> age  1993    1994    1995    1996    1997    1998    1999    2000    2001   
#>   1   127042  119775  170805  132111  284023   95931  106769  152187  127767
#>   2   232794  149721  156965  258892  229444  632999  197908  153901  227480
#>   3   186055  166043  116300  124669  180606  166580  471498  149855  157959
#>   4   103106   94195   86124   54650   46180   66895   68371  228268   99021
#>   5    61262   54134   48177   38231   20353   19359   30676   29938  115130
#>   6    39531   31020   26859   21120   14773    9146    9042   12194   13153
#>   7    16110   17752   13840   11087    7804    6086    3738    3273    5456
#>   8    15561    7022    7519    5742    4264    3320    2646    1591    1794
#>   9     3712    6966    3011    3263    2392    2033    1709    1351     915
#>   10    5732    4831    5910    4613    3989    3696    3368    2674    2426
#>     year
#> age  2002    2003    2004    2005    2006    2007    2008    2009    2010   
#>   1   348682  110635  348801  324855  307892  414490  394296  540194  765037
#>   2   193207  553417  196258  478999  339726  356652  620148  558899  512437
#>   3   201285  128553  391767  143925  344020  247374  279324  512672  447236
#>   4    82545   76998   57087  205480   73805  162571  125864  176123  362244
#>   5    44322   35789   44261   34394  104875   35776   91908   82343  108101
#>   6    49377   20658   22103   26684   17751   56086   23107   65088   51340
#>   7     5720   20388   10152   10971   12871    9459   36552   17192   43705
#>   8     2562    2078    8156    4487    4930    5834    4708   20005    8669
#>   9      753     825     834    3789    1815    1678    2068    1933    8345
#>   10    1908    1621    2050    2356    3814    3338    3932    6130    7879
#>     year
#> age  2011    2012    2013    2014    2015    2016    2017   
#>   1   618570  347028  452386  757392  476868  483202  421050
#>   2   645515  698137  594770  742372  830186  411851  504259
#>   3   449586  604165  706881  579645  647106  696275  374830
#>   4   329524  296114  394150  487695  411946  447919  450592
#>   5   213894  188581  181493  246656  291305  238021  262269
#>   6    63142  116579  104190  101172  136637  163171  136272
#>   7    29542   29587   55019   55812   58337   74965   78062
#>   8    20308   13067   15368   32228   30430   26537   29135
#>   9     3864    9981    7625    8896   15280   11836    9580
#>   10   14372   14805   21396   25935   29166   33911   33146
#> 
#> units:  1000 
index(ple4, timing=0)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957    1958    1959    1960    1961    1962    1963    1964    1965   
#>   1   152472  245611  319707  304497  335807  198657  132561  373840   96228
#>   2   148382  252373  408457  507371  456062  473734  309791  287174 1128614
#>   3   306199  192852  286216  404963  485155  441333  482883  341643  343934
#>   4   223134  251892  133751  178043  255179  301097  245564  236578  166551
#>   5   122745  131775  152099   82672  114385  156170  168685  136843  144958
#>   6    41399   78651   82091   92736   51410   69077   90312   99973   87212
#>   7    45711   30530   48087   45836   52971   28570   34863   42939   48793
#>   8    36442   30633   20033   31567   30034   31177   14820   17796   23742
#>   9    20584   25202   23505   15024   21473   18011   18195    9318   12163
#>   10   44122   52806   56705   50722   41108   42431   45160   48101   41786
#>     year
#> age  1966    1967    1968    1969    1970    1971    1972    1973    1974   
#>   1   105154  109763  155106  319237  348538  215029  159414  557907  467345
#>   2   317892  303963  259859  273005  420100  394138  231829  186535  712446
#>   3  1437490  398024  346712  239678  214078  324931  315305  199304  163614
#>   4   189937  858461  233799  187395  131493  127689  198448  194134  119185
#>   5   108429  117592  516799  144716  125714   88725   77062  109735   99828
#>   6    93728   66200   72350  331474   96888   81606   50978   40498   53412
#>   7    45044   49540   36106   39118  181288   54150   43399   25920   19929
#>   8    29113   27020   29490   20574   22348  105896   30009   23580   14564
#>   9    16700   19578   17543   18467   13286   14718   64398   17825   14839
#>   10   38466   39125   43325   46355   52052   50896   43298   64786   50898
#>     year
#> age  1975    1976    1977    1978    1979    1980    1981    1982    1983   
#>   1   438557  486468  832772  643879  511139  421786  320709  695952  604000
#>   2   581608  448563  388108  572600  459103  449447  522910  506327  961895
#>   3   594821  444999  314340  252602  368443  308663  338604  431979  395886
#>   4    95487  354654  270674  182719  134691  173336  135365  153406  206296
#>   5    58894   48992  196917  153084   95982   60384   70936   56780   64436
#>   6    47133   29712   27283  108791   75050   40896   25929   33583   26365
#>   7    27131   26145   17042   14018   47960   31062   19016   13935   17822
#>   8    12090   17079   15424    8759    6545   21797   15112   10300    7837
#>   9     9897    8071   10095    8009    4323    3183   11113    8719    6354
#>   10   43279   34624   25457   19941   16235   12716   10712   15866   16489
#>     year
#> age  1984    1985    1986    1987    1988    1989    1990    1991    1992   
#>   1   661742  941652 2049904  804732  661284  421682  344607  307761  271296
#>   2   666381  627436  848457 2090622  942685  838261  551237  464632  431830
#>   3   738079  507629  458867  646405 1595050  687974  612812  430164  385993
#>   4   214367  427093  279650  245888  314933  701996  306498  297986  221212
#>   5    93597  105579  212338  142492  120680  150385  350681  163005  162745
#>   6    30666   46210   50301  100802   66581   55699   68750  159986   77446
#>   7    14223   16289   21568   21666   40825   24639   18873   23498   62013
#>   8    10760    8466    8431   10416    9688   16164    9489    8348   12211
#>   9     5029    6521    4657    4800    5677    4573    7649    5441    5511
#>   10   14125   11734   12035   12772   12314    9704    7217    9134   10678
#>     year
#> age  1993    1994    1995    1996    1997    1998    1999    2000    2001   
#>   1   169811  155454  214696  162784  347783  118199  133854  194243  163754
#>   2   380945  237197  242298  409863  385047 1062590  305296  224393  343474
#>   3   356962  315921  223309  261727  433954  398394  935245  261229  284877
#>   4   195533  181036  169553  112785  100819  149155  146875  462860  192093
#>   5   120117  106040   95553   77505   42420   41561   67219   64651  232564
#>   6    82753   63227   53667   42881   31158   19432   18350   23623   25315
#>   7    33490   35813   26337   21070   15360   11642    6408    5226    8979
#>   8    30577   13372   13282   10000    7640    5797    4174    2327    2615
#>   9     6806   12295    4941    5251    3943    3358    2666    1951    1238
#>   10   10510    8528    9699    7423    6575    6105    5254    3862    3284
#>     year
#> age  2002    2003    2004    2005    2006    2007    2008    2009    2010   
#>   1   446348  143568  466479  439136  401849  522345  496129  692300  970166
#>   2   322271  971114  324672  738964  510474  528252  884411  757630  669994
#>   3   413685  272903  729522  238706  552652  393107  420759  719298  604847
#>   4   154413  137671   96329  325648  110229  231890  174660  241281  493740
#>   5    81933   61386   71576   52294  148974   48100  119667  105859  139153
#>   6    94576   36997   35273   38705   24326   73900   29367   80454   62816
#>   7    10118   34880   15127   14778   16686   12074   45720   20967   52248
#>   8     3826    3020   10947    5650    6014    7001    5581   23438   10044
#>   9      974    1036    1027    4546    2114    1915    2337    2178    9398
#>   10    2467    2037    2522    2826    4444    3809    4444    6907    8874
#>     year
#> age  2011    2012    2013    2014    2015    2016    2017   
#>   1   749193  409559  540670  942828  603806  588193  490035
#>   2   831744  899489  773584  975945 1091659  534280  641939
#>   3   611291  837139  980853  793858  878864  948914  515997
#>   4   448410  403008  536777  664080  559741  606014  606406
#>   5   278359  248003  238447  321336  377424  308787  341944
#>   6    77856  145491  130025  125175  168275  201725  169957
#>   7    34916   34826   64924   66217   69371   88837   91943
#>   8    23308   14926   17580   37031   35000   30366   33098
#>   9     4352   11244    8587   10008   17151   13243   10684
#>   10   16188   16678   24096   29176   32738   37944   36968
#> 
#> units:  1000