Computes the mean natural mortality acros the fully selected ages

Equivalent to the mean fishing mortality metric returned by 'fbar', 'mbar' calculates the mean natural mortality across the ages inside the range defined by 'minfbar' and 'maxfbar'.

Usage

mbar(object, ...)

Arguments

object

An object of class 'FLStock'.

Value

An object of class 'FLQuant'.

See also

Author

The FLR Team, proposal by H. Winker.

Examples

data(ple4)
mbar(ple4)
#> An x of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1 
#>      year
#> age   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1 
#>      year
#> age   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1 
#>      year
#> age   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1 
#>      year
#> age   2013 2014 2015 2016 2017
#>   all 0.1  0.1  0.1  0.1  0.1 
#> 
#> units:  m