Method show

Standard complete display of an object contents in an interactive session. Objects of class FLArray with length > 1 alonArray sixth dimension (iter) are output in a summarised form, as median(mad), where mad is the median absolute deviation. See mad.

Usage

# S4 method for class 'FLArray'
show(object)

# S4 method for class 'FLArray'
print(x)

Details

The same format is used for objects of class FLPar with length > 1 on the last dimension (iter).

Generic function

show(object)

show(object)

See also

Author

The FLR Team

Examples

# no 'iter'
  flq <- FLQuant(rnorm(80), dim=c(4,20), quant='age', units='kg')
  flq
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1       2       3       4       5       6       7       8       9      
#>   1 -0.7683 -0.8635 -1.0393  0.8814 -1.5713  0.1656 -0.7367  1.9859  0.1321
#>   2 -0.0928 -0.4314 -0.7416  0.1343  0.3898 -0.8074  1.8014 -1.1418 -0.9801
#>   3  0.5561 -0.5547  1.7454  0.0664 -0.4147 -1.5721 -1.1692 -0.0613  0.6202
#>   4  0.3939 -1.6194  1.5964  0.7333  1.4237 -0.6937 -0.2360 -0.2946 -0.1128
#>    year
#> age 10      11      12      13      14      15      16      17      18     
#>   1 -1.0302 -1.9053 -0.6399  1.6206  1.4134  0.9338  2.4380  2.0637 -0.4482
#>   2 -0.4088 -0.2007  0.7351 -1.0549 -1.0139 -0.9474  1.2289  0.5816 -0.6957
#>   3  1.1144  1.2800  1.9829 -0.6860  0.9816 -1.0705 -1.0820  1.8016  0.4753
#>   4 -0.6219 -1.5845 -1.0265 -1.7087 -0.5048  0.9192 -0.2072 -0.7493 -0.9596
#>    year
#> age 19      20     
#>   1 -0.6403  1.6205
#>   2 -0.3484 -0.4081
#>   3  0.9789 -0.3577
#>   4  0.2700 -1.0707
#> 
#> units:  kg 

# with 'iter'
  flq <- FLQuant(rnorm(800), dim=c(4,20,1,1,1,10), quant='age', units='kg')
  flq
#> An object of class "FLQuant"
#> iters:  10 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1                2                3                4               
#>   1  3.32e-01(1.667) -1.05e-01(1.340) -1.22e-01(0.902)  1.93e-01(1.329)
#>   2 -2.00e-01(0.945)  1.47e-01(0.922) -3.55e-01(1.881) -3.95e-01(0.721)
#>   3  8.88e-01(0.933) -3.53e-01(1.108) -3.16e-02(0.728)  3.88e-01(0.602)
#>   4  1.13e-01(0.799)  2.06e-01(0.417) -1.74e-01(1.073) -2.11e-02(0.328)
#>    year
#> age 5                6                7                8               
#>   1 -3.54e-01(1.314)  7.20e-01(1.091)  1.63e-01(0.841)  2.41e-01(0.530)
#>   2 -7.06e-01(0.856)  4.68e-01(0.623) -4.19e-01(1.366) -2.38e-01(1.401)
#>   3 -1.58e-01(0.530)  2.59e-02(0.805)  5.66e-01(0.872)  2.21e-01(1.520)
#>   4 -1.81e-01(1.351)  1.18e-01(0.873)  7.42e-01(0.636) -2.44e-02(0.788)
#>    year
#> age 9                10               11               12              
#>   1  4.68e-01(0.790)  9.17e-02(0.348) -2.26e-01(0.593) -3.94e-01(1.051)
#>   2 -8.18e-01(0.647) -5.28e-01(1.192)  2.80e-01(1.406)  1.56e-01(0.999)
#>   3 -9.02e-02(0.240)  5.92e-01(0.958)  2.61e-01(0.729) -2.16e-02(0.564)
#>   4  5.28e-01(0.701) -4.68e-02(0.822) -1.30e-01(1.079)  6.76e-02(0.486)
#>    year
#> age 13               14               15               16              
#>   1 -2.95e-01(0.313) -4.46e-01(0.842)  2.68e-01(0.937)  4.32e-01(0.636)
#>   2 -3.21e-01(0.856) -2.23e-01(0.644) -4.66e-01(0.598) -7.05e-05(0.668)
#>   3 -1.50e-01(0.974)  5.65e-01(0.437)  7.22e-01(0.537) -2.90e-01(1.397)
#>   4 -1.91e-01(0.242)  5.23e-01(0.635)  6.49e-02(1.119) -1.42e-01(0.999)
#>    year
#> age 17               18               19               20              
#>   1  3.20e-02(0.544) -1.62e-01(0.681) -4.74e-01(0.652)  3.28e-01(1.504)
#>   2  3.25e-01(0.866) -2.14e-01(0.653) -4.15e-01(0.501) -2.77e-01(0.549)
#>   3 -1.80e-01(1.166) -1.16e-01(1.019) -4.22e-02(1.141) -7.75e-01(0.538)
#>   4  1.47e-01(0.972)  9.33e-02(0.451)  6.77e-02(0.661)  5.74e-02(0.508)
#> 
#> units:  kg 

# no 'iter'
  flq <- FLQuant(rnorm(80), dim=c(4,20), quant='age', units='kg')
  print(flq)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1        2        3        4        5        6        7        8       
#>   1  0.35792  0.14314  0.13500  0.47246 -0.93493 -1.05931 -0.06643  0.16307
#>   2  0.31297 -0.56692 -1.31361 -1.62311  2.05337  1.07333 -1.63066 -1.61676
#>   3 -0.42846 -0.53705 -0.41579  1.00470 -0.12086 -0.58736 -0.99563  0.14657
#>   4 -1.20183  0.01234 -0.56465 -0.17441  0.29930  0.94404 -0.38025 -0.27040
#>    year
#> age 9        10       11       12       13       14       15       16      
#>   1  0.49622  0.99456 -0.81332  0.06408  1.45493 -0.34793  0.94599  0.56966
#>   2 -0.58427  0.39066  2.23443 -0.51628  0.47571  1.04122 -0.13877  0.96278
#>   3  0.08536 -0.27317 -0.01571  1.02562 -0.22306  0.44452 -0.54147 -1.21953
#>   4 -0.14937  0.23710 -0.89101 -0.42974 -0.76841 -2.27069  0.94408 -2.22199
#>    year
#> age 17       18       19       20      
#>   1  0.51996 -1.67117 -0.53261  2.15958
#>   2 -0.74746 -1.17643 -2.61846  0.00763
#>   3  0.08496 -0.21698  1.73671  1.23511
#>   4  0.87498  0.64437  0.83591 -0.16481
#> 
#> units:  kg 

# with 'iter'
  flq <- FLQuant(rnorm(800), dim=c(4,20,1,1,1,10), quant='age', units='kg')
  print(flq)
#> An object of class "FLQuant"
#> iters:  10 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1               2               3               4              
#>   1 -0.80751(0.812) -0.05616(0.641)  0.11569(0.707)  0.46866(0.730)
#>   2  0.20593(1.230)  0.19095(0.767)  0.26942(0.668) -0.65405(0.360)
#>   3  0.00156(0.604)  0.24380(0.642) -0.32949(0.761) -0.26554(1.184)
#>   4  0.40428(0.650) -0.27740(1.307) -0.88371(0.484) -0.17511(0.391)
#>    year
#> age 5               6               7               8              
#>   1 -0.81367(0.627)  0.06233(1.305) -0.54924(0.833)  0.33269(1.059)
#>   2  0.21549(0.789) -0.15101(1.297) -0.30782(1.120) -0.05947(0.885)
#>   3  0.20096(1.016) -0.31567(1.500)  0.04768(1.302)  0.27583(1.578)
#>   4  0.75097(0.923)  0.00923(0.869) -0.23566(1.310) -0.19389(0.683)
#>    year
#> age 9               10              11              12             
#>   1 -0.09150(0.827) -0.03163(1.342) -0.10790(0.956)  0.56864(0.870)
#>   2  0.14169(0.670) -0.15553(1.495)  0.06160(1.160)  0.46916(1.020)
#>   3  0.19309(0.528) -0.36745(0.417)  0.06643(1.009) -0.14145(1.051)
#>   4 -0.26118(1.066)  0.21743(0.936)  0.14658(1.710)  0.48043(0.951)
#>    year
#> age 13              14              15              16             
#>   1  0.25811(0.473) -0.27875(1.016)  0.38644(0.761) -0.19465(0.654)
#>   2 -0.39623(1.219)  0.35855(0.942) -0.11907(0.896)  0.16393(0.609)
#>   3 -0.12738(0.995) -0.06890(0.695) -0.14660(0.976) -0.00721(0.381)
#>   4  0.29468(1.185)  0.16376(0.781) -0.02563(0.882)  0.39290(0.751)
#>    year
#> age 17              18              19              20             
#>   1  0.43330(0.890)  0.19069(1.078)  0.31578(0.613)  0.12848(0.683)
#>   2  0.34238(0.879)  0.03490(1.340)  0.17179(0.693)  0.35240(0.517)
#>   3  0.18807(0.700) -0.00716(1.250)  0.29667(0.911)  0.25504(0.869)
#>   4  0.76575(0.688) -0.15963(0.615) -0.01298(1.253)  0.10248(0.951)
#> 
#> units:  kg