Calculate or return the Spawning Stock Biomass

The calculated Spawning Stock Biomass (SSB) of a fish population is returned by this method. SSB is the combined weight of all individuals in a fish stock that are capable of reproducing. In some classes this is calculated from information stored in different slots, while in others ssb() is simply an slot accessor. When the later is the case, the corresponding replacement method also exists.

ssb(object, ...)

# S4 method for FLStock
ssb(object, ...)

# S4 method for FLBiol
ssb(object, ...)

Arguments

object
Object on which ssb is calculated or extracted.

Value

An object, generally of class FLQuant.

Details

For an object of class FLStock, the calculation of SSB depends on the value of the 'units' attribute in the harvest slot. If this is in terms of fishing mortality (units(harvest(object)) == 'f'), and assuming an object structured by age, then SSB is calculated as $$SSB_{y} = \sum\nolimits_{a} N_{a,y} \cdot e^{-(F_{a,y} \cdot Hs_{a,y} + M_{a,y} \cdot Ms_{a,y})} \cdot W_{a,y} \cdot T_{a,y} $$ where \(N_{a,y}\) is the abundance in numbers (stock.n) by age (a) and year (y), \(F_{a,y}\) is the fishing mortality (harvest), \(Hs_{a,y}\) is the proportion of fishing mortality before spawning (harvest.spwn), \(M_{a,y}\) is the natural mortality (m), \(Ms_{a,y}\) is the proportion of natural mortality before spawning (m.spwn), \(W_{a,y}\) is the mean weight at age in the stock (m), and \(T_{a,y}\) is the proportion mature at age in the stock (mat). For FLStock objects with other dimensions (area, unit), the calculation is carried out along those dimensions too. To obtain a global value please use the corresponding summing method.

If the harvest slot contains estimates in terms of harvest rates (units(harvest(object)) == "hr"), SSB will be computed as $$SSB_{y} = \sum\nolimits_{a} N_{a,y} \cdot (1 - H_{a,y} \cdot Hs_{a,y}) \cdot e^{-(M_{a,y} \cdot Ms_{a,y})} \cdot W_{a,y} \cdot T_{a,y} $$ where \(H_{a,y}\) is the harvest rate (proportion of catch in weight over total biomass).

For an object of class FLBiol, the calculation of SSB is as follows: $$SSB_{a,y} = \sum\nolimits_{a} N_{a,y} \cdot W_{a,y} \cdot e^{-S_{a,y} \cdot M_{a,y}}$$ where \(S_{a,y}\) is the fraction of natural mortality before spawning (spwn). The method for this class does not correct the estimate, based on abundances at the start of the time period, for fishing mortality ocurring between that time and spawning, as the class holds no information on fishing mortality.

See also

Examples

data(ple4)
ssb(ple4)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 274205 288540 296825 308164 321354 372863 370373 363077 344013 361549
#>      year
#> age   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 416563 402521 377432 333933 316343 319062 268714 278648 293136 310954
#>      year
#> age   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 316929 303433 297122 272416 262061 263998 314021 326341 348675 375392
#>      year
#> age   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 445855 391254 408489 368969 335747 269528 228668 193093 174408 173903
#>      year
#> age   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 185308 211327 184733 208393 234078 162725 179158 151508 167531 173783
#>      year
#> age   2007   2008  
#>   all 166061 206480
#> 
#> units:  t