Method summary

Outputs a general summary of the structure and content of the object. The particular output obtained depends on the class of the argument object.

# S4 method for FLArray
summary(object, ...)

Generic function

summary(object)

See also

summary

Examples

flq <- FLQuant(rlnorm(90), dim=c(3,10), units='kg')
summary(flq)

#> An object of class "FLQuant" with:
#> dim  :  3 10 1 1 1 1 
#> quant:  quant 
#> units:  kg 
#> 
#> Min    :  0.2058224 
#> 1st Qu.:  0.5695396 
#> Mean   :  1.695865 
#> Median :  0.8955741 
#> 3rd Qu.:  2.161326 
#> Max    :  8.377503 
#> NAs    :  0 %

data(ple4)
summary(ple4)

#> An object of class "FLStock"
#> 
#> Name: Plaice in IV 
#> Description: Imported from a VPA file. ( N:\Projecten\ICES WG\Demersale werkgroep WGNSS [...] 
#> Quant: age 
#> Dims:  age 	year	unit	season	area	iter
#> 	10	52	1	1	1	1	
#> 
#> Range:  min	max	pgroup	minyear	maxyear	minfbar	maxfbar 
#> 	1	10	10	1957	2008	2	6	
#> 
#> catch         : [ 1 52 1 1 1 1 ], units =  t 
#> catch.n       : [ 10 52 1 1 1 1 ], units =  10^3 
#> catch.wt      : [ 10 52 1 1 1 1 ], units =  kg 
#> discards      : [ 1 52 1 1 1 1 ], units =  t 
#> discards.n    : [ 10 52 1 1 1 1 ], units =  10^3 
#> discards.wt   : [ 10 52 1 1 1 1 ], units =  kg 
#> landings      : [ 1 52 1 1 1 1 ], units =  t 
#> landings.n    : [ 10 52 1 1 1 1 ], units =  10^3 
#> landings.wt   : [ 10 52 1 1 1 1 ], units =  kg 
#> stock         : [ 1 52 1 1 1 1 ], units =  t 
#> stock.n       : [ 10 52 1 1 1 1 ], units =  10^3 
#> stock.wt      : [ 10 52 1 1 1 1 ], units =  kg 
#> m             : [ 10 52 1 1 1 1 ], units =  m 
#> mat           : [ 10 52 1 1 1 1 ], units =  NA 
#> harvest       : [ 10 52 1 1 1 1 ], units =  f 
#> harvest.spwn  : [ 10 52 1 1 1 1 ], units =  NA 
#> m.spwn        : [ 10 52 1 1 1 1 ], units =  NA 

data(nsher)
summary(nsher)

#> An object of class "FLSR"
#> 
#> Name:  
#> Description:  
#> Quant: age 
#> Dims:  age 	year	unit	season	area	iter
#> 	1	45	1	1	1	1	
#> 
#> Range:  min	minyear	max	maxyear 
#> 	0	1960	0	2004	
#> 
#> rec           : [ 1 45 1 1 1 1 ], units =  10^3 
#> ssb           : [ 1 45 1 1 1 1 ], units =  t*10^3 
#> residuals     : [ 1 45 1 1 1 1 ], units =  NA 
#> fitted        : [ 1 45 1 1 1 1 ], units =  10^3 
#> 
#> Model: 	rec ~ a * ssb * exp(-b * ssb)
#> <environment: 0x679e118>
#> Parameters: 
#>     params
#> iter     a        b
#>    1 119.4 0.009451
#> 
#> Log-likelihood:  15.862(0) 
#> Variance-covariance:    
#>               a            b
#>   a 255.3388181 1.808870e-02
#>   b   0.0180887 1.992659e-06