Calculate the survivors of a stock to the next year.

An FLStock object containing estimates of adundance at age ('stock.n') and harvest level at age ('harvest'), is used to bring forward the population by applying the total mortality at age ('z'). No calculation is made on recruitment, so abundances for the first age will be set as 'NA'.

survivors(object)

Arguments

object
An FLStock with estimated harvest and abundances

Value

The abundances at age of the survivors, 'FLQuant'.

Examples

data(ple4)
stock.n(ple4[, ac(2002:2006)])
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  2002      2003      2004      2005      2006     
#>   1  1792880.0  557844.0 1235790.0  863893.0  875191.0
#>   2   481832.0 1362670.0  417612.0  894667.0  618023.0
#>   3   413685.0  272903.0  729522.0  238706.0  552652.0
#>   4   181506.0  185801.0  118238.0  365617.0  136058.0
#>   5    98489.7   90505.8   97419.1   66113.3  219194.0
#>   6   106479.0   49763.0   49696.3   57107.9   41505.0
#>   7    13278.2   51718.8   26043.7   29565.3   37777.6
#>   8     7752.1    7045.7   28478.9   16720.1   21233.8
#>   9     3675.4    4964.1    4649.5   20533.5   12941.5
#>   10    9308.9    9758.3   11423.5   12764.9   27199.2
#> 
#> units:  1000 
survivors(ple4[, ac(2002:2006)])
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  2003      2004      2005      2006      2007     
#>   1         NA        NA        NA        NA        NA
#>   2  1362673.1  417612.2  894668.0  618023.2  651006.8
#>   3   272903.6  729521.1  238706.3  552652.0  393106.5
#>   4   185801.5  118237.7  365617.1  136057.6  326369.0
#>   5    90505.7   97418.9   66113.3  219193.4   87128.7
#>   6    49763.0   49696.3   57107.9   41505.0  148406.1
#>   7    51718.9   26043.7   29565.3   37777.6   29245.6
#>   8     7045.7   28478.9   16720.1   21233.8   28310.9
#>   9     4964.1    4649.5   20533.5   12941.5   17026.1
#>   10    9758.3   11423.5   12764.9   27199.1   33875.0
#> 
#> units:  1000