Total mortality z

Returns the calculation of total mortality, z, usually as the sum of fishing mortality, f, and natural mortality, m.

z(object, ...)

# S4 method for FLS
z(object, ...)

Arguments

object
Object to calculate on.

Value

An object of the corresponding class, usually FLQuant.

See also

Examples

data(ple4)

z(ple4)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957    1958    1959    1960    1961    1962    1963    1964    1965   
#>   1  0.17718 0.20501 0.25239 0.20830 0.19687 0.19593 0.24948 0.13160 0.16844
#>   2  0.32858 0.34987 0.41033 0.41809 0.38915 0.41867 0.46387 0.49933 0.36687
#>   3  0.35534 0.40217 0.45488 0.45332 0.44447 0.47328 0.51845 0.54816 0.49698
#>   4  0.40399 0.45799 0.47643 0.48386 0.45679 0.49836 0.53373 0.56874 0.51241
#>   5  0.44650 0.47385 0.51165 0.46602 0.51720 0.53400 0.52337 0.64005 0.45530
#>   6  0.30846 0.42141 0.48340 0.51851 0.43019 0.52630 0.57437 0.58762 0.60822
#>   7  0.37360 0.36802 0.45047 0.48263 0.46063 0.46231 0.55042 0.50310 0.51680
#>   8  0.41429 0.39072 0.40949 0.45863 0.53695 0.45007 0.55248 0.48976 0.45196
#>   9  0.39008 0.42325 0.46735 0.48307 0.48148 0.49541 0.54836 0.55941 0.51022
#>   10 0.39008 0.42325 0.46735 0.48307 0.48148 0.49541 0.54836 0.55941 0.51022
#>     year
#> age  1966    1967    1968    1969    1970    1971    1972    1973    1974   
#>   1  0.17146 0.15437 0.29731 0.24853 0.32289 0.29566 0.33209 0.21323 0.32065
#>   2  0.45613 0.45202 0.38721 0.41267 0.53489 0.43191 0.48100 0.49442 0.49863
#>   3  0.48766 0.50505 0.44373 0.42692 0.59197 0.48777 0.50095 0.53320 0.59091
#>   4  0.53037 0.50828 0.46082 0.44059 0.60476 0.48820 0.51338 0.64176 0.61532
#>   5  0.57722 0.57634 0.46623 0.41467 0.56235 0.50734 0.51911 0.64516 0.69646
#>   6  0.44343 0.60445 0.42307 0.52798 0.60431 0.49509 0.54345 0.51301 0.55227
#>   7  0.60594 0.40982 0.50953 0.39454 0.69352 0.52758 0.50800 0.48714 0.49428
#>   8  0.50894 0.53549 0.38859 0.49927 0.36070 0.51176 0.57823 0.58016 0.56476
#>   9  0.53459 0.52825 0.45063 0.45642 0.56672 0.50726 0.53384 0.57509 0.58632
#>   10 0.53459 0.52825 0.45063 0.45642 0.56672 0.50726 0.53384 0.57509 0.58632
#>     year
#> age  1975    1976    1977    1978    1979    1980    1981    1982    1983   
#>   1  0.45519 0.43307 0.42330 0.40437 0.52678 0.33792 0.27729 0.34179 0.33736
#>   2  0.60120 0.50722 0.57145 0.52849 0.73668 0.56921 0.58466 0.61506 0.61983
#>   3  0.63052 0.52605 0.59456 0.56399 0.76529 0.76383 0.67661 0.79420 0.66222
#>   4  0.65647 0.53220 0.59934 0.57083 0.77418 0.72169 0.69472 0.77451 0.84488
#>   5  0.69969 0.48316 0.76513 0.56063 0.78266 0.60665 0.68104 0.69270 0.70586
#>   6  0.71776 0.53345 0.51986 0.61915 0.80679 0.61640 0.54822 0.61912 0.61389
#>   7  0.57718 0.61796 0.54077 0.56115 0.80339 0.59151 0.60328 0.57867 0.55776
#>   8  0.60290 0.55223 0.63270 0.52630 0.70439 0.60047 0.63194 0.63353 0.56980
#>   9  0.65291 0.54527 0.61343 0.56922 0.77722 0.62931 0.63383 0.66187 0.66060
#>   10 0.65291 0.54527 0.61343 0.56922 0.77722 0.62931 0.63383 0.66187 0.66060
#>     year
#> age  1984    1985    1986    1987    1988    1989    1990    1991    1992   
#>   1  0.39986 0.36206 0.38581 0.32080 0.33190 0.31133 0.26185 0.33973 0.31486
#>   2  0.65295 0.57236 0.70826 0.74783 0.73631 0.68104 0.57478 0.70861 0.65775
#>   3  0.68503 0.59368 0.73063 0.77781 0.77353 0.73822 0.67201 0.76347 0.75883
#>   4  0.69190 0.80085 0.73806 0.82484 0.77139 0.74337 0.71438 0.79774 0.78068
#>   5  0.73513 0.56833 0.82264 0.88161 0.80098 0.73364 0.83094 0.85319 0.89287
#>   6  0.64294 0.59954 0.76972 0.74772 0.79431 0.69848 0.67593 0.93672 0.88202
#>   7  0.56179 0.60882 0.65558 0.79090 0.75961 0.72222 0.57832 0.74713 0.88626
#>   8  0.66728 0.55604 0.75212 0.58338 0.82766 0.72605 0.57894 0.67337 0.73520
#>   9  0.66198 0.62867 0.75038 0.76857 0.79385 0.72735 0.67796 0.80491 0.84076
#>   10 0.66198 0.62867 0.75038 0.76857 0.79385 0.72735 0.67796 0.80491 0.84076
#>     year
#> age  1993    1994    1995    1996    1997    1998    1999    2000    2001   
#>   1  0.32426 0.26665 0.22263 0.19654 0.16686 0.25523 0.27509 0.22808 0.17344
#>   2  0.59040 0.59759 0.57205 0.65411 0.90317 0.60944 0.58560 0.47156 0.90660
#>   3  0.71357 0.71945 0.77625 0.82215 1.05455 1.12425 0.64599 0.44600 1.11402
#>   4  0.76010 0.83283 0.89196 0.90498 0.92785 1.27016 1.36305 0.74735 1.05401
#>   5  0.86583 0.75929 0.89020 0.89599 1.06751 0.89247 0.89007 0.77714 1.09924
#>   6  0.80772 0.77791 0.74174 0.83780 0.92630 0.81531 0.87850 0.84770 0.70449
#>   7  0.82355 1.00789 0.81354 0.97591 0.84746 0.83314 0.78731 0.70436 0.93998
#>   8  0.80099 0.91497 0.72691 0.86184 0.85565 0.76150 0.87318 0.73526 0.66449
#>   9  0.81923 0.86485 0.80926 0.96529 1.03517 1.02880 1.09151 0.87571 0.83719
#>   10 0.81923 0.86485 0.80926 0.96529 1.03517 1.02880 1.09151 0.87571 0.83719
#>     year
#> age  2002    2003    2004    2005    2006    2007    2008   
#>   1  0.32660 0.25329 0.33788 0.25313 0.43283 0.19106 0.28528
#>   2  0.73370 0.80690 0.81806 0.67229 0.72109 0.68590 0.53596
#>   3  0.75474 0.81713 0.69474 0.67584 0.68440 0.63729 0.47826
#>   4  0.79014 0.82606 0.77992 0.64192 0.65990 0.43598 0.49447
#>   5  0.90234 0.91559 0.57851 0.87919 0.53225 0.57639 0.37777
#>   6  0.80722 0.96415 0.82954 0.74791 0.64057 0.48403 0.39510
#>   7  1.06795 1.21249 0.68955 0.91941 0.56078 0.46091 0.50496
#>   8  0.97470 1.04032 0.69879 0.75505 0.95375 0.60470 0.49337
#>   9  0.90174 0.92421 0.68284 0.68364 0.75957 0.54649 0.25002
#>   10 0.90174 0.92421 0.68284 0.68364 0.75957 0.54649 0.25002
#> 
#> units:  z