Arithmetic operators for FLCore classes

Overloaded arithmetic operators for FLCore classes

# S4 method for numeric,FLArray
Arith(e1, e2)

# S4 method for FLArray,numeric
Arith(e1, e2)

# S4 method for FLArray,FLArray
Arith(e1, e2)

# S4 method for FLPar,FLPar
Arith(e1, e2)

# S4 method for FLArray,FLPar
Arith(e1, e2)

# S4 method for FLPar,FLArray
Arith(e1, e2)

Details

These methods apply the standard arithmetic operators included in the Arith group ("+", "-", "*", "^", "%%", "%/%", and "/"), so that they return an object of the appropriate class.

When the operation involves objects of two classes (e.g. FLPar and FLQuant), the class is the returned object is that of the more complexs object, in this case FLQuant.

See also

Author

The FLR Team

Examples

flq <- FLQuant(rlnorm(90), dim=c(3,10), units='kg')
flp <- FLPar(a=99)

# FLQuant and numeric
flq * 25
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1        2        3        4        5        6        7        8       
#>     1   3.4180  29.3547  13.5309   8.5768  11.4102  22.4622  49.6507   2.2945
#>     2  60.0344 110.1478   4.2137  29.9429  11.1538  27.2760  34.5170  18.0131
#>     3  23.6438  12.6448  20.9067  28.0722  45.8577   8.2772  19.2940  28.8275
#>      year
#> quant 9        10      
#>     1  22.0649  15.7642
#>     2   5.9338  58.6280
#>     3   9.4143 208.4768
#> 
#> units:  kg 
# Two FLQuant objects
flq + flq
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1        2        3        4        5        6        7        8       
#>     1  0.27344  2.34837  1.08247  0.68615  0.91282  1.79698  3.97206  0.18356
#>     2  4.80275  8.81182  0.33710  2.39543  0.89230  2.18208  2.76136  1.44105
#>     3  1.89151  1.01159  1.67253  2.24578  3.66861  0.66218  1.54352  2.30620
#>      year
#> quant 9        10      
#>     1  1.76519  1.26114
#>     2  0.47470  4.69024
#>     3  0.75314 16.67814
#> 
#> units:  kg 

# FLQuant and FLPar
flq / flp
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1          2          3          4          5          6         
#>     1 0.00138102 0.01186048 0.00546704 0.00346538 0.00461018 0.00907565
#>     2 0.02425632 0.04450415 0.00170252 0.01209814 0.00450657 0.01102063
#>     3 0.00955306 0.00510903 0.00844713 0.01134230 0.01852835 0.00334434
#>      year
#> quant 7          8          9          10        
#>     1 0.02006090 0.00092708 0.00891511 0.00636937
#>     2 0.01394627 0.00727804 0.00239749 0.02368808
#>     3 0.00779557 0.01164749 0.00380375 0.08423305
#> 
#> units:  kg