Extract

Extract or replace parts of an FLR Object

# S4 method for FLArray,ANY,ANY,ANY
[(x, i, j, k, l, m, n, ..., drop = FALSE)

# S4 method for FLArray,array,missing,missing
[(x, i)

# S4 method for FLArray,ANY,ANY,ANY
[(x, i, j, k, l, m, n, ...) <- value

# S4 method for FLArray,ANY,ANY,FLArray
[(x, i, j, k, l, m, n, ...) <- value

Arguments

x

object from which to extract or replace element(s)

i, j, k, l, m, n, ...

indices specifying elements to extract or replace.

drop

If 'TRUE' the result is coerced to the lowest possible dimension, and so might change class (e.g. drop='TRUE' on an FLQuant might return an array of less dimensions, a matrix or a vector.

value

An object of a similar or simpler class than 'x'.

name

See Extract for further details.

Details

Operators acting on FLQuant, FLCohort, FLPar, FLComp, and derived classes to extract or replace sections of an object.

Please note the differences between referencing sections of an object by position using values of class numeric, or by using dimnames of class character. See examples below.

All classes that are derived from FLComp (for example, FLStock and FLBiol) can be subset along the six dimensions of their FLQuant slots.

Classes that are derived from FLlst (for example, FLStocks and FLBiols) can be subset in a similar way to ordinary list objects.

Generic function

[x,i,j,drop]

[<-(x,i,j,value)

[[<-(x,i,j,value)

\$<-(x,name,value)

See also

Extract

Examples

flq <- FLQuant(rnorm(200), dimnames=list(age=0:4, year=1991:2000,
  season=1:4))

# Extracting by position...
  flq[1,]

#> An object of class "FLQuant"
#> , , unit = unique, season = 1, area = unique
#> 
#>    year
#> age 1991      1992      1993      1994      1995      1996      1997     
#>   0 -1.541030 -1.482813  0.264662 -0.812275  1.016713 -0.681273 -0.046401
#>    year
#> age 1998      1999      2000     
#>   0 -0.919332  0.128688  1.399004
#> 
#> , , unit = unique, season = 2, area = unique
#> 
#>    year
#> age 1991      1992      1993      1994      1995      1996      1997     
#>   0  0.549900 -1.037146 -0.961277 -0.452426  0.542144  1.060396  2.169344
#>    year
#> age 1998      1999      2000     
#>   0 -0.416199 -0.218696  1.608624
#> 
#> , , unit = unique, season = 3, area = unique
#> 
#>    year
#> age 1991      1992      1993      1994      1995      1996      1997     
#>   0 -1.744239  1.392394  1.214707  0.743228 -0.506697  0.521976 -0.812119
#>    year
#> age 1998      1999      2000     
#>   0  0.530264 -1.122882 -0.680955
#> 
#> , , unit = unique, season = 4, area = unique
#> 
#>    year
#> age 1991      1992      1993      1994      1995      1996      1997     
#>   0 -0.797153  0.539466  0.164970  0.189213 -0.556918  0.900040 -1.377776
#>    year
#> age 1998      1999      2000     
#>   0  1.675011 -0.415795  0.309581
#> 
#> units:  NA 

  flq[,1:5]

#> An object of class "FLQuant"
#> , , unit = unique, season = 1, area = unique
#> 
#>    year
#> age 1991      1992      1993      1994      1995     
#>   0 -1.541030 -1.482813  0.264662 -0.812275  1.016713
#>   1 -0.310310  0.445750  0.303848 -1.640582  0.121621
#>   2 -0.020108  1.369776 -0.183388  0.507922 -1.078067
#>   3 -2.390200 -0.020110  0.559650  1.754337 -1.143566
#>   4  0.889865 -0.109218 -0.186554  0.592400 -0.529644
#> 
#> , , unit = unique, season = 2, area = unique
#> 
#>    year
#> age 1991      1992      1993      1994      1995     
#>   0  0.549900 -1.037146 -0.961277 -0.452426  0.542144
#>   1  0.867817  0.830148  0.255082 -0.439291 -2.290785
#>   2  0.684360  0.607347 -0.545402 -0.616223  0.310352
#>   3 -0.162680 -0.122186  0.930361  0.441635  1.404079
#>   4 -1.784365  0.933125 -0.537651  0.482597  1.377118
#> 
#> , , unit = unique, season = 3, area = unique
#> 
#>    year
#> age 1991      1992      1993      1994      1995     
#>   0 -1.744239  1.392394  1.214707  0.743228 -0.506697
#>   1 -0.970074  1.360076  0.506959  0.557154 -0.611523
#>   2 -0.229027 -0.365416 -2.094971 -1.248583  2.411659
#>   3 -0.974846 -0.868937  0.034079 -0.207874 -0.164958
#>   4  2.692372 -0.506557  0.852729  0.423969 -0.440406
#> 
#> , , unit = unique, season = 4, area = unique
#> 
#>    year
#> age 1991      1992      1993      1994      1995     
#>   0 -0.797153  0.539466  0.164970  0.189213 -0.556918
#>   1 -0.915921 -1.061234 -0.727408  0.248949  0.302427
#>   2  0.073267 -0.540738 -0.537837  1.046643  0.222266
#>   3  1.238171 -0.715205 -0.599665 -1.289301 -0.962167
#>   4  0.718371 -0.178920  1.182088  0.375116  0.052032
#> 
#> units:  NA 

  flq[1:2,,,c(1,3)]

#> An object of class "FLQuant"
#> , , unit = unique, season = 1, area = unique
#> 
#>    year
#> age 1991      1992      1993      1994      1995      1996      1997     
#>   0 -1.541030 -1.482813  0.264662 -0.812275  1.016713 -0.681273 -0.046401
#>   1 -0.310310  0.445750  0.303848 -1.640582  0.121621 -0.202448 -0.953629
#>    year
#> age 1998      1999      2000     
#>   0 -0.919332  0.128688  1.399004
#>   1  0.901335 -1.533065  0.372699
#> 
#> , , unit = unique, season = 3, area = unique
#> 
#>    year
#> age 1991      1992      1993      1994      1995      1996      1997     
#>   0 -1.744239  1.392394  1.214707  0.743228 -0.506697  0.521976 -0.812119
#>   1 -0.970074  1.360076  0.506959  0.557154 -0.611523 -1.918322  0.910748
#>    year
#> age 1998      1999      2000     
#>   0  0.530264 -1.122882 -0.680955
#>   1 -0.310975  2.053939 -1.488829
#> 
#> units:  NA 

# ...by dimnames
  flq['0',]

#> An object of class "FLQuant"
#> , , unit = unique, season = 1, area = unique
#> 
#>    year
#> age 1991      1992      1993      1994      1995      1996      1997     
#>   0 -1.541030 -1.482813  0.264662 -0.812275  1.016713 -0.681273 -0.046401
#>    year
#> age 1998      1999      2000     
#>   0 -0.919332  0.128688  1.399004
#> 
#> , , unit = unique, season = 2, area = unique
#> 
#>    year
#> age 1991      1992      1993      1994      1995      1996      1997     
#>   0  0.549900 -1.037146 -0.961277 -0.452426  0.542144  1.060396  2.169344
#>    year
#> age 1998      1999      2000     
#>   0 -0.416199 -0.218696  1.608624
#> 
#> , , unit = unique, season = 3, area = unique
#> 
#>    year
#> age 1991      1992      1993      1994      1995      1996      1997     
#>   0 -1.744239  1.392394  1.214707  0.743228 -0.506697  0.521976 -0.812119
#>    year
#> age 1998      1999      2000     
#>   0  0.530264 -1.122882 -0.680955
#> 
#> , , unit = unique, season = 4, area = unique
#> 
#>    year
#> age 1991      1992      1993      1994      1995      1996      1997     
#>   0 -0.797153  0.539466  0.164970  0.189213 -0.556918  0.900040 -1.377776
#>    year
#> age 1998      1999      2000     
#>   0  1.675011 -0.415795  0.309581
#> 
#> units:  NA 

  flq[,'1991']

#> An object of class "FLQuant"
#> , , unit = unique, season = 1, area = unique
#> 
#>    year
#> age 1991     
#>   0 -1.541030
#>   1 -0.310310
#>   2 -0.020108
#>   3 -2.390200
#>   4  0.889865
#> 
#> , , unit = unique, season = 2, area = unique
#> 
#>    year
#> age 1991     
#>   0  0.549900
#>   1  0.867817
#>   2  0.684360
#>   3 -0.162680
#>   4 -1.784365
#> 
#> , , unit = unique, season = 3, area = unique
#> 
#>    year
#> age 1991     
#>   0 -1.744239
#>   1 -0.970074
#>   2 -0.229027
#>   3 -0.974846
#>   4  2.692372
#> 
#> , , unit = unique, season = 4, area = unique
#> 
#>    year
#> age 1991     
#>   0 -0.797153
#>   1 -0.915921
#>   2  0.073267
#>   3  1.238171
#>   4  0.718371
#> 
#> units:  NA 

  flq[,as.character(1991:1995),,'1']

#> An object of class "FLQuant"
#> , , unit = unique, season = 1, area = unique
#> 
#>    year
#> age 1991      1992      1993      1994      1995     
#>   0 -1.541030 -1.482813  0.264662 -0.812275  1.016713
#>   1 -0.310310  0.445750  0.303848 -1.640582  0.121621
#>   2 -0.020108  1.369776 -0.183388  0.507922 -1.078067
#>   3 -2.390200 -0.020110  0.559650  1.754337 -1.143566
#>   4  0.889865 -0.109218 -0.186554  0.592400 -0.529644
#> 
#> units:  NA 

# Dimensions of length one can be drop
  flq[1, drop=TRUE]

#>       season
#> year         1       2       3       4
#>   1991 -1.5410  0.5499 -1.7442 -0.7972
#>   1992 -1.4828 -1.0371  1.3924  0.5395
#>   1993  0.2647 -0.9613  1.2147  0.1650
#>   1994 -0.8123 -0.4524  0.7432  0.1892
#>   1995  1.0167  0.5421 -0.5067 -0.5569
#>   1996 -0.6813  1.0604  0.5220  0.9000
#>   1997 -0.0464  2.1693 -0.8121 -1.3778
#>   1998 -0.9193 -0.4162  0.5303  1.6750
#>   1999  0.1287 -0.2187 -1.1229 -0.4158
#>   2000  1.3990  1.6086 -0.6810  0.3096

# Replacing part of the object
  flq['0',,,1]<-0