Class FLComp

A virtual class that forms the basis for most FLR classes composed of slots of class FLQuant. No objects of this class can be constructed.

Validity

Dimensions

All FLQuant slots must have iters equal to 1 or 'n'.

Iters

The dimname for iter[1] should be '1'.

Dimnames

The name of the quant dimension must be the same for all FLQuant slots.

Slots

name

A character vector for the object name.

desc

A textual description of the object contents.

range

A named numeric vector with various values of quant and year ranges, plusgroup, fishing mortality ranges, etc. Elements are specific to each child class.

See also

Author

The FLR Team