Class FLPar

A class for storing parameters of a model.

FLPar(object, ...)

Details

The FLPar class is based on the array class which can store Monte Carlo samples and the names of the relevant parameter vectors.

Methods for this class include subsetting and replacement as for the FLQuant class. There are methods for extracting statistics of the sample (mean, median etc.) and for plotting the parameter samples.

Slots

.Data

Describe slot. array.

units

Units of measurement. character.

See also

Author

The FLR Team

Examples

FLPar(rnorm(4), params=c('a','b','c','sigma2'))
#> An object of class "FLPar"
#> params
#>      a      b      c sigma2 
#> -1.383  1.298  2.713 -0.196 
#> units:  NA NA NA NA 

FLPar(rnorm(20), dimnames=list(params=c('a','b'), year=1990:1999, iter=1),
  units='NA')
#> An object of class "FLPar"
#>       year
#> params 1990    1991    1992    1993    1994    1995    1996    1997    1998   
#>      a  1.9671 -0.3550  0.0794 -0.9378  0.6859 -0.3132 -0.2542  1.3587 -1.3267
#>      b -2.0277  1.2428 -1.7046  0.6188 -1.9386  1.1990 -0.2265 -1.0878 -2.2980
#>       year
#> params 1999   
#>      a -0.3214
#>      b  0.7867
#> units:  NA 

# with iters
  FLPar(rnorm(80), params=c('a', 'b'), iter=1:40)
#> An object of class "FLPar"
#> iters:  40 
#> 
#> params
#>              a              b 
#> -0.19217(1.11)  0.10107(0.88) 
#> units:  NA NA