Class FLPar

A class for storing parameters of a model.

FLPar(object, ...)

Details

The FLPar class is based on the array class which can store Monte Carlo samples and the names of the relevant parameter vectors.

Methods for this class include subsetting and replacement as for the FLQuant class. There are methods for extracting statistics of the sample (mean, median etc.) and for plotting the parameter samples.

Slots

.Data

Describe slot. array.

units

Units of measurement. character.

See also

[, [<-, as.data.frame, densityplot, histogram, iter, iter<-, mean, median, plot, splom, summary, units,FLPar-method, units<-,FLPar,character-method, var

Examples

FLPar(rnorm(4), params=c('a','b','c','sigma2'))

#> An object of class "FLPar"
#> params
#>       a       b       c  sigma2 
#> 0.90790 0.45600 0.89975 1.53001 
#> units:  NA NA NA NA 

FLPar(rnorm(20), dimnames=list(params=c('a','b'), year=1990:1999, iter=1),
  units='NA')

#> An object of class "FLPar"
#>       year
#> params 1990      1991      1992      1993      1994      1995      1996     
#>      a -0.094239  0.875055 -0.117643  0.401663 -0.415283 -0.090910 -0.086230
#>      b -1.134575  1.295103  0.534351 -0.353550  1.936239 -0.296635  0.270238
#>       year
#> params 1997      1998      1999     
#>      a  1.027804  0.366421 -1.136145
#>      b  0.257960  1.468820 -0.492673
#> units:  NA 

# with iters
  FLPar(rnorm(80), params=c('a', 'b'), iter=1:40)

#> An object of class "FLPar"
#> iters:  40 
#> 
#> params
#>              a              b 
#> 0.11920(0.872) 0.33913(0.658) 
#> units:  NA NA