A class for jackknifing fisheries data

This extended FLQuant class holds both a jackknifed FLQuant, one in which each iter is missing one element, and the original object, as a separate FLQuant in the orig slot.

# S4 method for ANY
FLQuantJK(object, orig)

# S4 method for FLQuantJK
orig(object)

# S4 method for FLQuants
orig(object)

Arguments

object
Input numeric object

...
Additonal arguments

Slots

.Data

Unnamed slot containing the jackknifed object(FLQuant).

orig

Original object, (FLQuant).

Validity

slot dims

.Data and orig slots must have the same dimensions 1-5.

slot dimnames

.Data and var slots must have the same dimnames 1-5.

You can inspect the class validity function by using getValidity(getClassDef('FLQuantJK'))

Accessors

All slots in the class have accessor and replacement methods defined that allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot. A numeric vector can also be used when replacing FLQuant slots, and the vector will be used to substitute the values in the slot, but not its other attributes.

Constructor

Objects of this class must be constructed from an FLQuant that is to be jackknifed, through the jackknife method.

Methods

All methods defined for the FLQuant class are available, but they will operate only on the jackknifed (.Data) slot. Please use orig() to apply them to the original object stored in the class.

See also

Examples

data(ple4)
fjk <- jackknife(stock(ple4))
# New object has as many iters as length of jackknifed dimension (defaults to 'year')
dim(fjk)

#> [1]  1 52  1  1  1 52