Class FLQuantPoint

The FLQuantPoint class summarizes the contents of an FLQuant object with multiple iterations along its sixth dimension using a number of descriptive statistics.

Usage

FLQuantPoint(object, ...)

# S4 method for missing
FLQuantPoint(..., units = "NA", n = 1)

# S4 method for FLQuant
FLQuantPoint(object, ..., probs = c(0.25, 0.75))

# S4 method for FLQuantPoint
n(object, ...)

Arguments

object

Input numeric object

...

Additonal arguments

Details

An object of this class has a set structure along its sixth dimension (iter), which will always be of length 5, and with dimnames mean, median, var, uppq and lowq. They refer, respectively, to the sample mean, sample median, variance, and lower (0.25) and upper (0.75) quantiles.

Objects of this class wil be typically created from an FLQuant. The various statistics are calculated along the iter dimension of the original FLQuant using apply.

Slots

.Data

The main array holding the computed statistics. array.

units

Units of measurement. character.

Accesors

mean,mean<-:

'mean' element on 6th dimension, arithmetic mean.

median,median<-:

'median' element on 6th dimension, median.

var,var<-:

'var' element on 6th dimension, variance.

lowq,lowq<-:

'lowq' element on 6th dimension, lower quantile (0.25 by default).

uppq,uppq<-:

'uppq' element on 6th dimension, upper quantile (0.75 by default).

quantile:

returns the 'lowq' or 'uppq' iter, depending on the value of 'probs' (0.25 or 0.75).

Constructor

Inputs can be of class:

FLQuant:

An FLQuant object with iters (i.e. dim6 > 1)

Validity

iter:

iter dimension is of length 5.

Dimnames:

iter dimnames are 'mean', 'median', 'var', 'uppq' and'lowq'

See also

Author

The FLR Team

Examples

flq <- FLQuant(rlnorm(2000), dim=c(10,20,1,1,1,200), units="kg")
flqp <- FLQuantPoint(flq)
flqp <- FLQuantPoint(flq, probs=c(0.05, 0.95))
summary(flqp)
#> An object of class "FLQuantPoint" with:
#> dim  :  10 20 1 1 1 5 
#> quant:  quant 
#> units:  kg 
#> 
#> 1st Qu.:  0.2798872 
#> Mean   :  1.652017 
#> Median :  1.072711 
#> Var    :  4.229131 
#> 3rd Qu.:  5.681206 
mean(flqp)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1       2       3       4       5       6       7       8       9      
#>    1  1.64518 3.29503 0.91802 2.28167 1.79148 1.21329 1.04418 1.31054 1.59807
#>    2  1.30500 2.70891 1.51304 0.65337 1.26083 2.03442 1.24816 1.84421 1.78849
#>    3  1.72608 1.28226 1.87914 0.90563 2.98732 1.09821 1.56460 1.74708 1.50160
#>    4  1.46081 1.68573 3.36316 1.14047 2.53486 1.33153 1.89569 0.83789 2.07323
#>    5  1.11749 1.92763 0.86117 1.18337 2.33317 1.50922 1.32713 1.04796 1.39750
#>    6  2.25480 2.32396 2.29371 1.38517 1.59304 2.20574 1.33174 1.14754 1.08192
#>    7  1.59368 1.02706 1.27412 2.13379 1.03769 2.39800 1.73798 1.41104 1.56261
#>    8  1.12317 1.20965 1.30224 1.60504 2.38692 1.54433 0.92678 2.15185 1.90446
#>    9  2.00305 1.31471 1.76567 2.12799 2.20142 1.83648 1.55606 1.49897 1.01159
#>    10 2.78900 1.57454 1.84430 1.03200 1.67608 1.44640 0.90141 2.08572 1.48879
#>      year
#> quant 10      11      12      13      14      15      16      17      18     
#>    1  1.56385 0.69410 1.84355 1.17989 0.78158 2.42777 1.65084 0.94388 2.09289
#>    2  1.10761 0.89736 1.22099 1.89994 1.76549 4.46254 2.02941 1.08578 1.65101
#>    3  1.73257 2.27441 2.05426 2.10051 0.94218 1.26533 1.35407 1.87850 2.98190
#>    4  1.27797 1.93892 1.19289 1.01922 1.61901 2.16790 1.68390 1.58832 1.03811
#>    5  0.92071 1.80061 1.95344 1.86233 1.21863 1.66589 1.40279 1.68220 1.61940
#>    6  1.72523 1.69635 1.14565 0.96003 2.07596 1.55076 0.96603 1.66420 0.82132
#>    7  2.36870 1.03941 2.61099 1.93239 1.36250 1.85875 3.44759 2.26322 1.84290
#>    8  1.37974 1.57554 1.53029 1.66965 3.04222 0.90560 1.36770 2.68153 2.11631
#>    9  0.86522 3.63644 1.06735 1.15863 1.99009 2.30144 1.29083 2.15369 1.14241
#>    10 1.32750 1.20672 1.79150 1.39005 1.80739 0.89218 1.79581 1.45571 0.82504
#>      year
#> quant 19      20     
#>    1  0.92426 1.09014
#>    2  1.88627 1.64884
#>    3  1.51402 1.35284
#>    4  1.36715 1.42840
#>    5  0.98822 2.28207
#>    6  4.54409 1.28512
#>    7  1.25515 0.91270
#>    8  1.19856 2.69694
#>    9  2.24212 2.63695
#>    10 1.17621 1.46418
#> 
#> units:  kg 
var(flqp)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1          2          3          4          5          6         
#>    1    0.773828  33.013582   0.228222   8.754118   9.362816   1.574253
#>    2    0.350902   3.593010   2.171765   0.132745   1.630882   4.395036
#>    3    4.202308   1.563639   2.142253   0.521406   4.429209   0.809702
#>    4    3.482882   2.885003  37.890217   0.728480   8.297854   4.574640
#>    5    0.606156   2.881837   0.491559   1.439986   4.687646   1.141950
#>    6    3.747210   3.940274  11.886978   2.625321   3.616634   5.002037
#>    7    0.620601   0.394341   1.677674   1.714849   1.000042   4.465067
#>    8    0.536856   1.124677   1.146664   1.535522   8.600610   2.507955
#>    9    3.370610   0.533425   4.943101   3.000160   8.153028   1.129520
#>    10   6.223457   1.315425   3.192170   0.673410   1.813376   1.317185
#>      year
#> quant 7          8          9          10         11         12        
#>    1    0.931391   1.329629   1.736199   1.371791   0.235844   4.142285
#>    2    1.396252   2.164757   1.011559   1.336895   0.963449   0.848331
#>    3    0.828194   1.675462   0.908875   5.756017  10.346841   2.733438
#>    4    6.455043   0.323461   4.674825   2.081296   3.852284   1.478778
#>    5    0.637611   0.190585   2.216282   0.982077   2.057045   4.307113
#>    6    0.272101   0.605174   1.346165   6.373578   1.262074   0.309458
#>    7    2.273265   0.995984   2.489106   6.969697   1.369903  18.785591
#>    8    2.678302   8.404959   3.326250   2.532460   2.024053   1.246508
#>    9    1.602273   0.887670   0.426192   0.320899  14.333967   1.322233
#>    10   0.521493   9.489667   1.572506   1.508843   0.757442   5.428854
#>      year
#> quant 13         14         15         16         17         18        
#>    1    0.847675   0.359355  13.193909   2.034201   0.268396   2.864472
#>    2    6.730991   1.750119  13.059929   1.491208   1.240315   1.244067
#>    3    3.370519   0.566554   1.775179   0.767225   3.671908   5.531114
#>    4    0.326756   2.823079   4.319991   5.439551   1.434438   0.645693
#>    5    1.813838   0.271813   2.077980   1.269502   3.237958   4.431929
#>    6    0.576954   5.800504   1.424055   0.808698   1.213016   0.209303
#>    7    5.524967   1.411770   2.656463  40.457672   4.629545   1.698692
#>    8    1.370653   8.769808   0.212540   1.318109  14.448975   1.382577
#>    9    1.482416   7.974917  14.192154   0.768127   7.079203   0.595614
#>    10   3.127598   4.492923   0.268376   2.294440   1.433046   0.258266
#>      year
#> quant 19         20        
#>    1    0.098087   0.679265
#>    2    3.594618   0.751188
#>    3    0.997108   1.313655
#>    4    1.335136   0.699928
#>    5    0.399927  15.457441
#>    6  112.211678   1.547067
#>    7    1.330136   0.188752
#>    8    0.843808  37.126654
#>    9    2.770756  23.788691
#>    10   1.374000   1.600941
#> 
#> units:  kg 
rnorm(200, flqp)
#> An object of class "FLQuant"
#> iters:  200 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1               2               3               4              
#>    1  1.67773( 1.019) 2.79742( 5.173) 0.92174( 0.451) 2.75226( 2.953)
#>    2  1.20156( 0.499) 2.62503( 1.805) 1.35318( 1.395) 0.69246( 0.392)
#>    3  1.25338( 1.852) 1.38833( 1.061) 1.81636( 1.330) 0.81357( 0.744)
#>    4  1.59239( 1.823) 1.67853( 1.733) 3.17443( 6.112) 1.07165( 0.748)
#>    5  1.07626( 0.794) 1.97588( 1.668) 0.81847( 0.588) 1.06652( 1.349)
#>    6  1.93827( 1.650) 2.29044( 2.097) 2.51356( 3.547) 1.32694( 1.439)
#>    7  1.59633( 0.760) 1.08860( 0.637) 1.30988( 1.394) 2.18072( 1.320)
#>    8  1.03983( 0.769) 1.27310( 0.968) 1.29644( 1.023) 1.49768( 1.340)
#>    9  2.00135( 2.085) 1.32907( 0.727) 1.44506( 2.129) 2.21526( 1.819)
#>    10 2.64249( 2.355) 1.68938( 1.227) 1.44610( 1.984) 0.98585( 0.781)
#>      year
#> quant 5               6               7               8              
#>    1  1.52225( 3.500) 1.15206( 1.120) 0.98372( 0.929) 1.37355( 1.149)
#>    2  1.23594( 1.338) 1.99427( 2.175) 1.25985( 1.010) 1.79425( 1.487)
#>    3  3.14167( 1.860) 1.14661( 0.877) 1.52139( 1.013) 1.67668( 1.370)
#>    4  2.40385( 2.587) 1.25102( 2.143) 1.78835( 2.340) 0.89011( 0.544)
#>    5  2.49375( 1.959) 1.61615( 1.045) 1.28612( 0.796) 1.05873( 0.446)
#>    6  1.37344( 1.728) 2.26892( 2.279) 1.32682( 0.474) 1.07545( 0.760)
#>    7  1.04509( 0.981) 2.28524( 2.105) 1.64015( 1.636) 1.43620( 0.970)
#>    8  2.75003( 3.190) 1.77417( 1.651) 1.01307( 1.718) 1.77110( 3.105)
#>    9  2.03563( 2.833) 1.81603( 0.995) 1.71698( 1.361) 1.55714( 1.020)
#>    10 1.47178( 1.456) 1.52543( 1.119) 0.95600( 0.714) 2.41489( 3.274)
#>      year
#> quant 9               10              11              12             
#>    1  1.52213( 1.266) 1.54261( 1.190) 0.63911( 0.512) 1.94004( 2.013)
#>    2  1.83411( 0.996) 1.21350( 1.070) 0.96055( 1.040) 1.30801( 1.051)
#>    3  1.58561( 0.977) 1.83113( 2.538) 1.88451( 3.056) 2.16063( 1.738)
#>    4  2.28693( 2.044) 1.24561( 1.506) 1.80938( 1.771) 1.21654( 1.239)
#>    5  1.44150( 1.481) 0.86161( 0.837) 1.74544( 1.480) 1.95573( 2.096)
#>    6  0.93106( 1.101) 1.68451( 2.635) 1.40890( 1.254) 1.19512( 0.499)
#>    7  1.63187( 1.629) 2.33292( 2.311) 1.05828( 1.210) 2.82641( 4.279)
#>    8  1.59665( 1.878) 1.42625( 1.574) 1.70571( 1.213) 1.42958( 1.029)
#>    9  1.03911( 0.690) 0.88049( 0.548) 2.88800( 4.185) 0.90842( 1.147)
#>    10 1.35472( 1.366) 1.18378( 1.353) 1.24765( 0.859) 1.81168( 2.149)
#>      year
#> quant 13              14              15              16             
#>    1  1.04557( 0.926) 0.73765( 0.545) 2.29027( 3.929) 1.61029( 1.513)
#>    2  2.03138( 2.859) 1.63707( 1.193) 3.95733( 3.334) 2.03255( 1.339)
#>    3  2.10061( 1.844) 0.90685( 0.882) 1.16351( 1.360) 1.44278( 0.927)
#>    4  1.03296( 0.613) 1.61995( 1.682) 1.96936( 2.389) 1.51504( 2.308)
#>    5  1.92218( 1.516) 1.19131( 0.518) 1.81635( 1.538) 1.29339( 1.019)
#>    6  1.01386( 0.867) 2.07077( 2.397) 1.42662( 1.332) 0.85362( 0.924)
#>    7  1.88584( 2.287) 1.45788( 1.276) 2.02199( 1.579) 3.27482( 5.897)
#>    8  1.73950( 0.987) 3.27551( 3.193) 0.91520( 0.428) 1.29996( 1.128)
#>    9  1.02784( 1.418) 2.28569( 3.053) 2.67612( 3.865) 1.20793( 0.848)
#>    10 1.54345( 2.026) 1.91687( 1.992) 0.80519( 0.526) 1.75438( 1.676)
#>      year
#> quant 17              18              19              20             
#>    1  0.92830( 0.486) 2.19043( 1.479) 0.88994( 0.358) 1.03843( 0.774)
#>    2  0.96903( 1.172) 1.47229( 1.215) 2.09426( 1.721) 1.65030( 0.773)
#>    3  1.75553( 1.958) 2.73600( 2.236) 1.60659( 1.151) 1.21382( 1.024)
#>    4  1.52115( 1.362) 0.98726( 0.769) 1.34391( 1.141) 1.50673( 0.865)
#>    5  2.07968( 2.104) 1.38206( 2.095) 0.97864( 0.644) 1.56586( 4.502)
#>    6  1.64590( 1.083) 0.86576( 0.378) 5.82077(10.229) 1.45342( 1.226)
#>    7  2.47404( 2.099) 2.00410( 1.352) 1.28344( 1.127) 0.98570( 0.424)
#>    8  2.67289( 3.564) 1.94825( 1.304) 1.05250( 1.027) 2.45178( 5.420)
#>    9  2.43746( 2.605) 1.17187( 0.706) 2.56057( 1.818) 2.39662( 5.306)
#>    10 1.52610( 0.980) 0.78980( 0.482) 1.20188( 1.196) 1.45607( 1.110)
#> 
#> units:  kg