Class FLQuantPoint

The FLQuantPoint class summarizes the contents of an FLQuant object with multiple iterations along its sixth dimension using a number of descriptive statistics.

Usage

FLQuantPoint(object, ...)

# S4 method for class 'missing'
FLQuantPoint(..., units = "NA", n = 1)

# S4 method for class 'FLQuant'
FLQuantPoint(object, ..., probs = c(0.25, 0.75))

# S4 method for class 'FLQuantPoint'
n(object, ...)

Arguments

object

Input numeric object

...

Additonal arguments

Details

An object of this class has a set structure along its sixth dimension (iter), which will always be of length 5, and with dimnames mean, median, var, uppq and lowq. They refer, respectively, to the sample mean, sample median, variance, and lower (0.25) and upper (0.75) quantiles.

Objects of this class wil be typically created from an FLQuant. The various statistics are calculated along the iter dimension of the original FLQuant using apply.

Slots

.Data

The main array holding the computed statistics. array.

units

Units of measurement. character.

Accesors

mean,mean<-:

'mean' element on 6th dimension, arithmetic mean.

median,median<-:

'median' element on 6th dimension, median.

var,var<-:

'var' element on 6th dimension, variance.

lowq,lowq<-:

'lowq' element on 6th dimension, lower quantile (0.25 by default).

uppq,uppq<-:

'uppq' element on 6th dimension, upper quantile (0.75 by default).

quantile:

returns the 'lowq' or 'uppq' iter, depending on the value of 'probs' (0.25 or 0.75).

Constructor

Inputs can be of class:

FLQuant:

An FLQuant object with iters (i.e. dim6 > 1)

Validity

iter:

iter dimension is of length 5.

Dimnames:

iter dimnames are 'mean', 'median', 'var', 'uppq' and'lowq'

See also

Author

The FLR Team

Examples

flq <- FLQuant(rlnorm(2000), dim=c(10,20,1,1,1,200), units="kg")
flqp <- FLQuantPoint(flq)
flqp <- FLQuantPoint(flq, probs=c(0.05, 0.95))
summary(flqp)
#> An object of class "FLQuantPoint" with:
#> dim  :  10 20 1 1 1 5 
#> quant:  quant 
#> units:  kg 
#> 
#> 1st Qu.:  0.2801594 
#> Mean   :  1.651237 
#> Median :  1.070952 
#> Var    :  4.235913 
#> 3rd Qu.:  5.688918 
mean(flqp)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1     2     3     4     5     6     7     8     9     10    11    12   
#>    1  2.324 2.294 1.385 1.593 2.206 1.332 1.148 1.082 1.725 1.696 1.146 0.960
#>    2  1.027 1.274 2.134 1.038 2.398 1.738 1.411 1.563 2.369 1.039 2.611 1.932
#>    3  1.210 1.302 1.605 2.387 1.544 0.927 2.152 1.904 1.380 1.576 1.530 1.670
#>    4  1.315 1.766 2.128 2.201 1.836 1.556 1.499 1.012 0.865 3.636 1.067 1.159
#>    5  1.575 1.844 1.032 1.676 1.446 0.901 2.086 1.489 1.328 1.207 1.792 1.390
#>    6  0.918 2.282 1.791 1.213 1.044 1.311 1.598 1.564 0.694 1.844 1.180 0.782
#>    7  1.513 0.653 1.261 2.034 1.248 1.844 1.788 1.108 0.897 1.221 1.900 1.765
#>    8  1.879 0.906 2.987 1.098 1.565 1.747 1.502 1.733 2.274 2.054 2.101 0.942
#>    9  3.363 1.140 2.535 1.332 1.896 0.838 2.073 1.278 1.939 1.193 1.019 1.619
#>    10 0.861 1.183 2.333 1.509 1.327 1.048 1.397 0.921 1.801 1.953 1.862 1.219
#>      year
#> quant 13    14    15    16    17    18    19    20   
#>    1  2.076 1.551 0.966 1.664 0.821 4.544 1.285 2.160
#>    2  1.362 1.859 3.448 2.263 1.843 1.255 0.913 1.581
#>    3  3.042 0.906 1.368 2.682 2.116 1.199 2.697 1.212
#>    4  1.990 2.301 1.291 2.154 1.142 2.242 2.637 2.010
#>    5  1.807 0.892 1.796 1.456 0.825 1.176 1.464 2.768
#>    6  2.428 1.651 0.944 2.093 0.924 1.090 1.718 3.273
#>    7  4.463 2.029 1.086 1.651 1.886 1.649 1.145 2.721
#>    8  1.265 1.354 1.878 2.982 1.514 1.353 1.561 1.104
#>    9  2.168 1.684 1.588 1.038 1.367 1.428 1.418 1.609
#>    10 1.666 1.403 1.682 1.619 0.988 2.282 1.467 2.013
#> 
#> units:  kg 
var(flqp)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1        2        3        4        5        6        7        8       
#>    1    3.9403  11.8870   2.6253   3.6166   5.0020   0.2721   0.6052   1.3462
#>    2    0.3943   1.6777   1.7148   1.0000   4.4651   2.2733   0.9960   2.4891
#>    3    1.1247   1.1467   1.5355   8.6006   2.5080   2.6783   8.4050   3.3263
#>    4    0.5334   4.9431   3.0002   8.1530   1.1295   1.6023   0.8877   0.4262
#>    5    1.3154   3.1922   0.6734   1.8134   1.3172   0.5215   9.4897   1.5725
#>    6    0.2282   8.7541   9.3628   1.5743   0.9314   1.3296   1.7362   1.3718
#>    7    2.1718   0.1327   1.6309   4.3950   1.3963   2.1648   1.0116   1.3369
#>    8    2.1423   0.5214   4.4292   0.8097   0.8282   1.6755   0.9089   5.7560
#>    9   37.8902   0.7285   8.2979   4.5746   6.4550   0.3235   4.6748   2.0813
#>    10   0.4916   1.4400   4.6876   1.1419   0.6376   0.1906   2.2163   0.9821
#>      year
#> quant 9        10       11       12       13       14       15       16      
#>    1    6.3736   1.2621   0.3095   0.5770   5.8005   1.4241   0.8087   1.2130
#>    2    6.9697   1.3699  18.7856   5.5250   1.4118   2.6565  40.4577   4.6295
#>    3    2.5325   2.0241   1.2465   1.3707   8.7698   0.2125   1.3181  14.4490
#>    4    0.3209  14.3340   1.3222   1.4824   7.9749  14.1922   0.7681   7.0792
#>    5    1.5088   0.7574   5.4289   3.1276   4.4929   0.2684   2.2944   1.4330
#>    6    0.2358   4.1423   0.8477   0.3594  13.1939   2.0342   0.2684   2.8645
#>    7    0.9634   0.8483   6.7310   1.7501  13.0599   1.4912   1.2403   1.2441
#>    8   10.3468   2.7334   3.3705   0.5666   1.7752   0.7672   3.6719   5.5311
#>    9    3.8523   1.4788   0.3268   2.8231   4.3200   5.4396   1.4344   0.6457
#>    10   2.0570   4.3071   1.8138   0.2718   2.0780   1.2695   3.2380   4.4319
#>      year
#> quant 17       18       19       20      
#>    1    0.2093 112.2117   1.5471   3.8180
#>    2    1.6987   1.3301   0.1888   0.6429
#>    3    1.3826   0.8438  37.1267   0.4886
#>    4    0.5956   2.7708  23.7887   3.3566
#>    5    0.2583   1.3740   1.6009   6.3302
#>    6    0.0981   0.6793   0.8340  33.1019
#>    7    3.5946   0.7512   0.3870   3.5406
#>    8    0.9971   1.3137   4.2232   1.5942
#>    9    1.3351   0.6999   3.5026   3.0523
#>    10   0.3999  15.4574   1.6665   2.6698
#> 
#> units:  kg 
rnorm(200, flqp)
#> An object of class "FLQuant"
#> iters:  200 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1             2             3             4             5            
#>    1  2.290( 2.097) 2.514( 3.547) 1.327( 1.439) 1.373( 1.728) 2.269( 2.279)
#>    2  1.089( 0.637) 1.310( 1.394) 2.181( 1.320) 1.045( 0.981) 2.285( 2.105)
#>    3  1.273( 0.968) 1.296( 1.023) 1.498( 1.340) 2.750( 3.190) 1.774( 1.651)
#>    4  1.329( 0.727) 1.445( 2.129) 2.215( 1.819) 2.036( 2.833) 1.816( 0.995)
#>    5  1.689( 1.227) 1.446( 1.984) 0.986( 0.781) 1.472( 1.456) 1.525( 1.119)
#>    6  0.922( 0.451) 2.752( 2.953) 1.522( 3.500) 1.152( 1.120) 0.984( 0.929)
#>    7  1.353( 1.395) 0.692( 0.392) 1.236( 1.338) 1.994( 2.175) 1.260( 1.010)
#>    8  1.816( 1.330) 0.814( 0.744) 3.142( 1.860) 1.147( 0.877) 1.521( 1.013)
#>    9  3.174( 6.112) 1.072( 0.748) 2.404( 2.587) 1.251( 2.143) 1.788( 2.340)
#>    10 0.818( 0.588) 1.067( 1.349) 2.494( 1.959) 1.616( 1.045) 1.286( 0.796)
#>      year
#> quant 6             7             8             9             10           
#>    1  1.327( 0.474) 1.075( 0.760) 0.931( 1.101) 1.685( 2.635) 1.409( 1.254)
#>    2  1.640( 1.636) 1.436( 0.970) 1.632( 1.629) 2.333( 2.311) 1.058( 1.210)
#>    3  1.013( 1.718) 1.771( 3.105) 1.597( 1.878) 1.426( 1.574) 1.706( 1.213)
#>    4  1.717( 1.361) 1.557( 1.020) 1.039( 0.690) 0.880( 0.548) 2.888( 4.185)
#>    5  0.956( 0.714) 2.415( 3.274) 1.355( 1.366) 1.184( 1.353) 1.248( 0.859)
#>    6  1.374( 1.149) 1.522( 1.266) 1.543( 1.190) 0.639( 0.512) 1.940( 2.013)
#>    7  1.794( 1.487) 1.834( 0.996) 1.214( 1.070) 0.961( 1.040) 1.308( 1.051)
#>    8  1.677( 1.370) 1.586( 0.977) 1.831( 2.538) 1.885( 3.056) 2.161( 1.738)
#>    9  0.890( 0.544) 2.287( 2.044) 1.246( 1.506) 1.809( 1.771) 1.217( 1.239)
#>    10 1.059( 0.446) 1.442( 1.481) 0.862( 0.837) 1.745( 1.480) 1.956( 2.096)
#>      year
#> quant 11            12            13            14            15           
#>    1  1.195( 0.499) 1.014( 0.867) 2.071( 2.397) 1.427( 1.332) 0.854( 0.924)
#>    2  2.826( 4.279) 1.886( 2.287) 1.458( 1.276) 2.022( 1.579) 3.275( 5.897)
#>    3  1.430( 1.029) 1.739( 0.987) 3.276( 3.193) 0.915( 0.428) 1.300( 1.128)
#>    4  0.908( 1.147) 1.028( 1.418) 2.286( 3.053) 2.676( 3.865) 1.208( 0.848)
#>    5  1.812( 2.149) 1.543( 2.026) 1.917( 1.992) 0.805( 0.526) 1.754( 1.676)
#>    6  1.046( 0.926) 0.738( 0.545) 2.290( 3.929) 1.610( 1.513) 0.928( 0.486)
#>    7  2.031( 2.859) 1.637( 1.193) 3.957( 3.334) 2.033( 1.339) 0.969( 1.172)
#>    8  2.101( 1.844) 0.907( 0.882) 1.164( 1.360) 1.443( 0.927) 1.756( 1.958)
#>    9  1.033( 0.613) 1.620( 1.682) 1.969( 2.389) 1.515( 2.308) 1.521( 1.362)
#>    10 1.922( 1.516) 1.191( 0.518) 1.816( 1.538) 1.293( 1.019) 2.080( 2.104)
#>      year
#> quant 16            17            18            19            20           
#>    1  1.646( 1.083) 0.866( 0.378) 5.821(10.229) 1.453( 1.226) 1.841( 1.665)
#>    2  2.474( 2.099) 2.004( 1.352) 1.283( 1.127) 0.986( 0.424) 1.584( 0.774)
#>    3  2.673( 3.564) 1.948( 1.304) 1.052( 1.027) 2.452( 5.420) 1.123( 0.730)
#>    4  2.437( 2.605) 1.172( 0.706) 2.561( 1.818) 2.397( 5.306) 2.009( 2.100)
#>    5  1.526( 0.980) 0.790( 0.482) 1.202( 1.196) 1.456( 1.110) 2.664( 2.381)
#>    6  2.190( 1.479) 0.890( 0.358) 1.038( 0.774) 1.752( 1.051) 2.730( 5.220)
#>    7  1.472( 1.215) 2.094( 1.721) 1.650( 0.773) 1.037( 0.512) 2.638( 1.792)
#>    8  2.736( 2.236) 1.607( 1.151) 1.214( 1.024) 1.087( 1.872) 1.211( 1.069)
#>    9  0.987( 0.769) 1.344( 1.141) 1.507( 0.865) 1.550( 1.844) 1.604( 1.780)
#>    10 1.382( 2.095) 0.979( 0.644) 1.566( 4.502) 1.396( 1.318) 2.059( 1.605)
#> 
#> units:  kg