Class FLQuantPoint

The FLQuantPoint class summarizes the contents of an FLQuant object with multiple iterations along its sixth dimension using a number of descriptive statistics.

FLQuantPoint(object, ...)

# S4 method for missing
FLQuantPoint(..., units = "NA")

# S4 method for FLQuant
FLQuantPoint(object, ..., units = "NA")

Arguments

object

Input numeric object

...

Additonal arguments

Details

An object of this class has a set structure along its sixth dimension (iter), which will always be of length 5, and with dimnames mean, median, var, uppq and lowq. They refer, respectively, to the sample mean, sample median, variance, and lower (0.25) and upper (0.75) quantiles.

Objects of this class wil be typically created from an FLQuant. The various statistics are calculated along the iter dimension of the original FLQuant using apply.

Slots

.Data

The main array holding the computed statistics. array.

units

Units of measurement. character.

Accesors

mean,mean<-:

'mean' element on 6th dimension, arithmetic mean.

median,median<-:

'median' element on 6th dimension, median.

var,var<-:

'var' element on 6th dimension, variance.

lowq,lowq<-:

'lowq' element on 6th dimension, lower quantile (0.25 by default).

uppq,uppq<-:

'uppq' element on 6th dimension, upper quantile (0.75 by default).

Constructor

Inputs can be of class:

FLQuant:

An FLQuant object with iters (i.e. dim[6] > 1)

Validity

iter:

iter dimension is of length 5.

Dimnames:

iter dimnames are 'mean', 'median', 'var', 'uppq' and'lowq'

See also

FLQuant

Examples

flq <- FLQuant(rnorm(2000), dim=c(10,20,1,1,1,200))
flqp <- FLQuantPoint(flq)
summary(flqp)

#> An object of class "FLQuantPoint" with:
#> dim  :  10 20 1 1 1 5 
#> quant:  quant 
#> units:  NA 
#> 
#> 1st Qu.:  -0.6080744 
#> Mean   :  0.005799676 
#> Median :  0.01695316 
#> Var    :  0.8649465 
#> 3rd Qu.:  0.6455818 

mean(flqp)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1           2           3           4           5           6          
#>    1   5.0133e-01 -3.5833e-01  1.8115e-01  5.1524e-01 -2.2942e-02 -8.9399e-02
#>    2  -2.9179e-01  1.9956e-01  2.5240e-01 -3.5200e-02 -6.2350e-01 -1.0954e-01
#>    3  -7.0900e-01  1.1594e-02  2.6655e-01  5.2948e-01 -1.4830e-01  8.0968e-02
#>    4  -7.1124e-01  1.4561e-01 -2.3662e-01 -8.1976e-02 -1.1345e-01  4.2359e-02
#>    5  -3.6346e-01  4.7894e-02  7.2466e-01  7.7792e-02 -1.5405e-02  5.1558e-01
#>    6  -1.8014e-01 -3.7238e-01  2.1824e-01 -7.8384e-02 -3.7050e-01  1.5222e-01
#>    7   3.4695e-01  9.1277e-02 -2.1417e-01  3.5253e-01  3.1950e-01 -4.3103e-01
#>    8  -4.3347e-01 -5.7056e-03  1.1422e-01 -1.3751e-01  1.1667e-01 -8.5598e-02
#>    9   1.1102e-01 -2.0753e-01  1.5272e-01  3.6092e-01 -6.2858e-02 -2.3405e-01
#>    10  1.0871e-01 -5.7593e-02 -5.0395e-02  1.1208e-01  7.7732e-02 -1.4630e-01
#>      year
#> quant 7           8           9           10          11          12         
#>    1  -7.8844e-01  3.9570e-01  2.8611e-01 -1.2713e-01 -1.6177e-01 -1.3063e-01
#>    2   1.9423e-01 -2.0710e-01 -4.1846e-02 -2.9364e-01  1.7836e-01  2.1456e-01
#>    3  -9.7930e-02 -4.1623e-01  4.4189e-01  1.4540e-02  7.2704e-02 -6.4551e-01
#>    4   2.8415e-01 -5.1169e-02  1.0418e-01  3.0803e-01  5.3791e-01  8.0193e-02
#>    5  -6.8133e-02 -1.1542e-01 -3.1315e-01  7.4601e-02  1.5573e-02 -2.8489e-01
#>    6  -2.2086e-01  5.7330e-01 -4.3704e-03 -1.7701e-01  3.5647e-01  1.3149e-01
#>    7   1.3391e-01  3.7699e-01  4.2074e-01 -5.5618e-01  9.7330e-02 -3.4289e-01
#>    8  -1.4122e-01  8.8100e-02  4.4722e-01 -3.6153e-01 -1.4851e-01 -2.3493e-01
#>    9  -9.1282e-02  3.9159e-01  2.0290e-01  7.6301e-02  4.3762e-01 -3.3602e-02
#>    10 -5.6633e-02 -1.1558e-01 -4.5259e-01 -2.9345e-02  4.4380e-01 -3.3683e-02
#>      year
#> quant 13          14          15          16          17          18         
#>    1   4.8375e-01 -1.1883e-02  9.5773e-02  8.2152e-03  1.1584e-01  3.1870e-01
#>    2  -2.5232e-01 -3.2388e-01 -1.1510e-01  2.1638e-01 -8.4155e-03 -7.4539e-01
#>    3  -6.1946e-01  7.6037e-01  1.7064e-01 -1.1670e-02 -4.6572e-01 -1.4402e-01
#>    4   1.7838e-01  4.5226e-01  6.3942e-01 -6.1361e-01  2.3179e-01  2.1836e-01
#>    5   5.6989e-01  2.5137e-02 -8.0102e-01  2.0742e-02  5.3753e-02  1.9440e-01
#>    6  -2.6323e-01 -9.6932e-02 -2.6087e-01  6.2091e-01 -3.5769e-01 -1.2964e-01
#>    7   8.6591e-02 -5.5488e-06  2.4112e-01  1.4743e-01 -3.5312e-01  7.3899e-02
#>    8   2.6403e-01 -2.2781e-01 -1.6480e-01 -1.0951e-01 -2.6304e-01  8.2995e-02
#>    9   4.5221e-01  3.2144e-01  2.1943e-01  7.0764e-02  1.7613e-01  2.5046e-01
#>    10 -5.4109e-01 -3.2907e-02  1.9025e-01  2.9118e-01  1.4061e-01 -5.2100e-01
#>      year
#> quant 19          20         
#>    1  -7.8434e-01  1.7355e-01
#>    2  -2.5484e-01 -9.8956e-01
#>    3   3.7803e-01  3.0009e-01
#>    4   1.5126e-01 -1.6683e-02
#>    5   1.1880e-01 -1.9306e-01
#>    6   1.6746e-01  3.2545e-01
#>    7   3.4278e-02  2.4156e-01
#>    8   7.1110e-01  2.2693e-01
#>    9  -5.6621e-02 -3.0468e-01
#>    10 -3.1916e-01 -8.5201e-02
#> 
#> units:  NA 

var(flqp)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1       2       3       4       5       6       7       8       9      
#>    1  1.67175 0.76383 0.79695 0.87020 0.36100 0.77922 0.52237 0.88481 0.91430
#>    2  0.41905 0.57433 1.48177 0.61381 1.10598 1.04511 0.73290 0.51127 0.99106
#>    3  1.31600 0.99626 0.75628 1.03981 0.51312 0.43828 0.78496 1.10352 0.57200
#>    4  0.49447 0.43876 1.18457 1.14504 0.46467 1.40701 1.40021 1.62763 0.96831
#>    5  0.91557 0.30654 0.18141 0.76990 0.41793 1.03135 1.13994 0.88273 0.79529
#>    6  0.18095 0.93069 2.33879 0.53444 0.26195 0.46461 0.95365 1.10114 0.87947
#>    7  0.77356 0.52484 0.59216 1.17804 0.84054 0.63858 1.34038 0.76187 0.91138
#>    8  0.46995 0.89278 1.05645 0.34490 0.59902 0.59897 0.59002 1.19009 0.85822
#>    9  0.61943 0.24240 0.50964 0.53259 0.77789 1.83206 0.91246 0.60263 1.53576
#>    10 1.28955 0.90708 1.31693 0.59697 0.69657 0.41671 1.23372 1.04371 0.22468
#>      year
#> quant 10      11      12      13      14      15      16      17      18     
#>    1  0.74389 0.26240 1.16335 1.10777 1.03857 0.96209 0.75730 1.57526 1.12476
#>    2  0.51555 0.24555 1.18109 0.85907 0.99170 0.51654 1.11054 0.59882 1.35466
#>    3  0.92603 0.77770 0.83606 1.71734 0.82385 0.82435 0.69446 1.13628 1.58625
#>    4  1.24207 1.76996 0.83963 0.89773 0.87136 0.40889 1.20453 0.72501 0.59350
#>    5  1.03676 0.51422 0.64184 1.09052 1.28202 0.89315 0.66111 0.84105 1.58385
#>    6  1.04102 0.42035 0.71142 0.55898 0.65306 1.61209 0.95035 0.40094 0.26434
#>    7  0.35587 1.57379 0.55993 1.32374 1.68527 0.72797 1.17478 1.34331 1.24803
#>    8  0.49565 0.39467 0.63804 1.68692 0.79551 1.39245 0.74213 0.52992 0.64845
#>    9  0.81484 0.32509 1.39053 1.57283 0.87168 0.97417 1.03173 0.91436 0.61441
#>    10 1.29034 1.22269 0.42185 0.70308 1.30226 0.80411 1.14720 0.34709 0.69855
#>      year
#> quant 19      20     
#>    1  0.60452 0.59375
#>    2  0.99953 0.31441
#>    3  1.12745 0.42068
#>    4  1.04378 1.04709
#>    5  1.26522 0.93187
#>    6  0.61302 0.27302
#>    7  0.83666 0.19525
#>    8  1.51964 0.60275
#>    9  0.77012 0.82506
#>    10 0.36139 0.51038
#> 
#> units:  NA 

rnorm(200, flqp)

#> An object of class "FLQuant"
#> iters:  200 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1                  2                  3                 
#>    1   0.48429334(1.216) -0.26456157(0.883)  0.21808780(0.928)
#>    2  -0.38324899(0.696)  0.20304339(0.738)  0.23632212(1.196)
#>    3  -0.64618795(1.286)  0.03121922(0.994)  0.14254459(0.843)
#>    4  -0.71861990(0.722)  0.19282938(0.691) -0.47919002(1.145)
#>    5  -0.54782341(0.855)  0.02636076(0.533)  0.78303680(0.423)
#>    6  -0.19079367(0.406) -0.50056996(0.946)  0.42840388(1.622)
#>    7   0.40048723(0.761)  0.05672344(0.633) -0.32766863(0.811)
#>    8  -0.41548171(0.703) -0.07120356(0.915)  0.03134401(0.952)
#>    9   0.04446126(0.757) -0.23234808(0.423)  0.08110050(0.805)
#>    10  0.07672307(1.116)  0.02584356(0.930) -0.07926469(1.042)
#>      year
#> quant 4                  5                  6                 
#>    1   0.52517750(0.921) -0.01442492(0.586) -0.13509168(0.904)
#>    2  -0.11031397(0.855) -0.45659143(1.103)  0.05783916(1.056)
#>    3   0.49298180(1.164) -0.20791383(0.797)  0.04997091(0.633)
#>    4  -0.12545815(1.154) -0.18123701(0.743)  0.12809305(1.162)
#>    5   0.00059064(1.013) -0.05374541(0.612)  0.41889036(1.039)
#>    6  -0.07986594(0.795) -0.37501896(0.524)  0.17481799(0.608)
#>    7   0.40523643(0.965)  0.37137564(0.959) -0.47628596(0.939)
#>    8  -0.15987697(0.550)  0.03938492(0.831) -0.18574514(0.734)
#>    9   0.41504632(0.630)  0.05154949(0.826) -0.25720848(1.314)
#>    10  0.04524955(0.709)  0.10130580(0.888) -0.13230200(0.578)
#>      year
#> quant 7                  8                  9                 
#>    1  -0.78093365(0.744)  0.39331120(0.850)  0.31817278(1.073)
#>    2   0.23937062(0.936) -0.38908323(0.758) -0.18480293(1.009)
#>    3  -0.02443401(0.974) -0.38394761(1.138)  0.46009308(0.755)
#>    4   0.29170762(1.257)  0.07469270(1.257) -0.02459456(1.076)
#>    5  -0.00605822(1.075) -0.19534598(0.983) -0.25581156(0.941)
#>    6  -0.25799168(1.006)  0.56222718(0.992)  0.02987170(0.883)
#>    7   0.02315892(1.193)  0.43600842(0.874)  0.49885573(0.976)
#>    8  -0.08388483(0.725)  0.20918143(1.038)  0.38989854(0.930)
#>    9  -0.05904375(1.030)  0.39807561(0.755)  0.05858040(1.040)
#>    10 -0.09555413(1.144) -0.23264094(0.933) -0.41341993(0.515)
#>      year
#> quant 10                 11                 12                
#>    1  -0.14475220(0.756) -0.13963314(0.547) -0.06771595(1.089)
#>    2  -0.22700823(0.694)  0.21220285(0.448)  0.11652043(0.982)
#>    3  -0.04981161(0.919) -0.06280235(0.933) -0.73362692(0.900)
#>    4   0.15247087(1.319)  0.64143019(1.325)  0.09511497(0.856)
#>    5  -0.05326950(1.049)  0.06727259(0.688) -0.34935185(0.766)
#>    6  -0.10137870(1.094)  0.39361333(0.751)  0.16742587(0.887)
#>    7  -0.60141113(0.604)  0.13224928(1.295) -0.29055005(0.740)
#>    8  -0.39913530(0.649) -0.12631880(0.646) -0.23699656(0.842)
#>    9   0.06526999(0.907)  0.42555204(0.581) -0.00986547(1.262)
#>    10 -0.29992024(1.276)  0.44869794(1.013)  0.04846091(0.699)
#>      year
#> quant 13                 14                 15                
#>    1   0.40573119(1.078)  0.01274499(1.031)  0.15131033(0.924)
#>    2  -0.15365754(0.761) -0.21500600(1.122) -0.12771234(0.754)
#>    3  -0.72524894(1.482)  0.71794506(0.994)  0.23546784(0.909)
#>    4   0.26056080(1.014)  0.47655798(0.887)  0.51533953(0.641)
#>    5   0.46693382(0.958) -0.01772149(1.194) -0.80164092(0.964)
#>    6  -0.28921681(0.740) -0.19588830(0.765) -0.31204056(1.352)
#>    7   0.04449005(1.382) -0.09920621(1.277)  0.30549589(0.907)
#>    8   0.26475583(1.292) -0.29635343(0.965) -0.25533232(1.087)
#>    9   0.28883452(1.184)  0.37918308(0.979)  0.06125105(0.867)
#>    10 -0.51325476(0.808) -0.15162123(1.300)  0.05860110(0.995)
#>      year
#> quant 16                 17                 18                
#>    1  -0.01416286(0.843)  0.23881652(1.283)  0.52588798(1.064)
#>    2   0.15420888(0.993) -0.02290371(0.742) -0.90463096(1.320)
#>    3  -0.08305867(0.809) -0.27802364(1.006) -0.18856212(1.152)
#>    4  -0.53764402(1.149)  0.21103183(0.739)  0.18174396(0.758)
#>    5   0.00760881(0.749)  0.09410819(0.836)  0.05021024(1.376)
#>    6   0.49402538(0.945) -0.45914771(0.680) -0.04834699(0.531)
#>    7   0.14826344(1.119) -0.49438091(1.055)  0.06807356(1.231)
#>    8  -0.12259299(0.910) -0.30776366(0.682)  0.02252119(0.832)
#>    9   0.29137269(1.176)  0.11729873(0.938)  0.22707817(0.807)
#>    10  0.29430204(1.054)  0.19354430(0.512) -0.40851248(0.850)
#>      year
#> quant 19                 20                
#>    1  -0.70448942(0.805)  0.30175778(0.826)
#>    2  -0.37499510(1.121) -0.97767765(0.533)
#>    3   0.49580270(1.116)  0.25747482(0.688)
#>    4   0.17362469(1.012) -0.00592904(0.915)
#>    5   0.06935114(1.095) -0.16773539(1.124)
#>    6   0.18169001(0.682)  0.20519517(0.456)
#>    7  -0.00410977(0.874)  0.12725414(0.387)
#>    8   0.76508637(1.324)  0.28848545(0.812)
#>    9  -0.24270205(0.955) -0.34636705(0.859)
#>    10 -0.23782320(0.680) -0.23383161(0.599)
#> 
#> units:  NA