Class FLStock

A class for modelling a fish stock.

FLStock(object, ...)

# S4 method for FLQuant
FLStock(object, plusgroup = dims(object)$max, ...)

# S4 method for missing
FLStock(object, ...)

# S4 method for FLQuants
FLStock(object, ...)

Arguments

object
FLQuant object used for sizing

...
Other objects to be assigned by name to the class slots

plusgroup
Plusgroup age, to be stored in range

Details

The FLStock object contains a representation of a fish stock as constructed for the purposes of scientific analysis and advice. This includes information on removals (i.e. catches, landings and discards), maturity, natural mortality and the results of an analytical assessment (i.e. estimates of abundance and removal rates) .

Slots

catch

Total catch weight (FLQuant).

catch.n

Catch numbers (FLQuant).

catch.wt

Mean catch weights (FLQuant).

discards

Total discards weight (FLQuant).

discards.n

Discard numbers (FLQuant).

discards.wt

Mean discard weights (FLQuant).

landings

Total landings weight (FLQuant).

landings.n

Landing numbers (FLQuant).

landings.wt

Landing weights (FLQuant).

stock

Total stock weight (FLQuant).

stock.n

Stock numbers (FLQuant).

stock.wt

Mean stock weights (FLQuant).

m

Natural mortality (FLQuant).

mat

Proportion mature (FLQuant).

harvest

Harvest rate or fishing mortality. The units of this slot should be set to 'hr' or 'f' accordingly (FLQuant).

harvest.spwn

Proportion of harvest/fishing mortality before spawning (FLQuant).

m.spwn

Proportion of natural mortality before spawning (FLQuant).

name

Name of the stock (character).

desc

Description of the stock (character).

range

Named numeric vector containing the quant and year ranges, the plusgroup and the quant range that the average fishing mortality should be calculated over (numeric).

Accessors

All slots in the class have accessor and replacement methods defined that allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot. A numeric vector can also be used when replacing FLQuant slots, and the vector will be used to substitute the values in the slot, but not its other attributes.

Constructor

A construction method exists for this class that can take named arguments for any of its slots. All slots are then created to match the requirements of the class validity. If an unnamed FLQuant object is provided, this is used for sizing but not stored in any slot.

See also

Examples

data(ple4)
summary(ple4)

#> An object of class "FLStock"
#> 
#> Name: PLE 
#> Description: Plaice in IV. ICES WGNSSK 2018. FLAAP 
#> Quant: age 
#> Dims:  age 	year	unit	season	area	iter
#> 	10	61	1	1	1	1	
#> 
#> Range:  min	max	pgroup	minyear	maxyear	minfbar	maxfbar 
#> 	1	10	10	1957	2017	2	6	
#> 
#> catch         : [ 1 61 1 1 1 1 ], units =  t 
#> catch.n       : [ 10 61 1 1 1 1 ], units =  1000 
#> catch.wt      : [ 10 61 1 1 1 1 ], units =  kg 
#> discards      : [ 1 61 1 1 1 1 ], units =  t 
#> discards.n    : [ 10 61 1 1 1 1 ], units =  1000 
#> discards.wt   : [ 10 61 1 1 1 1 ], units =  kg 
#> landings      : [ 1 61 1 1 1 1 ], units =  t 
#> landings.n    : [ 10 61 1 1 1 1 ], units =  1000 
#> landings.wt   : [ 10 61 1 1 1 1 ], units =  kg 
#> stock         : [ 1 61 1 1 1 1 ], units =  t 
#> stock.n       : [ 10 61 1 1 1 1 ], units =  1000 
#> stock.wt      : [ 10 61 1 1 1 1 ], units =  kg 
#> m             : [ 10 61 1 1 1 1 ], units =  m 
#> mat           : [ 10 61 1 1 1 1 ], units =   
#> harvest       : [ 10 61 1 1 1 1 ], units =  f 
#> harvest.spwn  : [ 10 61 1 1 1 1 ], units =   
#> m.spwn        : [ 10 61 1 1 1 1 ], units =   

# get the landings slot and assign values to it
  landings(ple4)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all  70926  74157  78178  88764  85267  90305 103162 111121 105424  98334
#>      year
#> age   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 103947 121020 122661 111783 117301 130443 133768 115181  94458 122166
#>      year
#> age   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 108524 128366 119827 150640 151304 145669 143690 162681 182374 166633
#>      year
#> age   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 155005 168118 187666 174414 147844 134793 141800 126195 109681  93642
#>      year
#> age   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all  82875  74217  97340 100318  66596  87138  74770  82529  61501  62161
#>      year
#> age   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all  59713  64574  68842  74246  74437  82307  97123  86424  91033  86381
#>      year
#> age   2017  
#>   all  84750
#> 
#> units:  t 

  landings(ple4) <- apply(landings.n(ple4)*landings.wt(ple4),2,sum)

# perform similar calculation as the preceding apply function
  discards(ple4) <- computeDiscards(ple4)
  catch(ple4) <- computeCatch(ple4)
  catch(ple4) <- computeCatch(ple4, slot="all")

# set the units of the harvest slot of an FLStock object
  harvest(ple4) <- 'f'

# subset and trim the FLStock
  ple4[,1]

#> An object of class "FLStock"
#> Slot "catch":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957 
#>   all 78360
#> 
#> units:  t 
#> 
#> Slot "catch.n":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957   
#>   1  42702.7
#>   2  40141.3
#>   3  79353.0
#>   4  56560.1
#>   5  31888.3
#>   6  10988.0
#>   7  12049.6
#>   8   9595.8
#>   9   5495.2
#>   10 11779.1
#> 
#> units:  1000 
#> 
#> Slot "catch.wt":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957    
#>   1  0.043999
#>   2  0.109801
#>   3  0.194296
#>   4  0.256615
#>   5  0.349100
#>   6  0.455454
#>   7  0.532988
#>   8  0.588982
#>   9  0.395964
#>   10 0.997959
#> 
#> units:  kg 
#> 
#> Slot "discards":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957  
#>   all 7434.2
#> 
#> units:  t 
#> 
#> Slot "discards.n":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957      
#>   1  4.2702e+04
#>   2  3.6324e+04
#>   3  9.6720e+03
#>   4  1.0172e+03
#>   5  8.4596e+02
#>   6  2.8323e+01
#>   7  4.9188e-01
#>   8  4.7852e-01
#>   9  4.9744e-01
#>   10 4.8836e-01
#> 
#> units:  1000 
#> 
#> Slot "discards.wt":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957 
#>   1  0.044
#>   2  0.104
#>   3  0.146
#>   4  0.181
#>   5  0.206
#>   6  0.244
#>   7  0.244
#>   8  0.231
#>   9  0.000
#>   10 0.000
#> 
#> units:  kg 
#> 
#> Slot "landings":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957 
#>   all 70926
#> 
#> units:  t 
#> 
#> Slot "landings.n":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957      
#>   1  6.5789e-01
#>   2  3.8177e+03
#>   3  6.9681e+04
#>   4  5.5543e+04
#>   5  3.1042e+04
#>   6  1.0960e+04
#>   7  1.2049e+04
#>   8  9.5953e+03
#>   9  5.4948e+03
#>   10 1.1779e+04
#> 
#> units:  1000 
#> 
#> Slot "landings.wt":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957 
#>   1  0.000
#>   2  0.165
#>   3  0.201
#>   4  0.258
#>   5  0.353
#>   6  0.456
#>   7  0.533
#>   8  0.589
#>   9  0.396
#>   10 0.998
#> 
#> units:  kg 
#> 
#> Slot "stock":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957  
#>   all 402414
#> 
#> units:  t 
#> 
#> Slot "stock.n":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957  
#>   1  477074
#>   2  268185
#>   3  361589
#>   4  223134
#>   5  147904
#>   6   60609
#>   7   62627
#>   8   49418
#>   9   31997
#>   10  68585
#> 
#> units:  1000 
#> 
#> Slot "stock.wt":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957 
#>   1  0.038
#>   2  0.102
#>   3  0.157
#>   4  0.242
#>   5  0.325
#>   6  0.485
#>   7  0.719
#>   8  0.682
#>   9  0.844
#>   10 0.918
#> 
#> units:  kg 
#> 
#> Slot "m":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957
#>   1  0.1 
#>   2  0.1 
#>   3  0.1 
#>   4  0.1 
#>   5  0.1 
#>   6  0.1 
#>   7  0.1 
#>   8  0.1 
#>   9  0.1 
#>   10 0.1 
#> 
#> units:  m 
#> 
#> Slot "mat":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957
#>   1  0.0 
#>   2  0.5 
#>   3  0.5 
#>   4  1.0 
#>   5  1.0 
#>   6  1.0 
#>   7  1.0 
#>   8  1.0 
#>   9  1.0 
#>   10 1.0 
#> 
#> units:   
#> 
#> Slot "harvest":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957    
#>   1  0.098696
#>   2  0.170862
#>   3  0.261509
#>   4  0.308815
#>   5  0.256284
#>   6  0.210938
#>   7  0.225401
#>   8  0.227728
#>   9  0.198666
#>   10 0.198666
#> 
#> units:  f 
#> 
#> Slot "harvest.spwn":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957
#>   1  0   
#>   2  0   
#>   3  0   
#>   4  0   
#>   5  0   
#>   6  0   
#>   7  0   
#>   8  0   
#>   9  0   
#>   10 0   
#> 
#> units:   
#> 
#> Slot "m.spwn":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957
#>   1  0   
#>   2  0   
#>   3  0   
#>   4  0   
#>   5  0   
#>   6  0   
#>   7  0   
#>   8  0   
#>   9  0   
#>   10 0   
#> 
#> units:   
#> 
#> Slot "name":
#> [1] "PLE"
#> 
#> Slot "desc":
#> [1] "Plaice in IV. ICES WGNSSK 2018. FLAAP"
#> 
#> Slot "range":
#>       min       max plusgroup   minyear   maxyear   minfbar   maxfbar 
#>         1        10        10      1957      1957         2         6 
#> 

  trim(ple4, age=2:6, year=1980:1990)

#> An object of class "FLStock"
#> Slot "catch":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1980   1981   1982   1983   1984   1985   1986   1987   1988   1989  
#>   all 184690 184494 192439 212632 228265 247071 279228 308480 315245 292035
#>      year
#> age   1990  
#>   all 250604
#> 
#> units:  t 
#> 
#> Slot "catch.n":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1980    1981    1982    1983    1984    1985    1986    1987    1988   
#>   2  218847  269603  249567  462799  315003  294918  417134 1047256  478976
#>   3  139742  158998  195147  177915  330543  227034  213479  302191  746685
#>   4   80323   65083   69631   91099   93428  186569  128385  114725  148045
#>   5   28828   35922   27199   29609   41553   46044   96350   65923   56420
#>   6   19874   13526   16783   12739   14251   20730   23033   46662   31129
#>    year
#> age 1989    1990   
#>   2  438712  298563
#>   3  331436  307676
#>   4  337986  152615
#>   5   71205  168825
#>   6   26394   33026
#> 
#> units:  1000 
#> 
#> Slot "catch.wt":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1980    1981    1982    1983    1984    1985    1986    1987    1988   
#>   2 0.19700 0.18344 0.15200 0.15269 0.14991 0.16913 0.14140 0.11340 0.10180
#>   3 0.34296 0.32683 0.30825 0.27476 0.26546 0.26797 0.27541 0.21861 0.19749
#>   4 0.39937 0.41473 0.42061 0.37913 0.32104 0.33350 0.31052 0.34336 0.27587
#>   5 0.47118 0.50169 0.51192 0.50695 0.47045 0.44388 0.40240 0.37542 0.41513
#>   6 0.54244 0.55558 0.60590 0.60200 0.58835 0.56154 0.47181 0.47066 0.47669
#>    year
#> age 1989    1990   
#>   2 0.11764 0.12966
#>   3 0.21508 0.21065
#>   4 0.29130 0.29049
#>   5 0.36431 0.35963
#>   6 0.51227 0.43976
#> 
#> units:  kg 
#> 
#> Slot "discards":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1980   1981   1982   1983   1984   1985   1986   1987   1988   1989  
#>   all  34050  33190  46770  68942  65584  64697 112595 153474 147127 104369
#>      year
#> age   1990  
#>   all  76190
#> 
#> units:  t 
#> 
#> Slot "discards.n":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1980       1981       1982       1983       1984       1985      
#>   2 142175.000 175655.000 201855.000 363172.000 261939.000 223822.000
#>   3    949.720   2302.600   4162.390  37418.400  52242.200  44268.400
#>   4    692.682    405.554   1068.880   1982.810  26231.600   2033.170
#>   5    216.694    417.773    200.244    783.681   1442.790    242.048
#>   6     93.389     58.620     96.820     70.866     63.955     94.690
#>    year
#> age 1986       1987       1988       1989       1990      
#>   2 367088.000 986149.000 467295.000 400157.000 266361.000
#>   3  48280.400 182665.000 480953.000 212393.000 189772.000
#>   4  25565.200   1340.070  61806.300  89459.900  28992.200
#>   5   1356.000   1075.540    818.479   7795.260   4849.140
#>   6    120.006    258.605    174.467    139.947    194.602
#> 
#> units:  1000 
#> 
#> Slot "discards.wt":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1980  1981  1982  1983  1984  1985  1986  1987  1988  1989  1990 
#>   2 0.149 0.135 0.126 0.126 0.127 0.139 0.124 0.105 0.098 0.102 0.111
#>   3 0.191 0.179 0.182 0.180 0.172 0.177 0.181 0.166 0.153 0.163 0.157
#>   4 0.212 0.212 0.203 0.205 0.211 0.197 0.196 0.205 0.185 0.181 0.186
#>   5 0.231 0.220 0.231 0.211 0.205 0.231 0.220 0.220 0.220 0.196 0.212
#>   6 0.000 0.000 0.244 0.244 0.000 0.244 0.244 0.231 0.244 0.000 0.231
#> 
#> units:  kg 
#> 
#> Slot "landings":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1980   1981   1982   1983   1984   1985   1986   1987   1988   1989  
#>   all 150640 151304 145669 143690 162681 182374 166633 155005 168118 187666
#>      year
#> age   1990  
#>   all 174414
#> 
#> units:  t 
#> 
#> Slot "landings.n":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1980   1981   1982   1983   1984   1985   1986   1987   1988   1989  
#>   2  76672  93948  47712  99627  53064  71096  50046  61107  11681  38555
#>   3 138792 156695 190985 140497 278301 182766 165199 119526 265732 119043
#>   4  79631  64677  68562  89117  67196 184536 102820 113385  86239 248526
#>   5  28611  35504  26999  28826  40110  45802  94994  64848  55601  63410
#>   6  19781  13468  16686  12668  14187  20635  22913  46403  30955  26254
#>    year
#> age 1990  
#>   2  32202
#>   3 117904
#>   4 123623
#>   5 163976
#>   6  32831
#> 
#> units:  1000 
#> 
#> Slot "landings.wt":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1980  1981  1982  1983  1984  1985  1986  1987  1988  1989  1990 
#>   2 0.286 0.274 0.262 0.250 0.263 0.264 0.269 0.249 0.254 0.280 0.284
#>   3 0.344 0.329 0.311 0.300 0.283 0.290 0.303 0.299 0.278 0.308 0.297
#>   4 0.401 0.416 0.424 0.383 0.364 0.335 0.339 0.345 0.341 0.331 0.315
#>   5 0.473 0.505 0.514 0.515 0.480 0.445 0.405 0.378 0.418 0.385 0.364
#>   6 0.545 0.558 0.608 0.604 0.591 0.563 0.473 0.472 0.478 0.515 0.441
#> 
#> units:  kg 
#> 
#> Slot "stock":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1980   1981   1982   1983   1984   1985   1986   1987   1988   1989  
#>   all 465713 425416 477746 541397 579572 608773 728868 757282 689841 630289
#>      year
#> age   1990  
#>   all 554640
#> 
#> units:  t 
#> 
#> Slot "stock.n":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1980    1981    1982    1983    1984    1985    1986    1987    1988   
#>   2  603278  759629  728751 1391630  948596  864871 1177750 2866780 1308470
#>   3  308663  338604  431979  422976  820717  559870  503176  670566 1602170
#>   4  188898  147121  156042  206296  214367  429744  291704  253291  320928
#>   5   74240   94918   71554   75328  100486  105579  212338  142492  120680
#>   6   54396   39883   51872   38992   40129   51598   51971  101004   66592
#>    year
#> age 1989    1990   
#>   2 1251000  898824
#>   3  730349  716388
#>   4  743636  347377
#>   5  150385  353208
#>   6   55846   68750
#> 
#> units:  1000 
#> 
#> Slot "stock.wt":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1980  1981  1982  1983  1984  1985  1986  1987  1988  1989  1990 
#>   2 0.163 0.140 0.128 0.128 0.129 0.146 0.126 0.105 0.097 0.101 0.108
#>   3 0.289 0.239 0.250 0.242 0.216 0.232 0.245 0.200 0.172 0.192 0.176
#>   4 0.428 0.421 0.351 0.381 0.413 0.320 0.311 0.383 0.264 0.247 0.261
#>   5 0.444 0.473 0.490 0.494 0.464 0.452 0.440 0.401 0.426 0.362 0.343
#>   6 0.524 0.536 0.589 0.559 0.571 0.536 0.533 0.503 0.467 0.484 0.422
#> 
#> units:  kg 
#> 
#> Slot "m":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1980 1981 1982 1983 1984 1985 1986 1987 1988 1989 1990
#>   2 0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1 
#>   3 0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1 
#>   4 0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1 
#>   5 0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1 
#>   6 0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1  0.1 
#> 
#> units:  m 
#> 
#> Slot "mat":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1980 1981 1982 1983 1984 1985 1986 1987 1988 1989 1990
#>   2 0.5  0.5  0.5  0.5  0.5  0.5  0.5  0.5  0.5  0.5  0.5 
#>   3 0.5  0.5  0.5  0.5  0.5  0.5  0.5  0.5  0.5  0.5  0.5 
#>   4 1.0  1.0  1.0  1.0  1.0  1.0  1.0  1.0  1.0  1.0  1.0 
#>   5 1.0  1.0  1.0  1.0  1.0  1.0  1.0  1.0  1.0  1.0  1.0 
#>   6 1.0  1.0  1.0  1.0  1.0  1.0  1.0  1.0  1.0  1.0  1.0 
#> 
#> units:   
#> 
#> Slot "harvest":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1980    1981    1982    1983    1984    1985    1986    1987    1988   
#>   2 0.47755 0.46445 0.44402 0.42806 0.42728 0.44164 0.46324 0.48183 0.48309
#>   3 0.64099 0.67471 0.63907 0.57963 0.54699 0.55196 0.58640 0.63690 0.66756
#>   4 0.58819 0.62080 0.62827 0.61929 0.60823 0.60501 0.61645 0.64140 0.65802
#>   5 0.52136 0.50423 0.50711 0.52975 0.56654 0.60877 0.64302 0.66071 0.67054
#>   6 0.48192 0.43865 0.41374 0.41874 0.46480 0.54520 0.62235 0.65939 0.67044
#>    year
#> age 1989    1990   
#>   2 0.45748 0.42744
#>   3 0.64311 0.59619
#>   4 0.64450 0.61494
#>   5 0.68272 0.69197
#>   6 0.68093 0.69696
#> 
#> units:  f 
#> 
#> Slot "harvest.spwn":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1980 1981 1982 1983 1984 1985 1986 1987 1988 1989 1990
#>   2 0    0    0    0    0    0    0    0    0    0    0   
#>   3 0    0    0    0    0    0    0    0    0    0    0   
#>   4 0    0    0    0    0    0    0    0    0    0    0   
#>   5 0    0    0    0    0    0    0    0    0    0    0   
#>   6 0    0    0    0    0    0    0    0    0    0    0   
#> 
#> units:   
#> 
#> Slot "m.spwn":
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1980 1981 1982 1983 1984 1985 1986 1987 1988 1989 1990
#>   2 0    0    0    0    0    0    0    0    0    0    0   
#>   3 0    0    0    0    0    0    0    0    0    0    0   
#>   4 0    0    0    0    0    0    0    0    0    0    0   
#>   5 0    0    0    0    0    0    0    0    0    0    0   
#>   6 0    0    0    0    0    0    0    0    0    0    0   
#> 
#> units:   
#> 
#> Slot "name":
#> [1] "PLE"
#> 
#> Slot "desc":
#> [1] "Plaice in IV. ICES WGNSSK 2018. FLAAP"
#> 
#> Slot "range":
#>       min       max plusgroup   minyear   maxyear   minfbar   maxfbar 
#>         2         6        NA      1980      1990         2         6 
#> 

# Calculate SSB, and SSB per recruit at zero fishing mortality
  ssb(ple4)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 342223 355375 362119 380052 391386 482245 440658 430475 383583 404516
#>      year
#> age   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 473938 458977 402865 370472 361610 366129 302365 298094 301757 328726
#>      year
#> age   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 329116 327544 302271 319090 290780 284211 339238 367251 394666 407972
#>      year
#> age   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 470952 424696 448230 396458 356947 311431 279962 233481 222203 203391
#>      year
#> age   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 204948 237863 219149 230902 234211 221902 248312 233478 253737 284447
#>      year
#> age   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all 293330 371837 453026 554245 575459 617539 709948 823276 774157 836453
#>      year
#> age   2017  
#>   all 913290
#> 
#> units:  t 

  ssbpurec(ple4)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957  
#>   all 5.8363
#> 
#> units:  NA 

# Coerce an FLStock to an FLBiol
  biol <- as(ple4, "FLBiol")

# Initialise an FLSR object from an FLStock
  flsr <- as.FLSR(ple4)