Short aliases for most FLCore classes construction methods

When working interactively, the naming convention of the FLCore classes can become tiresome. This set of short, lowercase aliases make calling the class construction methods a bit simpler.

flq(object, ...)
flqp(object, ...)
flqd(object, var, ...)
flc(object, ...)
flp(object, ...)
fls(object, ...)
flsl(object, ...)
fli(object, ...)
flib(object, ...)
flsr(model, ...)
flqs(object, ...)
flcs(object, ...)
flss(object, ...)
flis(object, ...)
flps(object, ...)
flb(object, ...)
flbs(object, ...)
flms(object, ...)
flmss(object, ...)
pm(object, ...)

Value

An object of the requested class

Details

We recomend you use the full name of classes and methods when developing code that you intend to keep or distribute. A script written using these aliases can be later 'corrected' by substituting the aliases used in your editor.

The aliases' equivalences are as follows:

flq(...)

For FLQuant(...)

flqp(...)

For FLQuantPoint(...)

flqd(...)

For FLQuantDistr(...)

flc(...)

For FLCohort(...)

flp(...)

For FLPar(...)

fls(...)

For FLStock(...)

flsl(...)

For FLStockLen(...)

fli(...)

For FLIndex(...)

flib(...)

For FLIndexBiomass(...)

flsr(...)

For FLSR(...)

flqs(...)

For FLQuants(...)

flcs(...)

For FLCohorts(...)

flss(...)

For FLStocks(...)

flis(...)

For FLIndices(...)

flps(...)

For FLPars(...)

flb(...)

For FLBiol(...)

flbs(...)

For FLBiols(...)

flms(...)

For FLModelSim(...)

flmss(...)

For FLModelSims(...)

pm(...)

For predictModel(...)

See also

FLQuant, FLQuantPoint, FLQuantDistr, FLCohort, FLPar, FLStock, FLStockLen, FLIndex, FLIndexBiomass, FLSR, FLQuants, FLCohorts, FLStocks, FLIndices, FLPars, FLBiol, FLBiols, FLModelSim, FLModelSims, predictModel

Examples

flq <- flc(1:10, units="kg")