Append objects along the year dimension

Method to append objects along the year dimensions, by extending, combining and substituting sections of them.

Usage

# S4 method for class 'FLQuant,FLQuant'
append(x, values, after = dims(values)$minyear - 1)

# S4 method for class 'FLStock,FLStock'
append(x, values, after = dims(values)$minyear - 1)

Arguments

x

the object to which the values are to be appended to.

values

to be included in the modified object.

after

a year dimname after with the values are to be appended.

Value

An object of the same class as x with values appended.

Details

FLR objects are commonly manipulated along the year dimension, and the append method offers a simple interface for substituting parts of an object with another, or combine them into one, extending them when necessary. The object to be included or added to the first will be placed as defined by the year dimnames, unless the after input argument specifies otherwise.

Attributes like dimnames and units will always be taken from the first argument, unless the necessary chnages to dimnames$year

See also

Author

The FLR Team

Examples

# append(FLQuant, FLQuant)
fq1 <- FLQuant(1, dimnames=list(age=1:3, year=2000:2010))
fq2 <- FLQuant(2, dimnames=list(age=1:3, year=2011:2012))
fq3 <- FLQuant(2, dimnames=list(age=1:3, year=2014:2016))

# Appends by dimnames$year
append(fq1, fq2)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   1 1    1    1    1    1    1    1    1    1    1    1    2    2   
#>   2 1    1    1    1    1    1    1    1    1    1    1    2    2   
#>   3 1    1    1    1    1    1    1    1    1    1    1    2    2   
#> 
#> units:  NA 
# Appends by dimnames$year with gap (2011:2013)
append(fq1, fq3)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012 2013 2014
#>   1  1    1    1    1    1    1    1    1    1    1    1   NA   NA   NA    2  
#>   2  1    1    1    1    1    1    1    1    1    1    1   NA   NA   NA    2  
#>   3  1    1    1    1    1    1    1    1    1    1    1   NA   NA   NA    2  
#>    year
#> age 2015 2016
#>   1  2    2  
#>   2  2    2  
#>   3  2    2  
#> 
#> units:  NA 
# Appends inside x
append(fq1, fq2, after=2009)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011
#>   1 1    1    1    1    1    1    1    1    1    1    2    2   
#>   2 1    1    1    1    1    1    1    1    1    1    2    2   
#>   3 1    1    1    1    1    1    1    1    1    1    2    2   
#> 
#> units:  NA 
# Appends after end of x
append(fq1, fq2, after=2013)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012 2013 2014
#>   1  1    1    1    1    1    1    1    1    1    1    1   NA   NA   NA    2  
#>   2  1    1    1    1    1    1    1    1    1    1    1   NA   NA   NA    2  
#>   3  1    1    1    1    1    1    1    1    1    1    1   NA   NA   NA    2  
#>    year
#> age 2015
#>   1  2  
#>   2  2  
#>   3  2  
#> 
#> units:  NA 

# append(FLStock, FLStock)
data(ple4)
fs1 <- window(ple4, end=2001)
fs2 <- window(ple4, start=2002)
fs3 <- window(ple4, start=2005)

# Appends by dimnames$year
stock.n(append(fs1, fs2))
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957    1958    1959    1960    1961    1962    1963    1964    1965   
#>   1   477074  710748  874712  797702  870799  615691  610017 2449900  664500
#>   2   268185  391105  574019  696497  629331  688002  494583  502620 2057370
#>   3   361589  204551  286216  404963  485155  441333  482883  341643  343934
#>   4   223134  251892  135739  182546  255875  308828  276136  283950  189574
#>   5   147904  148258  164040   87018  116368  163035  195022  170312  170953
#>   6    60609  103574  100146  107321   55977   74520  103605  121552  104036
#>   7    62627   44412   73005   68029   71202   36670   47899   64387   73564
#>   8    49418   45232   32054   52466   48327   49595   24892   31672   42048
#>   9    31997   35609   32755   23309   38248   35143   35587   17426   21773
#>   10   68585   74612   79021   78691   73222   82791   88328   89950   74802
#>     year
#> age  1966    1967    1968    1969    1970    1971    1972    1973    1974   
#>   1   579075  428110  418228  666902  671454  433599  367450 1391430 1074920
#>   2   561599  486279  351300  328477  499525  495583  322847  274502 1025520
#>   3  1437490  402660  346712  239678  214078  324931  328032  212726  175441
#>   4   194255  862270  249951  214175  146130  130830  198448  194134  119185
#>   5   113379  117592  533809  158259  137189   92885   80660  115448  105314
#>   6   103728   69241   72676  333787   99805   86638   57736   47858   64234
#>   7    63414   64736   43516   44971  204065   61698   54135   35371   28087
#>   8    48695   42848   43754   28686   28864  130295   39537   34377   21998
#>   9    29157   34460   30486   30610   19556   19274   85654   25584   21900
#>   10   67159   68866   75288   76834   76615   66653   57589   92987   75118
#>     year
#> age  1975    1976    1977    1978    1979    1980    1981    1982    1983   
#>   1   787372  674010 1033740  879043  915553 1078660  999968 1935350 1375880
#>   2   767410  534441  432591  640577  548706  603278  759629  728751 1391630
#>   3   635424  469588  324352  256771  368443  308663  338604  431979  422976
#>   4    95487  354654  270674  186122  141111  188898  147121  156042  206296
#>   5    62528   51566  199395  153084  102391   74240   94918   71554   75328
#>   6    56842   34796   29688  113431   83439   54396   39883   51872   38992
#>   7    36941   33525   20971   17442   63121   45293   30398   23273   31033
#>   8    17263   22880   20930   12952   10501   37093   26405   18034   14363
#>   9    13875   10882   14513   13395    8319    6669   23029   16239   11328
#>   10   60678   46683   36598   33352   31240   26644   22198   29551   29399
#>     year
#> age  1984    1985    1986    1987    1988    1989    1990    1991    1992   
#>   1  1302060 1792220 4303680 1910200 1774940 1250510 1083810  981356  854841
#>   2   948596  864871 1177750 2866780 1308470 1251000  898824  785747  709583
#>   3   820717  559870  503176  670566 1602170  730349  716388  530410  465814
#>   4   214367  429744  291704  253291  320928  743636  347377  357105  268732
#>   5   100486  105579  212338  142492  120680  150385  353208  169944  177919
#>   6    40129   51598   51971  101004   66592   55846   68750  159986   77446
#>   7    23210   22812   27066   25238   47265   30819   25576   30985   70810
#>   8    19676   14467   13364   14671   12950   23965   16156   13837   16301
#>   9     9270   12766    9167    8060    8304    7096   13682    9708    8133
#>   10   26037   22968   23691   21446   18014   15058   12910   16298   15761
#>     year
#> age  1993    1994    1995    1996    1997    1998    1999    2000    2001   
#>   1   550376  566448  932162  893056 2431310  778427  683151  857525  634808
#>   2   614021  398693  423980  722992  708162 1941390  617294  531403  653888
#>   3   412376  355244  239117  261727  433954  398394 1091830  381345  349511
#>   4   230658  199926  173831  115825  114808  163844  151199  510105  205661
#>   5   133599  113279   96740   81896   51782   48218   67219   64651  232564
#>   6    82753   63227   53667   45202   37346   22898   20632   28115   27484
#>   7    33854   36416   28660   24871   20579   16298    9911    9398   13485
#>   8    33881   15014   16697   14022   12186    9697    7928    5446    5587
#>   9     8552   15996    7340    8872    7571    6375    5220    4778    3568
#>   10   13207   11094   14408   12542   12624   11590   10286    9460    9464
#>     year
#> age  2002    2003    2004    2005    2006    2007    2008    2009    2010   
#>   1  1792880  557844 1235790  863893  875191 1379750 1135050 1088820 1444570
#>   2   481832 1362670  417612  894667  618023  651007 1067080  879336  826492
#>   3   413685  272903  729522  238706  552652  393107  420759  719298  627121
#>   4   181506  185801  118238  365617  136058  326369  234965  266894  493740
#>   5    98490   90506   97419   66113  219194   87129  219954  163268  188123
#>   6   106479   49763   49696   57108   41505  148406   62707  164048  123503
#>   7    13278   51719   26044   29565   37778   29246  109232   48044  129629
#>   8     7752    7046   28479   16720   21234   28311   22299   85183   38535
#>   9     3675    4964    4649   20534   12942   17026   23120   18461   71435
#>   10    9309    9758   11424   12765   27199   33875   43959   58555   67452
#>     year
#> age  2011    2012    2013    2014    2015    2016    2017   
#>   1  1608190 1278010 1455050 1640700  895620 1211320 1823000
#>   2  1109460 1299830 1063130 1193580 1286330  689025  973599
#>   3   613580  837139  980853  793858  880738  948914  515997
#>   4   448410  436124  582663  681616  559741  626799  672732
#>   5   349991  318435  309658  413410  483694  398150  447979
#>   6   142101  261180  234879  228655  308142  362736  298156
#>   7    98703  112594  204189  183635  180488  244483  286575
#>   8   106302   81974   93938  169885  151867  148885  202450
#>   9    32718   91210   70713   80902  145583  130001  128177
#>   10  121698  135291  198423  235842  277893  372467  443492
#> 
#> units:  1000 

# Appends by dimnames$year with gap (2011:2013)
stock.n(append(fs1, fs3))
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957    1958    1959    1960    1961    1962    1963    1964    1965   
#>   1   477074  710748  874712  797702  870799  615691  610017 2449900  664500
#>   2   268185  391105  574019  696497  629331  688002  494583  502620 2057370
#>   3   361589  204551  286216  404963  485155  441333  482883  341643  343934
#>   4   223134  251892  135739  182546  255875  308828  276136  283950  189574
#>   5   147904  148258  164040   87018  116368  163035  195022  170312  170953
#>   6    60609  103574  100146  107321   55977   74520  103605  121552  104036
#>   7    62627   44412   73005   68029   71202   36670   47899   64387   73564
#>   8    49418   45232   32054   52466   48327   49595   24892   31672   42048
#>   9    31997   35609   32755   23309   38248   35143   35587   17426   21773
#>   10   68585   74612   79021   78691   73222   82791   88328   89950   74802
#>     year
#> age  1966    1967    1968    1969    1970    1971    1972    1973    1974   
#>   1   579075  428110  418228  666902  671454  433599  367450 1391430 1074920
#>   2   561599  486279  351300  328477  499525  495583  322847  274502 1025520
#>   3  1437490  402660  346712  239678  214078  324931  328032  212726  175441
#>   4   194255  862270  249951  214175  146130  130830  198448  194134  119185
#>   5   113379  117592  533809  158259  137189   92885   80660  115448  105314
#>   6   103728   69241   72676  333787   99805   86638   57736   47858   64234
#>   7    63414   64736   43516   44971  204065   61698   54135   35371   28087
#>   8    48695   42848   43754   28686   28864  130295   39537   34377   21998
#>   9    29157   34460   30486   30610   19556   19274   85654   25584   21900
#>   10   67159   68866   75288   76834   76615   66653   57589   92987   75118
#>     year
#> age  1975    1976    1977    1978    1979    1980    1981    1982    1983   
#>   1   787372  674010 1033740  879043  915553 1078660  999968 1935350 1375880
#>   2   767410  534441  432591  640577  548706  603278  759629  728751 1391630
#>   3   635424  469588  324352  256771  368443  308663  338604  431979  422976
#>   4    95487  354654  270674  186122  141111  188898  147121  156042  206296
#>   5    62528   51566  199395  153084  102391   74240   94918   71554   75328
#>   6    56842   34796   29688  113431   83439   54396   39883   51872   38992
#>   7    36941   33525   20971   17442   63121   45293   30398   23273   31033
#>   8    17263   22880   20930   12952   10501   37093   26405   18034   14363
#>   9    13875   10882   14513   13395    8319    6669   23029   16239   11328
#>   10   60678   46683   36598   33352   31240   26644   22198   29551   29399
#>     year
#> age  1984    1985    1986    1987    1988    1989    1990    1991    1992   
#>   1  1302060 1792220 4303680 1910200 1774940 1250510 1083810  981356  854841
#>   2   948596  864871 1177750 2866780 1308470 1251000  898824  785747  709583
#>   3   820717  559870  503176  670566 1602170  730349  716388  530410  465814
#>   4   214367  429744  291704  253291  320928  743636  347377  357105  268732
#>   5   100486  105579  212338  142492  120680  150385  353208  169944  177919
#>   6    40129   51598   51971  101004   66592   55846   68750  159986   77446
#>   7    23210   22812   27066   25238   47265   30819   25576   30985   70810
#>   8    19676   14467   13364   14671   12950   23965   16156   13837   16301
#>   9     9270   12766    9167    8060    8304    7096   13682    9708    8133
#>   10   26037   22968   23691   21446   18014   15058   12910   16298   15761
#>     year
#> age  1993    1994    1995    1996    1997    1998    1999    2000    2001   
#>   1   550376  566448  932162  893056 2431310  778427  683151  857525  634808
#>   2   614021  398693  423980  722992  708162 1941390  617294  531403  653888
#>   3   412376  355244  239117  261727  433954  398394 1091830  381345  349511
#>   4   230658  199926  173831  115825  114808  163844  151199  510105  205661
#>   5   133599  113279   96740   81896   51782   48218   67219   64651  232564
#>   6    82753   63227   53667   45202   37346   22898   20632   28115   27484
#>   7    33854   36416   28660   24871   20579   16298    9911    9398   13485
#>   8    33881   15014   16697   14022   12186    9697    7928    5446    5587
#>   9     8552   15996    7340    8872    7571    6375    5220    4778    3568
#>   10   13207   11094   14408   12542   12624   11590   10286    9460    9464
#>     year
#> age  2002    2003    2004    2005    2006    2007    2008    2009    2010   
#>   1       NA      NA      NA  863893  875191 1379750 1135050 1088820 1444570
#>   2       NA      NA      NA  894667  618023  651007 1067080  879336  826492
#>   3       NA      NA      NA  238706  552652  393107  420759  719298  627121
#>   4       NA      NA      NA  365617  136058  326369  234965  266894  493740
#>   5       NA      NA      NA   66113  219194   87129  219954  163268  188123
#>   6       NA      NA      NA   57108   41505  148406   62707  164048  123503
#>   7       NA      NA      NA   29565   37778   29246  109232   48044  129629
#>   8       NA      NA      NA   16720   21234   28311   22299   85183   38535
#>   9       NA      NA      NA   20534   12942   17026   23120   18461   71435
#>   10      NA      NA      NA   12765   27199   33875   43959   58555   67452
#>     year
#> age  2011    2012    2013    2014    2015    2016    2017   
#>   1  1608190 1278010 1455050 1640700  895620 1211320 1823000
#>   2  1109460 1299830 1063130 1193580 1286330  689025  973599
#>   3   613580  837139  980853  793858  880738  948914  515997
#>   4   448410  436124  582663  681616  559741  626799  672732
#>   5   349991  318435  309658  413410  483694  398150  447979
#>   6   142101  261180  234879  228655  308142  362736  298156
#>   7    98703  112594  204189  183635  180488  244483  286575
#>   8   106302   81974   93938  169885  151867  148885  202450
#>   9    32718   91210   70713   80902  145583  130001  128177
#>   10  121698  135291  198423  235842  277893  372467  443492
#> 
#> units:  1000 

# Appends inside x
stock.n(append(fs1, fs3, after=2000))
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957    1958    1959    1960    1961    1962    1963    1964    1965   
#>   1   477074  710748  874712  797702  870799  615691  610017 2449900  664500
#>   2   268185  391105  574019  696497  629331  688002  494583  502620 2057370
#>   3   361589  204551  286216  404963  485155  441333  482883  341643  343934
#>   4   223134  251892  135739  182546  255875  308828  276136  283950  189574
#>   5   147904  148258  164040   87018  116368  163035  195022  170312  170953
#>   6    60609  103574  100146  107321   55977   74520  103605  121552  104036
#>   7    62627   44412   73005   68029   71202   36670   47899   64387   73564
#>   8    49418   45232   32054   52466   48327   49595   24892   31672   42048
#>   9    31997   35609   32755   23309   38248   35143   35587   17426   21773
#>   10   68585   74612   79021   78691   73222   82791   88328   89950   74802
#>     year
#> age  1966    1967    1968    1969    1970    1971    1972    1973    1974   
#>   1   579075  428110  418228  666902  671454  433599  367450 1391430 1074920
#>   2   561599  486279  351300  328477  499525  495583  322847  274502 1025520
#>   3  1437490  402660  346712  239678  214078  324931  328032  212726  175441
#>   4   194255  862270  249951  214175  146130  130830  198448  194134  119185
#>   5   113379  117592  533809  158259  137189   92885   80660  115448  105314
#>   6   103728   69241   72676  333787   99805   86638   57736   47858   64234
#>   7    63414   64736   43516   44971  204065   61698   54135   35371   28087
#>   8    48695   42848   43754   28686   28864  130295   39537   34377   21998
#>   9    29157   34460   30486   30610   19556   19274   85654   25584   21900
#>   10   67159   68866   75288   76834   76615   66653   57589   92987   75118
#>     year
#> age  1975    1976    1977    1978    1979    1980    1981    1982    1983   
#>   1   787372  674010 1033740  879043  915553 1078660  999968 1935350 1375880
#>   2   767410  534441  432591  640577  548706  603278  759629  728751 1391630
#>   3   635424  469588  324352  256771  368443  308663  338604  431979  422976
#>   4    95487  354654  270674  186122  141111  188898  147121  156042  206296
#>   5    62528   51566  199395  153084  102391   74240   94918   71554   75328
#>   6    56842   34796   29688  113431   83439   54396   39883   51872   38992
#>   7    36941   33525   20971   17442   63121   45293   30398   23273   31033
#>   8    17263   22880   20930   12952   10501   37093   26405   18034   14363
#>   9    13875   10882   14513   13395    8319    6669   23029   16239   11328
#>   10   60678   46683   36598   33352   31240   26644   22198   29551   29399
#>     year
#> age  1984    1985    1986    1987    1988    1989    1990    1991    1992   
#>   1  1302060 1792220 4303680 1910200 1774940 1250510 1083810  981356  854841
#>   2   948596  864871 1177750 2866780 1308470 1251000  898824  785747  709583
#>   3   820717  559870  503176  670566 1602170  730349  716388  530410  465814
#>   4   214367  429744  291704  253291  320928  743636  347377  357105  268732
#>   5   100486  105579  212338  142492  120680  150385  353208  169944  177919
#>   6    40129   51598   51971  101004   66592   55846   68750  159986   77446
#>   7    23210   22812   27066   25238   47265   30819   25576   30985   70810
#>   8    19676   14467   13364   14671   12950   23965   16156   13837   16301
#>   9     9270   12766    9167    8060    8304    7096   13682    9708    8133
#>   10   26037   22968   23691   21446   18014   15058   12910   16298   15761
#>     year
#> age  1993    1994    1995    1996    1997    1998    1999    2000    2001   
#>   1   550376  566448  932162  893056 2431310  778427  683151  857525  634808
#>   2   614021  398693  423980  722992  708162 1941390  617294  531403  653888
#>   3   412376  355244  239117  261727  433954  398394 1091830  381345  349511
#>   4   230658  199926  173831  115825  114808  163844  151199  510105  205661
#>   5   133599  113279   96740   81896   51782   48218   67219   64651  232564
#>   6    82753   63227   53667   45202   37346   22898   20632   28115   27484
#>   7    33854   36416   28660   24871   20579   16298    9911    9398   13485
#>   8    33881   15014   16697   14022   12186    9697    7928    5446    5587
#>   9     8552   15996    7340    8872    7571    6375    5220    4778    3568
#>   10   13207   11094   14408   12542   12624   11590   10286    9460    9464
#>     year
#> age  2002    2003    2004    2005    2006    2007    2008    2009    2010   
#>   1       NA      NA      NA  863893  875191 1379750 1135050 1088820 1444570
#>   2       NA      NA      NA  894667  618023  651007 1067080  879336  826492
#>   3       NA      NA      NA  238706  552652  393107  420759  719298  627121
#>   4       NA      NA      NA  365617  136058  326369  234965  266894  493740
#>   5       NA      NA      NA   66113  219194   87129  219954  163268  188123
#>   6       NA      NA      NA   57108   41505  148406   62707  164048  123503
#>   7       NA      NA      NA   29565   37778   29246  109232   48044  129629
#>   8       NA      NA      NA   16720   21234   28311   22299   85183   38535
#>   9       NA      NA      NA   20534   12942   17026   23120   18461   71435
#>   10      NA      NA      NA   12765   27199   33875   43959   58555   67452
#>     year
#> age  2011    2012    2013    2014    2015    2016    2017   
#>   1  1608190 1278010 1455050 1640700  895620 1211320 1823000
#>   2  1109460 1299830 1063130 1193580 1286330  689025  973599
#>   3   613580  837139  980853  793858  880738  948914  515997
#>   4   448410  436124  582663  681616  559741  626799  672732
#>   5   349991  318435  309658  413410  483694  398150  447979
#>   6   142101  261180  234879  228655  308142  362736  298156
#>   7    98703  112594  204189  183635  180488  244483  286575
#>   8   106302   81974   93938  169885  151867  148885  202450
#>   9    32718   91210   70713   80902  145583  130001  128177
#>   10  121698  135291  198423  235842  277893  372467  443492
#> 
#> units:  1000 
# Appends after end of x
stock.n(append(fs1, fs3, after=2005))
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957    1958    1959    1960    1961    1962    1963    1964    1965   
#>   1   477074  710748  874712  797702  870799  615691  610017 2449900  664500
#>   2   268185  391105  574019  696497  629331  688002  494583  502620 2057370
#>   3   361589  204551  286216  404963  485155  441333  482883  341643  343934
#>   4   223134  251892  135739  182546  255875  308828  276136  283950  189574
#>   5   147904  148258  164040   87018  116368  163035  195022  170312  170953
#>   6    60609  103574  100146  107321   55977   74520  103605  121552  104036
#>   7    62627   44412   73005   68029   71202   36670   47899   64387   73564
#>   8    49418   45232   32054   52466   48327   49595   24892   31672   42048
#>   9    31997   35609   32755   23309   38248   35143   35587   17426   21773
#>   10   68585   74612   79021   78691   73222   82791   88328   89950   74802
#>     year
#> age  1966    1967    1968    1969    1970    1971    1972    1973    1974   
#>   1   579075  428110  418228  666902  671454  433599  367450 1391430 1074920
#>   2   561599  486279  351300  328477  499525  495583  322847  274502 1025520
#>   3  1437490  402660  346712  239678  214078  324931  328032  212726  175441
#>   4   194255  862270  249951  214175  146130  130830  198448  194134  119185
#>   5   113379  117592  533809  158259  137189   92885   80660  115448  105314
#>   6   103728   69241   72676  333787   99805   86638   57736   47858   64234
#>   7    63414   64736   43516   44971  204065   61698   54135   35371   28087
#>   8    48695   42848   43754   28686   28864  130295   39537   34377   21998
#>   9    29157   34460   30486   30610   19556   19274   85654   25584   21900
#>   10   67159   68866   75288   76834   76615   66653   57589   92987   75118
#>     year
#> age  1975    1976    1977    1978    1979    1980    1981    1982    1983   
#>   1   787372  674010 1033740  879043  915553 1078660  999968 1935350 1375880
#>   2   767410  534441  432591  640577  548706  603278  759629  728751 1391630
#>   3   635424  469588  324352  256771  368443  308663  338604  431979  422976
#>   4    95487  354654  270674  186122  141111  188898  147121  156042  206296
#>   5    62528   51566  199395  153084  102391   74240   94918   71554   75328
#>   6    56842   34796   29688  113431   83439   54396   39883   51872   38992
#>   7    36941   33525   20971   17442   63121   45293   30398   23273   31033
#>   8    17263   22880   20930   12952   10501   37093   26405   18034   14363
#>   9    13875   10882   14513   13395    8319    6669   23029   16239   11328
#>   10   60678   46683   36598   33352   31240   26644   22198   29551   29399
#>     year
#> age  1984    1985    1986    1987    1988    1989    1990    1991    1992   
#>   1  1302060 1792220 4303680 1910200 1774940 1250510 1083810  981356  854841
#>   2   948596  864871 1177750 2866780 1308470 1251000  898824  785747  709583
#>   3   820717  559870  503176  670566 1602170  730349  716388  530410  465814
#>   4   214367  429744  291704  253291  320928  743636  347377  357105  268732
#>   5   100486  105579  212338  142492  120680  150385  353208  169944  177919
#>   6    40129   51598   51971  101004   66592   55846   68750  159986   77446
#>   7    23210   22812   27066   25238   47265   30819   25576   30985   70810
#>   8    19676   14467   13364   14671   12950   23965   16156   13837   16301
#>   9     9270   12766    9167    8060    8304    7096   13682    9708    8133
#>   10   26037   22968   23691   21446   18014   15058   12910   16298   15761
#>     year
#> age  1993    1994    1995    1996    1997    1998    1999    2000    2001   
#>   1   550376  566448  932162  893056 2431310  778427  683151  857525  634808
#>   2   614021  398693  423980  722992  708162 1941390  617294  531403  653888
#>   3   412376  355244  239117  261727  433954  398394 1091830  381345  349511
#>   4   230658  199926  173831  115825  114808  163844  151199  510105  205661
#>   5   133599  113279   96740   81896   51782   48218   67219   64651  232564
#>   6    82753   63227   53667   45202   37346   22898   20632   28115   27484
#>   7    33854   36416   28660   24871   20579   16298    9911    9398   13485
#>   8    33881   15014   16697   14022   12186    9697    7928    5446    5587
#>   9     8552   15996    7340    8872    7571    6375    5220    4778    3568
#>   10   13207   11094   14408   12542   12624   11590   10286    9460    9464
#>     year
#> age  2002    2003    2004    2005    2006    2007    2008    2009    2010   
#>   1       NA      NA      NA  863893  875191 1379750 1135050 1088820 1444570
#>   2       NA      NA      NA  894667  618023  651007 1067080  879336  826492
#>   3       NA      NA      NA  238706  552652  393107  420759  719298  627121
#>   4       NA      NA      NA  365617  136058  326369  234965  266894  493740
#>   5       NA      NA      NA   66113  219194   87129  219954  163268  188123
#>   6       NA      NA      NA   57108   41505  148406   62707  164048  123503
#>   7       NA      NA      NA   29565   37778   29246  109232   48044  129629
#>   8       NA      NA      NA   16720   21234   28311   22299   85183   38535
#>   9       NA      NA      NA   20534   12942   17026   23120   18461   71435
#>   10      NA      NA      NA   12765   27199   33875   43959   58555   67452
#>     year
#> age  2011    2012    2013    2014    2015    2016    2017    2018   
#>   1  1608190 1278010 1455050 1640700  895620 1211320 1823000      NA
#>   2  1109460 1299830 1063130 1193580 1286330  689025  973599      NA
#>   3   613580  837139  980853  793858  880738  948914  515997      NA
#>   4   448410  436124  582663  681616  559741  626799  672732      NA
#>   5   349991  318435  309658  413410  483694  398150  447979      NA
#>   6   142101  261180  234879  228655  308142  362736  298156      NA
#>   7    98703  112594  204189  183635  180488  244483  286575      NA
#>   8   106302   81974   93938  169885  151867  148885  202450      NA
#>   9    32718   91210   70713   80902  145583  130001  128177      NA
#>   10  121698  135291  198423  235842  277893  372467  443492      NA
#> 
#> units:  1000