apply method for FLCore classes

Applies a function over the margins of an array-based FLCore class

# S4 method for FLArray,numeric,`function`
apply(X, MARGIN, FUN, ..., simplify = TRUE)

# S4 method for FLPar,ANY,ANY
apply(X, MARGIN, FUN, ..., simplify = TRUE)

# S4 method for FLQuantJK,numeric,`function`
apply(X, MARGIN, FUN, ..., simplify = TRUE)

# S4 method for FLParJK,numeric,`function`
apply(X, MARGIN, FUN, ..., simplify = TRUE)

Details

These methods call R's base::apply on an FLArray the standard arithmetic operators included in the Arith group ("+", "-", "*", `"^", "%%", "%/%", and "/"), so that they return an object of the appropriate class.

When the operation involves objects of two classes (e.g. FLPar and FLQuant), the class is the returned object is that of the more complexs object, in this case FLQuant.

See also

Author

The FLR Team

Examples

flq <- FLQuant(rlnorm(90), dim=c(3,10), units='kg')
flp <- FLPar(a=99)

# FLQuant and numeric
flq * 25
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1         2         3         4         5         6         7        
#>     1   19.8703   14.4465   24.6354   12.2230    5.8750   14.3226   23.5864
#>     2  105.7686   60.8900   11.7624  182.5817 1221.9580   98.5148   67.4380
#>     3    5.1542   96.1651   16.6089    5.9470   10.5272  168.5980   76.8502
#>      year
#> quant 8         9         10       
#>     1   14.4176   16.8488   15.7627
#>     2   56.0641   33.6592   35.9396
#>     3  105.3440    1.6677   97.9107
#> 
#> units:  kg 
# Two FLQuant objects
flq + flq
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1        2        3        4        5        6        7        8       
#>     1  1.58963  1.15572  1.97083  0.97784  0.47000  1.14581  1.88691  1.15341
#>     2  8.46149  4.87120  0.94099 14.60654 97.75664  7.88118  5.39504  4.48513
#>     3  0.41233  7.69321  1.32871  0.47576  0.84218 13.48784  6.14802  8.42752
#>      year
#> quant 9        10      
#>     1  1.34791  1.26102
#>     2  2.69274  2.87517
#>     3  0.13342  7.83286
#> 
#> units:  kg