apply method for FLCore classes

Applies a function over the margins of an array-based FLCore class

# S4 method for FLArray,numeric,`function`
apply(X, MARGIN, FUN, ...)

# S4 method for FLPar,ANY,ANY
apply(X, MARGIN, FUN, ...)

# S4 method for FLQuantJK,numeric,`function`
apply(X, MARGIN, FUN, ...)

# S4 method for FLParJK,numeric,`function`
apply(X, MARGIN, FUN, ...)

Details

These methods call R's base::apply on an FLArray the standard arithmetic operators included in the Arith group ("+", "-", "*", `"^", " "/"), so that they return an object of the appropriate class.

When the operation involves objects of two classes (e.g. FLPar and FLQuant), the class is the returned object is that of the more complexs object, in this case FLQuant.

See also

base::apply

Examples

flq <- FLQuant(rlnorm(90), dim=c(3,10), units='kg')
flp <- FLPar(a=99)

# FLQuant and numeric
flq * 25

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1        2        3        4        5        6        7        8       
#>     1   6.7537 133.8963  12.0807  39.4671  11.8366   5.8128  19.3728   6.4374
#>     2   2.4368  26.8593 208.8267  50.4469  20.5601  11.4609 101.1428   8.0637
#>     3  76.9089  51.9008  19.6924  13.0810  23.4884  27.1847  77.7664  38.1011
#>      year
#> quant 9        10      
#>     1  15.2718  15.0328
#>     2  34.0636  17.4996
#>     3 115.8129   2.3655
#> 
#> units:  kg 

# Two FLQuant objects
flq + flq

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1        2        3        4        5        6        7        8       
#>     1  0.54030 10.71171  0.96645  3.15737  0.94693  0.46502  1.54982  0.51499
#>     2  0.19495  2.14874 16.70613  4.03575  1.64481  0.91688  8.09143  0.64509
#>     3  6.15271  4.15206  1.57539  1.04648  1.87907  2.17478  6.22131  3.04809
#>      year
#> quant 9        10      
#>     1  1.22174  1.20263
#>     2  2.72509  1.39997
#>     3  9.26503  0.18924
#> 
#> units:  kg