Convert Objects Between Classes

Objects of various FLCore classes can be converted into other classes, both basic R ones, like data.frame, and defined in the package. For the specifics of the precise calculations carried out for each pair of classes, see Details.

Arguments

object
Object to be converted.

Class
Name of the class to convert the object to, character.

Value

An object of the requested class.

FLArray to data.frame

The six dimensions of an `FLArray` are converted into seven columns, named `quant` (or any other name given to the first dimension in the object), `year`, `unit`, `season`, `area`, `iter` and `data`. The last one contains the actual numbers stored in the array. `units` are stored as an attribute to the `data.frame`. The `year` and `data` columns are of type `numeric`, while all others are `factor`.

FLPar to data.frame

The two or more dimensions of an *FLPar* objects are converted into three or more columns. For a 2D objects, they are named *params*, *iter* and *data*. The last one contains the actual numbers stored in the array, in a column type `numeric`, while all others are `factor`.

data.frame to FLQuant

A *data.frame* with the right column names is converted into an *FLQuant* object with missing values being added. Missing columns are assumed to contain the default dimnames in *FLQuant*.

*FLBiol* to *FLStock*

- *n* = *stock.n* - *wt* = *stock.wt* - *m* = *m* - *mat* = *mat()* - *m.spwm*, *harvest.spwn* = *spwn*

See also

Examples

# from FLQuant to data.frame
as(FLQuant(rnorm(100), dim=c(5, 20)), "data.frame")

#>     quant year   unit season   area iter         data
#> 1       1    1 unique    all unique    1 -0.211436037
#> 2       2    1 unique    all unique    1  0.737402928
#> 3       3    1 unique    all unique    1  1.232897866
#> 4       4    1 unique    all unique    1 -0.960233033
#> 5       5    1 unique    all unique    1  0.760276228
#> 6       1    2 unique    all unique    1 -0.454219888
#> 7       2    2 unique    all unique    1  1.160524994
#> 8       3    2 unique    all unique    1 -2.040130752
#> 9       4    2 unique    all unique    1  0.499544069
#> 10      5    2 unique    all unique    1  0.227548897
#> 11      1    3 unique    all unique    1 -1.040504156
#> 12      2    3 unique    all unique    1  0.591474274
#> 13      3    3 unique    all unique    1  0.157713338
#> 14      4    3 unique    all unique    1 -0.591422024
#> 15      5    3 unique    all unique    1  0.244214097
#> 16      1    4 unique    all unique    1 -1.185068040
#> 17      2    4 unique    all unique    1 -0.081557987
#> 18      3    4 unique    all unique    1  0.379779714
#> 19      4    4 unique    all unique    1 -0.842286320
#> 20      5    4 unique    all unique    1  0.985510173
#> 21      1    5 unique    all unique    1 -1.159435490
#> 22      2    5 unique    all unique    1  0.991196331
#> 23      3    5 unique    all unique    1 -0.660593997
#> 24      4    5 unique    all unique    1 -0.389136475
#> 25      5    5 unique    all unique    1 -0.407945959
#> 26      1    6 unique    all unique    1 -1.538989067
#> 27      2    6 unique    all unique    1 -0.778746009
#> 28      3    6 unique    all unique    1 -1.117638078
#> 29      4    6 unique    all unique    1 -0.390339222
#> 30      5    6 unique    all unique    1 -0.070143896
#> 31      1    7 unique    all unique    1 -0.126470497
#> 32      2    7 unique    all unique    1  0.806825573
#> 33      3    7 unique    all unique    1  0.214014949
#> 34      4    7 unique    all unique    1 -1.208925194
#> 35      5    7 unique    all unique    1 -0.952026623
#> 36      1    8 unique    all unique    1 -0.004674233
#> 37      2    8 unique    all unique    1 -1.506323825
#> 38      3    8 unique    all unique    1 -0.216453113
#> 39      4    8 unique    all unique    1  0.321617563
#> 40      5    8 unique    all unique    1  0.830860930
#> 41      1    9 unique    all unique    1  2.362995851
#> 42      2    9 unique    all unique    1  0.929535255
#> 43      3    9 unique    all unique    1 -0.896632657
#> 44      4    9 unique    all unique    1  2.753657244
#> 45      5    9 unique    all unique    1 -0.155873248
#> 46      1   10 unique    all unique    1  0.136993590
#> 47      2   10 unique    all unique    1 -0.712519451
#> 48      3   10 unique    all unique    1  0.784674879
#> 49      4   10 unique    all unique    1  1.379028927
#> 50      5   10 unique    all unique    1 -0.542211554
#> 51      1   11 unique    all unique    1  0.478736667
#> 52      2   11 unique    all unique    1  0.183826152
#> 53      3   11 unique    all unique    1  1.265289420
#> 54      4   11 unique    all unique    1  0.019348475
#> 55      5   11 unique    all unique    1  0.939554767
#> 56      1   12 unique    all unique    1  0.881084680
#> 57      2   12 unique    all unique    1  1.029701670
#> 58      3   12 unique    all unique    1 -0.546640605
#> 59      4   12 unique    all unique    1 -0.472781995
#> 60      5   12 unique    all unique    1  0.806450268
#> 61      1   13 unique    all unique    1  0.850639802
#> 62      2   13 unique    all unique    1  1.174452946
#> 63      3   13 unique    all unique    1  0.365361910
#> 64      4   13 unique    all unique    1  0.630729751
#> 65      5   13 unique    all unique    1  1.599618041
#> 66      1   14 unique    all unique    1 -0.008955912
#> 67      2   14 unique    all unique    1  2.163959662
#> 68      3   14 unique    all unique    1  0.550850788
#> 69      4   14 unique    all unique    1  0.111095179
#> 70      5   14 unique    all unique    1  1.726431231
#> 71      1   15 unique    all unique    1  2.520699216
#> 72      2   15 unique    all unique    1 -0.138973519
#> 73      3   15 unique    all unique    1 -0.545281793
#> 74      4   15 unique    all unique    1 -0.165945399
#> 75      5   15 unique    all unique    1  0.168474982
#> 76      1   16 unique    all unique    1  1.755076321
#> 77      2   16 unique    all unique    1  0.605604359
#> 78      3   16 unique    all unique    1  1.267348512
#> 79      4   16 unique    all unique    1  0.783988655
#> 80      5   16 unique    all unique    1  0.136484638
#> 81      1   17 unique    all unique    1  0.717036524
#> 82      2   17 unique    all unique    1  1.260887226
#> 83      3   17 unique    all unique    1  0.112543467
#> 84      4   17 unique    all unique    1  0.780217369
#> 85      5   17 unique    all unique    1  1.760211098
#> 86      1   18 unique    all unique    1  0.047148352
#> 87      2   18 unique    all unique    1 -0.137586573
#> 88      3   18 unique    all unique    1 -2.332261954
#> 89      4   18 unique    all unique    1  0.928810617
#> 90      5   18 unique    all unique    1  1.052596748
#> 91      1   19 unique    all unique    1 -0.052039632
#> 92      2   19 unique    all unique    1 -1.358392859
#> 93      3   19 unique    all unique    1  0.806466033
#> 94      4   19 unique    all unique    1 -0.551700757
#> 95      5   19 unique    all unique    1  0.647689174
#> 96      1   20 unique    all unique    1  0.148964691
#> 97      2   20 unique    all unique    1 -0.204193545
#> 98      3   20 unique    all unique    1  2.486932606
#> 99      4   20 unique    all unique    1 -0.312109508
#> 100     5   20 unique    all unique    1 -0.370296485

# from FLPar to data.frame
as(FLPar(phi=rnorm(10), rho=rlnorm(10)), "data.frame")

#>    params iter       data
#> 1     phi    1 -0.9527676
#> 2     rho    1  2.2016681
#> 3     phi    2  0.3849457
#> 4     rho    2  0.6726127
#> 5     phi    3  2.0796986
#> 6     rho    3  2.9344150
#> 7     phi    4 -1.0847689
#> 8     rho    4  0.1417246
#> 9     phi    5 -0.6360062
#> 10    rho    5  1.2064770
#> 11    phi    6  0.9782658
#> 12    rho    6  0.5266847
#> 13    phi    7  1.5172862
#> 14    rho    7  0.6157452
#> 15    phi    8  2.0746376
#> 16    rho    8  0.9741508
#> 17    phi    9  0.2280332
#> 18    rho    9  2.9277690
#> 19    phi   10 -0.9689558
#> 20    rho   10  0.6280911

# from data.frame to FLQuant
as(data.frame(age=rep(1:4, each=3), year=2011:2013, data=rnorm(12)), "FLQuant")

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 2011      2012      2013     
#>   1  1.549528 -0.956717 -0.193851
#>   2 -0.085214 -0.198701 -0.030413
#>   3 -0.402241 -0.243652 -0.576173
#>   4  1.398777  0.707484  0.277812
#> 
#> units:  NA