Methods to compute quantities

Methods to compute total quant-aggregated catch, landings, discards and stock biomass from age or length-structured numbers and mean weights.

computeLandings(object, ...)

computeDiscards(object, ...)

computeCatch(object, ...)

computeStock(object, ...)

# S4 method for FLS
computeLandings(object, na.rm = TRUE)

# S4 method for FLS
computeDiscards(object, na.rm = TRUE)

# S4 method for FLS
computeCatch(object, slot = "catch", na.rm = TRUE)

# S4 method for FLS
computeStock(object, na.rm = TRUE)

Details

These methods compute the total catch, landings, discards and stock biomass from the quant-structured values in numbers and weight per individual. The calculation for landings, discards and stock involves the product of the landings/discards/stock in numbers (landings.n, discards.n or stock.n) by the individual weight-at-quant (landings.wt, discards.wt or stock.wt), as in

$$L=L_n * L_{wt}$$

By selecting slot="catch", computeCatch can calculate in the same way the total catch from the catch-at-quant and weight in the catch. Those two values (in slots catch.n and catch.wt) can also be calculated (from landings and discards) by specifying slot="n" and slot="wt" respectively. Calling computeCatch with option slot="all" will carry out the three calculations. In this case, the returned object will be of class FLQuants, with element names catch, catch.n and catch.wt, which can then be passed directly to the catch<- replacement method.

Generic function

computeCatch(object, ...)

computeLandings(object, ...)

computeDiscards(object, ...)

computeStock(object, ...)

See also

FLComp

Examples

data(ple4)
summary(computeLandings(ple4))

#> An object of class "FLQuant" with:
#> dim  :  1 52 1 1 1 1 
#> quant:  age 
#> units:  t 
#> 
#> Min    :  48874 
#> 1st Qu.:  81541.75 
#> Mean   :  108403.7 
#> Median :  110466 
#> 3rd Qu.:  132758.5 
#> Max    :  169818 
#> NAs    :  0 %

summary(computeCatch(ple4, slot="all"))

#> An object of class "FLQuants"
#> 
#> Elements: catch.wt catch.n catch 
#> 
#> Name: catch.wt 
#> 	dim  :  10 52 1 1 1 1 
#> 	quant:  age 
#> 	units:  kg 
#> 
#> 	Min    :  0.034 
#> 	1st Qu.:  0.2206675 
#> 	Mean   :  0.4551756 
#> 	Median :  0.4372537 
#> 	3rd Qu.:  0.6629051 
#> 	Max    :  1.108295 
#> 	NAs    :  0 %
#> Name: catch.n 
#> 	dim  :  10 52 1 1 1 1 
#> 	quant:  age 
#> 	units:  1000 
#> 
#> 	Min    :  412 
#> 	1st Qu.:  7027.75 
#> 	Mean   :  81702.88 
#> 	Median :  29044 
#> 	3rd Qu.:  101951.5 
#> 	Max    :  1459325 
#> 	NAs    :  0 %
#> Name: catch 
#> 	dim  :  1 52 1 1 1 1 
#> 	quant:  age 
#> 	units:  t 
#> 
#> 	Min    :  78422.95 
#> 	1st Qu.:  126077.3 
#> 	Mean   :  165127 
#> 	Median :  151975.4 
#> 	3rd Qu.:  182754.8 
#> 	Max    :  342985.1 
#> 	NAs    :  0 %

stock(ple4) <- computeStock(ple4)
landings(ple4) <- computeLandings(ple4)
catch.n(ple4) <- computeCatch(ple4, slot="n")
catch(ple4) <- computeCatch(ple4, slot="all")