Computes fishing mortality from abundances, catches and natural mortality

Objects or class 'FLStock' already contain a 'harvest' slot to store estimates of fishing mortality at age, for example those obtained from a stock assessment method. Fishing mortality at age can be recalculated using two methods:

Usage

# S4 method for class 'FLStock,missing'
computeHarvest(object, units = NULL)

Arguments

units

Harvest to be computed as 'f' or 'hr', 'character'.

x

An object of class 'FLStock'.

Value

An 'FLQuant' with the calculated fishing mortalities at age.

See also

Author

The FLR Team

Examples

data(ple4)
# Compute 'f' from stock.n and Baranov
computeHarvest(ple4)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   1  0.0987 0.1137 0.1278 0.1371 0.1356 0.1190 0.0937 0.0746 0.0682 0.0746
#>   2  0.1709 0.2122 0.2489 0.2616 0.2549 0.2540 0.2699 0.2794 0.2585 0.2327
#>   3  0.2615 0.3101 0.3497 0.3591 0.3517 0.3689 0.4310 0.4890 0.4713 0.4111
#>   4  0.3088 0.3289 0.3446 0.3502 0.3507 0.3597 0.3833 0.4074 0.4140 0.4020
#>   5  0.2563 0.2923 0.3243 0.3412 0.3457 0.3534 0.3728 0.3929 0.3996 0.3931
#>   6  0.2109 0.2498 0.2867 0.3103 0.3230 0.3420 0.3757 0.4022 0.3951 0.3715
#>   7  0.2254 0.2261 0.2304 0.2420 0.2616 0.2874 0.3137 0.3261 0.3126 0.2920
#>   8  0.2277 0.2227 0.2186 0.2161 0.2186 0.2319 0.2566 0.2747 0.2661 0.2458
#>   9  0.1987 0.2328 0.2510 0.2315 0.1974 0.1891 0.2203 0.2615 0.2633 0.2355
#>   10 0.1987 0.2328 0.2510 0.2315 0.1974 0.1891 0.2203 0.2615 0.2633 0.2355
#>     year
#> age  1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   1  0.0977 0.1416 0.1890 0.2037 0.1949 0.1916 0.2051 0.2370 0.2875 0.3435
#>   2  0.2383 0.2823 0.3281 0.3300 0.3126 0.3172 0.3477 0.3787 0.3912 0.3994
#>   3  0.3768 0.3817 0.3948 0.3924 0.3931 0.4246 0.4793 0.5083 0.4831 0.4509
#>   4  0.3795 0.3570 0.3454 0.3531 0.3837 0.4417 0.5116 0.5451 0.5161 0.4759
#>   5  0.3812 0.3695 0.3610 0.3596 0.3755 0.4220 0.4863 0.5167 0.4861 0.4521
#>   6  0.3645 0.3800 0.3921 0.3810 0.3703 0.3900 0.4329 0.4532 0.4280 0.4063
#>   7  0.2917 0.3167 0.3434 0.3486 0.3450 0.3541 0.3749 0.3867 0.3791 0.3711
#>   8  0.2404 0.2573 0.2832 0.3038 0.3195 0.3353 0.3509 0.3608 0.3615 0.3552
#>   9  0.2166 0.2197 0.2382 0.2666 0.3002 0.3321 0.3565 0.3693 0.3681 0.3529
#>   10 0.2166 0.2197 0.2382 0.2666 0.3002 0.3321 0.3565 0.3693 0.3681 0.3529
#>     year
#> age  1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   1  0.3786 0.3713 0.3172 0.2506 0.2164 0.2298 0.2719 0.3091 0.3198 0.3063
#>   2  0.4216 0.4531 0.4753 0.4775 0.4645 0.4440 0.4281 0.4273 0.4416 0.4632
#>   3  0.4554 0.4986 0.5681 0.6410 0.6747 0.6391 0.5796 0.5470 0.5520 0.5864
#>   4  0.4699 0.4976 0.5422 0.5882 0.6208 0.6283 0.6193 0.6082 0.6050 0.6165
#>   5  0.4641 0.5069 0.5325 0.5214 0.5042 0.5071 0.5297 0.5665 0.6088 0.6430
#>   6  0.4319 0.4861 0.5110 0.4819 0.4386 0.4137 0.4187 0.4648 0.5452 0.6224
#>   7  0.3819 0.4074 0.4316 0.4396 0.4221 0.3827 0.3557 0.3727 0.4347 0.5124
#>   8  0.3463 0.3428 0.3541 0.3767 0.3861 0.3650 0.3379 0.3326 0.3562 0.4056
#>   9  0.3269 0.3031 0.2952 0.3059 0.3256 0.3431 0.3474 0.3300 0.3110 0.3267
#>   10 0.3269 0.3031 0.2952 0.3059 0.3256 0.3431 0.3474 0.3300 0.3110 0.3267
#>     year
#> age  1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   1  0.2783 0.2498 0.2302 0.2216 0.2243 0.2309 0.2224 0.1897 0.1541 0.1320
#>   2  0.4818 0.4831 0.4575 0.4274 0.4228 0.4427 0.4472 0.4112 0.3824 0.4105
#>   3  0.6369 0.6676 0.6431 0.5962 0.5799 0.6029 0.6240 0.6147 0.6249 0.7240
#>   4  0.6414 0.6580 0.6445 0.6149 0.5967 0.5989 0.6111 0.6259 0.6526 0.7050
#>   5  0.6607 0.6705 0.6827 0.6920 0.6859 0.6655 0.6481 0.6471 0.6609 0.6852
#>   6  0.6594 0.6704 0.6809 0.6970 0.7151 0.7275 0.7209 0.6912 0.6691 0.6869
#>   7  0.5672 0.5792 0.5458 0.5143 0.5423 0.6371 0.7131 0.6798 0.6149 0.6134
#>   8  0.4691 0.5016 0.4605 0.4093 0.4314 0.5450 0.6506 0.6157 0.5323 0.5164
#>   9  0.3935 0.4584 0.4400 0.3896 0.4008 0.4929 0.5737 0.5313 0.4504 0.4285
#>   10 0.3935 0.4584 0.4400 0.3896 0.4008 0.4929 0.5737 0.5313 0.4504 0.4285
#>     year
#> age  1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   1  0.1250 0.1319 0.1512 0.1711 0.1757 0.1744 0.1895 0.2230 0.2349 0.1959
#>   2  0.4752 0.4755 0.3816 0.3190 0.3578 0.4685 0.5248 0.4593 0.3817 0.3524
#>   3  0.8740 0.8688 0.6610 0.5175 0.5552 0.7004 0.7364 0.5908 0.4622 0.4267
#>   4  0.7675 0.7910 0.7496 0.6855 0.6363 0.5959 0.5457 0.4813 0.4116 0.3457
#>   5  0.7160 0.7489 0.7717 0.7554 0.6812 0.5827 0.4995 0.4341 0.3656 0.2900
#>   6  0.7292 0.7373 0.6863 0.6347 0.6275 0.6221 0.5475 0.4193 0.3132 0.2501
#>   7  0.6524 0.6207 0.4989 0.4201 0.4536 0.5337 0.4967 0.3432 0.2310 0.1885
#>   8  0.5480 0.5194 0.4063 0.3228 0.3188 0.3457 0.3157 0.2271 0.1562 0.1208
#>   9  0.4553 0.4576 0.3942 0.3084 0.2364 0.1856 0.1537 0.1304 0.1023 0.0697
#>   10 0.4553 0.4576 0.3942 0.3084 0.2364 0.1856 0.1537 0.1304 0.1023 0.0697
#>     year
#> age  2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   1  0.1570 0.1553 0.1757 0.1639 0.1129 0.0841 0.0981 0.1433 0.1622 0.1185
#>   2  0.3365 0.2944 0.2380 0.1979 0.1816 0.1816 0.1921 0.2040 0.2042 0.1892
#>   3  0.4146 0.3552 0.2763 0.2354 0.2414 0.2624 0.2640 0.2494 0.2401 0.2440
#>   4  0.2946 0.2640 0.2498 0.2441 0.2423 0.2425 0.2432 0.2430 0.2406 0.2359
#>   5  0.2289 0.1933 0.1791 0.1806 0.1927 0.2043 0.2033 0.1939 0.1878 0.1892
#>   6  0.2065 0.1663 0.1355 0.1242 0.1327 0.1462 0.1461 0.1366 0.1314 0.1357
#>   7  0.1712 0.1487 0.1206 0.0984 0.0857 0.0812 0.0839 0.0899 0.0925 0.0887
#>   8  0.1025 0.0889 0.0760 0.0636 0.0531 0.0478 0.0494 0.0544 0.0555 0.0498
#>   9  0.0466 0.0359 0.0326 0.0321 0.0322 0.0323 0.0321 0.0309 0.0283 0.0249
#>   10 0.0466 0.0359 0.0326 0.0321 0.0322 0.0323 0.0321 0.0309 0.0283 0.0249
#>     year
#> age  2017  
#>   1  0.0686
#>   2  0.1682
#>   3  0.2551
#>   4  0.2300
#>   5  0.1948
#>   6  0.1454
#>   7  0.0819
#>   8  0.0417
#>   9  0.0213
#>   10 0.0213
#> 
#> units:  f 
# Recomputes all F at age by solving catch Baranov
recomputeHarvest(ple4)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   1  0.0987 0.1137 0.1278 0.1371 0.1356 0.1190 0.0937 0.0746 0.0682 0.0746
#>   2  0.1709 0.2122 0.2489 0.2616 0.2549 0.2540 0.2699 0.2794 0.2585 0.2327
#>   3  0.2615 0.3101 0.3497 0.3591 0.3517 0.3689 0.4310 0.4890 0.4713 0.4111
#>   4  0.3088 0.3289 0.3446 0.3503 0.3507 0.3597 0.3832 0.4074 0.4140 0.4019
#>   5  0.2563 0.2923 0.3243 0.3412 0.3457 0.3534 0.3728 0.3929 0.3996 0.3931
#>   6  0.2109 0.2498 0.2867 0.3103 0.3230 0.3420 0.3757 0.4022 0.3951 0.3714
#>   7  0.2254 0.2261 0.2304 0.2420 0.2616 0.2874 0.3137 0.3261 0.3126 0.2920
#>   8  0.2277 0.2227 0.2186 0.2161 0.2186 0.2319 0.2566 0.2748 0.2661 0.2458
#>   9  0.1987 0.2328 0.2510 0.2315 0.1974 0.1891 0.2203 0.2615 0.2633 0.2355
#>   10 0.1987 0.2328 0.2510 0.2315 0.1974 0.1891 0.2203 0.2615 0.2633 0.2355
#>     year
#> age  1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   1  0.0977 0.1416 0.1890 0.2037 0.1949 0.1916 0.2051 0.2370 0.2875 0.3435
#>   2  0.2383 0.2823 0.3281 0.3300 0.3126 0.3172 0.3477 0.3787 0.3912 0.3994
#>   3  0.3768 0.3817 0.3948 0.3924 0.3931 0.4246 0.4793 0.5083 0.4831 0.4509
#>   4  0.3795 0.3570 0.3454 0.3531 0.3836 0.4417 0.5116 0.5451 0.5161 0.4759
#>   5  0.3812 0.3695 0.3610 0.3596 0.3755 0.4220 0.4863 0.5167 0.4861 0.4521
#>   6  0.3645 0.3800 0.3921 0.3810 0.3703 0.3900 0.4329 0.4532 0.4280 0.4063
#>   7  0.2917 0.3167 0.3434 0.3486 0.3450 0.3541 0.3749 0.3867 0.3791 0.3711
#>   8  0.2404 0.2573 0.2832 0.3038 0.3195 0.3353 0.3509 0.3609 0.3615 0.3552
#>   9  0.2166 0.2197 0.2382 0.2666 0.3002 0.3321 0.3565 0.3693 0.3681 0.3529
#>   10 0.2166 0.2197 0.2382 0.2666 0.3002 0.3321 0.3565 0.3693 0.3681 0.3529
#>     year
#> age  1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   1  0.3786 0.3713 0.3172 0.2506 0.2164 0.2298 0.2719 0.3091 0.3199 0.3063
#>   2  0.4216 0.4531 0.4753 0.4775 0.4645 0.4440 0.4281 0.4273 0.4416 0.4632
#>   3  0.4554 0.4986 0.5681 0.6410 0.6747 0.6391 0.5796 0.5470 0.5520 0.5864
#>   4  0.4699 0.4976 0.5422 0.5882 0.6208 0.6283 0.6193 0.6082 0.6050 0.6164
#>   5  0.4641 0.5069 0.5325 0.5214 0.5042 0.5071 0.5297 0.5665 0.6088 0.6430
#>   6  0.4319 0.4861 0.5110 0.4819 0.4386 0.4137 0.4187 0.4648 0.5452 0.6223
#>   7  0.3819 0.4074 0.4316 0.4396 0.4221 0.3826 0.3557 0.3727 0.4347 0.5124
#>   8  0.3463 0.3428 0.3541 0.3767 0.3861 0.3650 0.3379 0.3326 0.3562 0.4056
#>   9  0.3269 0.3031 0.2952 0.3059 0.3256 0.3431 0.3474 0.3300 0.3110 0.3267
#>   10 0.3269 0.3031 0.2952 0.3059 0.3256 0.3431 0.3474 0.3300 0.3110 0.3267
#>     year
#> age  1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   1  0.2783 0.2498 0.2302 0.2216 0.2243 0.2309 0.2224 0.1897 0.1541 0.1320
#>   2  0.4818 0.4831 0.4575 0.4274 0.4228 0.4427 0.4472 0.4112 0.3824 0.4105
#>   3  0.6369 0.6676 0.6431 0.5962 0.5799 0.6028 0.6240 0.6147 0.6249 0.7241
#>   4  0.6414 0.6580 0.6445 0.6149 0.5967 0.5989 0.6111 0.6259 0.6527 0.7051
#>   5  0.6607 0.6706 0.6827 0.6920 0.6859 0.6655 0.6481 0.6470 0.6609 0.6852
#>   6  0.6594 0.6705 0.6809 0.6970 0.7151 0.7275 0.7209 0.6912 0.6691 0.6869
#>   7  0.5672 0.5792 0.5458 0.5143 0.5423 0.6371 0.7131 0.6798 0.6149 0.6134
#>   8  0.4691 0.5016 0.4605 0.4093 0.4314 0.5450 0.6506 0.6156 0.5323 0.5164
#>   9  0.3935 0.4584 0.4400 0.3896 0.4008 0.4929 0.5737 0.5313 0.4504 0.4285
#>   10 0.3935 0.4584 0.4400 0.3896 0.4008 0.4929 0.5737 0.5313 0.4504 0.4285
#>     year
#> age  1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   1  0.1250 0.1319 0.1512 0.1711 0.1757 0.1744 0.1895 0.2230 0.2349 0.1959
#>   2  0.4752 0.4756 0.3816 0.3190 0.3578 0.4685 0.5248 0.4593 0.3817 0.3525
#>   3  0.8740 0.8688 0.6610 0.5175 0.5552 0.7004 0.7364 0.5908 0.4622 0.4267
#>   4  0.7675 0.7909 0.7496 0.6855 0.6363 0.5959 0.5457 0.4813 0.4116 0.3457
#>   5  0.7160 0.7489 0.7717 0.7554 0.6812 0.5827 0.4995 0.4341 0.3656 0.2900
#>   6  0.7292 0.7374 0.6863 0.6347 0.6274 0.6221 0.5475 0.4193 0.3132 0.2501
#>   7  0.6524 0.6207 0.4989 0.4200 0.4536 0.5337 0.4967 0.3432 0.2310 0.1885
#>   8  0.5480 0.5194 0.4063 0.3228 0.3188 0.3457 0.3157 0.2271 0.1562 0.1208
#>   9  0.4553 0.4576 0.3942 0.3084 0.2364 0.1856 0.1537 0.1304 0.1023 0.0697
#>   10 0.4553 0.4576 0.3942 0.3084 0.2364 0.1856 0.1537 0.1304 0.1023 0.0697
#>     year
#> age  2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   1  0.1570 0.1553 0.1757 0.1639 0.1129 0.0841 0.0981 0.1433 0.1622 0.1185
#>   2  0.3365 0.2944 0.2380 0.1979 0.1816 0.1816 0.1921 0.2040 0.2042 0.1892
#>   3  0.4146 0.3552 0.2763 0.2354 0.2414 0.2624 0.2640 0.2494 0.2401 0.2440
#>   4  0.2946 0.2640 0.2498 0.2441 0.2423 0.2425 0.2432 0.2430 0.2406 0.2359
#>   5  0.2289 0.1932 0.1791 0.1806 0.1927 0.2043 0.2033 0.1939 0.1878 0.1892
#>   6  0.2065 0.1664 0.1355 0.1242 0.1327 0.1462 0.1461 0.1365 0.1314 0.1357
#>   7  0.1712 0.1487 0.1206 0.0984 0.0857 0.0812 0.0839 0.0899 0.0925 0.0887
#>   8  0.1025 0.0889 0.0760 0.0636 0.0531 0.0478 0.0494 0.0544 0.0555 0.0498
#>   9  0.0466 0.0359 0.0326 0.0321 0.0322 0.0323 0.0321 0.0309 0.0283 0.0249
#>   10 0.0466 0.0359 0.0326 0.0321 0.0322 0.0323 0.0321 0.0309 0.0283 0.0249
#>     year
#> age  2017  
#>   1  0.0686
#>   2  0.1682
#>   3  0.2551
#>   4  0.2300
#>   5  0.1948
#>   6  0.1454
#>   7  0.0819
#>   8  0.0417
#>   9  0.0213
#>   10 0.0213
#> 
#> units:  f