cpue, a method to generate an observation of a CPUE index of abundance

The observation of stock abundance by CPUE series from commercial fleets is an important step in the generation of management advice that needs to replicated on an Operating Model during any simulation exercise. This method gemnerates an observation of biomass or numbers-at-age from an FLstock being used as OM.

Description: Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque eleifend odio ac rutrum luctus. Aenean placerat porttitor commodo. Pellentesque eget porta libero. Pellentesque molestie mi sed orci feugiat, non mollis enim tristique.

cpue(object, ...)

# S4 method for FLStock
cpue(object, sel.pattern = harvest(object),
  effort = units(harvest(object)), mass = TRUE)

survey(object, ...)

# S4 method for FLStock
survey(object, sel = stock.n(object) %=% 1,
  timing = 0.5, mass = FALSE)

Arguments

object
The object from which to generate the observation.

effort
Units of index to use to mimic effort series in the fishery, "f" or "hr"

mass
Is the index to be in weight at age?

sel
The selectivity of the survey, defaults to be 1 for all ages.

object
The object on which to draw the observation

Value

An FLQuant for the index of abundance, age-disaggregated

An FLQuant for the index of abundance

Details

Details: Aliquam sagittis feugiat felis eget consequat. Praesent eleifend dolor massa, vitae faucibus justo lacinia a. Cras sed erat et magna pharetra bibendum quis in mi. Sed sodales mollis arcu, sit amet venenatis lorem fringilla vel. Vivamus vitae ipsum sem. Donec malesuada purus at libero bibendum accumsan. Donec ipsum sapien, feugiat blandit arcu in, dapibus dictum felis.

See also

FLComp FLComp

Examples

data(ple4)

cpue(ple4)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957      1958      1959      1960      1961      1962      1963     
#>   1     427.69   1060.98   2552.43   1022.82    968.96    554.68   1575.09
#>   2    4631.19   6606.38  14557.09  17002.21  12155.53  14724.63  13036.31
#>   3   13982.60   8428.97  14336.58  20977.35  21946.74  21219.90  26234.70
#>   4   14686.64  19587.44   9981.76  15678.47  21485.68  24029.86  27563.15
#>   5   16320.92  14726.89  18058.97   8452.25  16813.54  21718.25  22571.17
#>   6    3496.59   9464.78  11594.70  15299.71   6291.97  11831.05  19433.58
#>   7    6650.21   3724.16   8005.76   8924.76   9343.31   4635.54   8706.61
#>   8    6910.07   4737.33   3436.52   6039.76   8351.05   5422.64   4972.11
#>   9    2361.99   4350.12   5054.57   3859.77   5427.09   4001.54   5446.85
#>   10  12998.53  13964.57  16525.36  17368.78  17504.81  14137.62  18441.14
#>     year
#> age  1964      1965      1966      1967      1968      1969      1970     
#>   1     151.36    308.44    275.82     97.22   2691.46   2020.09   2939.31
#>   2   17779.29  22457.28  15434.91  12551.03   8082.65  13043.81  19991.74
#>   3   20458.85  22252.57  60137.91  21105.31  20151.27  16621.91  21329.22
#>   4   26806.83  18943.98  22662.77  59783.98  19352.53  16104.01  20751.00
#>   5   26553.69  15664.00  19811.61  20051.11  55338.46  12490.02  16948.12
#>   6   19004.09  18786.13  11636.01  14525.07   8776.10  55770.18  15917.81
#>   7    8606.52  10692.77  11909.85   7108.36   7648.30   5395.61  47461.05
#>   8    3867.02   5174.37   7023.25   5482.49   5318.56   4887.96   2245.96
#>   9    2896.13   3252.54   4949.33   5052.20   2889.81   5900.19   2952.05
#>   10  16754.52  16580.39  15451.80  11122.00  13066.51  15735.79  12544.71
#>     year
#> age  1971      1972      1973      1974      1975      1976      1977     
#>   1    2097.64   2805.42   1536.89   5706.82  11018.60  13555.70  12067.93
#>   2   17626.98  17738.11  14330.08  32362.87  41535.04  30761.30  32083.10
#>   3   25845.82  27645.61  20543.46  21030.34  49609.22  40256.29  29067.60
#>   4   14698.66  24591.81  36337.96  18640.62  14806.02  42079.83  30936.48
#>   5   12954.44  12013.94  25578.47  24799.95  14345.64   6704.20  48080.55
#>   6   12576.20  10207.66   6956.55  10435.54  14580.09   6289.42   4213.16
#>   7   10598.23   8912.42   4697.44   4061.69   6822.28   7456.09   3310.72
#>   8   20092.33   8415.65   6692.29   4301.95   3964.63   5017.39   3787.59
#>   9    4937.15  13577.66   4416.99   4045.43   3365.59   2394.31   2927.45
#>   10  16583.66  15926.17  21642.24  20906.30  21003.96  13489.52   9063.47
#>     year
#> age  1978      1979      1980      1981      1982      1983      1984     
#>   1   10636.77  13021.57   5626.90   2234.29   9673.61   5782.43   8822.73
#>   2   36693.83  48574.89  31106.56  48587.53  32811.85  73954.30  52825.13
#>   3   27252.77  51401.45  55930.20  44309.50  76762.39  37991.49  85950.83
#>   4   23098.09  22713.57  35421.22  27087.88  33653.13  48235.78  25360.19
#>   5   21050.28  20929.54  10957.65  19974.04  14711.31  15616.30  20107.24
#>   6   18758.30  19915.67   9233.02   5893.67   8886.29   6770.31   7695.60
#>   7    3432.59  13900.78   7255.79   5508.17   4051.16   4293.51   3456.36
#>   8    2020.00   2517.59   5112.47   5235.75   3489.94   2561.02   3859.25
#>   9    1812.37   2244.14   1389.62   3573.98   3121.48   2321.29   2165.50
#>   10   6848.03   8957.91   5259.79   6013.20   8033.23   7706.78   6461.41
#>     year
#> age  1985      1986      1987      1988      1989      1990      1991     
#>   1   11304.25  25068.08   5046.64   5267.07   3522.85   2252.26   3683.98
#>   2   45410.38  72767.32 152517.70  59723.07  58632.31  32544.78  43705.24
#>   3   44285.58  55832.99  62537.30 139005.28  65727.73  53485.03  43439.95
#>   4   81917.79  36330.98  40460.21  36807.72  94742.87  41236.90  40945.99
#>   5   13125.58  39530.28  29191.68  22133.20  22343.30  72179.61  28306.61
#>   6    8349.46  11909.19  16140.09  13048.97   9802.10  11465.06  40198.46
#>   7    4407.89   5267.75   6545.43   9809.86   6312.33   3850.27   6894.00
#>   8    2256.51   3321.62   2167.61   3991.22   5049.41   2317.63   2724.44
#>   9    2132.55   2290.30   1694.02   2601.07   1677.05   2517.58   2238.15
#>   10   5821.65   6917.71   5145.64   6397.65   5254.11   3665.15   4909.57
#>     year
#> age  1992      1993      1994      1995      1996      1997      1998     
#>   1    2621.18   2260.79   1088.26   1677.59    822.13    426.53    949.37
#>   2   30988.24  24897.56  19559.81  15145.88  39932.97  65847.87  38318.14
#>   3   37978.72  34103.85  33573.79  27328.18  24345.49  63219.17  64715.54
#>   4   28447.88  25853.87  30679.96  31102.42  18929.09  15015.70  44944.53
#>   5   28774.58  21535.90  15590.28  17243.33  14736.94  11353.44   7750.15
#>   6   15905.45  13149.87   9515.69   7731.27   7088.09   6653.76   3263.01
#>   7   17441.04   7616.14  10668.57   5423.07   6312.12   3247.55   2703.50
#>   8    3786.98   7520.50   4867.71   2463.42   2766.91   1914.72   1346.11
#>   9    2396.23   2851.05   4591.22   1791.37   2226.21   1576.25   1337.73
#>   10   6464.70   5462.20   5755.15   3560.75   4943.71   4092.35   3094.10
#>     year
#> age  1999      2000      2001      2002      2003      2004      2005     
#>   1    1616.93   1582.07    266.96   5928.84    951.71   4974.77   1640.40
#>   2   13936.26  15318.30  48398.91  22944.53  64262.29  26561.92  36624.04
#>   3   54014.64  11699.27  53919.46  27994.32  21341.49  45108.55  13983.64
#>   4   40118.29  71361.14  30624.98  17331.77  18280.55  13777.65  26622.14
#>   5   11121.85   8091.29  49245.08  14805.10  11386.54   6292.65   9491.65
#>   6    3725.50   6079.57   2285.94  16894.87   7425.72   5655.24   5519.25
#>   7    1537.88   1590.26   2484.12   3224.22  13666.15   2234.40   3291.94
#>   8    1511.45    808.61    509.44   1519.11   1264.81   2350.11   1495.97
#>   9    1447.30    953.13    422.13    635.45    524.89    296.06   1255.63
#>   10   3181.81   2512.29   1540.82   1606.25    773.87    423.26    509.48
#>     year
#> age  2006      2007      2008     
#>   1    8362.93    931.01   4132.53
#>   2   38528.58  31594.87  39413.53
#>   3   39869.16  27302.84  20465.23
#>   4   10825.29  13157.04  15689.70
#>   5   13849.10   6095.47   9596.70
#>   6    2725.94   7581.39   2236.15
#>   7    2117.50   1065.96   6350.19
#>   8    1628.89   1324.05    917.64
#>   9     792.23    308.49    181.45
#>   10   1199.50    749.70    240.21
#> 
#> units:  t 

# Am aggregated biomass CPUE
quantSums(cpue(ple4))

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all  82466  86652 104104 114626 120289 122276 147981 142878 134112 169293
#>      year
#> age   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 156879 143316 147970 163081 138011 141834 142732 146292 181051 168004
#>      year
#> age   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 175538 151603 204177 167293 168418 195194 205233 216704 219012 259236
#>      year
#> age   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 321446 298785 273064 225514 217046 174805 145252 135890 113467 122104
#>      year
#> age   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 173347 168422 132212 119996 189698 112884 139878 107675 100434 119899
#>      year
#> age   2007   2008  
#>   all  90111  99223
#> 
#> units:  t 

# NOT RUN {
plot(FLQuants(om=stock(ple4), cpue=quantSums(cpue(ple4)),
  hr=quantSums(cpue(ple4, effort="hr"))))
# }

# NOT RUN {
plot(FLQuants(om=stock(ple4), survey=quantSums(survey(ple4) * stock.wt(ple4)),
 cpue=quantSums(cpue(ple4)), hr=quantSums(cpue(ple4, effort="hr"))))
# }