cpue, a method to generate an observation of a CPUE index of abundance

The observation of stock abundance by CPUE series from commercial fleets is an important step in the generation of management advice that needs to replicated on an Operating Model during any simulation exercise. This method gemnerates an observation of biomass or numbers-at-age from an FLstock being used as OM.

Description: Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque eleifend odio ac rutrum luctus. Aenean placerat porttitor commodo. Pellentesque eget porta libero. Pellentesque molestie mi sed orci feugiat, non mollis enim tristique.

cpue(object, index, ...)

# S4 method for FLStock,missing
cpue(
  object,
  sel.pattern = harvest(object),
  effort = units(harvest(object)),
  mass = TRUE
)

survey(object, index, ...)

# S4 method for FLStock,FLIndex
survey(
  object,
  index,
  sel = sel.pattern(index),
  mass = FALSE,
  timing = mean(range(index, c("startf", "endf"))),
  index.q = index@index.q
)

# S4 method for FLStock,FLIndexBiomass
survey(
  object,
  index,
  sel = sel.pattern(index),
  ages = ac(seq(range(index, c("min")), range(index, c("max")))),
  timing = mean(range(index, c("startf", "endf"))),
  index.q = index@index.q
)

Arguments

object
The object on which to draw the observation

effort
Units of index to use to mimic effort series in the fishery, "f" or "hr"

mass
Is the index to be in weight at age?

sel
The selectivity of the survey, defaults to be 1 for all ages.

Value

An FLQuant for the index of abundance, age-disaggregated

An FLQuant for the index of abundance

Details

Details: Aliquam sagittis feugiat felis eget consequat. Praesent eleifend dolor massa, vitae faucibus justo lacinia a. Cras sed erat et magna pharetra bibendum quis in mi. Sed sodales mollis arcu, sit amet venenatis lorem fringilla vel. Vivamus vitae ipsum sem. Donec malesuada purus at libero bibendum accumsan. Donec ipsum sapien, feugiat blandit arcu in, dapibus dictum felis.

See also

Author

Laurie Kell & Iago Mosqueira, FLR Team.

The FLR Team

Examples

data(ple4)

cpue(ple4)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957       1958       1959       1960       1961       1962      
#>   1     767.288   1394.257   2097.161   1850.641   1930.711    980.453
#>   2    3116.027   5730.247  11584.189  14179.477  10732.472  11060.129
#>   3   16682.878  10970.286  17371.209  26291.968  26745.991  25913.379
#>   4   18545.854  20382.242  10902.494  15952.147  23465.355  26061.061
#>   5   11804.807  12387.105  15614.987   9524.687  13883.661  19362.358
#>   6    4367.922   8849.738  10363.473  12537.061   7411.198  11292.673
#>   7    5989.660   3566.158   6283.474   6247.191   7596.441   4265.738
#>   8    5325.447   4232.886   2628.253   4419.905   4079.079   4680.178
#>   9    1788.641   3916.313   4197.936   2371.484   3220.101   2525.734
#>   10   9662.826  12241.592  13503.917  12098.482   8131.236   8630.976
#>     year
#> age  1963       1964       1965       1966       1967       1968      
#>   1     637.017   1016.055    279.202    310.745    383.786   1261.501
#>   2    9012.993  10435.516  22612.450   7451.985   7440.388   7646.469
#>   3   33179.510  31937.879  32181.043  77798.570  24768.493  25290.766
#>   4   26698.218  25414.657  20532.798  21455.648  66565.164  19826.788
#>   5   23088.199  19951.856  21139.631  16339.018  16446.346  55839.680
#>   6   16865.815  19006.927  15486.123  15219.258  11483.576  12146.194
#>   7    6570.945   8882.439   8187.923   6906.189   7818.465   5947.007
#>   8    2508.339   3540.820   4263.537   4285.584   3991.601   4097.996
#>   9    2883.826   1869.299   2369.132   2616.917   2937.474   2438.489
#>   10  10291.574  13159.239   9945.066   8150.790   7779.671   7997.887
#>     year
#> age  1969       1970       1971       1972       1973       1974      
#>   1    2953.195   3252.183   2218.391   1961.231   5060.312   6026.773
#>   2   13761.037  15963.246  17285.033  13784.332  12152.598  29381.225
#>   3   21971.565  19230.982  29866.239  33433.341  27438.693  24262.975
#>   4   16054.793  14206.105  16158.219  30513.679  33938.070  21931.121
#>   5   16150.361  15418.292  12617.202  13638.925  21877.174  20632.661
#>   6   45421.046  13958.524  13224.473   9389.431   8278.709  11650.813
#>   7    6738.090  27456.042   8753.030   8311.309   5302.395   4216.361
#>   8    3435.182   4055.738  17963.834   5832.046   4731.164   3310.689
#>   9    2661.247   2383.460   2725.443  12953.329   3912.443   3508.290
#>   10   9065.957  11196.472  12014.460  12133.169  17686.561  15762.514
#>     year
#> age  1975       1976       1977       1978       1979       1980      
#>   1    8048.748  12902.538  18713.346  13815.851   9559.216   5884.949
#>   2   30557.942  27787.777  29098.591  39345.234  34144.698  37985.389
#>   3   49979.260  43957.652  28195.317  31012.641  48726.834  56679.239
#>   4   16189.238  48521.634  32306.143  25870.440  19661.827  34812.451
#>   5   11946.787   9443.498  28974.836  26093.686  17451.570  13065.687
#>   6   10268.915   6103.655   5277.747  20200.703  14636.168   9585.474
#>   7    5819.569   5401.185   3499.868   2830.249   9540.637   7338.150
#>   8    2755.334   3741.915   3254.115   1728.934   1358.233   5112.677
#>   9    2550.631   1848.393   2011.544   1608.177    906.479    730.450
#>   10  14517.316  10488.156   6939.100   4902.576   4106.028   3830.073
#>     year
#> age  1981       1982       1983       1984       1985       1986      
#>   1    3591.318   9233.451   8522.059  10364.456  14702.418  27873.289
#>   2   42491.682  31994.935  58725.454  38596.463  40015.636  46601.893
#>   3   64859.799  73023.127  55009.097  91810.325  60997.533  58804.679
#>   4   30997.645  34952.237  41524.877  34897.758  68379.047  41917.443
#>   5   16810.351  13412.671  15437.702  21185.524  22600.262  42522.558
#>   6    6098.026   7991.829   6234.338   7454.787  11528.241  11535.401
#>   7    4713.377   3409.326   4143.405   3327.319   4045.037   6054.552
#>   8    3702.764   2598.465   1987.530   2454.067   1954.586   2207.751
#>   9    2666.299   2164.846   1743.298   1316.323   1483.826   1162.363
#>   10   3569.877   5252.915   5462.943   4566.019   3377.989   3558.766
#>     year
#> age  1987       1988       1989       1990       1991       1992      
#>   1    8600.623   6380.705   4277.083   4007.650   3956.531   3499.170
#>   2   92886.893  37395.279  37975.062  27327.282  27255.569  25697.276
#>   3   68297.381 156268.527  73734.751  63815.358  45497.409  45176.512
#>   4   41013.806  42662.554 102056.356  45025.035  41622.136  31417.920
#>   5   26543.359  24931.448  28484.861  69386.163  32235.773  29378.010
#>   6   23507.236  15793.244  14807.670  16717.291  41884.724  18833.277
#>   7    5520.102  10778.173   6429.287   4891.432   5893.356  17742.432
#>   8    2909.001   2642.923   4179.377   2416.986   2216.789   3565.183
#>   9    1334.746   1714.704   1339.820   2063.711   1518.612   1764.109
#>   10   4223.905   4682.088   3404.899   2584.448   3170.581   4236.640
#>     year
#> age  1993       1994       1995       1996       1997       1998      
#>   1    2387.590   1883.820   2325.922   1163.786   2156.152    787.917
#>   2   24830.132  15806.506  15039.315  20818.626  22020.285  43259.267
#>   3   46798.350  41139.760  31385.173  38992.398  65533.754  55323.567
#>   4   30633.251  30856.538  29030.326  20701.343  22419.239  27978.211
#>   5   21992.989  20795.744  18448.360  17114.919  10403.105  11298.015
#>   6   19559.099  14523.896  11747.658  10655.332   8955.453   6332.807
#>   7    9885.370   9594.018   6136.212   5103.442   4674.150   3786.356
#>   8   10411.346   4145.511   3476.595   2402.560   2169.709   1811.613
#>   9    2415.822   4092.494   1405.330   1402.498   1040.925    970.357
#>   10   4435.468   3517.844   3405.339   2389.341   2569.860   2308.952
#>     year
#> age  1999       2000       2001       2002       2003       2004      
#>   1    1143.504   2384.554   2699.104   4552.675   2003.157   5987.320
#>   2   11332.726   9340.633  15960.892  17770.085  60616.006  17677.978
#>   3  102016.327  27422.168  27191.538  52144.752  39778.300  84972.005
#>   4   23468.116  78151.940  24115.919  22289.796  20257.399  12931.981
#>   5   16984.916  15833.915  39730.259  13306.678   9988.774  11196.256
#>   6    4288.543   6230.377   5511.741  21069.440   7411.654   6019.654
#>   7    1535.836   1254.347   2148.819   2213.014   7054.653   2625.741
#>   8    1006.075    477.072    568.593    810.133    638.532   1470.415
#>   9     736.188    481.548    235.635    140.867    138.354    117.940
#>   10   1559.487   1126.077    627.706    424.525    305.528    317.106
#>     year
#> age  2005       2006       2007       2008       2009       2010      
#>   1    7060.890   5267.526   5948.949   5364.404   8330.877  10739.440
#>   2   31393.091  25845.063  22915.588  37065.053  24435.863  18763.638
#>   3   21466.236  49757.539  39765.082  41472.432  49226.625  33982.958
#>   4   37341.665  11556.474  24501.899  17503.655  21267.667  39449.785
#>   5    6419.204  17181.745   4613.173  10886.683   8456.150  10779.481
#>   6    4122.330   2521.705   6903.409   2613.016   5793.695   3840.932
#>   7    1573.395   1499.944   1243.362   3091.561   1506.134   2680.311
#>   8     515.283    381.754    325.071    295.523    922.545    420.132
#>   9     310.359     92.209     68.165     70.008     56.810    239.575
#>   10    284.400    272.999    146.083    132.711    213.695    215.231
#>     year
#> age  2011       2012       2013       2014       2015       2016      
#>   1    4437.629   2082.370   3082.909   3643.054   2684.406   3687.525
#>   2   18472.581  17322.288  14779.077  19478.278  18488.298   9026.548
#>   3   32396.652  45589.105  48775.302  34842.742  34692.631  37842.260
#>   4   31208.633  30821.735  38042.492  42022.491  35275.623  36216.865
#>   5   21007.230  19260.960  18137.092  20823.322  24027.982  18610.917
#>   6    5109.838  10145.326   8679.589   7322.651   9313.008  10548.098
#>   7    1988.436   1665.396   3071.157   3113.768   3225.734   3719.977
#>   8     624.527    467.376    608.183   1227.349   1035.695    801.162
#>   9      89.290    240.448    196.958    220.775    318.195    211.139
#>   10    480.366    434.581    619.183    666.125    626.123    744.167
#>     year
#> age  2017      
#>   1    2026.661
#>   2   10446.692
#>   3   22559.350
#>   4   33313.078
#>   5   23282.985
#>   6   10870.966
#>   7    3867.666
#>   8     878.460
#>   9     176.397
#>   10    698.663
#> 
#> units:  t 

# Am aggregated biomass CPUE
quantSums(cpue(ple4))

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all  78051  83671  94547 105473 107196 114773 131736 135215 136997 160535
#>      year
#> age   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 149615 142493 138212 127121 132826 141951 140378 140683 152634 170196
#>      year
#> age   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 158271 167408 160092 175025 179501 184034 198791 215973 229085 242239
#>      year
#> age   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 274837 303250 276689 238235 205251 181311 173349 146356 122400 120744
#>      year
#> age   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 141943 153857 164072 142703 118790 134722 148192 143316 110487 114377
#>      year
#> age   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all 106431 118495 120210 121111 115815 128030 135992 133361 129688 121409
#>      year
#> age   2017  
#>   all 108121
#> 
#> units:  t 

if (FALSE) {
plot(FLQuants(om=stock(ple4), cpue=quantSums(cpue(ple4)),
  hr=quantSums(cpue(ple4, effort="hr"))))
}

if (FALSE) {
plot(FLQuants(om=stock(ple4), survey=quantSums(survey(ple4) * stock.wt(ple4)),
 cpue=quantSums(cpue(ple4)), hr=quantSums(cpue(ple4, effort="hr"))))
}
data(ple4)
data(ple4.index)
# survey(ple4, ple4.index)