FLCore datasets

Example datasets for the classes defined in FLCore.

Details

ple4, FLStockA dataset for North Sea (ICES Area IV) plaice. Catch, landings, discards, natural mortality, weight-at-age and maturity, together with the VPA estimated abundances and fishing mortalities.

ple4sex, FLStockA dataset of North Sea (ICES Area IV) plaice disaggregated by sex. Catch, yield, landings, discards, natural mortality, weight-at-age and maturity, together with the VPA estimated abundances and fishing mortalities.

ple4.index, FLIndexA dataset of North Sea (ICES Area IV) plaice survey catch per unit effort, index and index variance.

ple4.indices, FLIndicesA dataset of three North Sea (ICES Area IV) plaice survey catch per unit effort series. Index and index variance.

ple4.biol, FLBiolA dataset of the North Sea plaice population. Numbers, natural mortality, mass and fecundity-at-age.

nsher , FLSRStock and recruit data and fitted relationship for autumn spawning North Sea herring.

Datasets can be loaded by issuing the data command, like in data(ple4).

References

ICES.

See also

Examples

data(ple4)
summary(ple4)

#> An object of class "FLStock"
#> 
#> Name: Plaice in IV 
#> Description: Imported from a VPA file. ( N:\Projecten\ICES WG\Demersale werkgroep WGNSS [...] 
#> Quant: age 
#> Dims:  age 	year	unit	season	area	iter
#> 	10	52	1	1	1	1	
#> 
#> Range:  min	max	pgroup	minyear	maxyear	minfbar	maxfbar 
#> 	1	10	10	1957	2008	2	6	
#> 
#> catch         : [ 1 52 1 1 1 1 ], units =  t 
#> catch.n       : [ 10 52 1 1 1 1 ], units =  10^3 
#> catch.wt      : [ 10 52 1 1 1 1 ], units =  kg 
#> discards      : [ 1 52 1 1 1 1 ], units =  t 
#> discards.n    : [ 10 52 1 1 1 1 ], units =  10^3 
#> discards.wt   : [ 10 52 1 1 1 1 ], units =  kg 
#> landings      : [ 1 52 1 1 1 1 ], units =  t 
#> landings.n    : [ 10 52 1 1 1 1 ], units =  10^3 
#> landings.wt   : [ 10 52 1 1 1 1 ], units =  kg 
#> stock         : [ 1 52 1 1 1 1 ], units =  t 
#> stock.n       : [ 10 52 1 1 1 1 ], units =  10^3 
#> stock.wt      : [ 10 52 1 1 1 1 ], units =  kg 
#> m             : [ 10 52 1 1 1 1 ], units =  m 
#> mat           : [ 10 52 1 1 1 1 ], units =   
#> harvest       : [ 10 52 1 1 1 1 ], units =  f 
#> harvest.spwn  : [ 10 52 1 1 1 1 ], units =   
#> m.spwn        : [ 10 52 1 1 1 1 ], units =   

data(nsher)
is(nsher)

#> [1] "FLSR"    "FLModel" "FLComp"