Methods for binding objects of array classes along a given dimension

These methods can bind two or more objects of array-based classes (e.g. FLQuant), along the specified dimension.

dbind(x, y, ...)

# S4 method for FLArray,FLArray
dbind(x, y, ..., dim = 1)

qbind(...)

ybind(...)

ubind(...)

sbind(...)

abind(...)

ibind(...)

Arguments

x
First object to bind

y
Second object to bind

...
Other objects to bind

dim
Dimension to bind on, *numeric* or *character*.

Value

An object of the same class as the inputs

Details

The objects to bind must contain the same dimmames in all dimensions other than that used to bind, while dimnames in the selected one must differ. See the examples below for correct and incorrect uses.

Object are bound in the order they are provided, with no attempt to sort according to the dimnames of the chosen dimension.

See also

Examples

# By iter
x <- FLQuant(rnorm(80000), dim=c(4,20,1,1,1,1000))
y <- FLQuant(rnorm(80000), dim=c(4,20,1,1,1,1000))
  dimnames(y) <- list(iter=1001:2000)
ibind(x,y)

#> An object of class "FLQuant"
#> iters:  2000 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1                  2                  3                 
#>     1  0.04334098(1.023) -0.00185550(1.075) -0.04289988(0.990)
#>     2  0.00942455(0.963) -0.00512111(0.989) -0.05581679(0.971)
#>     3 -0.02744599(0.933)  0.04332583(1.010)  0.04142419(1.018)
#>     4 -0.03731320(1.028) -0.00082079(1.077) -0.02931749(1.014)
#>      year
#> quant 4                  5                  6                 
#>     1  0.00988258(0.995)  0.01563936(1.015)  0.00145826(1.029)
#>     2  0.02353123(1.019)  0.00580316(1.001) -0.04345542(1.027)
#>     3  0.00738681(1.010) -0.03555249(1.012)  0.00209192(0.959)
#>     4 -0.04581638(1.027) -0.02899838(0.964) -0.00434247(1.010)
#>      year
#> quant 7                  8                  9                 
#>     1  0.02509285(0.987)  0.01607275(1.013)  0.05322713(1.001)
#>     2 -0.01947600(1.040) -0.03083122(0.975)  0.01392328(1.030)
#>     3  0.05497119(1.014)  0.02938053(1.017)  0.01714675(0.987)
#>     4 -0.00058335(1.019)  0.00120857(0.963)  0.01168442(0.941)
#>      year
#> quant 10                 11                 12                
#>     1 -0.01745866(0.970)  0.00543992(0.992)  0.02547039(1.012)
#>     2 -0.02796371(0.960)  0.00147563(1.023) -0.00896929(1.000)
#>     3  0.02022801(0.996)  0.00922204(0.998)  0.01794089(0.985)
#>     4  0.01717331(0.975)  0.00202409(1.006) -0.00572637(0.965)
#>      year
#> quant 13                 14                 15                
#>     1 -0.07021344(0.993) -0.01188379(0.953)  0.03483894(1.023)
#>     2 -0.02267260(0.959) -0.01309799(0.998) -0.02254996(1.018)
#>     3  0.01929442(0.997)  0.04976219(0.964)  0.02580009(0.958)
#>     4  0.00286601(0.983)  0.02868997(1.017)  0.02986963(0.983)
#>      year
#> quant 16                 17                 18                
#>     1  0.01216912(0.961)  0.01394473(0.975)  0.02478644(1.005)
#>     2 -0.00481498(1.059)  0.00230548(0.983) -0.04162523(0.979)
#>     3 -0.05921841(0.948)  0.00635049(0.985)  0.00115406(1.007)
#>     4  0.01155197(0.987) -0.00744323(0.989)  0.01374626(1.007)
#>      year
#> quant 19                 20                
#>     1  0.02023334(1.030) -0.02250861(1.006)
#>     2 -0.03817214(0.968) -0.01110561(0.992)
#>     3 -0.00869417(0.991)  0.01156214(0.980)
#>     4 -0.02666777(1.073) -0.00883026(0.989)
#> 
#> units:  NA 

# By quant (age)
x <- FLQuant(1, dimnames=list(age=1:3, year=1:10))
y <- FLQuant(2, dimnames=list(age=4:12, year=1:10))
qbind(x, y)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1 2 3 4 5 6 7 8 9 10
#>   1  1 1 1 1 1 1 1 1 1 1 
#>   2  1 1 1 1 1 1 1 1 1 1 
#>   3  1 1 1 1 1 1 1 1 1 1 
#>   4  2 2 2 2 2 2 2 2 2 2 
#>   5  2 2 2 2 2 2 2 2 2 2 
#>   6  2 2 2 2 2 2 2 2 2 2 
#>   7  2 2 2 2 2 2 2 2 2 2 
#>   8  2 2 2 2 2 2 2 2 2 2 
#>   9  2 2 2 2 2 2 2 2 2 2 
#>   10 2 2 2 2 2 2 2 2 2 2 
#>   11 2 2 2 2 2 2 2 2 2 2 
#>   12 2 2 2 2 2 2 2 2 2 2 
#> 
#> units:  NA 

# By year
x <- FLQuant(1, dimnames=list(age=1:3, year=1:10))
y <- FLQuant(2, dimnames=list(age=1:3, year=11:20))
z <- FLQuant(3, dimnames=list(age=1:3, year=21:30))
ybind(x, y, z)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27 28
#>   1 1 1 1 1 1 1 1 1 1 1  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3  3  3  3 
#>   2 1 1 1 1 1 1 1 1 1 1  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3  3  3  3 
#>   3 1 1 1 1 1 1 1 1 1 1  2  2  2  2  2  2  2  2  2  2  3  3  3  3  3  3  3  3 
#>    year
#> age 29 30
#>   1 3  3 
#>   2 3  3 
#>   3 3  3 
#> 
#> units:  NA 

# By season
x <- FLQuant(1, dimnames=list(year=1:10, season=1:2))
y <- FLQuant(2, dimnames=list(year=1:10, season=3:4))
sbind(x, y)

#> An object of class "FLQuant"
#> , , unit = unique, season = 1, area = unique
#> 
#>      year
#> quant 1 2 3 4 5 6 7 8 9 10
#>   all 1 1 1 1 1 1 1 1 1 1 
#> 
#> , , unit = unique, season = 2, area = unique
#> 
#>      year
#> quant 1 2 3 4 5 6 7 8 9 10
#>   all 1 1 1 1 1 1 1 1 1 1 
#> 
#> , , unit = unique, season = 3, area = unique
#> 
#>      year
#> quant 1 2 3 4 5 6 7 8 9 10
#>   all 2 2 2 2 2 2 2 2 2 2 
#> 
#> , , unit = unique, season = 4, area = unique
#> 
#>      year
#> quant 1 2 3 4 5 6 7 8 9 10
#>   all 2 2 2 2 2 2 2 2 2 2 
#> 
#> units:  NA