Summaries by dimension

Methods to compute various summary calculations (sum, mean, variance) over selected dimensions of objects from any array-based classes (e.g. FLQuant). These methods return an object of the same dimensions as the input but with length one in the dimension chosen to operate along.

quantSums(x, ...)

yearSums(x, ...)

unitSums(x, ...)

seasonSums(x, ...)

areaSums(x, ...)

iterSums(x, ...)

dimSums(x, ...)

quantMeans(x, ...)

yearMeans(x, ...)

unitMeans(x, ...)

seasonMeans(x, ...)

areaMeans(x, ...)

iterMeans(x, ...)

dimMeans(x, ...)

quantVars(x, ...)

yearVars(x, ...)

unitVars(x, ...)

seasonVars(x, ...)

areaVars(x, ...)

iterVars(x, ...)

dimVars(x, ...)

iterMedians(x, ...)

iterCVs(x, ...)

iterProb(x, ...)

# S4 method for FLQuant
quantSums(x, na.rm = TRUE)

# S4 method for FLQuant
yearSums(x, na.rm = TRUE)

# S4 method for FLQuant
unitSums(x, na.rm = TRUE)

# S4 method for FLQuant
seasonSums(x, na.rm = TRUE)

# S4 method for FLQuant
areaSums(x, na.rm = TRUE)

# S4 method for FLQuant
iterSums(x, na.rm = TRUE)

# S4 method for FLQuant
quantMeans(x, na.rm = TRUE)

# S4 method for FLQuant
yearMeans(x, na.rm = TRUE)

# S4 method for FLQuant
unitMeans(x, na.rm = TRUE)

# S4 method for FLQuant
seasonMeans(x, na.rm = TRUE)

# S4 method for FLQuant
areaMeans(x, na.rm = TRUE)

# S4 method for FLQuant
iterMeans(x, na.rm = TRUE)

# S4 method for FLQuant
iterMedians(x, na.rm = TRUE)

# S4 method for FLQuant
quantVars(x, na.rm = TRUE)

# S4 method for FLQuant
yearVars(x, na.rm = TRUE)

# S4 method for FLQuant
unitVars(x, na.rm = TRUE)

# S4 method for FLQuant
seasonVars(x, na.rm = TRUE)

# S4 method for FLQuant
areaVars(x, na.rm = TRUE)

# S4 method for FLQuant
iterVars(x, na.rm = TRUE)

# S4 method for FLQuant
iterCVs(x, na.rm = TRUE)

# S4 method for FLQuant
iterProb(x, na.rm = TRUE)

# S4 method for FLQuantDistr
yearSums(x, na.rm = TRUE)

# S4 method for FLQuantDistr
unitSums(x, na.rm = TRUE)

# S4 method for FLQuantDistr
seasonSums(x, na.rm = TRUE)

# S4 method for FLQuantDistr
areaSums(x, na.rm = TRUE)

# S4 method for FLQuantDistr
yearMeans(x, na.rm = TRUE)

# S4 method for FLQuantDistr
unitMeans(x, na.rm = TRUE)

# S4 method for FLQuantDistr
seasonMeans(x, na.rm = TRUE)

# S4 method for FLQuantDistr
areaMeans(x, na.rm = TRUE)

# S4 method for FLQuantDistr
iterMeans(x, na.rm = TRUE)

# S4 method for FLQuantDistr
iterMedians(x, na.rm = TRUE)

# S4 method for FLQuantDistr
quantVars(x, na.rm = TRUE)

# S4 method for FLQuantDistr
yearVars(x, na.rm = TRUE)

# S4 method for FLQuantDistr
unitVars(x, na.rm = TRUE)

# S4 method for FLQuantDistr
seasonVars(x, na.rm = TRUE)

# S4 method for FLQuantDistr
areaVars(x, na.rm = TRUE)

# S4 method for FLQuantDistr
iterVars(x, na.rm = TRUE)

# S4 method for FLPar
iterMeans(x, na.rm = TRUE)

# S4 method for FLPar
iterMedians(x, na.rm = TRUE)

# S4 method for FLPar
iterVars(x, na.rm = TRUE)

# S4 method for FLPar
iterSums(x, na.rm = TRUE)

Arguments

x
An object.

na.rm
Should NAs be removed before calculation? Defaults to TRUE.

Details

This set of methods computes three different summaries (sum, mean and variance) of an FLQuant object along each of the six dimensions (quant, year, unit, season, area, or iter). Medians and CVs can also be computed along the sixth dimension, iter.

These methods encapsulate a call to apply with the corresponding dimensions and function: mean, median, var, and sum, while iterCVs are computed as sqrt(iterVars) / iterMeans.

In contrast with R standard behaviour, the sum of a dimension where all elements are NA will be NA and not 0. See example below.

Methods working along the iter dimension are also defined for objects of class FLPar.

Methods to operate over the first dimension refer to it as the quant dimension, regardless of the actual name used in the object.

Generic methods

quantSums(x), quantMeans(x), quantVars(x) yearSums(x), yearMeans(x), yearVars(x) unitSums(x), unitMeans(x), unitVars(x) seasonSums(x), seasonMeans(x), seasonVars(x) areaSums(x), areaMeans(x), areaVars(x) iterMeans(x), iterVars(x), iterMedians(x), iterSums(x) dimSums(x), dimMeans(x), dimVars(x)

See also

Examples

flq <- FLQuant(rnorm(4000), dim=c(5,10,2,2,2,10), quant='age')

quantSums(flq)

#> An object of class "FLQuant"
#> iters:  10 
#> 
#> , , unit = 1, season = 1, area = 1
#> 
#>      year
#> age   1                 2                 3                 4                
#>   all -0.9338745(2.555) -2.0089193(0.994)  0.1517140(2.346)  0.2562777(2.254)
#>      year
#> age   5                 6                 7                 8                
#>   all -0.5492879(3.204)  0.3480758(1.639)  0.2601805(2.973)  0.6677555(1.611)
#>      year
#> age   9                 10               
#>   all  1.0598731(2.849)  0.3877024(1.032)
#> 
#> , , unit = 2, season = 1, area = 1
#> 
#>      year
#> age   1                 2                 3                 4                
#>   all -0.4554380(1.667) -0.2983541(1.012)  0.7850723(1.947)  0.0642899(2.527)
#>      year
#> age   5                 6                 7                 8                
#>   all  1.4885348(2.176) -0.0101764(1.588)  1.0287146(3.378)  0.4912939(1.512)
#>      year
#> age   9                 10               
#>   all  0.6235399(2.773) -0.1605473(2.939)
#> 
#> , , unit = 1, season = 2, area = 1
#> 
#>      year
#> age   1                 2                 3                 4                
#>   all -0.7223510(2.933) -0.3637102(1.974) -0.8225070(1.818) -0.5046355(1.394)
#>      year
#> age   5                 6                 7                 8                
#>   all -0.6076298(2.154) -1.0937115(2.627)  0.2617980(1.856) -2.0789074(1.686)
#>      year
#> age   9                 10               
#>   all  0.7503456(2.021) -1.2373666(3.234)
#> 
#> , , unit = 2, season = 2, area = 1
#> 
#>      year
#> age   1                 2                 3                 4                
#>   all -0.1825185(2.102)  0.8473450(2.293) -1.0793745(2.322) -0.1647870(2.730)
#>      year
#> age   5                 6                 7                 8                
#>   all  1.3629925(2.195) -0.9285642(3.419) -0.2977798(2.612) -0.7733984(2.305)
#>      year
#> age   9                 10               
#>   all -0.1350689(0.617) -0.5763666(2.071)
#> 
#> , , unit = 1, season = 1, area = 2
#> 
#>      year
#> age   1                 2                 3                 4                
#>   all -0.1612749(1.643) -0.4745957(3.080)  0.6337623(1.127)  0.0972081(3.026)
#>      year
#> age   5                 6                 7                 8                
#>   all  0.6843306(2.397)  0.9344221(1.288)  0.5940123(1.107)  0.4885515(2.891)
#>      year
#> age   9                 10               
#>   all  0.6823425(1.157) -0.5918501(3.089)
#> 
#> , , unit = 2, season = 1, area = 2
#> 
#>      year
#> age   1                 2                 3                 4                
#>   all -0.4408111(1.236) -0.7158285(1.497) -1.0732748(1.526)  0.6974163(1.959)
#>      year
#> age   5                 6                 7                 8                
#>   all  0.9892682(1.812)  0.3535457(2.044) -0.2058570(2.014) -0.1265324(1.862)
#>      year
#> age   9                 10               
#>   all -0.7578979(1.439) -0.4226282(1.908)
#> 
#> , , unit = 1, season = 2, area = 2
#> 
#>      year
#> age   1                 2                 3                 4                
#>   all -0.0078141(1.557) -0.0851529(2.275) -0.6112672(1.113)  0.1164032(1.249)
#>      year
#> age   5                 6                 7                 8                
#>   all -0.6161097(2.045) -2.0955532(2.184)  0.1409789(2.252) -0.1538267(1.802)
#>      year
#> age   9                 10               
#>   all  1.0531059(0.895)  0.0898632(1.699)
#> 
#> , , unit = 2, season = 2, area = 2
#> 
#>      year
#> age   1                 2                 3                 4                
#>   all  0.8934814(2.468)  1.2542881(3.336)  0.1102171(1.692)  0.7939988(2.859)
#>      year
#> age   5                 6                 7                 8                
#>   all  0.4101261(1.567) -1.5767820(1.465)  0.5074394(1.656) -0.0286330(2.989)
#>      year
#> age   9                 10               
#>   all -0.2776967(1.003)  0.3695653(1.707)
#> 
#> units:  NA 

quantMeans(flq)

#> An object of class "FLQuant"
#> iters:  10 
#> 
#> , , unit = 1, season = 1, area = 1
#> 
#>      year
#> age   1                 2                 3                 4                
#>   all -0.1867749(0.511) -0.4017839(0.199)  0.0303428(0.469)  0.0512555(0.451)
#>      year
#> age   5                 6                 7                 8                
#>   all -0.1098576(0.641)  0.0696152(0.328)  0.0520361(0.595)  0.1335511(0.322)
#>      year
#> age   9                 10               
#>   all  0.2119746(0.570)  0.0775405(0.206)
#> 
#> , , unit = 2, season = 1, area = 1
#> 
#>      year
#> age   1                 2                 3                 4                
#>   all -0.0910876(0.333) -0.0596708(0.202)  0.1570145(0.389)  0.0128580(0.505)
#>      year
#> age   5                 6                 7                 8                
#>   all  0.2977070(0.435) -0.0020353(0.318)  0.2057429(0.676)  0.0982588(0.302)
#>      year
#> age   9                 10               
#>   all  0.1247080(0.555) -0.0321095(0.588)
#> 
#> , , unit = 1, season = 2, area = 1
#> 
#>      year
#> age   1                 2                 3                 4                
#>   all -0.1444702(0.587) -0.0727420(0.395) -0.1645014(0.364) -0.1009271(0.279)
#>      year
#> age   5                 6                 7                 8                
#>   all -0.1215260(0.431) -0.2187423(0.525)  0.0523596(0.371) -0.4157815(0.337)
#>      year
#> age   9                 10               
#>   all  0.1500691(0.404) -0.2474733(0.647)
#> 
#> , , unit = 2, season = 2, area = 1
#> 
#>      year
#> age   1                 2                 3                 4                
#>   all -0.0365037(0.420)  0.1694690(0.459) -0.2158749(0.464) -0.0329574(0.546)
#>      year
#> age   5                 6                 7                 8                
#>   all  0.2725985(0.439) -0.1857128(0.684) -0.0595560(0.522) -0.1546797(0.461)
#>      year
#> age   9                 10               
#>   all -0.0270138(0.123) -0.1152733(0.414)
#> 
#> , , unit = 1, season = 1, area = 2
#> 
#>      year
#> age   1                 2                 3                 4                
#>   all -0.0322550(0.329) -0.0949191(0.616)  0.1267525(0.225)  0.0194416(0.605)
#>      year
#> age   5                 6                 7                 8                
#>   all  0.1368661(0.479)  0.1868844(0.258)  0.1188025(0.221)  0.0977103(0.578)
#>      year
#> age   9                 10               
#>   all  0.1364685(0.231) -0.1183700(0.618)
#> 
#> , , unit = 2, season = 1, area = 2
#> 
#>      year
#> age   1                 2                 3                 4                
#>   all -0.0881622(0.247) -0.1431657(0.299) -0.2146550(0.305)  0.1394833(0.392)
#>      year
#> age   5                 6                 7                 8                
#>   all  0.1978536(0.362)  0.0707091(0.409) -0.0411714(0.403) -0.0253065(0.372)
#>      year
#> age   9                 10               
#>   all -0.1515796(0.288) -0.0845256(0.382)
#> 
#> , , unit = 1, season = 2, area = 2
#> 
#>      year
#> age   1                 2                 3                 4                
#>   all -0.0015628(0.311) -0.0170306(0.455) -0.1222534(0.223)  0.0232806(0.250)
#>      year
#> age   5                 6                 7                 8                
#>   all -0.1232219(0.409) -0.4191106(0.437)  0.0281958(0.450) -0.0307653(0.360)
#>      year
#> age   9                 10               
#>   all  0.2106212(0.179)  0.0179726(0.340)
#> 
#> , , unit = 2, season = 2, area = 2
#> 
#>      year
#> age   1                 2                 3                 4                
#>   all  0.1786963(0.494)  0.2508576(0.667)  0.0220434(0.338)  0.1587998(0.572)
#>      year
#> age   5                 6                 7                 8                
#>   all  0.0820252(0.313) -0.3153564(0.293)  0.1014879(0.331) -0.0057266(0.598)
#>      year
#> age   9                 10               
#>   all -0.0555393(0.201)  0.0739131(0.341)
#> 
#> units:  NA 

yearSums(flq)

#> An object of class "FLQuant"
#> iters:  10 
#> 
#> , , unit = 1, season = 1, area = 1
#> 
#>    year
#> age 1              
#>   1 -0.339091(2.28)
#>   2 -1.139518(2.25)
#>   3 -1.070050(2.99)
#>   4  0.821232(2.73)
#>   5 -0.524379(1.28)
#> 
#> , , unit = 2, season = 1, area = 1
#> 
#>    year
#> age 1              
#>   1  0.615854(3.83)
#>   2 -2.881023(3.95)
#>   3 -1.082845(3.10)
#>   4  0.911333(2.23)
#>   5  2.472307(2.75)
#> 
#> , , unit = 1, season = 2, area = 1
#> 
#>    year
#> age 1              
#>   1  0.901250(4.61)
#>   2  0.336043(2.26)
#>   3 -0.522240(3.36)
#>   4 -0.718176(2.45)
#>   5  0.105705(2.00)
#> 
#> , , unit = 2, season = 2, area = 1
#> 
#>    year
#> age 1              
#>   1 -1.095533(2.46)
#>   2  1.841297(4.15)
#>   3 -2.056828(4.88)
#>   4 -0.344726(2.43)
#>   5 -0.251685(2.79)
#> 
#> , , unit = 1, season = 1, area = 2
#> 
#>    year
#> age 1              
#>   1  0.899842(7.35)
#>   2  1.365758(3.32)
#>   3  1.007982(3.23)
#>   4 -1.466535(1.63)
#>   5  0.750336(4.12)
#> 
#> , , unit = 2, season = 1, area = 2
#> 
#>    year
#> age 1              
#>   1 -1.085653(1.25)
#>   2 -0.973950(1.74)
#>   3  2.062667(2.35)
#>   4 -0.517846(3.19)
#>   5  0.228035(3.89)
#> 
#> , , unit = 1, season = 2, area = 2
#> 
#>    year
#> age 1              
#>   1  0.908573(4.94)
#>   2 -0.045532(2.91)
#>   3  0.980736(3.57)
#>   4  0.280214(2.56)
#>   5 -2.166655(3.11)
#> 
#> , , unit = 2, season = 2, area = 2
#> 
#>    year
#> age 1              
#>   1  0.426732(2.46)
#>   2  2.247406(4.49)
#>   3 -1.685861(1.48)
#>   4  1.429054(2.18)
#>   5  0.414708(3.01)
#> 
#> units:  NA 

iterMeans(flq)

#> An object of class "FLQuant"
#> , , unit = 1, season = 1, area = 1
#> 
#>    year
#> age 1           2           3           4           5           6          
#>   1 -0.28853601 -0.35549891  0.45449594  0.42946883 -0.04050555  0.04987318
#>   2 -0.24220935 -0.45266022  0.42218588  0.14376152  0.20919021 -0.12398548
#>   3 -0.10496084 -0.82260733  0.04654552  0.17422846 -0.17392911  0.37467858
#>   4 -0.41350861  0.02035212  0.09024165 -0.19527559 -0.35670892  0.16541417
#>   5  0.43216265 -0.54129634  0.08155210 -0.08180354 -0.41533044 -0.22106310
#>    year
#> age 7           8           9           10         
#>   1  0.06411701  0.04442215 -0.06479908  0.05639802
#>   2 -0.41606765 -0.28018914  0.32268683 -0.21195715
#>   3 -0.07041711 -0.30034414 -0.29974876 -0.47267924
#>   4  0.25011545  0.42202877  0.45646764  0.03776451
#>   5  0.06915584  0.01489868 -0.02752050 -0.26589153
#> 
#> , , unit = 2, season = 1, area = 1
#> 
#>    year
#> age 1           2           3           4           5           6          
#>   1 -0.24956390 -0.18771830 -0.14169572  0.48371532  0.04572409  0.01706417
#>   2 -0.27666041 -0.65690136 -0.04572433 -0.56509427  0.03096841 -0.47634997
#>   3 -0.23527495 -0.21913376 -0.37854795  0.05188571  0.28892929 -0.12911768
#>   4  0.15259525  0.73937522 -0.48772323  0.01002245  0.47656621  0.30501210
#>   5  0.13151419  0.32318865  0.65333977 -0.06649872  0.58632843  0.01507484
#>    year
#> age 7           8           9           10         
#>   1  0.33658809  0.23602224 -0.09886717  0.27386598
#>   2  0.09659371 -0.28087192  0.31642743 -0.16684769
#>   3 -0.31340925 -0.19939724  0.34476036 -0.24860605
#>   4  0.13478352 -0.09405940  0.33462635  0.42050115
#>   5  0.17224039  0.43769369  0.24544761  0.06697560
#> 
#> , , unit = 1, season = 2, area = 1
#> 
#>    year
#> age 1           2           3           4           5           6          
#>   1 -0.15432770  0.07084112 -0.04610736 -0.35771662 -0.00334260 -0.06188135
#>   2 -0.15367893 -0.00303252  0.27270749 -0.19531893  0.25176466 -0.22840075
#>   3 -0.39694381  0.06916525 -0.21542834  0.75678995  0.17733911 -0.45137398
#>   4 -0.32486612 -0.29637213  0.03072374  0.37961178 -0.04030458 -0.15647811
#>   5 -0.09938732  0.12695203 -0.30369451 -0.23474186 -0.60575236  0.46230137
#>    year
#> age 7           8           9           10         
#>   1  0.77543169 -0.51092092 -0.01648968 -0.34092895
#>   2 -0.23315051 -0.08626177  0.80560205 -0.16748782
#>   3 -0.65933356 -0.16118744  0.06409812  0.25962070
#>   4 -0.12466939  0.52094031 -0.19661269 -0.31161064
#>   5  0.43278244 -0.75857721  0.00353249  0.15755642
#> 
#> , , unit = 2, season = 2, area = 1
#> 
#>    year
#> age 1           2           3           4           5           6          
#>   1 -0.36201532  0.24075531 -0.46409269  0.33842457 -0.46401624 -0.26267357
#>   2  0.09143376  0.61731677 -0.02648109  0.24199457  0.33717332 -0.05064146
#>   3 -0.71518943 -0.11672435 -0.32090763 -0.00977126  0.34459541 -0.16852553
#>   4  0.54104133 -0.16589815 -0.17217158 -0.56540839  0.54957988  0.05861887
#>   5  0.14769386 -0.30215387 -0.06331165 -0.20836935  0.38919662  0.15384756
#>    year
#> age 7           8           9           10         
#>   1 -0.40414749 -0.13643336 -0.22228298  0.08044327
#>   2  0.06337519 -0.04900141 -0.24393695  0.08954724
#>   3  0.01731732 -0.57671412 -0.37353632 -0.25107758
#>   4 -0.32685376 -0.29856336 -0.24487461  0.13900278
#>   5 -0.25399885 -0.02537515  0.60691849 -0.32984623
#> 
#> , , unit = 1, season = 1, area = 2
#> 
#>    year
#> age 1           2           3           4           5           6          
#>   1  0.22729295  0.00358577 -0.21127729  0.02790768 -0.29450382  0.29042774
#>   2 -0.25544053  0.09730876  0.38993749  0.34173516  0.48958921  0.43582586
#>   3  0.33314808 -0.06603526  0.27983189 -0.11613430  0.35298445  0.42553399
#>   4 -0.30259079 -1.06352307  0.52892350 -0.17857871  0.26969809 -0.02856976
#>   5 -0.77961000 -0.07073775 -0.30728055 -0.02603521  0.69636469 -0.10462514
#>    year
#> age 7           8           9           10         
#>   1  0.49958667  0.56207777 -0.18498834  0.09911257
#>   2 -0.20643459  0.42432682 -0.38971555 -0.44437661
#>   3 -0.39561860 -0.10461610 -0.07456501  0.14959089
#>   4 -0.45727965 -0.48452167  0.26402825 -0.12211496
#>   5 -0.18370895  0.46728029  0.53982565  0.18786859
#> 
#> , , unit = 2, season = 1, area = 2
#> 
#>    year
#> age 1           2           3           4           5           6          
#>   1  0.10809377  0.11887714 -0.60409903  0.12791230  0.27221984  0.02488546
#>   2  0.32363005  0.20606472 -0.13010164 -0.02621871 -0.36737720  0.13081933
#>   3  0.37628884  0.04713206 -0.14517888  0.56774637  0.17860818  0.12144423
#>   4 -0.01277666 -0.54065053 -0.12384307 -0.36808715  0.40503572 -0.26202485
#>   5 -0.45422909 -0.17797085 -0.03266467  0.35144097  0.43905320 -0.06948939
#>    year
#> age 7           8           9           10         
#>   1  0.12315988 -0.24738500 -0.07745162 -0.28232073
#>   2  0.17326515  0.08343160 -0.25927575 -0.08885720
#>   3  0.10794040 -0.05860399 -0.09123865  0.00064513
#>   4 -0.22243515  0.33898173 -0.69264267  0.07307662
#>   5 -0.00583221 -0.10647541 -0.09343733  0.18024513
#> 
#> , , unit = 1, season = 2, area = 2
#> 
#>    year
#> age 1           2           3           4           5           6          
#>   1 -0.11746731  0.27138202  0.18212103  0.02735545 -0.01843378  0.35968471
#>   2 -0.07562304 -0.10567845 -0.25987669 -0.20698807  0.28293515 -0.31445701
#>   3  0.23033668 -0.09348001  0.21371186  0.37432514 -0.10957492 -0.41246766
#>   4 -0.24103345 -0.13752825 -0.17594642  0.39986867  0.40918984 -0.71466485
#>   5  0.05098363 -0.37638910 -0.81972798 -0.40741102 -0.23939448 -0.40585002
#>    year
#> age 7           8           9           10         
#>   1 -0.44377611 -0.30069712  0.08953967  0.22945168
#>   2 -0.37573381  0.15693170  0.68653191 -0.05648069
#>   3 -0.24203474 -0.02607114  0.10394481  0.33112903
#>   4  0.01347861  0.13249075  0.29968058  0.33577536
#>   5 -0.13276793 -0.42495947 -0.02331732  0.16390213
#> 
#> , , unit = 2, season = 2, area = 2
#> 
#>    year
#> age 1           2           3           4           5           6          
#>   1  0.05002209  0.14517724 -0.08281008  0.07384202 -0.13702226 -0.34299728
#>   2  0.19963402  0.42204357 -0.29853996  0.14526814  0.26234784  0.05130764
#>   3  0.24173844 -0.30399252  0.07685937 -0.30272314  0.40032455 -0.68428265
#>   4 -0.13446496 -0.29820940 -0.10655204  0.39269187 -0.12657238  0.26006831
#>   5  0.00940656  0.79267713 -0.12321137  0.53735103  0.42595707 -0.26367884
#>    year
#> age 7           8           9           10         
#>   1  0.44338854  0.23199347 -0.12133240  0.01240352
#>   2  0.70585762 -0.01358996  0.28709757  0.31844634
#>   3 -0.60318586 -0.41442047  0.26950394 -0.05628311
#>   4  0.06376203 -0.33762131 -0.00737910  0.59828076
#>   5  0.16117556  0.03298059 -0.24566414 -0.20446834
#> 
#> units:  NA 

dim(quantSums(flq))

#> [1]  1 10  2  2  2 10

# NA dims stay as NA when summed along
x <- FLQuant(c(NA, NA, NA, rnorm(6)), dim=c(3, 3))
quantSums(x)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1        2        3       
#>   all       NA  0.99218 -0.86944
#> 
#> units:  NA 

# although in fact a sum of no elements (as na.rm=TRUE) is zero
apply(x, 2:6, sum, na.rm=TRUE)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1        2        3       
#>   all  0.00000  0.99218 -0.86944
#> 
#> units:  NA