Summaries by dimension

Methods to compute various summary calculations (sum, mean, variance) over selected dimensions of objects from any array-based classes (e.g. FLQuant, FLPar). These methods return an object of the same dimensions as the input but with of length one in the dimension chosen to operate along.

quantSums(x, ...)

yearSums(x, ...)

unitSums(x, ...)

seasonSums(x, ...)

areaSums(x, ...)

dimSums(x, ...)

quantMeans(x, ...)

yearMeans(x, ...)

unitMeans(x, ...)

seasonMeans(x, ...)

areaMeans(x, ...)

iterMeans(x, ...)

dimMeans(x, ...)

quantVars(x, ...)

yearVars(x, ...)

unitVars(x, ...)

seasonVars(x, ...)

areaVars(x, ...)

iterVars(x, ...)

dimVars(x, ...)

iterMedians(x, ...)

iterCVs(x, ...)

# S4 method for FLQuant
quantSums(x, na.rm = TRUE)

# S4 method for FLQuant
yearSums(x, na.rm = TRUE)

# S4 method for FLQuant
unitSums(x, na.rm = TRUE)

# S4 method for FLQuant
seasonSums(x, na.rm = TRUE)

# S4 method for FLQuant
areaSums(x, na.rm = TRUE)

Arguments

x

An object.

na.rm

Should NAs be removed before calculation? Defaults to TRUE.

Details

This set of methods computes three different summaries (sum, mean and variance) of an FLQuant object along each of the six dimensions (quant, year, unit, season, area, or iter). Medians can also be computed along the sixth dimension, iter.

These methods simply encapsulate a call to apply with the corresponding dimensions and function.

Methods along the iter dimension are also defined for objects of class FLPar.

Methods to operate over the first dimension refer to it as the quant dimension, regardless of the actual name used in the object.

Generic methods

quantSums(x), quantMeans(x), quantVars(x) yearSums(x), yearMeans(x), yearVars(x) unitSums(x), unitMeans(x), unitVars(x) seasonSums(x), seasonMeans(x), seasonVars(x) areaSums(x), areaMeans(x), areaVars(x) iterMeans(x), iterVars(x), iterMedians(x) dimSums(x), dimMeans(x), dimVars(x)

See also

FLQuant, sum, mean, var

Examples

 flq <- FLQuant(rnorm(4000), dim=c(5,10,2,2,2,10), quant='age')
 quantSums(flq)

#> An object of class "FLQuant"
#> iters:  10 
#> 
#> , , unit = 1, season = 1, area = 1
#> 
#>      year
#> age   1                2                3                4               
#>   all -0.987791(2.581)  0.213560(1.607) -0.494161(1.858) -0.318460(2.075)
#>      year
#> age   5                6                7                8               
#>   all -1.110861(2.814) -0.221582(1.129)  0.335390(2.023)  0.305295(2.145)
#>      year
#> age   9                10              
#>   all -0.311312(2.319)  0.521833(2.602)
#> 
#> , , unit = 2, season = 1, area = 1
#> 
#>      year
#> age   1                2                3                4               
#>   all  0.026852(1.893) -1.387536(2.185)  0.131893(1.574) -0.713206(1.697)
#>      year
#> age   5                6                7                8               
#>   all -0.188410(2.106) -0.774890(2.255) -0.100795(1.720)  0.445642(2.432)
#>      year
#> age   9                10              
#>   all  0.682647(1.465)  1.407702(2.388)
#> 
#> , , unit = 1, season = 2, area = 1
#> 
#>      year
#> age   1                2                3                4               
#>   all  0.377908(1.667)  0.846330(2.499)  0.646966(2.356)  0.283251(1.233)
#>      year
#> age   5                6                7                8               
#>   all -0.827412(1.640) -1.147421(2.296)  0.262089(1.335)  0.038831(3.235)
#>      year
#> age   9                10              
#>   all -0.269329(1.519) -0.348615(3.592)
#> 
#> , , unit = 2, season = 2, area = 1
#> 
#>      year
#> age   1                2                3                4               
#>   all  0.235213(1.344) -0.682821(1.522)  0.629375(0.707)  1.373450(1.720)
#>      year
#> age   5                6                7                8               
#>   all -0.690835(1.107)  0.087171(2.056)  0.233418(2.391)  0.550923(1.631)
#>      year
#> age   9                10              
#>   all -0.452555(2.318)  2.616438(2.969)
#> 
#> , , unit = 1, season = 1, area = 2
#> 
#>      year
#> age   1                2                3                4               
#>   all -0.719477(2.315) -0.113795(2.418) -0.586245(1.985)  0.986381(2.323)
#>      year
#> age   5                6                7                8               
#>   all  0.127572(1.240) -0.864002(1.294)  0.254014(2.265)  1.374489(3.045)
#>      year
#> age   9                10              
#>   all  0.459301(1.278)  1.318003(1.281)
#> 
#> , , unit = 2, season = 1, area = 2
#> 
#>      year
#> age   1                2                3                4               
#>   all -0.754852(1.973) -0.403929(1.216)  1.672242(1.371)  1.064110(1.799)
#>      year
#> age   5                6                7                8               
#>   all  1.628475(1.959) -0.449793(0.862)  0.474539(2.232)  0.164688(2.839)
#>      year
#> age   9                10              
#>   all -0.772440(1.114) -0.326151(3.175)
#> 
#> , , unit = 1, season = 2, area = 2
#> 
#>      year
#> age   1                2                3                4               
#>   all  0.119298(3.330)  0.965033(1.339) -0.736482(3.400)  0.355126(2.082)
#>      year
#> age   5                6                7                8               
#>   all  0.132607(2.959) -0.886310(2.453)  1.673936(1.395) -0.814037(1.909)
#>      year
#> age   9                10              
#>   all  0.220331(1.877)  0.064024(2.911)
#> 
#> , , unit = 2, season = 2, area = 2
#> 
#>      year
#> age   1                2                3                4               
#>   all  0.470041(1.964)  0.364391(3.236)  0.553386(4.116) -0.508998(2.312)
#>      year
#> age   5                6                7                8               
#>   all  0.787850(2.707) -0.390064(1.574) -1.105399(1.304) -1.348497(2.284)
#>      year
#> age   9                10              
#>   all -0.287552(2.439) -0.697438(2.257)
#> 
#> units:  NA 

 quantMeans(flq)

#> An object of class "FLQuant"
#> iters:  10 
#> 
#> , , unit = 1, season = 1, area = 1
#> 
#>      year
#> age   1                 2                 3                 4                
#>   all -0.1975581(0.516)  0.0427121(0.321) -0.0988323(0.372) -0.0636920(0.415)
#>      year
#> age   5                 6                 7                 8                
#>   all -0.2221722(0.563) -0.0443164(0.226)  0.0670781(0.405)  0.0610591(0.429)
#>      year
#> age   9                 10               
#>   all -0.0622624(0.464)  0.1043665(0.520)
#> 
#> , , unit = 2, season = 1, area = 1
#> 
#>      year
#> age   1                 2                 3                 4                
#>   all  0.0053703(0.379) -0.2775071(0.437)  0.0263786(0.315) -0.1426413(0.339)
#>      year
#> age   5                 6                 7                 8                
#>   all -0.0376821(0.421) -0.1549780(0.451) -0.0201590(0.344)  0.0891285(0.486)
#>      year
#> age   9                 10               
#>   all  0.1365295(0.293)  0.2815405(0.478)
#> 
#> , , unit = 1, season = 2, area = 1
#> 
#>      year
#> age   1                 2                 3                 4                
#>   all  0.0755815(0.333)  0.1692661(0.500)  0.1293933(0.471)  0.0566502(0.247)
#>      year
#> age   5                 6                 7                 8                
#>   all -0.1654824(0.328) -0.2294841(0.459)  0.0524179(0.267)  0.0077662(0.647)
#>      year
#> age   9                 10               
#>   all -0.0538658(0.304) -0.0697231(0.718)
#> 
#> , , unit = 2, season = 2, area = 1
#> 
#>      year
#> age   1                 2                 3                 4                
#>   all  0.0470426(0.269) -0.1365641(0.304)  0.1258751(0.141)  0.2746900(0.344)
#>      year
#> age   5                 6                 7                 8                
#>   all -0.1381670(0.221)  0.0174343(0.411)  0.0466836(0.478)  0.1101847(0.326)
#>      year
#> age   9                 10               
#>   all -0.0905110(0.464)  0.5232875(0.594)
#> 
#> , , unit = 1, season = 1, area = 2
#> 
#>      year
#> age   1                 2                 3                 4                
#>   all -0.1438954(0.463) -0.0227590(0.484) -0.1172489(0.397)  0.1972761(0.465)
#>      year
#> age   5                 6                 7                 8                
#>   all  0.0255145(0.248) -0.1728004(0.259)  0.0508029(0.453)  0.2748978(0.609)
#>      year
#> age   9                 10               
#>   all  0.0918601(0.256)  0.2636006(0.256)
#> 
#> , , unit = 2, season = 1, area = 2
#> 
#>      year
#> age   1                 2                 3                 4                
#>   all -0.1509704(0.395) -0.0807857(0.243)  0.3344485(0.274)  0.2128220(0.360)
#>      year
#> age   5                 6                 7                 8                
#>   all  0.3256950(0.392) -0.0899587(0.172)  0.0949077(0.446)  0.0329376(0.568)
#>      year
#> age   9                 10               
#>   all -0.1544880(0.223) -0.0652303(0.635)
#> 
#> , , unit = 1, season = 2, area = 2
#> 
#>      year
#> age   1                 2                 3                 4                
#>   all  0.0238597(0.666)  0.1930066(0.268) -0.1472964(0.680)  0.0710252(0.416)
#>      year
#> age   5                 6                 7                 8                
#>   all  0.0265214(0.592) -0.1772620(0.491)  0.3347872(0.279) -0.1628074(0.382)
#>      year
#> age   9                 10               
#>   all  0.0440661(0.375)  0.0128048(0.582)
#> 
#> , , unit = 2, season = 2, area = 2
#> 
#>      year
#> age   1                 2                 3                 4                
#>   all  0.0940081(0.393)  0.0728783(0.647)  0.1106771(0.823) -0.1017996(0.462)
#>      year
#> age   5                 6                 7                 8                
#>   all  0.1575700(0.541) -0.0780127(0.315) -0.2210797(0.261) -0.2696994(0.457)
#>      year
#> age   9                 10               
#>   all -0.0575104(0.488) -0.1394877(0.451)
#> 
#> units:  NA 

 yearSums(flq)

#> An object of class "FLQuant"
#> iters:  10 
#> 
#> , , unit = 1, season = 1, area = 1
#> 
#>    year
#> age 1              
#>   1 -1.706889(3.21)
#>   2  1.438605(1.22)
#>   3 -0.061889(2.42)
#>   4 -1.310727(3.00)
#>   5  1.101575(2.73)
#> 
#> , , unit = 2, season = 1, area = 1
#> 
#>    year
#> age 1              
#>   1 -0.022412(2.06)
#>   2 -1.267580(2.31)
#>   3  2.419793(2.40)
#>   4 -1.708131(2.21)
#>   5  0.209010(2.75)
#> 
#> , , unit = 1, season = 2, area = 1
#> 
#>    year
#> age 1              
#>   1 -0.417070(5.17)
#>   2  0.035239(1.06)
#>   3 -2.480688(1.62)
#>   4  0.907843(3.84)
#>   5 -0.675603(2.32)
#> 
#> , , unit = 2, season = 2, area = 1
#> 
#>    year
#> age 1              
#>   1  1.285892(2.51)
#>   2 -0.101510(3.16)
#>   3  1.021705(2.36)
#>   4  0.773374(2.19)
#>   5  0.055806(4.51)
#> 
#> , , unit = 1, season = 1, area = 2
#> 
#>    year
#> age 1              
#>   1 -1.173142(4.06)
#>   2  0.908960(3.18)
#>   3  2.167466(1.06)
#>   4  1.371880(3.02)
#>   5  0.003542(1.77)
#> 
#> , , unit = 2, season = 1, area = 2
#> 
#>    year
#> age 1              
#>   1 -0.715841(4.06)
#>   2  1.219893(1.89)
#>   3  0.804951(1.26)
#>   4  2.818628(6.04)
#>   5 -0.249429(2.97)
#> 
#> , , unit = 1, season = 2, area = 2
#> 
#>    year
#> age 1              
#>   1  0.145992(3.91)
#>   2 -2.051115(1.16)
#>   3 -0.422242(2.61)
#>   4  0.748606(2.55)
#>   5  1.005810(2.45)
#> 
#> , , unit = 2, season = 2, area = 2
#> 
#>    year
#> age 1              
#>   1  2.122277(3.00)
#>   2  0.219369(2.63)
#>   3  0.265386(2.48)
#>   4 -0.364834(3.51)
#>   5 -3.058010(2.99)
#> 
#> units:  NA 

 iterMeans(flq)

#> An object of class "FLQuant"
#> , , unit = 1, season = 1, area = 1
#> 
#>    year
#> age 1          2          3          4          5          6         
#>   1 -0.7196726 -0.1082600 -0.4013377 -0.7334973 -0.1678169 -0.0825538
#>   2 -0.1475784 -0.0769732  0.1055469 -0.4232474 -0.0314802  0.4406191
#>   3  0.3539685  0.3815547 -0.1378106  0.1788679  0.0261705  0.0712497
#>   4  0.0218472 -0.4264748 -0.0543675  0.0677891 -0.7066212  0.0960889
#>   5 -0.1510207  0.4013900  0.1414768  0.1126582 -0.2433991  0.0127677
#>    year
#> age 7          8          9          10        
#>   1  0.1885587 -0.1319666 -0.2276174  0.5476378
#>   2  0.2889383  0.2191715  0.3539848 -0.0277021
#>   3 -0.0412520 -0.1241227 -0.4591723 -0.2994402
#>   4 -0.1844280 -0.0794458 -0.2943030 -0.1246356
#>   5  0.0670415 -0.4083114  0.0100415  0.6948872
#> 
#> , , unit = 2, season = 1, area = 1
#> 
#>    year
#> age 1          2          3          4          5          6         
#>   1 -0.1328157 -0.1107164 -0.2327308 -0.6188605  0.6103669  0.1656805
#>   2  0.3083916  0.0284560 -0.1859519 -0.4399683 -0.5658308 -0.0122506
#>   3  0.3549124 -0.6184990  0.3090890  0.1087758 -0.3790950  0.2319421
#>   4 -0.2471863 -0.2403036 -0.0450496  0.2699772 -0.1846196 -0.1687602
#>   5 -0.3488072 -0.1618325  0.2829549  0.2447534  0.2910893 -0.3404010
#>    year
#> age 7          8          9          10        
#>   1  0.0785810  0.5305091  0.0767188  0.2322303
#>   2  0.4285320 -0.6054995  0.1147918  0.1184972
#>   3  0.2058746  0.1275841  0.4873012  0.5396239
#>   4 -0.4983775  0.3392127 -0.3709583  0.0772334
#>   5  0.1455445  0.0293778  0.3453931  0.0545104
#> 
#> , , unit = 1, season = 2, area = 1
#> 
#>    year
#> age 1          2          3          4          5          6         
#>   1 -0.3343590 -0.1809625  0.1622492  0.1413848 -0.0568059  0.1583259
#>   2 -0.1376295  0.4831070 -0.0304966  0.0155921 -0.2605551 -0.7056306
#>   3 -0.3033406  0.0639531 -0.0402401 -0.0269127 -0.1346963 -0.5693408
#>   4 -0.0239563  0.1285552 -0.2939850  0.1316185 -0.3478130  0.4653445
#>   5  0.6041640  0.4439105  0.0907007 -0.3179309 -0.3017421  0.3292828
#>    year
#> age 7          8          9          10        
#>   1 -0.0467191 -0.0921430  0.1635968 -0.2791126
#>   2  0.4950541  0.1259033  0.3434629 -0.3498108
#>   3 -0.1967773 -0.0375300 -0.4458452  0.0962406
#>   4  0.2882815  0.0817102  0.3441104 -0.2726580
#>   5  0.2213585 -0.4263616 -0.5203130 -0.1632962
#> 
#> , , unit = 2, season = 2, area = 1
#> 
#>    year
#> age 1          2          3          4          5          6         
#>   1  0.0785024  0.2270875  0.2244156  0.4134275  0.1529028 -0.2764313
#>   2 -0.0650031 -0.3628341 -0.0025524 -0.0774684 -0.1517260  0.3874053
#>   3  0.4052359 -0.0284637 -0.2749972  0.8626670  0.3498106  0.0843709
#>   4  0.2368207 -0.3747860  0.2481043  0.0922402 -0.1957714  0.0416140
#>   5 -0.0801572  0.0389207  0.1086998  0.1081099 -0.5154065 -0.0092443
#>    year
#> age 7          8          9          10        
#>   1  0.0831483  0.1922327 -0.3674401  0.5058783
#>   2 -0.4269900 -0.0125589  0.1133127  0.4011862
#>   3  0.3740668 -0.0428364  0.0705651  0.1626790
#>   4 -0.0559232 -0.2050554  0.2037005  0.9448058
#>   5  0.3579259  0.1934844 -0.0984808  0.1678478
#> 
#> , , unit = 1, season = 1, area = 2
#> 
#>    year
#> age 1          2          3          4          5          6         
#>   1 -0.0466322 -0.0394166 -0.1504173 -0.3759371 -0.3071735 -0.6160293
#>   2 -0.4023490  0.0929646  0.0672840 -0.1375949 -0.3028545  0.2163244
#>   3 -0.2421591  0.4490014 -0.5324224  0.1209214  0.1383545  0.1432410
#>   4  0.4294007  0.0268940 -0.1689938  0.2517951  0.2099669 -0.4584042
#>   5 -0.1967834 -0.1778682  0.3933022  0.5121549  0.0668063 -0.4608471
#>    year
#> age 7          8          9          10        
#>   1  0.0221300  0.4405551  0.7258139  0.1612050
#>   2  0.1481673  0.5690504  0.4516835  0.2572429
#>   3  0.5794641  0.6208048  0.1469833  0.5959911
#>   4  0.1225489 -0.2833320 -0.1502173  0.6424207
#>   5 -0.2050830  0.1763951  0.1143933  0.2081430
#> 
#> , , unit = 2, season = 1, area = 2
#> 
#>    year
#> age 1          2          3          4          5          6         
#>   1 -0.7785255  0.1649178 -0.0150986 -0.0569229  0.1224161  0.1639523
#>   2 -0.0296293  0.3097230 -0.0145242  0.1646614  0.2378665 -0.2697755
#>   3 -0.2111828 -0.5294446  0.7368124 -0.5588907 -0.0654425  0.0864006
#>   4 -0.1085656 -0.0878502  0.2809594  0.9374029  0.5162094 -0.5302125
#>   5  0.2803562 -0.2384073  0.3262744  0.3526674  0.0028686  0.2439619
#>    year
#> age 7          8          9          10        
#>   1 -0.1173254 -0.0976728  0.1811423 -0.2374993
#>   2  0.4221775 -0.2021488  0.2324923 -0.3919869
#>   3  0.2869577 -0.1268988  0.2126882  0.1773391
#>   4  0.4293489  0.0967922 -0.4575552  0.5028406
#>   5 -0.0931776  0.2496021 -0.4397010 -0.5408100
#> 
#> , , unit = 1, season = 2, area = 2
#> 
#>    year
#> age 1          2          3          4          5          6         
#>   1 -0.1645153 -0.1740653  0.1488634  0.1349799  0.0489651 -0.3123700
#>   2 -0.4607832 -0.2513488 -0.2475917 -0.2391113 -0.2001921 -0.1534936
#>   3  0.2014794  0.2322557  0.1928159 -0.0695767  0.3195847 -0.1084254
#>   4 -0.0299908  0.0888599  0.1703859  0.3583441  0.2657940 -0.4958943
#>   5 -0.3042343  0.2914777 -0.2751938 -0.1340555  0.0433412  0.1027628
#>    year
#> age 7          8          9          10        
#>   1  0.2003846  0.1611073  0.1883190  0.1376135
#>   2  0.6134129 -1.0080015 -0.1659071  0.4235011
#>   3 -0.1538952  0.1577127 -0.5246124 -0.1732292
#>   4 -0.1281008  0.1429765 -0.0742776  0.0637401
#>   5  0.9183433 -0.2828716  0.3965150 -0.0912937
#> 
#> , , unit = 2, season = 2, area = 2
#> 
#>    year
#> age 1          2          3          4          5          6         
#>   1  0.3788341  0.5247406 -0.1898197  0.0802087  0.0307193  0.1059786
#>   2 -0.2593276 -0.3557802  0.2745077  0.2267195  0.4146621 -0.1148075
#>   3 -0.1139285  0.0238212  0.1307720  0.4907945 -0.1887352  0.5271031
#>   4  0.3955672 -0.3631186  0.0693708 -0.6081555 -0.0784352 -0.1304509
#>   5  0.3205946 -0.0124865 -0.3834444 -0.2839389 -0.3785923  0.0517353
#>    year
#> age 7          8          9          10        
#>   1  0.1926397 -0.0455259 -0.1643695  0.0370705
#>   2 -0.4345475 -0.3760642  0.2887213  0.3156270
#>   3 -0.3614746 -0.4025924 -0.3458558 -0.2655661
#>   4  0.0275625 -0.2987451 -0.2582429  0.1831258
#>   5 -0.8609419  0.0067336  0.1308282 -0.7460923
#> 
#> units:  NA 

 dim(quantSums(flq))

#> [1]  1 10  2  2  2 10