Compute the ratio of discards to total catch in numbers or weight

A calculation is made of the proportion of discards over total catch at age, either as numbers (value = 'numbers') or weight (value = 'weight'), or for the total discards and catch in biomass (value = 'total').

Usage

discardsRatio(object, value = c("numbers", "weight", "total"))

Arguments

object

An object of class 'FLStock'

value

One of 'numbers' (default), 'weight' or 'total'.

Value

The discards ratio (between 0 and 1), 'FLQuant'

See also

Author

The FLR Team

Examples

data(ple4)
# Discards ratio at age in numbers
discardsRatio(ple4)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957     1958     1959     1960     1961     1962     1963     1964    
#>   1  1.00e+00 1.00e+00 1.00e+00 1.00e+00 1.00e+00 1.00e+00 1.00e+00 1.00e+00
#>   2  9.05e-01 9.07e-01 8.83e-01 9.66e-01 9.84e-01 9.89e-01 9.69e-01 9.12e-01
#>   3  1.22e-01 5.00e-01 5.99e-01 4.19e-01 7.24e-01 7.16e-01 8.56e-01 6.11e-01
#>   4  1.80e-02 3.75e-02 3.05e-01 1.91e-02 7.75e-03 2.52e-01 1.13e-01 3.42e-01
#>   5  2.65e-02 5.56e-02 1.01e-01 4.74e-02 4.63e-02 1.63e-02 3.67e-02 4.95e-02
#>   6  2.58e-03 3.13e-03 4.52e-03 1.85e-02 9.97e-03 9.27e-03 5.78e-03 8.58e-03
#>   7  4.08e-05 5.75e-05 3.53e-05 3.50e-05 3.35e-05 5.71e-05 3.82e-05 3.14e-05
#>   8  4.99e-05 5.88e-05 7.64e-05 5.51e-05 5.05e-05 5.72e-05 7.18e-05 7.52e-05
#>   9  9.05e-05 7.69e-05 7.42e-05 9.76e-05 8.21e-05 9.84e-05 6.96e-05 1.25e-04
#>   10 4.15e-05 3.58e-05 3.02e-05 3.28e-05 3.36e-05 4.04e-05 2.96e-05 2.94e-05
#>     year
#> age  1965     1966     1967     1968     1969     1970     1971     1972    
#>   1  1.00e+00 1.00e+00 1.00e+00 1.00e+00 1.00e+00 9.99e-01 1.00e+00 9.68e-01
#>   2  9.77e-01 9.75e-01 9.56e-01 8.81e-01 8.09e-01 9.00e-01 7.60e-01 5.96e-01
#>   3  5.83e-01 8.41e-01 7.39e-01 5.25e-01 2.59e-01 3.58e-01 4.59e-01 3.46e-01
#>   4  5.98e-02 2.15e-01 5.70e-01 4.03e-01 3.36e-01 2.30e-02 7.18e-02 9.51e-02
#>   5  2.06e-02 3.62e-02 1.46e-02 1.50e-01 2.01e-02 1.13e-01 1.56e-02 2.34e-02
#>   6  8.75e-03 4.48e-03 1.14e-02 3.31e-03 1.84e-02 4.70e-03 3.43e-03 4.19e-03
#>   7  2.71e-05 2.98e-05 3.03e-05 4.70e-05 4.54e-05 6.51e-06 2.86e-05 3.38e-05
#>   8  5.81e-05 4.55e-05 6.12e-05 4.81e-05 7.76e-05 8.13e-05 1.63e-05 4.46e-05
#>   9  1.18e-04 7.59e-05 7.29e-05 1.24e-04 5.77e-05 1.23e-04 7.55e-05 2.62e-05
#>   10 2.83e-05 3.29e-05 4.39e-05 3.63e-05 2.94e-05 3.46e-05 2.86e-05 3.11e-05
#>     year
#> age  1973     1974     1975     1976     1977     1978     1979     1980    
#>   1  9.90e-01 9.89e-01 9.96e-01 9.85e-01 9.88e-01 9.95e-01 9.96e-01 9.96e-01
#>   2  5.99e-01 9.28e-01 9.04e-01 8.03e-01 6.45e-01 7.25e-01 7.88e-01 6.50e-01
#>   3  6.06e-02 5.96e-02 7.44e-01 4.72e-01 6.47e-01 3.12e-01 3.27e-01 6.80e-03
#>   4  8.70e-03 6.49e-03 1.26e-01 1.54e-01 3.37e-01 1.71e-01 2.73e-01 8.62e-03
#>   5  2.82e-02 1.42e-02 9.30e-03 1.22e-02 1.00e-01 2.51e-02 1.22e-02 7.52e-03
#>   6  4.63e-03 5.34e-03 5.46e-03 4.07e-03 1.39e-02 1.64e-02 7.73e-03 4.70e-03
#>   7  5.36e-05 6.02e-05 3.96e-05 5.35e-05 8.44e-05 8.73e-05 2.07e-05 3.57e-05
#>   8  4.83e-05 7.63e-05 7.94e-05 7.63e-05 9.94e-05 1.55e-04 1.20e-04 5.80e-05
#>   9  7.82e-05 8.66e-05 1.15e-04 1.65e-04 1.23e-04 2.14e-04 1.78e-04 2.63e-04
#>   10 1.98e-05 2.19e-05 2.39e-05 3.88e-05 5.43e-05 6.94e-05 5.39e-05 9.09e-05
#>     year
#> age  1981     1982     1983     1984     1985     1986     1987     1988    
#>   1  9.98e-01 9.92e-01 9.95e-01 1.00e+00 1.00e+00 9.99e-01 1.00e+00 1.00e+00
#>   2  6.52e-01 8.09e-01 7.85e-01 8.32e-01 7.59e-01 8.80e-01 9.42e-01 9.76e-01
#>   3  1.45e-02 2.13e-02 2.10e-01 1.58e-01 1.95e-01 2.26e-01 6.04e-01 6.44e-01
#>   4  6.23e-03 1.54e-02 2.18e-02 2.81e-01 1.09e-02 1.99e-01 1.17e-02 4.17e-01
#>   5  1.16e-02 7.36e-03 2.65e-02 3.47e-02 5.26e-03 1.41e-02 1.63e-02 1.45e-02
#>   6  4.33e-03 5.77e-03 5.56e-03 4.49e-03 4.57e-03 5.21e-03 5.54e-03 5.60e-03
#>   7  5.10e-05 6.49e-05 6.07e-05 7.31e-05 6.72e-05 5.08e-05 3.69e-05 2.58e-05
#>   8  6.03e-05 9.01e-05 1.18e-04 8.94e-05 1.32e-04 1.06e-04 1.07e-04 8.07e-05
#>   9  9.99e-05 1.10e-04 1.66e-04 1.83e-04 1.78e-04 1.76e-04 2.13e-04 1.46e-04
#>   10 8.21e-05 5.70e-05 6.02e-05 7.63e-05 8.58e-05 7.06e-05 8.91e-05 7.81e-05
#>     year
#> age  1989     1990     1991     1992     1993     1994     1995     1996    
#>   1  9.94e-01 9.89e-01 9.92e-01 9.76e-01 9.66e-01 9.78e-01 9.39e-01 9.90e-01
#>   2  9.12e-01 8.92e-01 8.75e-01 8.34e-01 7.42e-01 6.82e-01 6.92e-01 8.84e-01
#>   3  6.41e-01 6.17e-01 6.19e-01 5.23e-01 3.25e-01 2.52e-01 1.60e-01 2.88e-01
#>   4  2.65e-01 1.90e-01 2.50e-01 2.97e-01 1.29e-01 1.18e-02 1.00e-02 6.91e-02
#>   5  1.09e-01 2.87e-02 6.77e-02 4.64e-02 2.52e-02 1.70e-02 1.37e-02 1.20e-02
#>   6  5.30e-03 5.89e-03 1.18e-02 2.29e-02 6.14e-03 5.19e-03 4.88e-03 6.04e-03
#>   7  4.31e-05 5.52e-05 3.24e-05 1.64e-05 3.39e-05 2.89e-05 3.39e-05 4.03e-05
#>   8  6.23e-05 1.07e-04 9.27e-05 6.92e-05 4.35e-05 7.51e-05 9.46e-05 8.50e-05
#>   9  2.15e-04 1.31e-04 1.55e-04 1.52e-04 1.35e-04 9.17e-05 2.00e-04 1.79e-04
#>   10 8.61e-05 1.19e-04 8.87e-05 7.17e-05 8.50e-05 9.16e-05 1.28e-04 1.00e-04
#>     year
#> age  1997     1998     1999     2000     2001     2002     2003     2004    
#>   1  9.93e-01 9.98e-01 9.96e-01 9.75e-01 8.70e-01 9.96e-01 9.90e-01 9.98e-01
#>   2  9.22e-01 9.55e-01 9.60e-01 9.16e-01 9.05e-01 9.18e-01 9.21e-01 9.38e-01
#>   3  6.86e-01 7.20e-01 5.94e-01 5.56e-01 7.61e-01 5.65e-01 5.28e-01 5.18e-01
#>   4  1.18e-02 4.61e-01 5.37e-01 2.74e-01 5.67e-01 2.01e-01 2.69e-01 2.14e-01
#>   5  8.48e-03 1.29e-02 8.69e-03 4.53e-02 3.48e-01 6.02e-02 1.12e-01 7.67e-02
#>   6  5.78e-03 3.65e-03 4.14e-03 1.88e-02 1.31e-02 1.64e-02 2.02e-02 3.27e-02
#>   7  5.25e-05 8.77e-05 1.24e-04 8.13e-02 9.51e-03 1.65e-02 2.41e-01 7.99e-03
#>   8  9.92e-05 1.74e-04 1.90e-04 1.29e-01 1.24e-03 6.64e-04 8.45e-04 3.14e-04
#>   9  1.64e-04 2.52e-04 3.13e-04 5.24e-04 7.68e-04 6.29e-04 7.25e-04 1.08e-03
#>   10 1.14e-04 1.49e-04 1.55e-04 2.36e-04 2.11e-04 3.05e-04 5.68e-04 8.77e-04
#>     year
#> age  2005     2006     2007     2008     2009     2010     2011     2012    
#>   1  9.70e-01 9.98e-01 9.82e-01 9.97e-01 9.90e-01 9.88e-01 9.97e-01 9.98e-01
#>   2  9.46e-01 9.18e-01 9.08e-01 9.51e-01 9.40e-01 9.01e-01 9.22e-01 9.81e-01
#>   3  6.10e-01 5.95e-01 6.35e-01 4.20e-01 5.27e-01 5.50e-01 5.80e-01 6.71e-01
#>   4  2.07e-01 2.22e-01 1.99e-01 1.06e-01 1.82e-01 2.46e-01 3.74e-01 3.47e-01
#>   5  1.82e-01 4.44e-02 7.26e-02 4.98e-02 7.46e-02 9.68e-02 2.01e-01 1.73e-01
#>   6  3.55e-01 3.03e-02 2.43e-01 4.26e-02 5.64e-02 1.21e-01 1.65e-01 4.61e-02
#>   7  4.64e-02 1.16e-01 5.80e-02 5.23e-01 3.91e-02 1.75e-01 2.11e-02 2.91e-02
#>   8  2.35e-03 7.83e-02 4.50e-01 4.75e-01 4.31e-01 1.32e-01 2.46e-01 2.20e-03
#>   9  2.17e-04 3.79e-04 8.27e-04 6.80e-04 8.51e-04 1.98e-04 3.96e-04 1.44e-04
#>   10 7.72e-04 3.55e-04 3.83e-04 5.34e-04 6.08e-04 3.94e-04 1.82e-04 2.60e-04
#>     year
#> age  2013     2014     2015     2016     2017    
#>   1  9.96e-01 1.00e+00 1.00e+00 1.00e+00 1.00e+00
#>   2  9.37e-01 9.73e-01 9.87e-01 9.72e-01 9.65e-01
#>   3  6.65e-01 7.41e-01 7.54e-01 7.73e-01 7.58e-01
#>   4  2.43e-01 4.15e-01 3.67e-01 3.63e-01 5.04e-01
#>   5  9.96e-02 2.68e-01 1.71e-01 2.57e-01 2.17e-01
#>   6  3.39e-02 1.06e-01 6.83e-02 1.58e-01 1.10e-01
#>   7  1.41e-02 3.56e-02 3.92e-02 9.75e-02 7.00e-02
#>   8  3.90e-03 7.98e-03 2.30e-02 1.94e-01 3.41e-02
#>   9  2.03e-04 2.02e-04 1.23e-04 9.41e-05 1.63e-04
#>   10 1.32e-04 1.45e-04 1.03e-04 9.75e-05 1.56e-04
#> 
#> units:   
# Total proportion of discards by year
discardsRatio(ple4, value="total")
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 0.0949 0.1648 0.2568 0.2476 0.2867 0.2849 0.2674 0.2468 0.3037 0.3940
#>      year
#> age   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 0.3271 0.1922 0.1609 0.1818 0.1694 0.1268 0.1171 0.2710 0.4289 0.3054
#>      year
#> age   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 0.3456 0.2795 0.3060 0.1844 0.1799 0.2430 0.3242 0.2873 0.2619 0.4032
#>      year
#> age   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 0.4975 0.4667 0.3574 0.3040 0.3224 0.3005 0.2104 0.1656 0.1730 0.2891
#>      year
#> age   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 0.4555 0.5666 0.4296 0.3129 0.4802 0.3941 0.5146 0.4107 0.4631 0.4443
#>      year
#> age   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all 0.4301 0.4305 0.4050 0.3752 0.3782 0.3759 0.3115 0.3816 0.3372 0.3417
#>      year
#> age   2017  
#>   all 0.3216
#> 
#> units: