Compute the inter-annual variability of a time series

The inter-annual variability of a time series stored in an FLQuant object, is computed as \(|x_y - x_{y-1}) / x_{y-1}|\). The resulting object will be one year shorter than the input. The first year will be missing as values are assigned to the final year of each pair.

iav(object)

Value

An object of the same class as object.

Author

The FLR Team

Examples

data(ple4)
# Compute inter-annual variability in catch
iav(catch(ple4))
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1958      1959      1960      1961      1962      1963      1964     
#>   all 0.1330372 0.1847259 0.1215789 0.0132751 0.0564642 0.1150058 0.0477615
#>      year
#> age   1965      1966      1967      1968      1969      1970      1971     
#>   all 0.0262144 0.0717180 0.0480213 0.0301291 0.0243120 0.0653893 0.0337214
#>      year
#> age   1972      1973      1974      1975      1976      1977      1978     
#>   all 0.0578063 0.0142249 0.0427618 0.0468241 0.0634222 0.0570740 0.0743023
#>      year
#> age   1979      1980      1981      1982      1983      1984      1985     
#>   all 0.0309460 0.0697257 0.0010639 0.0430649 0.1049299 0.0735235 0.0823874
#>      year
#> age   1986      1987      1988      1989      1990      1991      1992     
#>   all 0.1301530 0.1047575 0.0219299 0.0736247 0.1418697 0.1293683 0.1168393
#>      year
#> age   1993      1994      1995      1996      1997      1998      1999     
#>   all 0.0680792 0.1577328 0.1231082 0.0068560 0.1554476 0.1251388 0.0033755
#>      year
#> age   2000      2001      2002      2003      2004      2005      2006     
#>   all 0.1445205 0.1225443 0.1225586 0.0710821 0.0907169 0.1821103 0.0234528
#>      year
#> age   2007      2008      2009      2010      2011      2012      2013     
#>   all 0.0634175 0.0823385 0.0203357 0.0269798 0.0075226 0.1015176 0.0696424
#>      year
#> age   2014      2015      2016      2017     
#>   all 0.0092539 0.0172577 0.0445777 0.0479687
#> 
#> units: