Methods iter

Select or modify iterations of an FLR object

iter(obj, ...)

# S4 method for FLArray
iter(obj, iter)

# S4 method for vector
iter(obj, iter)

Details

To extract or modify a subset of the iterations contained in an FLR object, the iter and iter<- methods can be used.

In complex objects with various FLQuant slots, the iter method checks whether individual slots contain more than one iteration, i.e. dims(object)[6] > 1. If a particular slot contains a single iteration, that is returned, otherwise the chosen iteration is selected. This is in contrast with the subset operator [, which does not carry out this check.

For objects of class FLModel, iters are extracted for slots of classes FLQuant, FLCohort and FLPar.

Generic function

iter(object) iter<-(object,value)

See also

FLComp, FLQuant

Examples

# For an FLQuant
  flq <- FLQuant(rnorm(800), dim=c(4,10,2), iter=10)
  iter(flq, 2)

#> An object of class "FLQuant"
#> , , unit = 1, season = all, area = unique
#> 
#>      year
#> quant 1          2          3          4          5          6         
#>     1  0.5549896 -0.2114360  0.7602762  0.4995441  0.1577133 -0.0815580
#>     2 -1.1152055  0.7374029 -0.4542199  0.2275489 -0.5914220  0.3797797
#>     3  0.0359525  1.2328979  1.1605250 -1.0405042  0.2442141 -0.8422863
#>     4  0.1092659 -0.9602330 -2.0401308  0.5914743 -1.1850680  0.9855102
#>      year
#> quant 7          8          9          10        
#>     1 -1.1594355 -0.4079460 -0.3903392  0.2140149
#>     2  0.9911963 -1.5389891 -0.0701439 -1.2089252
#>     3 -0.6605940 -0.7787460 -0.1264705 -0.9520266
#>     4 -0.3891365 -1.1176381  0.8068256 -0.0046742
#> 
#> , , unit = 2, season = all, area = unique
#> 
#>      year
#> quant 1          2          3          4          5          6         
#>     1 -1.5063238  2.3629959 -0.1558732  1.3790289  1.2652894  1.0297017
#>     2 -0.2164531  0.9295353  0.1369936 -0.5422116  0.0193485 -0.5466406
#>     3  0.3216176 -0.8966327 -0.7125195  0.4787367  0.9395548 -0.4727820
#>     4  0.8308609  2.7536572  0.7846749  0.1838262  0.8810847  0.8064503
#>      year
#> quant 7          8          9          10        
#>     1  0.8506398  1.5996180  0.1110952 -0.5452818
#>     2  1.1744529 -0.0089559  1.7264312 -0.1659454
#>     3  0.3653619  2.1639597  2.5206992  0.1684750
#>     4  0.6307298  0.5508508 -0.1389735  1.7550763
#> 
#> units:  NA 

# For the more complex FLStock object
  fls <- FLStock(catch.n=flq, m=FLQuant(0.2, dim=c(4,10,2)))
  summary(fls)

#> An object of class "FLStock"
#> 
#> Name:  
#> Description:  
#> Quant: quant 
#> Dims:  quant 	year	unit	season	area	iter
#> 	quant	10	2	1	1	10	
#> 
#> Range:  min	max	pgroup	minyear	maxyear	minfbar	maxfbar 
#> 	1	4	4	1	10	1	4	
#> 
#> catch         : [ 1 10 2 1 1 10 ], units =  NA 
#> catch.n       : [ 4 10 2 1 1 10 ], units =  NA 
#> catch.wt      : [ 4 10 2 1 1 10 ], units =  NA 
#> discards      : [ 1 10 2 1 1 10 ], units =  NA 
#> discards.n    : [ 4 10 2 1 1 10 ], units =  NA 
#> discards.wt   : [ 4 10 2 1 1 10 ], units =  NA 
#> landings      : [ 1 10 2 1 1 10 ], units =  NA 
#> landings.n    : [ 4 10 2 1 1 10 ], units =  NA 
#> landings.wt   : [ 4 10 2 1 1 10 ], units =  NA 
#> stock         : [ 1 10 2 1 1 10 ], units =  NA 
#> stock.n       : [ 4 10 2 1 1 10 ], units =  NA 
#> stock.wt      : [ 4 10 2 1 1 10 ], units =  NA 
#> m             : [ 4 10 2 1 1 1 ], units =  NA 
#> mat           : [ 4 10 2 1 1 10 ], units =  prop 
#> harvest       : [ 4 10 2 1 1 10 ], units =  NA 
#> harvest.spwn  : [ 4 10 2 1 1 10 ], units =  prop 
#> m.spwn        : [ 4 10 2 1 1 10 ], units =  prop 

  # Extraction using iter...
    fls2 <- iter(fls, 2)
    summary(fls2)

#> An object of class "FLStock"
#> 
#> Name:  
#> Description:  
#> Quant: quant 
#> Dims:  quant 	year	unit	season	area	iter
#> 	quant	10	2	1	1	1	
#> 
#> Range:  min	max	pgroup	minyear	maxyear	minfbar	maxfbar 
#> 	1	4	4	1	10	1	4	
#> 
#> catch         : [ 1 10 2 1 1 1 ], units =  NA 
#> catch.n       : [ 4 10 2 1 1 1 ], units =  NA 
#> catch.wt      : [ 4 10 2 1 1 1 ], units =  NA 
#> discards      : [ 1 10 2 1 1 1 ], units =  NA 
#> discards.n    : [ 4 10 2 1 1 1 ], units =  NA 
#> discards.wt   : [ 4 10 2 1 1 1 ], units =  NA 
#> landings      : [ 1 10 2 1 1 1 ], units =  NA 
#> landings.n    : [ 4 10 2 1 1 1 ], units =  NA 
#> landings.wt   : [ 4 10 2 1 1 1 ], units =  NA 
#> stock         : [ 1 10 2 1 1 1 ], units =  NA 
#> stock.n       : [ 4 10 2 1 1 1 ], units =  NA 
#> stock.wt      : [ 4 10 2 1 1 1 ], units =  NA 
#> m             : [ 4 10 2 1 1 1 ], units =  NA 
#> mat           : [ 4 10 2 1 1 1 ], units =  prop 
#> harvest       : [ 4 10 2 1 1 1 ], units =  NA 
#> harvest.spwn  : [ 4 10 2 1 1 1 ], units =  prop 
#> m.spwn        : [ 4 10 2 1 1 1 ], units =  prop 

  # ...in contrast to using [ which returns an error
    ## Not run: fls[,,,,,2]