Methods iter

Select or modify iterations of an FLR object

iter(obj, ...)

# S4 method for FLArray
iter(obj, iter)

Details

To extract or modify a subset of the iterations contained in an FLR object, the iter and iter<- methods can be used.

In complex objects with various FLQuant slots, the iter method checks whether individual slots contain more than one iteration, i.e. dims(object)[6] > 1. If a particular slot contains a single iteration, that is returned, otherwise the chosen iteration is selected. This is in contrast with the subset operator [, which does not carry out this check.

For objects of class FLModel, iters are extracted for slots of classes FLQuant, FLCohort and FLPar.

Generic function

iter(object) iter<-(object,value)

See also

Author

The FLR Team

Examples

# For an FLQuant
  flq <- FLQuant(rnorm(800), dim=c(4,10,2), iter=10)
  iter(flq, 2)
#> An object of class "FLQuant"
#> , , unit = 1, season = all, area = unique
#> 
#>      year
#> quant 1         2         3         4         5         6         7        
#>     1 -1.352216  0.756363 -1.709766  1.450836 -0.271837 -1.382373  0.942496
#>     2 -1.940399  1.230329 -1.133418 -0.704376  1.400316  0.612837  2.038637
#>     3  1.450225  0.559721 -1.331987  0.181495  0.250366  0.259660 -0.491266
#>     4  0.592853 -0.028782 -0.044024 -1.677628 -1.699283  2.300851 -1.161327
#>      year
#> quant 8         9         10       
#>     1 -0.718673 -0.164734  1.853971
#>     2  1.702926  0.629675  0.686454
#>     3  0.541148  0.255370 -0.384555
#>     4  2.293740  1.003254 -0.978868
#> 
#> , , unit = 2, season = all, area = unique
#> 
#>      year
#> quant 1         2         3         4         5         6         7        
#>     1  1.704393 -0.662761  0.024426 -1.763256 -0.070950 -0.795883 -2.438448
#>     2  0.237118 -0.339683  0.013693  1.375092 -0.767328  0.492438  1.374253
#>     3  0.064994 -1.572120 -1.833836  1.680459  0.395532  0.554064  1.006782
#>     4 -0.796433 -0.328311  0.417248 -0.339481  0.460949  1.821016 -0.232798
#>      year
#> quant 8         9         10       
#>     1  2.012425  1.096831 -0.738891
#>     2 -1.412838  1.587253 -0.028496
#>     3 -0.848238 -0.088374  0.158780
#>     4 -0.132633 -0.780475  0.191180
#> 
#> units:  NA 
# For the more complex FLStock object
  data(ple4)
  fls <- propagate(ple4, 10)
  # Extraction using iter...
    fls2 <- iter(fls, 2)
    summary(fls2)
#> An object of class "FLStock"
#> 
#> Name: PLE 
#> Description: Plaice in IV. ICES WGNSSK 2018. FLAAP 
#> Quant: age 
#> Dims:  age 	year	unit	season	area	iter
#> 	10	61	1	1	1	1	
#> 
#> Range:  min	max	pgroup	minyear	maxyear	minfbar	maxfbar 
#> 	1	10	10	1957	2017	2	6	
#> 
#> Metrics: 
#>   rec: 367450 - 4303680  (1000) 
#>   ssb: 203391 - 913290  (t) 
#>   catch: 78360 - 315245  (t) 
#>   fbar: 0.20 - 0.72  (f)