Methods iter

Select or modify iterations of an FLR object

iter(obj, ...)

# S4 method for FLArray
iter(obj, iter)

Details

To extract or modify a subset of the iterations contained in an FLR object, the iter and iter<- methods can be used.

In complex objects with various FLQuant slots, the iter method checks whether individual slots contain more than one iteration, i.e. dims(object)[6] > 1. If a particular slot contains a single iteration, that is returned, otherwise the chosen iteration is selected. This is in contrast with the subset operator [, which does not carry out this check.

For objects of class FLModel, iters are extracted for slots of classes FLQuant, FLCohort and FLPar.

Generic function

iter(object) iter<-(object,value)

See also

Author

The FLR Team

Examples

# For an FLQuant
  flq <- FLQuant(rnorm(800), dim=c(4,10,2), iter=10)
  iter(flq, 2)
#> An object of class "FLQuant"
#> , , unit = 1, season = all, area = unique
#> 
#>      year
#> quant 1         2         3         4         5         6         7        
#>     1 -0.442000  2.007881  0.076999  0.114823 -0.750070 -0.514172 -0.588414
#>     2  0.623726 -0.673844  1.250163  0.241713  0.304727 -0.775902 -0.442828
#>     3  1.038149 -1.465461 -1.974589  0.243682 -1.076594  1.208575  2.135920
#>     4  0.958761  0.615369 -0.012112 -0.505401  1.918718 -1.594166  0.354880
#>      year
#> quant 8         9         10       
#>     1  0.326673 -0.230727 -0.611315
#>     2 -2.111447 -0.218737 -0.076286
#>     3 -1.223762 -1.207471  0.613740
#>     4 -0.891152 -0.548189 -0.466444
#> 
#> , , unit = 2, season = all, area = unique
#> 
#>      year
#> quant 1         2         3         4         5         6         7        
#>     1 -0.876738  0.791940  0.258755 -0.454094 -0.349878 -0.803043 -0.829343
#>     2  0.636365 -0.499413 -0.466678 -0.657722 -0.483631  0.530965  1.558066
#>     3 -0.459346  0.947241 -1.355376  0.183542  1.231126 -1.234978 -1.231379
#>     4 -0.322459 -0.756777 -0.168950 -0.937203  0.843273 -0.350118  0.070741
#>      year
#> quant 8         9         10       
#>     1 -0.067485  0.960857 -0.927393
#>     2  0.956091 -0.143713  0.339454
#>     3 -1.811261  0.459306 -0.407331
#>     4  1.819369 -1.210591  0.027214
#> 
#> units:  NA 
# For the more complex FLStock object
  data(ple4)
  fls <- propagate(ple4, 10)
  # Extraction using iter...
    fls2 <- iter(fls, 2)
    summary(fls2)
#> An object of class "FLStock"
#> 
#> Name: PLE 
#> Description: Plaice in IV. ICES WGNSSK 2018. FLAAP 
#> Quant: age 
#> Dims:  age 	year	unit	season	area	iter
#> 	10	61	1	1	1	1	
#> 
#> Range:  min	max	pgroup	minyear	maxyear	minfbar	maxfbar 
#> 	1	10	10	1957	2017	2	6	
#> 
#> Metrics: 
#>   rec: 367450 - 4303680  (1000) 
#>   ssb: 203391 - 913290  (t) 
#>   catch: 78360 - 315245  (t) 
#>   fbar: 0.20 - 0.72  (f)