Methods iter

Select or modify iterations of an FLR object

iter(obj, ...)

# S4 method for FLArray
iter(obj, iter)

# S4 method for vector
iter(obj, iter)

Details

To extract or modify a subset of the iterations contained in an FLR object, the iter and iter<- methods can be used.

In complex objects with various FLQuant slots, the iter method checks whether individual slots contain more than one iteration, i.e. dims(object)[6] > 1. If a particular slot contains a single iteration, that is returned, otherwise the chosen iteration is selected. This is in contrast with the subset operator [, which does not carry out this check.

For objects of class FLModel, iters are extracted for slots of classes FLQuant, FLCohort and FLPar.

Generic function

iter(object) iter<-(object,value)

See also

FLComp, FLQuant

Examples

# For an FLQuant
  flq <- FLQuant(rnorm(800), dim=c(4,10,2), iter=10)
  iter(flq, 2)

#> An object of class "FLQuant"
#> , , unit = 1, season = all, area = unique
#> 
#>      year
#> quant 1         2         3         4         5         6         7        
#>     1  0.090530 -0.079086 -1.149473  0.812626  0.920418 -0.253411  0.115474
#>     2  0.951428 -0.231611  2.287377  0.130429 -0.154438  1.417268  0.564436
#>     3 -1.497427  0.477244  2.897842  0.466637  1.988254  0.426489 -0.292125
#>     4 -1.663960 -1.334249 -0.430304 -1.053646 -0.089527 -0.151338  0.585202
#>      year
#> quant 8         9         10       
#>     1  0.375576 -0.872512 -0.651046
#>     2  1.086023 -1.878808  1.189476
#>     3  0.400441 -0.144040  0.316996
#>     4  0.258793 -0.537420  0.465542
#> 
#> , , unit = 2, season = all, area = unique
#> 
#>      year
#> quant 1         2         3         4         5         6         7        
#>     1 -0.110086  1.009875 -0.959959  0.171714  0.967551  1.194909  1.257798
#>     2  0.384122 -0.047082 -0.859461 -0.936854  1.391292 -0.963873 -0.523927
#>     3  0.608537 -1.020705  0.357915  0.089443 -0.266533  0.445499  0.058783
#>     4  0.137924 -0.207448  0.877584  0.276962  0.828046  0.301788  0.110979
#>      year
#> quant 8         9         10       
#>     1  1.803635 -1.781715 -0.232663
#>     2 -0.544628 -0.889818 -0.663944
#>     3  0.038323  0.752724  0.274931
#>     4 -0.282182 -0.855544 -0.153608
#> 
#> units:  NA 

# For the more complex FLStock object
  fls <- FLStock(catch.n=flq, m=FLQuant(0.2, dim=c(4,10,2)))
  summary(fls)

#> An object of class "FLStock"
#> 
#> Name:  
#> Description:  
#> Quant: quant 
#> Dims:  quant 	year	unit	season	area	iter
#> 	quant	10	2	1	1	10	
#> 
#> Range:  min	max	pgroup	minyear	maxyear	minfbar	maxfbar 
#> 	1	4	4	1	10	1	4	
#> 
#> catch         : [ 1 10 2 1 1 10 ], units =  NA 
#> catch.n       : [ 4 10 2 1 1 10 ], units =  NA 
#> catch.wt      : [ 4 10 2 1 1 10 ], units =  NA 
#> discards      : [ 1 10 2 1 1 10 ], units =  NA 
#> discards.n    : [ 4 10 2 1 1 10 ], units =  NA 
#> discards.wt   : [ 4 10 2 1 1 10 ], units =  NA 
#> landings      : [ 1 10 2 1 1 10 ], units =  NA 
#> landings.n    : [ 4 10 2 1 1 10 ], units =  NA 
#> landings.wt   : [ 4 10 2 1 1 10 ], units =  NA 
#> stock         : [ 1 10 2 1 1 10 ], units =  NA 
#> stock.n       : [ 4 10 2 1 1 10 ], units =  NA 
#> stock.wt      : [ 4 10 2 1 1 10 ], units =  NA 
#> m             : [ 4 10 2 1 1 1 ], units =  NA 
#> mat           : [ 4 10 2 1 1 10 ], units =  prop 
#> harvest       : [ 4 10 2 1 1 10 ], units =  NA 
#> harvest.spwn  : [ 4 10 2 1 1 10 ], units =  prop 
#> m.spwn        : [ 4 10 2 1 1 10 ], units =  prop 

  # Extraction using iter...
    fls2 <- iter(fls, 2)
    summary(fls2)

#> An object of class "FLStock"
#> 
#> Name:  
#> Description:  
#> Quant: quant 
#> Dims:  quant 	year	unit	season	area	iter
#> 	quant	10	2	1	1	1	
#> 
#> Range:  min	max	pgroup	minyear	maxyear	minfbar	maxfbar 
#> 	1	4	4	1	10	1	4	
#> 
#> catch         : [ 1 10 2 1 1 1 ], units =  NA 
#> catch.n       : [ 4 10 2 1 1 1 ], units =  NA 
#> catch.wt      : [ 4 10 2 1 1 1 ], units =  NA 
#> discards      : [ 1 10 2 1 1 1 ], units =  NA 
#> discards.n    : [ 4 10 2 1 1 1 ], units =  NA 
#> discards.wt   : [ 4 10 2 1 1 1 ], units =  NA 
#> landings      : [ 1 10 2 1 1 1 ], units =  NA 
#> landings.n    : [ 4 10 2 1 1 1 ], units =  NA 
#> landings.wt   : [ 4 10 2 1 1 1 ], units =  NA 
#> stock         : [ 1 10 2 1 1 1 ], units =  NA 
#> stock.n       : [ 4 10 2 1 1 1 ], units =  NA 
#> stock.wt      : [ 4 10 2 1 1 1 ], units =  NA 
#> m             : [ 4 10 2 1 1 1 ], units =  NA 
#> mat           : [ 4 10 2 1 1 1 ], units =  prop 
#> harvest       : [ 4 10 2 1 1 1 ], units =  NA 
#> harvest.spwn  : [ 4 10 2 1 1 1 ], units =  prop 
#> m.spwn        : [ 4 10 2 1 1 1 ], units =  prop 

  # ...in contrast to using [ which returns an error
    ## Not run: fls[,,,,,2]