Methods iter

Select or modify iterations of an FLR object

iter(obj, ...)

# S4 method for FLArray
iter(obj, iter)

Details

To extract or modify a subset of the iterations contained in an FLR object, the iter and iter<- methods can be used.

In complex objects with various FLQuant slots, the iter method checks whether individual slots contain more than one iteration, i.e. dims(object)[6] > 1. If a particular slot contains a single iteration, that is returned, otherwise the chosen iteration is selected. This is in contrast with the subset operator [, which does not carry out this check.

For objects of class FLModel, iters are extracted for slots of classes FLQuant, FLCohort and FLPar.

Generic function

iter(object) iter<-(object,value)

See also

Author

The FLR Team

Examples

# For an FLQuant
  flq <- FLQuant(rnorm(800), dim=c(4,10,2), iter=10)
  iter(flq, 2)

#> An object of class "FLQuant"
#> , , unit = 1, season = all, area = unique
#> 
#>      year
#> quant 1          2          3          4          5          6         
#>     1  0.9036330  0.4479273 -0.9975230  0.0149466  0.6344955 -0.8772473
#>     2 -2.2526568 -0.1700767  0.4486126 -1.8073206 -3.1210439  1.0489363
#>     3  1.3916326 -1.3379917 -0.1288426 -0.4498878 -0.0741168 -0.9426929
#>     4  0.0021686  0.9058720 -0.4817287  0.4187975  0.6536831  0.4300361
#>      year
#> quant 7          8          9          10        
#>     1 -0.7387653  1.3264455 -0.4749096  1.2635706
#>     2  0.3529074 -0.7998410 -1.1784852 -0.8299078
#>     3 -0.8374302 -0.7282673 -2.0924444 -0.9804823
#>     4 -0.8901320  0.2307829 -1.0345330  0.0510746
#> 
#> , , unit = 2, season = all, area = unique
#> 
#>      year
#> quant 1          2          3          4          5          6         
#>     1 -0.0640060 -0.6805505 -0.4361111 -0.6580589 -0.3492219  0.8510244
#>     2  0.8517549  1.5046112  2.9364849  0.2891020 -0.3599562  0.5712133
#>     3  0.9328557 -0.0082422  1.3491986 -0.2551301  0.1257492 -0.0961339
#>     4 -0.6754396 -1.1969618  0.5892612  0.1710397 -2.2179958  0.7698861
#>      year
#> quant 7          8          9          10        
#>     1 -0.2846050  0.6568769 -0.7696865 -0.2988305
#>     2 -1.0979271 -0.6450367  0.1959713  1.3195509
#>     3 -0.4645786 -1.8681936 -1.7682876 -0.2843006
#>     4  0.6567013 -0.9572913  0.9828270  0.9512391
#> 
#> units:  NA 

# For the more complex FLStock object
  fls <- FLStock(catch.n=flq, m=FLQuant(0.2, dim=c(4,10,2)))
  summary(fls)

#> An object of class "FLStock"
#> 
#> Name:  
#> Description:  
#> Quant: quant 
#> Dims:  quant 	year	unit	season	area	iter
#> 	4	10	2	1	1	10	
#> 
#> Range:  min	max	pgroup	minyear	maxyear	minfbar	maxfbar 
#> 	1	4	4	1	10	1	4	
#> 
#> catch         : [ 1 10 2 1 1 1 ], units =  NA 
#> catch.n       : [ 4 10 2 1 1 10 ], units =  NA 
#> catch.wt      : [ 4 10 2 1 1 1 ], units =  NA 
#> discards      : [ 1 10 2 1 1 1 ], units =  NA 
#> discards.n    : [ 4 10 2 1 1 1 ], units =  NA 
#> discards.wt   : [ 4 10 2 1 1 1 ], units =  NA 
#> landings      : [ 1 10 2 1 1 1 ], units =  NA 
#> landings.n    : [ 4 10 2 1 1 1 ], units =  NA 
#> landings.wt   : [ 4 10 2 1 1 1 ], units =  NA 
#> stock         : [ 1 10 2 1 1 1 ], units =  NA 
#> stock.n       : [ 4 10 2 1 1 1 ], units =  NA 
#> stock.wt      : [ 4 10 2 1 1 1 ], units =  NA 
#> m             : [ 4 10 2 1 1 1 ], units =  NA 
#> mat           : [ 4 10 2 1 1 1 ], units =   
#> harvest       : [ 4 10 2 1 1 1 ], units =  NA 
#> harvest.spwn  : [ 4 10 2 1 1 1 ], units =   
#> m.spwn        : [ 4 10 2 1 1 1 ], units =   

  # Extraction using iter...
    fls2 <- iter(fls, 2)
    summary(fls2)

#> An object of class "FLStock"
#> 
#> Name:  
#> Description:  
#> Quant: quant 
#> Dims:  quant 	year	unit	season	area	iter
#> 	4	10	2	1	1	1	
#> 
#> Range:  min	max	pgroup	minyear	maxyear	minfbar	maxfbar 
#> 	1	4	4	1	10	1	4	
#> 
#> catch         : [ 1 10 2 1 1 1 ], units =  NA 
#> catch.n       : [ 4 10 2 1 1 1 ], units =  NA 
#> catch.wt      : [ 4 10 2 1 1 1 ], units =  NA 
#> discards      : [ 1 10 2 1 1 1 ], units =  NA 
#> discards.n    : [ 4 10 2 1 1 1 ], units =  NA 
#> discards.wt   : [ 4 10 2 1 1 1 ], units =  NA 
#> landings      : [ 1 10 2 1 1 1 ], units =  NA 
#> landings.n    : [ 4 10 2 1 1 1 ], units =  NA 
#> landings.wt   : [ 4 10 2 1 1 1 ], units =  NA 
#> stock         : [ 1 10 2 1 1 1 ], units =  NA 
#> stock.n       : [ 4 10 2 1 1 1 ], units =  NA 
#> stock.wt      : [ 4 10 2 1 1 1 ], units =  NA 
#> m             : [ 4 10 2 1 1 1 ], units =  NA 
#> mat           : [ 4 10 2 1 1 1 ], units =   
#> harvest       : [ 4 10 2 1 1 1 ], units =  NA 
#> harvest.spwn  : [ 4 10 2 1 1 1 ], units =   
#> m.spwn        : [ 4 10 2 1 1 1 ], units =   

  # ...in contrast to using [ which returns an error
    if (FALSE) fls[,,,,,2]