Calculate quantile(s) of length distribution

z = (k * (linf - lmean)) / (lmean - lc) lmean = sum(naa * len) / sum(naa) lc, length at first capture

Usage

indicators.len(
  object,
  indicators = "lbar",
  model = vonbert,
  params,
  cv = 0.1,
  lmax = 1.25,
  bin = 1,
  n = 500,
  metric = catch.n,
  ...
)

lenquantile(x, quantile = 0.5)

lmax5(x)

l95(x)

l25(x)

lc50(x)

lmode(x)

lbar(x)

lmean(x)

lmaxy(x, lenwt)

pmega(x, linf, lopt = linf * 2/3)

bheqz(x, linf, k, t0, lc = lc50(x))

References

Kell, L.T., Minto, C., Gerritsen, H.D. 2022. Evaluation of the skill of length-based indicators to identify stock status and trends. ICES Journal of Marine Science. doiu: 10.1093/icesjms/fsac043.

ICES. 2015. Report of the Fifth Workshop on the Development of Quantitative Assessment Methodologies based on Life-history Traits, Exploitation Characteristics and other Relevant Parameters for Data-limited Stocks (WKLIFE V), 5–9 October 2015, Lisbon, Portugal. ICES CM 2015/ACOM:56. 157 pp.

ICES. 2020. Tenth Workshop on the Development of Quantitative Assessment Methodologies based on LIFE-history traits, exploitation characteristics, and other relevant parameters for data-limited stocks (WKLIFE X). ICES Scientific Reports. 2:98. 72 pp. http://doi.org/10.17895/ices.pub.5985

Examples

data(ple4)
indicators.len(ple4, indicators=c('lbar', 'lmaxy'),
  params=FLPar(linf=132, k=0.080, t0=-0.35), metric='catch.n',
  lenwt=FLPar(a=0.01030, b=2.975))
#> $lbar
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 34.9 34.9 31.0 30.5 31.2 32.0 35.7 31.5 30.4 33.7 36.4 37.0 36.6 32.2
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 34.4 35.2 28.5 26.4 27.2 28.9 25.9 26.4 27.7 26.5 27.1 23.7 24.6 25.0
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 24.6 20.5 24.2 27.2 29.1 30.7 30.1 31.2 30.9 31.8 29.5 28.1 25.1 26.5
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 29.2 30.8 30.3 26.5 26.6 26.2 25.9 27.8 25.8 25.4 26.5 26.3 29.6 30.5
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 30.5 28.3 30.3 31.6 32.5
#> 
#> units:  cm 
#> 
#> $lmaxy
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 79   70   71   63   72   46   69   51   63   31   41   46   50   56  
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 70   70   70   62   61   68   43   53   51   74   70   75   75   90  
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 71   46   23   31   41   49   48   59   54   58   61   74   33   59  
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 32   41   45   48   24   55   71   46   57   55   32   64   51   50  
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 61   38   58   61   77  
#> 
#> units:  cm 
#> 
indicators.len(ple4, indicators=c('pmega'),
  params=FLPar(linf=60, k=2.29e-01, t0=-1.37), metric='catch.n')
#> $pmega
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957  1958  1959  1960  1961  1962  1963  1964  1965  1966  1967  1968 
#>   all 0.298 0.310 0.228 0.210 0.228 0.266 0.280 0.272 0.148 0.210 0.312 0.374
#>      year
#> len   1969  1970  1971  1972  1973  1974  1975  1976  1977  1978  1979  1980 
#>   all 0.370 0.298 0.326 0.274 0.190 0.142 0.096 0.150 0.202 0.186 0.138 0.136
#>      year
#> len   1981  1982  1983  1984  1985  1986  1987  1988  1989  1990  1991  1992 
#>   all 0.124 0.086 0.082 0.108 0.114 0.100 0.078 0.116 0.162 0.216 0.218 0.210
#>      year
#> len   1993  1994  1995  1996  1997  1998  1999  2000  2001  2002  2003  2004 
#>   all 0.224 0.224 0.218 0.128 0.114 0.068 0.124 0.202 0.226 0.186 0.146 0.108
#>      year
#> len   2005  2006  2007  2008  2009  2010  2011  2012  2013  2014  2015  2016 
#>   all 0.136 0.154 0.132 0.124 0.118 0.132 0.196 0.192 0.194 0.186 0.192 0.246
#>      year
#> len   2017 
#>   all 0.280
#> 
#> units:  NA 
#> 
data(ple4.index)
indicators.len(ple4.index, indicators=c('lbar', 'lmean'),
  params=FLPar(linf=132, k=0.080, t0=-0.35), metric='index')
#> $lbar
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009
#>   all 21.6 17.1 21.1 22.4 21.1 22.5 16.5 22.2 22.3 22.2 21.7 19.2 21.3 21.6
#>      year
#> len   2010 2011 2012 2013 2014 2015 2016 2017
#>   all 22.4 20.3 25.1 24.5 23.1 27.1 25.4 21.8
#> 
#> units:  cm 
#> 
#> $lmean
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009
#>   all 25.6 23.3 27.0 25.3 24.0 25.4 22.3 25.8 28.2 25.1 25.1 25.7 24.2 25.5
#>      year
#> len   2010 2011 2012 2013 2014 2015 2016 2017
#>   all 30.4 24.3 29.5 27.8 26.7 31.5 29.9 29.5
#> 
#> units:  cm 
#> 
#
ialk <- invALK(params=FLPar(linf = 60, k = 2.29e-01, t0 = -1.37e+00),
  model=vonbert, age=1:10, lmax=1.2)
samps <- lenSamples(catch.n(ple4), invALK=ialk, n=250)
lenquantile(samps, 0.50)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 39.0 38.5 36.0 35.0 35.0 36.0 39.0 35.5 35.0 38.0 40.0 40.0 39.5 34.0
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 36.0 39.0 34.0 32.0 34.0 37.0 31.0 32.0 33.0 33.0 34.0 31.0 33.0 34.0
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 33.0 27.5 32.0 35.0 36.0 36.0 37.0 37.0 36.0 36.0 35.0 34.0 33.0 33.0
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 37.0 38.0 37.0 34.0 33.0 34.0 34.0 35.0 33.0 33.0 32.0 33.0 34.0 37.0
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 36.0 34.0 36.0 38.0 38.0
#> 
#> units:  NA 
lmax5(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 59.1 58.5 58.4 55.5 57.4 57.1 58.8 54.8 55.9 54.5 54.1 57.7 59.6 57.6
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 61.3 60.0 57.0 57.3 51.4 56.8 54.2 53.7 55.6 52.3 52.6 54.1 52.0 54.5
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 53.8 49.8 53.4 50.0 52.8 53.6 52.1 55.0 54.6 54.8 53.7 54.5 49.0 49.3
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 50.8 52.3 52.9 53.8 53.3 50.3 51.7 51.8 50.4 51.1 51.4 51.8 54.5 54.3
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 53.5 52.5 55.6 55.8 58.0
#> 
#> units:  cm 
l95(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 54.0 54.5 54.5 52.0 55.0 53.0 56.0 51.5 52.0 49.6 51.0 53.5 54.5 54.0
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 56.5 57.0 55.0 54.0 47.0 52.5 51.0 49.0 53.0 49.0 49.0 49.1 49.0 50.0
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 49.1 46.0 47.5 46.0 50.0 49.5 50.0 52.0 51.0 52.5 49.5 51.0 46.5 44.0
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 48.5 51.0 50.0 49.5 49.0 46.5 48.0 48.5 48.0 47.0 49.0 47.0 48.5 50.5
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 51.0 48.5 51.5 53.0 53.5
#> 
#> units:  cm 
l25(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 34.0 31.0 30.0 29.0 30.0 32.0 33.0 27.0 31.0 34.0 34.2 34.0 31.0 29.0
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 29.0 33.0 26.0 27.0 29.0 29.0 26.0 27.0 26.0 27.0 28.0 26.0 27.0 27.2
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 27.0 25.0 29.0 29.0 30.0 29.0 30.2 31.0 30.0 30.0 29.0 29.0 26.0 30.2
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 33.0 32.0 31.0 27.0 30.0 26.0 29.0 30.0 27.2 28.0 27.0 27.0 28.0 32.0
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 30.0 29.0 31.0 31.2 31.0
#> 
#> units:  cm 
lc50(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 17.5 17.5 13.0 17.0 16.5 18.5 19.5 13.0 16.5 19.5 18.5 19.5 22.5 16.0
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 18.5 19.5 12.0 13.0 19.0 19.0 13.0 13.5 12.0 12.5 16.5 12.0 18.5 17.0
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 12.0 12.5 15.5 18.0 17.0 18.5 21.0 16.0 19.0 18.5 13.5 16.0 18.5 16.0
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 18.5 21.0 16.0 19.0 16.5 16.5 17.5 17.0 13.0 17.5 13.0 12.5 17.0 15.0
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 22.0 16.0 17.5 19.0 19.5
#> 
#> units:  cm 
lmode(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 35   35   26   34   33   37   39   26   33   39   37   39   45   32  
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 37   39   24   26   38   38   26   27   24   25   33   24   37   34  
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 24   25   31   36   34   37   42   32   38   37   27   32   37   32  
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 37   42   32   38   33   33   35   34   26   35   26   25   34   30  
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 44   32   35   38   39  
#> 
#> units:  cm 
lbar(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 39.6 38.7 37.0 36.0 36.5 37.2 40.3 36.0 36.8 38.0 39.4 40.0 39.1 36.3
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 37.9 39.8 35.5 34.2 34.7 36.3 33.0 34.3 34.0 33.6 34.8 32.5 33.3 34.2
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 34.1 30.2 33.5 34.9 36.1 36.1 37.1 37.4 36.0 36.6 35.8 34.5 33.1 34.1
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 37.0 37.8 37.0 34.2 34.5 33.8 34.6 35.2 34.1 33.5 33.7 34.3 34.8 37.6
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 36.2 34.9 36.8 38.2 38.3
#> 
#> units:  cm 
lmean(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 44.6 45.5 39.6 42.3 42.1 45.1 48.0 39.6 41.2 44.7 44.1 46.6 51.6 43.6
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 48.6 47.9 38.0 37.2 44.3 45.0 37.2 38.1 36.3 36.1 41.2 34.8 44.2 41.0
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 35.6 33.8 38.0 41.6 42.2 44.4 47.6 42.4 44.9 44.8 38.9 40.1 42.7 38.0
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 41.8 47.0 41.4 44.6 40.0 40.0 41.7 41.3 36.6 41.7 36.5 36.5 41.9 40.1
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 49.9 40.5 43.1 45.6 47.1
#> 
#> units:  cm 
# Linf(ple4) = 60
lmean(samps) / (0.75 * lc50(samps) + 0.25 * 60) #
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 1.59 1.62 1.60 1.53 1.54 1.56 1.62 1.60 1.51 1.51 1.53 1.57 1.62 1.61
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 1.68 1.62 1.58 1.50 1.51 1.54 1.50 1.52 1.51 1.48 1.51 1.45 1.53 1.48
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 1.48 1.39 1.43 1.46 1.52 1.54 1.55 1.57 1.54 1.55 1.55 1.49 1.48 1.41
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 1.45 1.53 1.53 1.53 1.46 1.46 1.48 1.49 1.48 1.48 1.48 1.50 1.51 1.53
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 1.58 1.50 1.53 1.56 1.59
#> 
#> units:   
lenwt <- FLPar(a=0.01030, b=2.975)
lmaxy(samps, lenwt)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 51   49   58   41   56   37   52   48   36   39   43   45   45   51  
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 56   41   56   42   38   41   41   47   55   43   40   36   37   34  
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 38   39   33   39   38   45   42   43   38   46   42   42   37   35  
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 38   45   48   38   33   42   35   40   40   35   48   44   34   47  
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 44   41   44   53   47  
#> 
#> units:  cm 
pmega(samps, linf=60)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957  1958  1959  1960  1961  1962  1963  1964  1965  1966  1967  1968 
#>   all 0.320 0.332 0.216 0.188 0.196 0.204 0.332 0.276 0.196 0.180 0.276 0.376
#>      year
#> len   1969  1970  1971  1972  1973  1974  1975  1976  1977  1978  1979  1980 
#>   all 0.364 0.268 0.284 0.316 0.244 0.156 0.124 0.208 0.152 0.192 0.164 0.132
#>      year
#> len   1981  1982  1983  1984  1985  1986  1987  1988  1989  1990  1991  1992 
#>   all 0.144 0.104 0.116 0.128 0.136 0.072 0.084 0.088 0.172 0.208 0.232 0.252
#>      year
#> len   1993  1994  1995  1996  1997  1998  1999  2000  2001  2002  2003  2004 
#>   all 0.164 0.224 0.200 0.124 0.084 0.056 0.108 0.276 0.212 0.152 0.136 0.092
#>      year
#> len   2005  2006  2007  2008  2009  2010  2011  2012  2013  2014  2015  2016 
#>   all 0.120 0.148 0.140 0.104 0.152 0.176 0.152 0.224 0.220 0.136 0.224 0.260
#>      year
#> len   2017 
#>   all 0.320
#> 
#> units:  NA 
linf <- 60
k <- 2.29e-01
t0 <- -1.37e+00
bheqz(samps, linf = 60, k = 2.29e-01, t0 = -1.37e+00)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957  1958  1959  1960  1961  1962  1963  1964  1965  1966  1967  1968 
#>   all 0.211 0.231 0.220 0.290 0.270 0.280 0.218 0.239 0.263 0.273 0.226 0.223
#>      year
#> len   1969  1970  1971  1972  1973  1974  1975  1976  1977  1978  1979  1980 
#>   all 0.290 0.268 0.261 0.228 0.238 0.280 0.369 0.314 0.308 0.283 0.270 0.286
#>      year
#> len   1981  1982  1983  1984  1985  1986  1987  1988  1989  1990  1991  1992 
#>   all 0.316 0.307 0.411 0.343 0.269 0.386 0.338 0.341 0.288 0.311 0.325 0.242
#>      year
#> len   1993  1994  1995  1996  1997  1998  1999  2000  2001  2002  2003  2004 
#>   all 0.322 0.295 0.248 0.315 0.423 0.328 0.285 0.304 0.252 0.387 0.324 0.346
#>      year
#> len   2005  2006  2007  2008  2009  2010  2011  2012  2013  2014  2015  2016 
#>   all 0.340 0.313 0.282 0.379 0.290 0.270 0.324 0.226 0.382 0.303 0.276 0.260
#>      year
#> len   2017 
#>   all 0.265
#> 
#> units:  z