Calculate quantile(s) of length distribution

z = (k * (linf - lmean)) / (lmean - lc) lmean = sum(naa * len) / sum(naa) lc, length at first capture

indicators.len(
  object,
  indicators = "lbar",
  model = vonbert,
  params,
  cv = 0.1,
  lmax = 1.25,
  bin = 1,
  n = 500,
  metric = catch.n
)

lenquantile(x, quantile = 0.5)

lmax5(x)

l95(x)

l25(x)

lc50(x)

lmode(x)

lbar(x)

lmean(x)

lmaxy(x, lenwt)

pmega(x, linf, lopt = linf * 2/3)

bheqz(x, linf, k, t0, lc = lc50(x))

References

- Kell, L.T., Minto, C., Gerritsen, H.D. 2022. Evaluation of the skill of length-based indicators to identify stock status and trends. ICES Journal of Marine Science. doiu: 10.1093/icesjms/fsac043. - ICES. 2015. Report of the Fifth Workshop on the Development of Quantitative Assessment Methodologies based on Life-history Traits, Exploitation Characteristics and other Relevant Parameters for Data-limited Stocks (WKLIFE V), 5–9 October 2015, Lisbon, Portugal. ICES CM 2015/ACOM:56. 157 pp. - ICES. 2020. Tenth Workshop on the Development of Quantitative Assessment Methodologies based on LIFE-history traits, exploitation characteristics, and other relevant parameters for data-limited stocks (WKLIFE X). ICES Scientific Reports. 2:98. 72 pp. http://doi.org/10.17895/ices.pub.5985

Examples

data(ple4)
indicators.len(ple4, indicators=c('lbar', 'lmean'),
  params=FLPar(linf=132, k=0.080, t0=-0.35), metric='catch.n')
#> $lbar
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 34.902 34.080 30.872 30.800 30.798 32.412 34.130 31.888 30.566 33.432
#>      year
#> len   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 34.910 36.952 34.292 33.490 34.318 36.260 28.086 25.864 26.608 26.868
#>      year
#> len   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 25.540 26.624 26.658 26.880 27.298 23.954 24.698 25.450 24.812 20.846
#>      year
#> len   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 24.500 27.134 29.148 29.600 31.782 30.760 31.428 31.662 29.674 29.346
#>      year
#> len   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 25.118 26.008 29.802 30.010 30.124 26.996 27.504 26.572 25.896 26.722
#>      year
#> len   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all 25.758 26.348 26.602 27.142 28.414 29.778 29.846 28.772 31.206 32.162
#>      year
#> len   2017  
#>   all 31.758
#> 
#> units:  cm 
#> 
#> $lmean
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 44.552 37.858 33.290 32.836 34.078 38.772 38.454 34.288 37.934 39.891
#>      year
#> len   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 45.427 52.133 37.858 35.175 39.992 49.162 35.554 29.044 30.341 29.570
#>      year
#> len   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 29.247 31.018 31.942 29.440 30.698 26.867 33.977 29.827 30.118 26.388
#>      year
#> len   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 30.922 29.935 35.443 32.902 38.378 32.667 38.132 33.908 33.469 38.521
#>      year
#> len   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 28.815 29.122 40.833 32.828 36.078 32.030 33.676 31.352 27.683 29.850
#>      year
#> len   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all 28.206 28.425 29.749 30.164 31.100 35.556 32.678 32.406 39.183 35.411
#>      year
#> len   2017  
#>   all 34.826
#> 
#> units:  cm 
#> 
indicators.len(ple4, indicators=c('lbar', lmean),
  params=FLPar(linf=132, k=0.080, t0=-0.35), metric='catch.n')
#> $lbar
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 35.902 33.132 31.582 31.358 31.132 32.832 34.318 31.522 29.656 32.920
#>      year
#> len   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 34.824 37.134 35.708 32.556 34.234 35.574 29.468 26.492 27.612 27.646
#>      year
#> len   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 25.912 25.874 26.704 26.798 27.576 24.806 24.648 25.594 25.112 22.122
#>      year
#> len   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 23.956 27.806 28.960 29.720 29.802 31.212 31.468 31.312 30.200 28.416
#>      year
#> len   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 25.452 26.554 29.592 30.450 30.578 26.744 26.818 25.762 26.608 27.130
#>      year
#> len   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all 26.516 26.220 26.292 26.638 28.460 30.194 30.706 29.134 30.312 32.744
#>      year
#> len   2017  
#>   all 33.102
#> 
#> units:  cm 
#> 
#> $lmean
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 39.082 36.721 33.995 38.087 34.084 40.343 39.471 34.537 36.409 38.920
#>      year
#> len   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 48.225 38.481 38.578 35.126 41.294 47.471 37.085 30.670 30.105 31.559
#>      year
#> len   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 31.744 30.768 29.490 30.261 29.615 27.704 26.633 29.311 28.496 29.042
#>      year
#> len   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 30.196 39.043 34.767 31.583 31.775 38.972 33.668 33.258 32.988 35.648
#>      year
#> len   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 27.801 32.208 38.005 32.447 36.904 29.700 33.906 28.659 28.560 30.369
#>      year
#> len   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all 28.621 33.147 29.823 32.185 32.292 38.068 33.479 31.766 35.541 35.107
#>      year
#> len   2017  
#>   all 35.900
#> 
#> units:  cm 
#> 
data(ple4.index)
indicators.len(ple4.index, indicators=c('lbar', 'lmean'),
  params=FLPar(linf=132, k=0.080, t0=-0.35), metric='index')
#> $lbar
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1996   1997   1998   1999   2000   2001   2002   2003   2004   2005  
#>   all 21.390 17.034 21.444 21.996 21.044 21.730 17.072 21.760 22.040 21.720
#>      year
#> len   2006   2007   2008   2009   2010   2011   2012   2013   2014   2015  
#>   all 21.018 20.418 21.548 21.740 22.632 20.070 25.104 24.500 21.418 27.966
#>      year
#> len   2016   2017  
#>   all 25.630 22.344
#> 
#> units:  cm 
#> 
#> $lmean
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1996   1997   1998   1999   2000   2001   2002   2003   2004   2005  
#>   all 25.614 22.571 30.328 24.901 24.353 24.508 23.131 25.130 29.142 26.737
#>      year
#> len   2006   2007   2008   2009   2010   2011   2012   2013   2014   2015  
#>   all 27.781 27.196 24.480 25.422 30.585 26.377 27.898 29.283 24.603 31.232
#>      year
#> len   2016   2017  
#>   all 31.991 25.924
#> 
#> units:  cm 
#> 
#
ialk <- invALK(params=FLPar(linf = 60, k = 2.29e-01, t0 = -1.37e+00),
  model=vonbert, age=1:10, lmax=1.2)
samps <- lenSamples(catch.n(ple4), invALK=ialk, n=250)
lenquantile(samps, 0.50)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 40.0 37.0 35.0 35.0 36.0 37.5 38.0 38.0 35.0 38.0 41.0 40.0 40.0 36.5
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 37.0 38.5 31.0 32.0 34.0 35.0 32.0 32.5 33.0 33.0 34.0 32.0 33.0 35.0
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 33.0 28.0 33.0 36.0 35.0 36.0 37.0 36.0 37.0 37.0 36.0 34.0 32.0 34.0
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 37.0 38.0 36.0 32.0 33.0 34.5 34.0 34.0 34.0 33.0 35.0 36.0 34.0 36.0
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 37.0 35.0 37.0 38.0 40.0
#> 
#> units:  NA 
lmax5(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 58.462 60.133 56.231 58.154 54.125 55.286 56.750 55.412 54.786 55.929
#>      year
#> len   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 55.588 57.800 58.077 58.071 60.643 58.923 57.429 55.714 52.385 55.308
#>      year
#> len   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 54.462 54.385 56.000 52.000 52.500 51.688 53.235 51.000 47.615 49.615
#>      year
#> len   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 48.733 50.562 52.769 52.250 52.611 54.571 57.846 53.471 55.000 53.571
#>      year
#> len   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 51.769 50.529 51.071 52.067 54.250 53.643 55.429 50.625 51.786 51.600
#>      year
#> len   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all 52.353 51.231 52.385 54.000 53.154 52.846 54.267 54.067 56.000 54.688
#>      year
#> len   2017  
#>   all 57.062
#> 
#> units:  cm 
l95(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957  1958  1959  1960  1961  1962  1963  1964  1965  1966  1967  1968 
#>   all 54.55 57.00 52.55 54.55 52.00 52.00 54.00 52.00 53.00 53.00 53.00 55.00
#>      year
#> len   1969  1970  1971  1972  1973  1974  1975  1976  1977  1978  1979  1980 
#>   all 55.55 55.00 57.00 55.55 53.00 52.00 47.55 50.55 50.55 50.55 52.55 48.00
#>      year
#> len   1981  1982  1983  1984  1985  1986  1987  1988  1989  1990  1991  1992 
#>   all 49.00 48.00 49.00 47.55 45.55 45.10 46.00 47.00 48.55 50.00 50.00 51.00
#>      year
#> len   1993  1994  1995  1996  1997  1998  1999  2000  2001  2002  2003  2004 
#>   all 52.55 50.00 51.00 51.00 48.55 45.00 47.00 50.00 51.00 50.00 52.00 48.00
#>      year
#> len   2005  2006  2007  2008  2009  2010  2011  2012  2013  2014  2015  2016 
#>   all 49.00 48.00 48.00 47.55 49.55 50.00 50.10 50.55 50.00 52.00 52.00 53.00
#>      year
#> len   2017 
#>   all 53.00
#> 
#> units:  cm 
l25(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957  1958  1959  1960  1961  1962  1963  1964  1965  1966  1967  1968 
#>   all 33.25 29.00 28.00 30.00 30.00 31.00 32.00 28.00 31.00 35.00 36.00 34.00
#>      year
#> len   1969  1970  1971  1972  1973  1974  1975  1976  1977  1978  1979  1980 
#>   all 30.00 30.00 32.00 32.25 25.00 27.00 28.00 28.00 26.00 26.00 26.00 28.00
#>      year
#> len   1981  1982  1983  1984  1985  1986  1987  1988  1989  1990  1991  1992 
#>   all 29.25 26.00 27.00 29.00 27.00 25.00 29.00 29.25 29.00 31.00 30.25 30.00
#>      year
#> len   1993  1994  1995  1996  1997  1998  1999  2000  2001  2002  2003  2004 
#>   all 31.00 31.00 29.00 29.00 26.00 30.25 33.00 32.00 31.00 26.00 30.00 26.00
#>      year
#> len   2005  2006  2007  2008  2009  2010  2011  2012  2013  2014  2015  2016 
#>   all 29.00 29.00 26.00 27.25 28.00 28.25 28.00 31.00 31.00 29.00 30.00 32.00
#>      year
#> len   2017 
#>   all 31.00
#> 
#> units:  cm 
lc50(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 21.0 16.0 12.0 16.0 22.5 15.5 18.0 13.0 17.5 19.0 19.5 22.0 22.5 18.5
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 16.5 18.5 12.5 17.0 17.5 16.5 13.0 15.5 12.5 15.5 17.0 12.0 13.5 17.5
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 13.5 12.5 16.5 18.0 17.5 16.0 20.5 16.0 15.5 19.0 18.5 16.0 12.0 16.0
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 20.0 20.0 18.0 13.0 15.5 18.0 14.5 17.5 19.0 13.0 12.0 18.0 13.0 17.0
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 20.0 17.0 21.0 20.5 21.0
#> 
#> units:  cm 
lmode(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 42   32   24   32   45   31   36   26   35   38   39   44   45   37  
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 33   37   25   34   35   33   26   31   25   31   34   24   27   35  
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 27   25   33   36   35   32   41   32   31   38   37   32   24   32  
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 40   40   36   26   31   36   29   35   38   26   24   36   26   34  
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 40   34   42   41   42  
#> 
#> units:  cm 
lbar(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 39.600 37.848 36.068 36.904 36.520 37.808 38.336 36.956 36.480 39.016
#>      year
#> len   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 40.540 39.988 39.276 37.420 38.916 39.240 34.116 33.516 34.328 35.432
#>      year
#> len   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 33.752 33.920 34.092 33.936 34.856 32.876 33.820 34.564 33.116 30.280
#>      year
#> len   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 33.360 35.452 35.516 36.568 36.892 36.612 37.764 37.060 35.900 35.356
#>      year
#> len   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 32.908 34.536 36.888 37.312 36.672 33.732 34.788 33.712 34.496 34.636
#>      year
#> len   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all 34.244 33.980 34.836 35.804 34.648 36.168 36.784 35.852 36.780 37.832
#>      year
#> len   2017  
#>   all 38.708
#> 
#> units:  cm 
lmean(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 49.000 44.133 37.810 43.101 50.432 41.416 44.543 40.215 42.873 44.557
#>      year
#> len   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 45.858 50.405 50.783 46.613 43.600 45.949 37.747 42.682 41.792 41.819
#>      year
#> len   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 37.365 40.571 37.260 40.227 41.852 34.706 37.383 41.301 36.475 33.593
#>      year
#> len   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 39.327 42.108 42.008 41.067 47.437 41.671 41.739 44.779 44.582 40.466
#>      year
#> len   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 35.151 38.601 44.869 45.619 43.975 37.022 38.949 42.253 38.556 41.764
#>      year
#> len   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all 44.838 36.638 36.601 43.726 37.374 41.208 46.418 42.493 48.825 47.960
#>      year
#> len   2017  
#>   all 48.216
#> 
#> units:  cm 
# Linf(ple4) = 60
lmean(samps) / (0.75 * lc50(samps) + 0.25 * 60) #
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 1.5935 1.6346 1.5754 1.5963 1.5822 1.5555 1.5629 1.6248 1.5244 1.5233
#>      year
#> len   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 1.5480 1.6002 1.5932 1.6143 1.5927 1.5913 1.5486 1.5381 1.4860 1.5276
#>      year
#> len   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 1.5097 1.5238 1.5286 1.5109 1.5082 1.4461 1.4879 1.4685 1.4517 1.3782
#>      year
#> len   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 1.4366 1.4775 1.4936 1.5210 1.5617 1.5434 1.5677 1.5309 1.5440 1.4987
#>      year
#> len   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 1.4646 1.4297 1.4956 1.5206 1.5430 1.4958 1.4629 1.4825 1.4901 1.4849
#>      year
#> len   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all 1.5329 1.4803 1.5250 1.5342 1.5100 1.4850 1.5473 1.5313 1.5878 1.5789
#>      year
#> len   2017  
#>   all 1.5680
#> 
#> units:   
lenwt <- FLPar(a=0.01030, b=2.975)
lmaxy(samps, lenwt)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 42   51   47   40   45   43   41   45   35   38   44   44   48   52  
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 55   54   51   34   43   47   41   51   41   38   47   37   36   40  
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 36   37   33   38   43   42   41   46   49   38   37   35   37   32  
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 40   46   47   44   39   39   42   41   38   40   40   45   49   36  
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 40   40   42   41   47  
#> 
#> units:  cm 
pmega(samps, linf=60)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957  1958  1959  1960  1961  1962  1963  1964  1965  1966  1967  1968 
#>   all 0.312 0.308 0.240 0.228 0.228 0.256 0.260 0.292 0.172 0.200 0.364 0.356
#>      year
#> len   1969  1970  1971  1972  1973  1974  1975  1976  1977  1978  1979  1980 
#>   all 0.416 0.272 0.268 0.296 0.200 0.148 0.132 0.188 0.160 0.156 0.164 0.144
#>      year
#> len   1981  1982  1983  1984  1985  1986  1987  1988  1989  1990  1991  1992 
#>   all 0.156 0.108 0.128 0.120 0.088 0.064 0.092 0.136 0.140 0.216 0.232 0.224
#>      year
#> len   1993  1994  1995  1996  1997  1998  1999  2000  2001  2002  2003  2004 
#>   all 0.280 0.208 0.204 0.168 0.120 0.084 0.124 0.252 0.216 0.168 0.128 0.120
#>      year
#> len   2005  2006  2007  2008  2009  2010  2011  2012  2013  2014  2015  2016 
#>   all 0.136 0.124 0.144 0.128 0.144 0.192 0.168 0.180 0.216 0.200 0.212 0.232
#>      year
#> len   2017 
#>   all 0.336
#> 
#> units:  NA 
linf <- 60
k <- 2.29e-01
t0 <- -1.37e+00
bheqz(samps, linf = 60, k = 2.29e-01, t0 = -1.37e+00)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957    1958    1959    1960    1961    1962    1963    1964    1965   
#>   all 0.25116 0.23219 0.22771 0.25301 0.38352 0.22781 0.24395 0.22028 0.28378
#>      year
#> len   1966    1967    1968    1969    1970    1971    1972    1973    1974   
#>   all 0.24007 0.21180 0.25477 0.28289 0.27330 0.21539 0.22922 0.27422 0.36721
#>      year
#> len   1975    1976    1977    1978    1979    1980    1981    1982    1983   
#>   all 0.34935 0.29717 0.28965 0.32423 0.27477 0.32375 0.32247 0.29754 0.29504
#>      year
#> len   1984    1985    1986    1987    1988    1989    1990    1991    1992   
#>   all 0.34135 0.31385 0.38278 0.36184 0.32211 0.31121 0.26089 0.32282 0.25984
#>      year
#> len   1993    1994    1995    1996    1997    1998    1999    2000    2001   
#>   all 0.22871 0.29088 0.31718 0.29156 0.29673 0.31459 0.31340 0.30011 0.28610
#>      year
#> len   2002    2003    2004    2005    2006    2007    2008    2009    2010   
#>   all 0.29015 0.29933 0.38314 0.29208 0.33896 0.38691 0.28401 0.25235 0.31122
#>      year
#> len   2011    2012    2013    2014    2015    2016    2017   
#>   all 0.26818 0.28472 0.31676 0.29333 0.33697 0.29290 0.27535
#> 
#> units:  z