Calculate quantile(s) of length distribution

z = (k * (linf - lmean)) / (lmean - lc) lmean = sum(naa * len) / sum(naa) lc, length at first capture

Usage

indicators.len(
  object,
  indicators = "lbar",
  model = vonbert,
  params,
  cv = 0.1,
  lmax = 1.25,
  bin = 1,
  n = 500,
  metric = catch.n,
  ...
)

lenquantile(x, quantile = 0.5)

lmax5(x)

l95(x)

l25(x)

lc50(x)

lmode(x)

lbar(x)

lmean(x)

lmaxy(x, lenwt)

pmega(x, linf, lopt = linf * 2/3)

bheqz(x, linf, k, t0, lc = lc50(x))

References

Kell, L.T., Minto, C., Gerritsen, H.D. 2022. Evaluation of the skill of length-based indicators to identify stock status and trends. ICES Journal of Marine Science. doiu: 10.1093/icesjms/fsac043.

ICES. 2015. Report of the Fifth Workshop on the Development of Quantitative Assessment Methodologies based on Life-history Traits, Exploitation Characteristics and other Relevant Parameters for Data-limited Stocks (WKLIFE V), 5–9 October 2015, Lisbon, Portugal. ICES CM 2015/ACOM:56. 157 pp.

ICES. 2020. Tenth Workshop on the Development of Quantitative Assessment Methodologies based on LIFE-history traits, exploitation characteristics, and other relevant parameters for data-limited stocks (WKLIFE X). ICES Scientific Reports. 2:98. 72 pp. http://doi.org/10.17895/ices.pub.5985

Examples

data(ple4)
indicators.len(ple4, indicators=c('lbar', 'lmaxy'),
  params=FLPar(linf=132, k=0.080, t0=-0.35), metric='catch.n',
  lenwt=FLPar(a=0.01030, b=2.975))
#> $lbar
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 35.550 33.606 32.102 29.352 30.964 33.018 35.150 32.692 29.956 33.084
#>      year
#> len   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 36.796 37.600 35.584 32.714 35.242 33.678 28.788 26.812 26.800 27.938
#>      year
#> len   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 26.366 26.542 26.316 26.494 26.544 23.156 24.220 25.302 25.314 21.378
#>      year
#> len   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 25.054 27.338 29.332 30.662 29.980 30.346 31.560 31.676 29.336 28.334
#>      year
#> len   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 25.476 26.092 29.632 29.824 30.736 25.978 26.878 26.526 26.298 26.606
#>      year
#> len   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all 25.320 26.226 27.014 26.302 28.692 29.774 30.448 28.302 30.774 31.854
#>      year
#> len   2017  
#>   all 32.812
#> 
#> units:  cm 
#> 
#> $lmaxy
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 60   63   86   57   76   73   81   74   77   33   77   50   55   59  
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 72   68   83   56   74   76   42   53   54   68   75   76   39   34  
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 37   65   24   32   38   43   56   58   61   47   49   52   31   23  
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 31   41   48   42   65   34   73   46   43   51   51   35   47   52  
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 58   69   48   77   59  
#> 
#> units:  cm 
#> 
indicators.len(ple4, indicators=c('pmega'),
  params=FLPar(linf=60, k=2.29e-01, t0=-1.37), metric='catch.n')
#> $pmega
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957  1958  1959  1960  1961  1962  1963  1964  1965  1966  1967  1968 
#>   all 0.338 0.312 0.242 0.242 0.234 0.246 0.272 0.290 0.182 0.160 0.290 0.388
#>      year
#> len   1969  1970  1971  1972  1973  1974  1975  1976  1977  1978  1979  1980 
#>   all 0.394 0.332 0.274 0.318 0.188 0.118 0.118 0.220 0.146 0.136 0.178 0.124
#>      year
#> len   1981  1982  1983  1984  1985  1986  1987  1988  1989  1990  1991  1992 
#>   all 0.132 0.112 0.082 0.116 0.128 0.106 0.070 0.138 0.162 0.232 0.190 0.218
#>      year
#> len   1993  1994  1995  1996  1997  1998  1999  2000  2001  2002  2003  2004 
#>   all 0.236 0.216 0.238 0.160 0.088 0.088 0.106 0.204 0.246 0.156 0.122 0.130
#>      year
#> len   2005  2006  2007  2008  2009  2010  2011  2012  2013  2014  2015  2016 
#>   all 0.138 0.134 0.120 0.122 0.158 0.144 0.152 0.180 0.200 0.184 0.218 0.254
#>      year
#> len   2017 
#>   all 0.286
#> 
#> units:  NA 
#> 
data(ple4.index)
indicators.len(ple4.index, indicators=c('lbar', 'lmean'),
  params=FLPar(linf=132, k=0.080, t0=-0.35), metric='index')
#> $lbar
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1996   1997   1998   1999   2000   2001   2002   2003   2004   2005  
#>   all 21.474 16.932 22.240 21.960 21.178 22.754 16.140 22.016 21.782 21.138
#>      year
#> len   2006   2007   2008   2009   2010   2011   2012   2013   2014   2015  
#>   all 21.384 20.892 20.382 21.486 21.592 20.920 26.442 24.016 22.316 26.860
#>      year
#> len   2016   2017  
#>   all 25.614 21.344
#> 
#> units:  cm 
#> 
#> $lmean
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1996   1997   1998   1999   2000   2001   2002   2003   2004   2005  
#>   all 23.959 23.189 29.558 27.101 24.385 27.282 19.093 24.127 27.929 24.024
#>      year
#> len   2006   2007   2008   2009   2010   2011   2012   2013   2014   2015  
#>   all 24.731 28.636 26.214 27.919 28.730 27.146 30.636 27.339 26.050 29.328
#>      year
#> len   2016   2017  
#>   all 29.053 27.904
#> 
#> units:  cm 
#> 
#
ialk <- invALK(params=FLPar(linf = 60, k = 2.29e-01, t0 = -1.37e+00),
  model=vonbert, age=1:10, lmax=1.2)
samps <- lenSamples(catch.n(ple4), invALK=ialk, n=250)
lenquantile(samps, 0.50)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 39.0 40.0 36.0 35.0 35.0 38.0 38.0 35.0 35.0 38.0 40.0 40.0 41.0 36.0
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 38.0 39.0 32.0 33.0 35.0 35.0 33.0 33.0 33.0 35.0 34.0 31.0 33.0 34.0
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 33.5 28.0 32.0 36.0 35.0 36.0 35.0 35.0 36.0 37.0 36.0 35.0 33.0 34.0
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 37.0 38.0 36.0 35.0 34.0 34.0 32.0 33.0 32.0 34.0 34.0 32.0 33.0 36.5
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 36.0 37.0 37.0 37.0 38.0
#> 
#> units:  NA 
lmax5(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 59.923 59.929 57.357 58.154 53.000 56.714 56.071 56.714 54.692 50.471
#>      year
#> len   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 55.267 56.643 58.111 57.167 59.062 59.647 59.846 56.333 53.769 57.154
#>      year
#> len   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 54.059 54.467 53.533 55.643 55.692 53.786 54.692 47.842 54.000 50.929
#>      year
#> len   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 48.923 51.000 49.000 53.188 54.067 51.750 53.625 55.769 57.846 53.000
#>      year
#> len   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 55.214 49.308 50.667 53.929 53.308 55.769 53.692 52.000 51.286 50.357
#>      year
#> len   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all 49.429 52.214 52.000 52.692 51.235 55.429 55.846 55.333 54.857 55.571
#>      year
#> len   2017  
#>   all 56.600
#> 
#> units:  cm 
l95(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957  1958  1959  1960  1961  1962  1963  1964  1965  1966  1967  1968 
#>   all 56.55 57.00 55.00 55.55 49.55 52.00 52.00 52.00 51.55 48.00 52.00 55.00
#>      year
#> len   1969  1970  1971  1972  1973  1974  1975  1976  1977  1978  1979  1980 
#>   all 56.00 55.00 57.00 56.00 54.55 51.00 49.55 50.55 51.00 52.00 51.00 53.00
#>      year
#> len   1981  1982  1983  1984  1985  1986  1987  1988  1989  1990  1991  1992 
#>   all 50.55 48.00 49.10 45.00 49.55 49.00 46.10 48.00 47.00 50.00 51.00 49.00
#>      year
#> len   1993  1994  1995  1996  1997  1998  1999  2000  2001  2002  2003  2004 
#>   all 51.00 52.55 51.55 51.00 51.00 45.55 47.00 49.00 50.55 52.10 48.55 47.10
#>      year
#> len   2005  2006  2007  2008  2009  2010  2011  2012  2013  2014  2015  2016 
#>   all 49.00 47.00 47.00 49.00 48.00 47.55 48.00 53.00 50.55 52.00 53.00 53.00
#>      year
#> len   2017 
#>   all 53.00
#> 
#> units:  cm 
l25(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957  1958  1959  1960  1961  1962  1963  1964  1965  1966  1967  1968 
#>   all 33.00 31.00 29.00 30.00 28.00 33.00 34.00 27.00 31.00 34.00 34.00 32.00
#>      year
#> len   1969  1970  1971  1972  1973  1974  1975  1976  1977  1978  1979  1980 
#>   all 32.00 28.25 32.00 32.00 26.00 27.00 29.00 27.25 26.00 26.00 26.25 28.00
#>      year
#> len   1981  1982  1983  1984  1985  1986  1987  1988  1989  1990  1991  1992 
#>   all 28.00 26.00 29.00 27.00 26.00 25.00 28.00 31.00 30.00 30.00 29.00 30.00
#>      year
#> len   1993  1994  1995  1996  1997  1998  1999  2000  2001  2002  2003  2004 
#>   all 30.00 30.00 29.00 30.00 28.00 30.00 32.00 31.00 31.00 28.00 30.00 27.00
#>      year
#> len   2005  2006  2007  2008  2009  2010  2011  2012  2013  2014  2015  2016 
#>   all 27.25 27.00 27.00 28.25 28.00 27.00 28.00 32.00 30.00 29.00 30.00 29.00
#>      year
#> len   2017 
#>   all 31.00
#> 
#> units:  cm 
lc50(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 22.0 21.5 18.5 16.0 18.0 20.5 18.0 12.5 15.5 18.5 20.5 21.5 22.5 13.5
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 19.0 21.0 13.0 13.0 18.5 13.5 13.5 12.0 13.0 17.0 16.5 12.0 17.0 14.0
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 13.0 12.5 16.0 18.0 19.0 18.5 21.0 17.0 16.0 20.0 16.5 17.5 13.0 16.0
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 18.5 20.5 17.5 19.0 16.0 17.5 17.0 12.5 13.0 17.5 17.0 13.5 15.5 16.5
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 18.0 13.5 18.5 19.0 17.0
#> 
#> units:  cm 
lmode(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 44   43   37   32   36   41   36   25   31   37   41   43   45   27  
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 38   42   26   26   37   27   27   24   26   34   33   24   34   28  
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 26   25   32   36   38   37   42   34   32   40   33   35   26   32  
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 37   41   35   38   32   35   34   25   26   35   34   27   31   33  
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 36   27   37   38   34  
#> 
#> units:  cm 
lbar(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 39.332 39.572 37.012 37.044 35.240 38.536 39.280 35.888 36.504 37.748
#>      year
#> len   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 39.492 39.780 40.476 37.484 39.324 39.844 34.932 34.444 34.644 35.436
#>      year
#> len   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 34.336 34.552 34.228 35.368 34.676 32.860 33.776 33.644 33.924 30.616
#>      year
#> len   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 33.252 35.972 35.224 35.972 36.064 35.716 36.320 36.912 36.608 35.740
#>      year
#> len   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 34.508 34.296 36.112 36.856 36.520 34.904 35.208 34.020 33.548 33.484
#>      year
#> len   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all 33.452 34.888 34.720 33.800 34.300 37.580 36.488 36.988 37.116 36.736
#>      year
#> len   2017  
#>   all 38.076
#> 
#> units:  cm 
lmean(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 51.060 51.023 46.538 43.079 43.577 47.686 44.551 38.249 39.960 42.477
#>      year
#> len   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 46.622 49.931 51.652 41.600 49.000 50.208 39.374 37.350 44.096 38.979
#>      year
#> len   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 39.530 36.729 37.695 42.437 40.852 34.767 41.304 37.796 37.341 33.994
#>      year
#> len   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 38.893 42.396 43.080 43.910 47.762 42.209 41.312 46.659 42.824 42.771
#>      year
#> len   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 37.137 38.626 42.422 46.122 42.478 45.013 39.389 41.476 41.862 35.900
#>      year
#> len   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all 35.795 42.430 41.711 37.657 39.229 41.880 43.508 40.091 45.188 45.580
#>      year
#> len   2017  
#>   all 44.797
#> 
#> units:  cm 
# Linf(ple4) = 60
lmean(samps) / (0.75 * lc50(samps) + 0.25 * 60) #
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 1.6209 1.6393 1.6117 1.5955 1.5290 1.5699 1.5632 1.5692 1.5009 1.4710
#>      year
#> len   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 1.5349 1.6042 1.6205 1.6557 1.6752 1.6328 1.5909 1.5091 1.5271 1.5514
#>      year
#> len   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 1.5734 1.5304 1.5230 1.5292 1.4923 1.4486 1.4884 1.4822 1.5087 1.3946
#>      year
#> len   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 1.4405 1.4876 1.4728 1.5207 1.5532 1.5211 1.5301 1.5553 1.5644 1.5208
#>      year
#> len   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 1.5005 1.4306 1.4692 1.5184 1.5103 1.5389 1.4589 1.4747 1.5085 1.4728
#>      year
#> len   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all 1.4463 1.5086 1.5031 1.4988 1.4734 1.5299 1.5266 1.5957 1.5649 1.5583
#>      year
#> len   2017  
#>   all 1.6143
#> 
#> units:   
lenwt <- FLPar(a=0.01030, b=2.975)
lmaxy(samps, lenwt)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 44   56   56   53   44   48   47   46   47   39   41   51   56   49  
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 53   56   46   46   37   43   51   50   55   44   41   44   34   43  
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 38   50   37   41   43   37   42   48   39   40   45   51   44   32  
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 41   41   43   40   37   35   45   40   39   46   44   44   48   45  
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 44   48   44   38   46  
#> 
#> units:  cm 
pmega(samps, linf=60)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957  1958  1959  1960  1961  1962  1963  1964  1965  1966  1967  1968 
#>   all 0.320 0.352 0.264 0.232 0.196 0.284 0.320 0.228 0.196 0.156 0.288 0.348
#>      year
#> len   1969  1970  1971  1972  1973  1974  1975  1976  1977  1978  1979  1980 
#>   all 0.436 0.320 0.328 0.356 0.232 0.168 0.128 0.172 0.196 0.204 0.176 0.212
#>      year
#> len   1981  1982  1983  1984  1985  1986  1987  1988  1989  1990  1991  1992 
#>   all 0.120 0.124 0.100 0.080 0.148 0.100 0.084 0.152 0.128 0.176 0.220 0.192
#>      year
#> len   1993  1994  1995  1996  1997  1998  1999  2000  2001  2002  2003  2004 
#>   all 0.188 0.212 0.240 0.180 0.148 0.084 0.136 0.192 0.204 0.140 0.144 0.096
#>      year
#> len   2005  2006  2007  2008  2009  2010  2011  2012  2013  2014  2015  2016 
#>   all 0.128 0.128 0.120 0.164 0.156 0.148 0.128 0.236 0.212 0.248 0.260 0.220
#>      year
#> len   2017 
#>   all 0.288
#> 
#> units:  NA 
linf <- 60
k <- 2.29e-01
t0 <- -1.37e+00
bheqz(samps, linf = 60, k = 2.29e-01, t0 = -1.37e+00)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957    1958    1959    1960    1961    1962    1963    1964    1965   
#>   all 0.27308 0.25885 0.28437 0.24981 0.32889 0.27252 0.22297 0.23609 0.25617
#>      year
#> len   1966    1967    1968    1969    1970    1971    1972    1973    1974   
#>   all 0.26474 0.24728 0.25330 0.24872 0.21498 0.23297 0.24494 0.26174 0.27291
#>      year
#> len   1975    1976    1977    1978    1979    1980    1981    1982    1983   
#>   all 0.35967 0.25643 0.28206 0.25841 0.27802 0.30710 0.31906 0.29794 0.35797
#>      year
#> len   1984    1985    1986    1987    1988    1989    1990    1991    1992   
#>   all 0.30725 0.28539 0.37144 0.35505 0.30617 0.34971 0.31493 0.36387 0.29713
#>      year
#> len   1993    1994    1995    1996    1997    1998    1999    2000    2001   
#>   all 0.26687 0.31263 0.26640 0.30458 0.27142 0.32172 0.31060 0.32404 0.28270
#>      year
#> len   2002    2003    2004    2005    2006    2007    2008    2009    2010   
#>   all 0.36135 0.29557 0.36013 0.36606 0.28937 0.29726 0.33073 0.32670 0.29556
#>      year
#> len   2011    2012    2013    2014    2015    2016    2017   
#>   all 0.31305 0.24356 0.29123 0.22436 0.28150 0.30038 0.23821
#> 
#> units:  z