Calculate quantile(s) of length distribution

z = (k * (linf - lmean)) / (lmean - lc) lmean = sum(naa * len) / sum(naa) lc, length at first capture

indicators.len(
  object,
  indicators = "lbar",
  model = vonbert,
  params,
  cv = 0.1,
  lmax = 1.25,
  bin = 1,
  n = 500,
  metric = catch.n
)

lenquantile(x, quantile = 0.5)

lmax5(x)

l95(x)

l25(x)

lc50(x)

lmode(x)

lbar(x)

lmean(x)

lmaxy(x, lenwt)

pmega(x, linf, lopt = linf * 2/3)

bheqz(x, linf, k, t0, lc = lc50(x))

References

- Kell, L.T., Minto, C., Gerritsen, H.D. 2022. Evaluation of the skill of length-based indicators to identify stock status and trends. ICES Journal of Marine Science. doiu: 10.1093/icesjms/fsac043. - ICES. 2015. Report of the Fifth Workshop on the Development of Quantitative Assessment Methodologies based on Life-history Traits, Exploitation Characteristics and other Relevant Parameters for Data-limited Stocks (WKLIFE V), 5–9 October 2015, Lisbon, Portugal. ICES CM 2015/ACOM:56. 157 pp. - ICES. 2020. Tenth Workshop on the Development of Quantitative Assessment Methodologies based on LIFE-history traits, exploitation characteristics, and other relevant parameters for data-limited stocks (WKLIFE X). ICES Scientific Reports. 2:98. 72 pp. http://doi.org/10.17895/ices.pub.5985

Examples

data(ple4)
indicators.len(ple4, indicators=c('lbar', 'lmean'),
  params=FLPar(linf=132, k=0.080, t0=-0.35), metric='catch.n')
#> $lbar
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 35.646 33.806 32.304 31.196 29.202 31.482 34.956 31.138 30.694 32.442
#>      year
#> len   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 35.604 36.304 36.404 33.712 35.656 35.452 28.090 26.098 27.784 29.174
#>      year
#> len   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 25.398 26.950 27.526 27.230 26.632 23.430 24.722 26.462 25.940 21.728
#>      year
#> len   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 23.980 26.650 29.084 29.490 30.302 29.792 30.344 31.762 29.966 28.116
#>      year
#> len   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 24.476 27.040 29.164 30.690 30.696 26.340 26.388 26.002 26.914 27.484
#>      year
#> len   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all 26.706 25.362 26.294 26.264 28.258 30.068 30.006 28.954 30.752 32.482
#>      year
#> len   2017  
#>   all 33.618
#> 
#> units:  cm 
#> 
#> $lmean
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 48.401 37.964 41.788 33.314 32.860 38.151 39.318 34.137 38.779 44.169
#>      year
#> len   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 49.545 38.424 38.581 36.170 44.590 55.053 35.564 28.518 30.865 32.342
#>      year
#> len   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 31.890 29.761 31.505 31.529 33.611 26.786 27.618 28.552 30.491 25.558
#>      year
#> len   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 31.495 41.150 35.910 31.092 32.291 37.445 34.570 34.594 33.493 33.410
#>      year
#> len   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 27.095 34.174 38.000 33.786 37.143 30.545 32.988 28.712 35.969 30.431
#>      year
#> len   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all 29.312 29.044 30.337 28.778 30.361 37.403 32.825 36.289 37.929 45.354
#>      year
#> len   2017  
#>   all 36.604
#> 
#> units:  cm 
#> 
indicators.len(ple4, indicators=c('lbar', lmean),
  params=FLPar(linf=132, k=0.080, t0=-0.35), metric='catch.n')
#> $lbar
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 35.510 33.110 31.188 29.944 29.950 34.310 34.532 32.202 30.766 33.448
#>      year
#> len   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 36.164 36.120 34.054 34.066 33.268 34.560 29.494 26.324 27.302 27.458
#>      year
#> len   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 26.484 28.598 26.510 26.962 27.282 24.474 24.336 26.566 25.386 22.058
#>      year
#> len   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 24.460 27.298 29.212 30.658 31.326 29.754 31.350 31.198 28.998 29.440
#>      year
#> len   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 25.186 26.064 29.238 29.360 30.766 27.274 27.580 27.428 26.502 27.544
#>      year
#> len   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all 25.724 26.720 26.338 26.648 27.778 30.638 30.606 30.556 31.884 31.134
#>      year
#> len   2017  
#>   all 32.442
#> 
#> units:  cm 
#> 
#> $lmean
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 46.981 36.931 35.442 33.012 33.055 38.716 48.092 36.215 36.900 40.821
#>      year
#> len   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 43.038 55.581 37.087 36.856 42.293 36.322 37.395 28.917 29.251 30.675
#>      year
#> len   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 30.501 31.207 31.080 29.894 30.276 27.559 27.786 29.890 28.217 25.354
#>      year
#> len   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 30.532 29.863 34.398 33.188 32.965 32.296 37.195 34.536 33.106 36.384
#>      year
#> len   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 27.656 31.801 39.463 32.589 37.088 30.230 33.850 30.000 35.621 29.290
#>      year
#> len   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all 29.375 35.132 29.689 29.517 30.074 37.289 32.636 34.191 38.789 33.269
#>      year
#> len   2017  
#>   all 34.710
#> 
#> units:  cm 
#> 
data(ple4.index)
indicators.len(ple4.index, indicators=c('lbar', 'lmean'),
  params=FLPar(linf=132, k=0.080, t0=-0.35), metric='index')
#> $lbar
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1996   1997   1998   1999   2000   2001   2002   2003   2004   2005  
#>   all 21.344 16.972 21.390 21.796 21.006 22.068 16.382 21.220 21.854 21.398
#>      year
#> len   2006   2007   2008   2009   2010   2011   2012   2013   2014   2015  
#>   all 21.688 19.942 21.918 23.464 20.852 19.908 25.556 24.030 22.298 26.672
#>      year
#> len   2016   2017  
#>   all 25.792 22.342
#> 
#> units:  cm 
#> 
#> $lmean
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1996   1997   1998   1999   2000   2001   2002   2003   2004   2005  
#>   all 25.427 19.812 28.364 26.348 23.970 27.387 19.053 24.497 25.648 24.337
#>      year
#> len   2006   2007   2008   2009   2010   2011   2012   2013   2014   2015  
#>   all 27.650 26.527 26.732 26.764 24.556 26.008 27.939 27.674 26.173 30.100
#>      year
#> len   2016   2017  
#>   all 32.514 26.247
#> 
#> units:  cm 
#> 
#
ialk <- invALK(params=FLPar(linf = 60, k = 2.29e-01, t0 = -1.37e+00),
  model=vonbert, age=1:10, lmax=1.2)
samps <- lenSamples(catch.n(ple4), invALK=ialk, n=250)
lenquantile(samps, 0.50)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 39.0 38.0 35.0 35.0 36.0 38.0 40.0 37.0 34.0 38.0 40.0 41.0 38.0 35.0
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 36.0 39.0 33.0 32.0 34.0 35.0 31.0 32.0 34.0 33.0 33.0 32.0 31.5 33.0
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 31.5 28.0 32.0 35.0 36.0 34.0 36.0 36.0 37.0 36.0 35.0 36.0 32.0 33.0
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 37.0 37.0 37.0 33.0 34.0 34.0 33.0 35.0 33.0 32.0 34.5 33.0 34.0 35.0
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 36.5 35.0 37.0 37.0 37.0
#> 
#> units:  NA 
lmax5(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 57.071 56.625 58.133 57.846 58.267 57.800 58.562 56.857 56.941 54.333
#>      year
#> len   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 57.077 59.462 57.833 58.538 59.214 58.562 59.308 55.125 53.786 56.333
#>      year
#> len   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 53.333 51.667 57.500 54.429 53.400 52.267 52.118 51.769 52.462 50.600
#>      year
#> len   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 50.231 51.154 51.500 52.733 54.231 56.000 55.714 55.733 55.143 52.714
#>      year
#> len   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 49.231 49.357 52.143 51.188 54.765 51.538 53.333 49.500 51.308 51.467
#>      year
#> len   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all 52.769 52.077 51.385 54.615 53.769 52.294 52.846 53.294 56.308 52.312
#>      year
#> len   2017  
#>   all 56.643
#> 
#> units:  cm 
l95(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957  1958  1959  1960  1961  1962  1963  1964  1965  1966  1967  1968 
#>   all 54.00 53.00 55.00 54.55 55.00 55.00 55.00 54.00 53.00 51.00 53.55 54.55
#>      year
#> len   1969  1970  1971  1972  1973  1974  1975  1976  1977  1978  1979  1980 
#>   all 56.00 54.55 58.00 56.00 56.55 51.00 50.00 52.00 51.00 49.00 53.00 51.00
#>      year
#> len   1981  1982  1983  1984  1985  1986  1987  1988  1989  1990  1991  1992 
#>   all 49.00 49.00 46.00 47.10 48.55 46.00 45.55 46.55 49.00 50.00 51.55 52.55
#>      year
#> len   1993  1994  1995  1996  1997  1998  1999  2000  2001  2002  2003  2004 
#>   all 52.00 51.00 53.00 51.00 45.55 47.00 48.00 48.00 52.00 48.55 50.00 46.00
#>      year
#> len   2005  2006  2007  2008  2009  2010  2011  2012  2013  2014  2015  2016 
#>   all 48.55 49.00 47.55 48.55 47.55 49.55 50.55 49.00 50.55 50.00 50.55 50.00
#>      year
#> len   2017 
#>   all 54.00
#> 
#> units:  cm 
l25(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957  1958  1959  1960  1961  1962  1963  1964  1965  1966  1967  1968 
#>   all 34.00 31.25 28.00 30.00 30.00 32.00 34.00 28.00 31.00 34.00 36.00 35.00
#>      year
#> len   1969  1970  1971  1972  1973  1974  1975  1976  1977  1978  1979  1980 
#>   all 29.00 29.00 30.00 33.00 26.00 28.00 29.00 27.00 26.00 26.00 27.00 26.00
#>      year
#> len   1981  1982  1983  1984  1985  1986  1987  1988  1989  1990  1991  1992 
#>   all 27.00 25.00 27.00 27.00 26.00 25.00 28.00 29.00 30.00 29.00 30.00 30.00
#>      year
#> len   1993  1994  1995  1996  1997  1998  1999  2000  2001  2002  2003  2004 
#>   all 31.00 30.25 30.00 30.00 27.00 31.00 32.00 31.00 32.00 26.00 31.00 27.00
#>      year
#> len   2005  2006  2007  2008  2009  2010  2011  2012  2013  2014  2015  2016 
#>   all 27.00 30.00 28.00 27.00 27.00 27.00 27.25 30.25 31.25 28.00 31.00 29.00
#>      year
#> len   2017 
#>   all 30.00
#> 
#> units:  cm 
lc50(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 19.5 16.0 13.5 14.0 17.0 20.5 20.0 12.5 17.0 19.5 20.0 18.5 12.5 16.0
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 18.0 20.0 12.5 17.0 16.0 13.5 13.0 12.5 14.0 12.5 16.0 12.5 14.0 12.5
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 12.5 12.5 13.0 17.5 17.5 16.5 21.0 13.0 17.5 19.5 16.5 18.5 16.0 16.0
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 18.5 18.5 16.5 12.5 16.0 12.5 13.0 15.0 16.5 13.5 12.0 13.5 12.5 15.5
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 18.5 16.5 19.0 17.5 18.0
#> 
#> units:  cm 
lmode(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 39   32   27   28   34   41   40   25   34   39   40   37   25   32  
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 36   40   25   34   32   27   26   25   28   25   32   25   28   25  
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 25   25   26   35   35   33   42   26   35   39   33   37   32   32  
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 37   37   33   25   32   25   26   30   33   27   24   27   25   31  
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 37   33   38   35   36  
#> 
#> units:  cm 
lbar(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 39.132 38.616 36.600 37.044 37.396 38.752 39.908 37.160 36.600 38.340
#>      year
#> len   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 40.816 40.360 38.592 37.296 37.652 39.356 35.556 33.728 34.356 35.276
#>      year
#> len   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 33.396 33.484 35.016 33.912 34.304 32.720 32.648 33.440 32.880 30.364
#>      year
#> len   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 33.172 34.848 35.568 35.560 36.536 36.844 37.368 36.708 36.364 36.016
#>      year
#> len   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 32.984 34.520 36.640 36.796 37.752 33.544 35.516 33.616 33.768 35.488
#>      year
#> len   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all 34.288 33.488 34.436 33.912 34.748 36.112 36.556 35.408 37.072 36.676
#>      year
#> len   2017  
#>   all 37.116
#> 
#> units:  cm 
lmean(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 46.470 43.190 39.563 40.005 43.706 48.144 47.773 39.819 43.273 45.860
#>      year
#> len   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 46.694 46.119 41.292 44.302 46.203 48.084 39.036 43.262 39.971 39.141
#>      year
#> len   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 37.563 36.894 39.892 36.623 40.682 35.941 36.730 35.875 35.766 33.880
#>      year
#> len   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 35.141 41.074 41.805 41.881 48.466 38.823 43.310 45.607 43.014 44.424
#>      year
#> len   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 38.992 38.399 43.019 43.403 42.256 36.096 39.408 36.176 36.303 39.131
#>      year
#> len   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all 41.213 36.584 35.950 38.091 37.490 39.912 43.785 41.614 45.710 43.390
#>      year
#> len   2017  
#>   all 44.244
#> 
#> units:  cm 
# Linf(ple4) = 60
lmean(samps) / (0.75 * lc50(samps) + 0.25 * 60) #
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 1.5686 1.5996 1.5746 1.5688 1.5750 1.5850 1.5924 1.6336 1.5594 1.5480
#>      year
#> len   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 1.5565 1.5972 1.6940 1.6408 1.6212 1.6028 1.6015 1.5590 1.4804 1.5578
#>      year
#> len   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 1.5177 1.5136 1.5644 1.5025 1.5067 1.4745 1.4404 1.4718 1.4673 1.3900
#>      year
#> len   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 1.4199 1.4604 1.4864 1.5299 1.5761 1.5686 1.5399 1.5395 1.5713 1.5385
#>      year
#> len   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 1.4441 1.4222 1.4898 1.5031 1.5436 1.4809 1.4595 1.4841 1.4668 1.4907
#>      year
#> len   2007   2008   2009   2010   2011   2012   2013   2014   2015   2016  
#>   all 1.5055 1.4561 1.4979 1.5161 1.5381 1.4990 1.5164 1.5202 1.5627 1.5427
#>      year
#> len   2017  
#>   all 1.5524
#> 
#> units:   
lenwt <- FLPar(a=0.01030, b=2.975)
lmaxy(samps, lenwt)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 39   51   39   45   36   41   45   44   34   39   43   46   56   50  
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 52   46   52   34   35   44   43   49   49   44   44   41   34   40  
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 41   46   34   40   43   46   42   47   42   39   48   37   35   37  
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 37   44   50   44   36   38   39   41   44   37   46   41   47   45  
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 40   41   38   48   46  
#> 
#> units:  cm 
pmega(samps, linf=60)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957  1958  1959  1960  1961  1962  1963  1964  1965  1966  1967  1968 
#>   all 0.284 0.308 0.220 0.240 0.248 0.264 0.328 0.308 0.200 0.204 0.336 0.376
#>      year
#> len   1969  1970  1971  1972  1973  1974  1975  1976  1977  1978  1979  1980 
#>   all 0.368 0.324 0.260 0.320 0.236 0.136 0.132 0.196 0.164 0.164 0.192 0.160
#>      year
#> len   1981  1982  1983  1984  1985  1986  1987  1988  1989  1990  1991  1992 
#>   all 0.168 0.108 0.080 0.104 0.124 0.084 0.072 0.124 0.160 0.196 0.200 0.228
#>      year
#> len   1993  1994  1995  1996  1997  1998  1999  2000  2001  2002  2003  2004 
#>   all 0.228 0.208 0.252 0.196 0.088 0.080 0.160 0.216 0.244 0.140 0.132 0.088
#>      year
#> len   2005  2006  2007  2008  2009  2010  2011  2012  2013  2014  2015  2016 
#>   all 0.136 0.148 0.168 0.108 0.132 0.156 0.192 0.188 0.184 0.188 0.212 0.248
#>      year
#> len   2017 
#>   all 0.244
#> 
#> units:  NA 
linf <- 60
k <- 2.29e-01
t0 <- -1.37e+00
bheqz(samps, linf = 60, k = 2.29e-01, t0 = -1.37e+00)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957    1958    1959    1960    1961    1962    1963    1964    1965   
#>   all 0.24342 0.21653 0.23197 0.22813 0.25379 0.26659 0.23112 0.21210 0.27340
#>      year
#> len   1966    1967    1968    1969    1970    1971    1972    1973    1974   
#>   all 0.26328 0.21105 0.20574 0.18789 0.24414 0.26042 0.24424 0.24279 0.35965
#>      year
#> len   1975    1976    1977    1978    1979    1980    1981    1982    1983   
#>   all 0.31992 0.26000 0.29870 0.28937 0.27224 0.27901 0.32148 0.30896 0.33589
#>      year
#> len   1984    1985    1986    1987    1988    1989    1990    1991    1992   
#>   all 0.29046 0.30473 0.37991 0.30456 0.33202 0.30966 0.29364 0.34586 0.22239
#>      year
#> len   1993    1994    1995    1996    1997    1998    1999    2000    2001   
#>   all 0.26086 0.30996 0.27249 0.31356 0.36426 0.31506 0.29490 0.29043 0.23973
#>      year
#> len   2002    2003    2004    2005    2006    2007    2008    2009    2010   
#>   all 0.28789 0.28729 0.28613 0.28925 0.27398 0.33101 0.30374 0.26093 0.29268
#>      year
#> len   2011    2012    2013    2014    2015    2016    2017   
#>   all 0.25992 0.26540 0.29733 0.29784 0.29053 0.27854 0.27414
#> 
#> units:  z