Calculate quantile(s) of length distribution

z = (k * (linf - lmean)) / (lmean - lc) lmean = sum(naa * len) / sum(naa) lc, length at first capture

Usage

indicators.len(
  object,
  indicators = "lbar",
  model = vonbert,
  params,
  cv = 0.1,
  lmax = 1.25,
  bin = 1,
  n = 500,
  metric = catch.n,
  ...
)

lenquantile(x, quantile = 0.5)

lmax5(x)

l95(x)

l25(x)

lc50(x)

lmode(x)

lbar(x)

lmean(x)

lmaxy(x, lenwt)

pmega(x, linf, lopt = linf * 2/3)

bheqz(x, linf, k, t0, lc = lc50(x))

References

Kell, L.T., Minto, C., Gerritsen, H.D. 2022. Evaluation of the skill of length-based indicators to identify stock status and trends. ICES Journal of Marine Science. doiu: 10.1093/icesjms/fsac043.

ICES. 2015. Report of the Fifth Workshop on the Development of Quantitative Assessment Methodologies based on Life-history Traits, Exploitation Characteristics and other Relevant Parameters for Data-limited Stocks (WKLIFE V), 5–9 October 2015, Lisbon, Portugal. ICES CM 2015/ACOM:56. 157 pp.

ICES. 2020. Tenth Workshop on the Development of Quantitative Assessment Methodologies based on LIFE-history traits, exploitation characteristics, and other relevant parameters for data-limited stocks (WKLIFE X). ICES Scientific Reports. 2:98. 72 pp. http://doi.org/10.17895/ices.pub.5985

Examples

data(ple4)
indicators.len(ple4, indicators=c('lbar', 'lmaxy'),
  params=FLPar(linf=132, k=0.080, t0=-0.35), metric='catch.n',
  lenwt=FLPar(a=0.01030, b=2.975))
#> $lbar
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 35.6 33.6 32.0 29.4 31.0 32.8 35.3 32.7 29.9 33.2 36.7 37.6 35.7 32.7
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 35.2 33.9 28.9 26.7 26.7 27.8 26.4 26.6 26.2 26.5 26.6 23.2 24.1 25.3
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 25.2 21.4 25.0 27.3 29.4 30.7 29.9 30.4 31.6 31.6 29.3 28.3 25.5 26.0
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 29.6 29.8 30.8 25.9 26.9 26.5 26.3 26.6 25.3 26.2 27.0 26.3 28.8 29.7
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 30.5 28.3 30.8 31.9 32.8
#> 
#> units:  cm 
#> 
#> $lmaxy
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 76   75   86   57   76   73   81   74   77   33   77   50   55   59  
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 72   68   83   56   74   76   42   53   54   68   75   76   39   34  
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 35   65   24   32   38   43   56   58   61   47   49   52   31   23  
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 31   41   48   42   65   34   73   46   43   51   51   35   47   52  
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 58   69   48   77   59  
#> 
#> units:  cm 
#> 
indicators.len(ple4, indicators=c('pmega'),
  params=FLPar(linf=60, k=2.29e-01, t0=-1.37), metric='catch.n')
#> $pmega
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957  1958  1959  1960  1961  1962  1963  1964  1965  1966  1967  1968 
#>   all 0.340 0.312 0.244 0.242 0.232 0.248 0.272 0.288 0.182 0.160 0.292 0.388
#>      year
#> len   1969  1970  1971  1972  1973  1974  1975  1976  1977  1978  1979  1980 
#>   all 0.396 0.334 0.274 0.316 0.188 0.120 0.120 0.220 0.146 0.136 0.176 0.124
#>      year
#> len   1981  1982  1983  1984  1985  1986  1987  1988  1989  1990  1991  1992 
#>   all 0.132 0.112 0.082 0.114 0.128 0.106 0.070 0.140 0.162 0.230 0.194 0.218
#>      year
#> len   1993  1994  1995  1996  1997  1998  1999  2000  2001  2002  2003  2004 
#>   all 0.232 0.218 0.236 0.160 0.088 0.088 0.106 0.206 0.248 0.156 0.122 0.132
#>      year
#> len   2005  2006  2007  2008  2009  2010  2011  2012  2013  2014  2015  2016 
#>   all 0.144 0.132 0.116 0.124 0.158 0.142 0.152 0.180 0.200 0.182 0.222 0.258
#>      year
#> len   2017 
#>   all 0.282
#> 
#> units:  NA 
#> 
data(ple4.index)
indicators.len(ple4.index, indicators=c('lbar', 'lmean'),
  params=FLPar(linf=132, k=0.080, t0=-0.35), metric='index')
#> $lbar
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009
#>   all 21.4 16.9 22.3 21.9 21.2 22.8 16.1 22.0 21.7 21.3 21.4 20.7 20.5 21.4
#>      year
#> len   2010 2011 2012 2013 2014 2015 2016 2017
#>   all 21.5 20.9 26.5 24.0 22.3 26.8 25.6 21.5
#> 
#> units:  cm 
#> 
#> $lmean
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009
#>   all 23.8 23.1 29.7 27.0 24.5 27.3 19.0 24.1 27.9 24.1 24.8 28.4 26.3 27.8
#>      year
#> len   2010 2011 2012 2013 2014 2015 2016 2017
#>   all 28.5 27.1 30.6 27.4 26.1 29.3 29.1 28.0
#> 
#> units:  cm 
#> 
#
ialk <- invALK(params=FLPar(linf = 60, k = 2.29e-01, t0 = -1.37e+00),
  model=vonbert, age=1:10, lmax=1.2)
samps <- lenSamples(catch.n(ple4), invALK=ialk, n=250)
lenquantile(samps, 0.50)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 39.0 39.5 36.0 35.0 35.0 38.5 38.0 35.0 35.0 38.0 40.0 40.0 41.0 36.0
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 38.0 39.0 32.0 33.0 35.0 35.0 33.0 33.0 33.0 35.0 34.0 31.0 33.0 34.0
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 33.0 28.0 32.0 36.0 35.0 36.0 35.0 35.0 36.0 37.0 36.0 35.0 33.0 34.0
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 37.0 38.0 36.0 35.0 34.0 34.5 32.0 33.0 32.5 34.0 34.0 32.0 33.0 36.5
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 36.0 37.0 37.0 37.0 38.0
#> 
#> units:  NA 
lmax5(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 59.9 59.9 57.4 58.2 53.0 56.7 56.1 56.8 54.7 50.5 55.3 56.6 58.1 58.1
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 59.1 59.6 59.8 56.3 53.8 57.2 54.1 54.5 53.5 55.6 55.7 53.8 54.7 47.8
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 54.0 50.9 48.9 51.0 49.0 53.2 54.1 51.8 53.6 55.8 57.8 53.0 55.2 49.3
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 50.4 53.9 53.3 55.8 53.7 52.0 51.5 50.4 49.4 50.9 52.0 52.7 51.2 55.4
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 55.8 55.3 54.9 55.6 56.6
#> 
#> units:  cm 
l95(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 56.5 57.0 55.0 55.6 49.5 52.0 52.0 51.5 51.5 48.0 52.0 55.0 56.0 55.5
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 57.0 56.0 54.5 51.0 49.5 50.5 51.0 52.0 51.0 53.0 50.5 48.0 49.1 45.0
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 49.5 49.0 46.1 48.0 47.0 50.0 51.0 49.0 51.0 52.5 51.5 51.0 51.0 45.5
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 47.0 49.0 50.5 52.1 48.5 47.1 48.5 47.0 47.0 48.0 48.0 47.5 48.0 53.0
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 50.5 52.0 53.0 53.0 53.0
#> 
#> units:  cm 
l25(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 33.0 31.0 29.0 30.0 28.0 33.0 34.0 27.0 31.0 34.0 34.0 32.0 32.0 28.2
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 32.0 32.0 26.0 27.0 29.0 27.2 26.0 26.2 26.2 28.0 28.0 26.0 29.0 27.0
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 26.0 25.0 28.0 31.0 30.0 30.0 29.0 30.0 30.0 30.0 29.0 30.0 28.0 30.0
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 32.0 31.0 31.0 28.0 30.0 27.0 27.0 27.0 27.0 28.0 28.0 27.0 28.0 32.0
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 30.0 29.0 30.0 29.0 31.0
#> 
#> units:  cm 
lc50(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 22.0 21.5 18.5 16.0 18.0 20.5 18.0 12.5 15.5 18.5 20.5 21.5 22.5 13.5
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 19.0 21.0 13.0 13.0 18.5 13.5 13.5 12.0 13.0 17.0 16.5 12.0 17.0 14.0
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 13.0 12.5 16.0 18.0 19.0 18.5 21.0 17.0 16.0 20.0 16.5 17.5 13.0 16.0
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 18.5 20.5 17.5 19.0 16.0 17.5 17.0 12.5 13.0 17.5 17.0 13.5 15.5 16.5
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 18.0 13.5 18.5 19.0 17.0
#> 
#> units:  cm 
lmode(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 44   43   37   32   36   41   36   25   31   37   41   43   45   27  
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 38   42   26   26   37   27   27   24   26   34   33   24   34   28  
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 26   25   32   36   38   37   42   34   32   40   33   35   26   32  
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 37   41   35   38   32   35   34   25   26   35   34   27   31   33  
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 36   27   37   38   34  
#> 
#> units:  cm 
lbar(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 39.3 39.5 37.1 37.0 35.3 38.6 39.3 35.9 36.4 37.8 39.5 39.7 40.5 37.6
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 39.2 40.0 34.7 34.3 34.7 35.4 34.3 34.6 34.2 35.4 34.7 32.8 33.8 33.7
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 33.9 30.6 33.2 36.0 35.3 36.0 36.0 35.7 36.3 36.9 36.6 35.7 34.5 34.3
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 36.2 36.8 36.5 34.9 35.2 34.1 33.5 33.5 33.5 34.7 34.7 33.8 34.3 37.6
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 36.5 37.0 37.1 36.7 38.1
#> 
#> units:  cm 
lmean(samps)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 51.1 51.0 46.6 43.1 43.5 47.7 44.5 38.2 39.9 42.5 46.7 49.9 51.7 41.7
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 48.9 50.3 39.2 37.3 44.1 39.0 39.5 36.8 37.7 42.4 40.9 34.8 41.3 37.8
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 37.3 34.0 38.9 42.4 43.1 43.9 47.8 42.2 41.3 46.6 42.8 42.8 37.2 38.6
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 42.5 46.1 42.5 45.0 39.4 41.5 41.8 35.9 35.9 42.2 41.7 37.7 39.2 41.9
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 43.5 40.1 45.2 45.6 44.8
#> 
#> units:  cm 
# Linf(ple4) = 60
lmean(samps) / (0.75 * lc50(samps) + 0.25 * 60) #
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 1.62 1.64 1.61 1.60 1.53 1.57 1.56 1.57 1.50 1.47 1.54 1.60 1.62 1.66
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 1.67 1.63 1.58 1.51 1.53 1.55 1.57 1.53 1.52 1.53 1.49 1.45 1.49 1.48
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 1.51 1.39 1.44 1.49 1.47 1.52 1.56 1.52 1.53 1.55 1.56 1.52 1.50 1.43
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 1.47 1.52 1.51 1.54 1.46 1.47 1.51 1.47 1.45 1.50 1.50 1.50 1.47 1.53
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 1.53 1.60 1.56 1.56 1.61
#> 
#> units:   
lenwt <- FLPar(a=0.01030, b=2.975)
lmaxy(samps, lenwt)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970
#>   all 44   56   56   53   44   48   47   46   47   37   41   43   56   49  
#>      year
#> len   1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#>   all 53   56   46   46   37   43   51   50   55   44   41   44   34   43  
#>      year
#> len   1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998
#>   all 38   50   37   41   43   37   42   48   39   40   45   51   44   32  
#>      year
#> len   1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010 2011 2012
#>   all 41   41   43   40   37   35   45   40   39   46   44   44   48   45  
#>      year
#> len   2013 2014 2015 2016 2017
#>   all 44   48   44   38   46  
#> 
#> units:  cm 
pmega(samps, linf=60)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957  1958  1959  1960  1961  1962  1963  1964  1965  1966  1967  1968 
#>   all 0.320 0.356 0.264 0.232 0.196 0.292 0.316 0.228 0.192 0.156 0.292 0.344
#>      year
#> len   1969  1970  1971  1972  1973  1974  1975  1976  1977  1978  1979  1980 
#>   all 0.432 0.324 0.324 0.364 0.224 0.164 0.128 0.172 0.196 0.204 0.176 0.212
#>      year
#> len   1981  1982  1983  1984  1985  1986  1987  1988  1989  1990  1991  1992 
#>   all 0.120 0.124 0.100 0.080 0.148 0.100 0.084 0.152 0.132 0.176 0.216 0.192
#>      year
#> len   1993  1994  1995  1996  1997  1998  1999  2000  2001  2002  2003  2004 
#>   all 0.188 0.212 0.240 0.180 0.148 0.084 0.140 0.188 0.204 0.140 0.144 0.096
#>      year
#> len   2005  2006  2007  2008  2009  2010  2011  2012  2013  2014  2015  2016 
#>   all 0.124 0.128 0.120 0.156 0.156 0.148 0.128 0.236 0.212 0.248 0.260 0.220
#>      year
#> len   2017 
#>   all 0.288
#> 
#> units:  NA 
linf <- 60
k <- 2.29e-01
t0 <- -1.37e+00
bheqz(samps, linf = 60, k = 2.29e-01, t0 = -1.37e+00)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> len   1957  1958  1959  1960  1961  1962  1963  1964  1965  1966  1967  1968 
#>   all 0.275 0.260 0.283 0.252 0.326 0.270 0.223 0.236 0.258 0.264 0.246 0.256
#>      year
#> len   1969  1970  1971  1972  1973  1974  1975  1976  1977  1978  1979  1980 
#>   all 0.249 0.214 0.236 0.240 0.267 0.275 0.359 0.257 0.282 0.258 0.278 0.307
#>      year
#> len   1981  1982  1983  1984  1985  1986  1987  1988  1989  1990  1991  1992 
#>   all 0.320 0.299 0.357 0.306 0.286 0.372 0.355 0.305 0.349 0.315 0.365 0.297
#>      year
#> len   1993  1994  1995  1996  1997  1998  1999  2000  2001  2002  2003  2004 
#>   all 0.267 0.312 0.268 0.305 0.270 0.323 0.309 0.325 0.284 0.360 0.296 0.359
#>      year
#> len   2005  2006  2007  2008  2009  2010  2011  2012  2013  2014  2015  2016 
#>   all 0.369 0.289 0.296 0.336 0.327 0.296 0.313 0.244 0.291 0.224 0.282 0.300
#>      year
#> len   2017 
#>   all 0.238
#> 
#> units:  z