Extract simply-defined metrics from compex objects

Time series summaries of complex objects are commonly needed, for example for plotting the inputs and outputs of a class like FLStock. These methods allow for simple specification of those metrics by means of function calls and formulas.

metrics(object, metrics, ...)

# S4 method for FLComp,list
metrics(object, metrics)

# S4 method for FLS,missing
metrics(object, metrics, ...)

Arguments

object

A complex FLR object from which to extract time series metrics.

Value

An object, generally of class FLQuants.

See also

FLComp

Examples

data(ple4)
# missing
metrics(ple4)

#> $ Rec 
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1957    1958    1959    1960    1961    1962    1963    1964    1965   
#>   1  457973  698110  863386  757299  860577  589154  688367 2231504  694575
#>    year
#> age 1966    1967    1968    1969    1970    1971    1972    1973    1974   
#>   1  586779  401298  434281  648877  650584  410281  366633 1312097 1132831
#>    year
#> age 1975    1976    1977    1978    1979    1980    1981    1982    1983   
#>   1  864875  692849  988889  913474  891160 1128822  869640 2029493 1306601
#>    year
#> age 1984    1985    1986    1987    1988    1989    1990    1991    1992   
#>   1 1261067 1849179 4732214 1918256 1770637 1184055 1033216  910370  773003
#>    year
#> age 1993    1994    1995    1996    1997    1998    1999    2000    2001   
#>   1  522410  434986 1153325 1283485 2105676  765785  836929  927442  516739
#>    year
#> age 2002    2003    2004    2005    2006    2007    2008   
#>   1 1612473  505292 1159019  714344  820006  949341  844041
#> 
#> units:  10^3 
#> 
#> $ SSB 
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 274205 288540 296825 308164 321354 372863 370373 363077 344013 361549
#>      year
#> age   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 416563 402521 377432 333933 316343 319062 268714 278648 293136 310954
#>      year
#> age   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 316929 303433 297122 272416 262061 263998 314021 326341 348675 375392
#>      year
#> age   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 445855 391254 408489 368969 335747 269528 228668 193093 174408 173903
#>      year
#> age   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 185308 211327 184733 208393 234078 162725 179158 151508 167531 173783
#>      year
#> age   2007   2008  
#>   all 166061 206480
#> 
#> units:  t 
#> 
#> $ Catch 
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all  78423  88240 109238 117138 118331 125272 148170 147357 139820 166784
#>      year
#> age   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 163178 139503 142896 160026 136932 142495 143883 157804 195154 167089
#>      year
#> age   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 176691 159727 213422 171235 172671 204286 218424 226930 220928 296876
#>      year
#> age   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 342985 311635 277738 228734 229607 183284 152242 134392 120316 133797
#>      year
#> age   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 179957 175002 151708 126142 182578 125884 145390 117702 111060 121205
#>      year
#> age   2007   2008  
#>   all  90283  96040
#> 
#> units:  t 
#> 
#> $ F 
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957    1958    1959    1960    1961    1962    1963    1964    1965   
#>   all 0.26857 0.32106 0.36734 0.36796 0.34756 0.39012 0.42276 0.46878 0.38796
#>      year
#> age   1966    1967    1968    1969    1970    1971    1972    1973    1974   
#>   all 0.39896 0.42923 0.33621 0.34457 0.47965 0.38206 0.41158 0.46551 0.49072
#>      year
#> age   1975    1976    1977    1978    1979    1980    1981    1982    1983   
#>   all 0.56113 0.41641 0.51007 0.46862 0.67312 0.55555 0.53705 0.59912 0.58934
#>      year
#> age   1984    1985    1986    1987    1988    1989    1990    1991    1992   
#>   all 0.58159 0.52695 0.65386 0.69596 0.67530 0.61895 0.59361 0.71195 0.69443
#>      year
#> age   1993    1994    1995    1996    1997    1998    1999    2000    2001   
#>   all 0.64752 0.63741 0.67444 0.72301 0.87588 0.84233 0.77264 0.55795 0.87567
#>      year
#> age   2002    2003    2004    2005    2006    2007    2008   
#>   all 0.69763 0.76597 0.64015 0.62343 0.54764 0.46392 0.35631
#> 
#> units:  f 
#> 

# metrics = function
metrics(ple4, metrics=function(x) FLQuants(SSB=ssb(x), REC=rec(x),
  F=fbar(x), SSBREC=ssb(x) / rec(x)))

#> $ SSB 
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 274205 288540 296825 308164 321354 372863 370373 363077 344013 361549
#>      year
#> age   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 416563 402521 377432 333933 316343 319062 268714 278648 293136 310954
#>      year
#> age   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 316929 303433 297122 272416 262061 263998 314021 326341 348675 375392
#>      year
#> age   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 445855 391254 408489 368969 335747 269528 228668 193093 174408 173903
#>      year
#> age   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 185308 211327 184733 208393 234078 162725 179158 151508 167531 173783
#>      year
#> age   2007   2008  
#>   all 166061 206480
#> 
#> units:  t 
#> 
#> $ REC 
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1957    1958    1959    1960    1961    1962    1963    1964    1965   
#>   1  457973  698110  863386  757299  860577  589154  688367 2231504  694575
#>    year
#> age 1966    1967    1968    1969    1970    1971    1972    1973    1974   
#>   1  586779  401298  434281  648877  650584  410281  366633 1312097 1132831
#>    year
#> age 1975    1976    1977    1978    1979    1980    1981    1982    1983   
#>   1  864875  692849  988889  913474  891160 1128822  869640 2029493 1306601
#>    year
#> age 1984    1985    1986    1987    1988    1989    1990    1991    1992   
#>   1 1261067 1849179 4732214 1918256 1770637 1184055 1033216  910370  773003
#>    year
#> age 1993    1994    1995    1996    1997    1998    1999    2000    2001   
#>   1  522410  434986 1153325 1283485 2105676  765785  836929  927442  516739
#>    year
#> age 2002    2003    2004    2005    2006    2007    2008   
#>   1 1612473  505292 1159019  714344  820006  949341  844041
#> 
#> units:  10^3 
#> 
#> $ F 
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957    1958    1959    1960    1961    1962    1963    1964    1965   
#>   all 0.26857 0.32106 0.36734 0.36796 0.34756 0.39012 0.42276 0.46878 0.38796
#>      year
#> age   1966    1967    1968    1969    1970    1971    1972    1973    1974   
#>   all 0.39896 0.42923 0.33621 0.34457 0.47965 0.38206 0.41158 0.46551 0.49072
#>      year
#> age   1975    1976    1977    1978    1979    1980    1981    1982    1983   
#>   all 0.56113 0.41641 0.51007 0.46862 0.67312 0.55555 0.53705 0.59912 0.58934
#>      year
#> age   1984    1985    1986    1987    1988    1989    1990    1991    1992   
#>   all 0.58159 0.52695 0.65386 0.69596 0.67530 0.61895 0.59361 0.71195 0.69443
#>      year
#> age   1993    1994    1995    1996    1997    1998    1999    2000    2001   
#>   all 0.64752 0.63741 0.67444 0.72301 0.87588 0.84233 0.77264 0.55795 0.87567
#>      year
#> age   2002    2003    2004    2005    2006    2007    2008   
#>   all 0.69763 0.76597 0.64015 0.62343 0.54764 0.46392 0.35631
#> 
#> units:  f 
#> 
#> $ SSBREC 
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957     1958     1959     1960     1961     1962     1963     1964    
#>   all 0.598737 0.413316 0.343792 0.406925 0.373417 0.632879 0.538046 0.162705
#>      year
#> age   1965     1966     1967     1968     1969     1970     1971     1972    
#>   all 0.495286 0.616158 1.038041 0.926867 0.581670 0.513283 0.771041 0.870250
#>      year
#> age   1973     1974     1975     1976     1977     1978     1979     1980    
#>   all 0.204797 0.245975 0.338935 0.448805 0.320489 0.332174 0.333410 0.241327
#>      year
#> age   1981     1982     1983     1984     1985     1986     1987     1988    
#>   all 0.301344 0.130081 0.240334 0.258781 0.188557 0.079327 0.232427 0.220968
#>      year
#> age   1989     1990     1991     1992     1993     1994     1995     1996    
#>   all 0.344991 0.357108 0.368802 0.348677 0.437716 0.443906 0.151222 0.135493
#>      year
#> age   1997     1998     1999     2000     2001     2002     2003     2004    
#>   all 0.088004 0.275961 0.220727 0.224697 0.452991 0.100917 0.354564 0.130721
#>      year
#> age   2005     2006     2007     2008    
#>   all 0.234524 0.211929 0.174922 0.244632
#> 
#> units:  kg 
#> 

# metrics = formula
metrics(ple4, metrics=SSB~ssb)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 274205 288540 296825 308164 321354 372863 370373 363077 344013 361549
#>      year
#> age   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 416563 402521 377432 333933 316343 319062 268714 278648 293136 310954
#>      year
#> age   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 316929 303433 297122 272416 262061 263998 314021 326341 348675 375392
#>      year
#> age   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 445855 391254 408489 368969 335747 269528 228668 193093 174408 173903
#>      year
#> age   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 185308 211327 184733 208393 234078 162725 179158 151508 167531 173783
#>      year
#> age   2007   2008  
#>   all 166061 206480
#> 
#> units:  t 

metrics(ple4, metrics=~ssb)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 274205 288540 296825 308164 321354 372863 370373 363077 344013 361549
#>      year
#> age   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 416563 402521 377432 333933 316343 319062 268714 278648 293136 310954
#>      year
#> age   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 316929 303433 297122 272416 262061 263998 314021 326341 348675 375392
#>      year
#> age   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 445855 391254 408489 368969 335747 269528 228668 193093 174408 173903
#>      year
#> age   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 185308 211327 184733 208393 234078 162725 179158 151508 167531 173783
#>      year
#> age   2007   2008  
#>   all 166061 206480
#> 
#> units:  t 

# metrics = list
metrics(ple4, metrics=list(SSB=ssb, REC=rec, F=fbar))

#> $ SSB 
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 274205 288540 296825 308164 321354 372863 370373 363077 344013 361549
#>      year
#> age   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 416563 402521 377432 333933 316343 319062 268714 278648 293136 310954
#>      year
#> age   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 316929 303433 297122 272416 262061 263998 314021 326341 348675 375392
#>      year
#> age   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 445855 391254 408489 368969 335747 269528 228668 193093 174408 173903
#>      year
#> age   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 185308 211327 184733 208393 234078 162725 179158 151508 167531 173783
#>      year
#> age   2007   2008  
#>   all 166061 206480
#> 
#> units:  t 
#> 
#> $ REC 
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1957    1958    1959    1960    1961    1962    1963    1964    1965   
#>   1  457973  698110  863386  757299  860577  589154  688367 2231504  694575
#>    year
#> age 1966    1967    1968    1969    1970    1971    1972    1973    1974   
#>   1  586779  401298  434281  648877  650584  410281  366633 1312097 1132831
#>    year
#> age 1975    1976    1977    1978    1979    1980    1981    1982    1983   
#>   1  864875  692849  988889  913474  891160 1128822  869640 2029493 1306601
#>    year
#> age 1984    1985    1986    1987    1988    1989    1990    1991    1992   
#>   1 1261067 1849179 4732214 1918256 1770637 1184055 1033216  910370  773003
#>    year
#> age 1993    1994    1995    1996    1997    1998    1999    2000    2001   
#>   1  522410  434986 1153325 1283485 2105676  765785  836929  927442  516739
#>    year
#> age 2002    2003    2004    2005    2006    2007    2008   
#>   1 1612473  505292 1159019  714344  820006  949341  844041
#> 
#> units:  10^3 
#> 
#> $ F 
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957    1958    1959    1960    1961    1962    1963    1964    1965   
#>   all 0.26857 0.32106 0.36734 0.36796 0.34756 0.39012 0.42276 0.46878 0.38796
#>      year
#> age   1966    1967    1968    1969    1970    1971    1972    1973    1974   
#>   all 0.39896 0.42923 0.33621 0.34457 0.47965 0.38206 0.41158 0.46551 0.49072
#>      year
#> age   1975    1976    1977    1978    1979    1980    1981    1982    1983   
#>   all 0.56113 0.41641 0.51007 0.46862 0.67312 0.55555 0.53705 0.59912 0.58934
#>      year
#> age   1984    1985    1986    1987    1988    1989    1990    1991    1992   
#>   all 0.58159 0.52695 0.65386 0.69596 0.67530 0.61895 0.59361 0.71195 0.69443
#>      year
#> age   1993    1994    1995    1996    1997    1998    1999    2000    2001   
#>   all 0.64752 0.63741 0.67444 0.72301 0.87588 0.84233 0.77264 0.55795 0.87567
#>      year
#> age   2002    2003    2004    2005    2006    2007    2008   
#>   all 0.69763 0.76597 0.64015 0.62343 0.54764 0.46392 0.35631
#> 
#> units:  f 
#> 

metrics(ple4, metrics=list(SSB=~ssb, REC=rec, F=fbar))

#> $ SSB 
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957   1958   1959   1960   1961   1962   1963   1964   1965   1966  
#>   all 274205 288540 296825 308164 321354 372863 370373 363077 344013 361549
#>      year
#> age   1967   1968   1969   1970   1971   1972   1973   1974   1975   1976  
#>   all 416563 402521 377432 333933 316343 319062 268714 278648 293136 310954
#>      year
#> age   1977   1978   1979   1980   1981   1982   1983   1984   1985   1986  
#>   all 316929 303433 297122 272416 262061 263998 314021 326341 348675 375392
#>      year
#> age   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all 445855 391254 408489 368969 335747 269528 228668 193093 174408 173903
#>      year
#> age   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all 185308 211327 184733 208393 234078 162725 179158 151508 167531 173783
#>      year
#> age   2007   2008  
#>   all 166061 206480
#> 
#> units:  t 
#> 
#> $ REC 
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1957    1958    1959    1960    1961    1962    1963    1964    1965   
#>   1  457973  698110  863386  757299  860577  589154  688367 2231504  694575
#>    year
#> age 1966    1967    1968    1969    1970    1971    1972    1973    1974   
#>   1  586779  401298  434281  648877  650584  410281  366633 1312097 1132831
#>    year
#> age 1975    1976    1977    1978    1979    1980    1981    1982    1983   
#>   1  864875  692849  988889  913474  891160 1128822  869640 2029493 1306601
#>    year
#> age 1984    1985    1986    1987    1988    1989    1990    1991    1992   
#>   1 1261067 1849179 4732214 1918256 1770637 1184055 1033216  910370  773003
#>    year
#> age 1993    1994    1995    1996    1997    1998    1999    2000    2001   
#>   1  522410  434986 1153325 1283485 2105676  765785  836929  927442  516739
#>    year
#> age 2002    2003    2004    2005    2006    2007    2008   
#>   1 1612473  505292 1159019  714344  820006  949341  844041
#> 
#> units:  10^3 
#> 
#> $ F 
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1957    1958    1959    1960    1961    1962    1963    1964    1965   
#>   all 0.26857 0.32106 0.36734 0.36796 0.34756 0.39012 0.42276 0.46878 0.38796
#>      year
#> age   1966    1967    1968    1969    1970    1971    1972    1973    1974   
#>   all 0.39896 0.42923 0.33621 0.34457 0.47965 0.38206 0.41158 0.46551 0.49072
#>      year
#> age   1975    1976    1977    1978    1979    1980    1981    1982    1983   
#>   all 0.56113 0.41641 0.51007 0.46862 0.67312 0.55555 0.53705 0.59912 0.58934
#>      year
#> age   1984    1985    1986    1987    1988    1989    1990    1991    1992   
#>   all 0.58159 0.52695 0.65386 0.69596 0.67530 0.61895 0.59361 0.71195 0.69443
#>      year
#> age   1993    1994    1995    1996    1997    1998    1999    2000    2001   
#>   all 0.64752 0.63741 0.67444 0.72301 0.87588 0.84233 0.77264 0.55795 0.87567
#>      year
#> age   2002    2003    2004    2005    2006    2007    2008   
#>   all 0.69763 0.76597 0.64015 0.62343 0.54764 0.46392 0.35631
#> 
#> units:  f 
#>