Methods quantTotals

Methods to compute totals over selected dimensions of FLQuant objects These methods return an object of same dimensions as the input but with the sums along the first (yearTotals) or second dimension (quantTotals). Although the names might appear contradictory, it must be noted that what each method really returns are the totals over the selected dimension.

quantTotals(x, ...)

Generic function

quantTotals(x)

yearTotals(x)

See also

Author

The FLR Team

Examples

flq <- FLQuant(rlnorm(100), dim=c(10,10))
quantTotals(flq)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1       2       3       4       5       6       7       8       9      
#>    1  17.2889 17.2889 17.2889 17.2889 17.2889 17.2889 17.2889 17.2889 17.2889
#>    2  14.8610 14.8610 14.8610 14.8610 14.8610 14.8610 14.8610 14.8610 14.8610
#>    3  17.5973 17.5973 17.5973 17.5973 17.5973 17.5973 17.5973 17.5973 17.5973
#>    4   8.0644  8.0644  8.0644  8.0644  8.0644  8.0644  8.0644  8.0644  8.0644
#>    5  13.0276 13.0276 13.0276 13.0276 13.0276 13.0276 13.0276 13.0276 13.0276
#>    6  25.0089 25.0089 25.0089 25.0089 25.0089 25.0089 25.0089 25.0089 25.0089
#>    7  12.6895 12.6895 12.6895 12.6895 12.6895 12.6895 12.6895 12.6895 12.6895
#>    8  16.7151 16.7151 16.7151 16.7151 16.7151 16.7151 16.7151 16.7151 16.7151
#>    9   8.3470  8.3470  8.3470  8.3470  8.3470  8.3470  8.3470  8.3470  8.3470
#>    10 10.8745 10.8745 10.8745 10.8745 10.8745 10.8745 10.8745 10.8745 10.8745
#>      year
#> quant 10     
#>    1  17.2889
#>    2  14.8610
#>    3  17.5973
#>    4   8.0644
#>    5  13.0276
#>    6  25.0089
#>    7  12.6895
#>    8  16.7151
#>    9   8.3470
#>    10 10.8745
#> 
#> units:  NA 
# See how the values obtained by yearSums are being replicated
  yearSums(flq)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1      
#>    1  17.2889
#>    2  14.8610
#>    3  17.5973
#>    4   8.0644
#>    5  13.0276
#>    6  25.0089
#>    7  12.6895
#>    8  16.7151
#>    9   8.3470
#>    10 10.8745
#> 
#> units:  NA 
# Get the proportions by quant
  flq / quantTotals(flq)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1         2         3         4         5         6         7        
#>    1  0.2597342 0.0511107 0.1419546 0.2180459 0.0590937 0.0201879 0.0405648
#>    2  0.0931921 0.2553599 0.0881916 0.0045337 0.1584443 0.1405888 0.0536341
#>    3  0.5151607 0.0597723 0.0204680 0.0225236 0.0165413 0.0340659 0.0307040
#>    4  0.0405549 0.1307376 0.0625243 0.1422617 0.0787110 0.1709977 0.0662583
#>    5  0.0231109 0.1134609 0.3531494 0.0701214 0.0345239 0.0715590 0.1754652
#>    6  0.1073183 0.2413557 0.0229623 0.1802784 0.0643052 0.0563176 0.0824409
#>    7  0.0277045 0.0538221 0.0681274 0.1504058 0.0258461 0.2224044 0.1367029
#>    8  0.2882413 0.0874534 0.0143723 0.0774136 0.0706689 0.0362239 0.3590108
#>    9  0.0401996 0.1410952 0.1198158 0.2161245 0.0246974 0.1711854 0.1475119
#>    10 0.0213840 0.1675990 0.2035925 0.1398002 0.0736434 0.0216467 0.1142355
#>      year
#> quant 8         9         10       
#>    1  0.0967872 0.0150631 0.0974579
#>    2  0.0787974 0.0357018 0.0915563
#>    3  0.1693326 0.0259208 0.1055107
#>    4  0.0681502 0.1164570 0.1233475
#>    5  0.0827421 0.0393659 0.0365013
#>    6  0.0074707 0.1774019 0.0601490
#>    7  0.0810629 0.1757176 0.0582064
#>    8  0.0169581 0.0231852 0.0264724
#>    9  0.0251437 0.0393509 0.0748755
#>    10 0.0412097 0.0729741 0.1439148
#> 
#> units:  NA 
# or year
  flq / yearTotals(flq)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1         2         3         4         5         6         7        
#>    1  0.1871783 0.0454460 0.1737758 0.2175329 0.1151000 0.0294550 0.0409415
#>    2  0.0577279 0.1951714 0.0927999 0.0038879 0.2652721 0.1763195 0.0465303
#>    3  0.3778755 0.0540957 0.0255032 0.0228715 0.0327933 0.0505904 0.0315420
#>    4  0.0136325 0.0542237 0.0357021 0.0662020 0.0715115 0.1163765 0.0311932
#>    5  0.0125499 0.0760195 0.3257572 0.0527137 0.0506700 0.0786736 0.1334447
#>    6  0.1118737 0.3104334 0.0406614 0.2601649 0.1811793 0.1188616 0.1203608
#>    7  0.0146540 0.0351254 0.0612124 0.1101335 0.0369493 0.2381717 0.1012674
#>    8  0.2008286 0.0751801 0.0170101 0.0746685 0.1330777 0.0510984 0.3503199
#>    9  0.0139867 0.0605705 0.0708141 0.1040991 0.0232248 0.1205874 0.0718799
#>    10 0.0096930 0.0937343 0.1567638 0.0877261 0.0902220 0.0198658 0.0725203
#>      year
#> quant 8         9         10       
#>    1  0.1741502 0.0239469 0.1498007
#>    2  0.1218704 0.0487871 0.1209666
#>    3  0.3101174 0.0419432 0.1650719
#>    4  0.0571978 0.0863586 0.0884368
#>    5  0.1121834 0.0471575 0.0422767
#>    6  0.0194443 0.4079624 0.1337374
#>    7  0.1070547 0.2050346 0.0656667
#>    8  0.0295003 0.0356359 0.0393399
#>    9  0.0218425 0.0302034 0.0555651
#>    10 0.0466391 0.0729704 0.1391382
#> 
#> units:  NA