Methods quantTotals

Methods to compute totals over selected dimensions of FLQuant objects These methods return an object of same dimensions as the input but with the sums along the first (yearTotals) or second dimension (quantTotals). Although the names might appear contradictory, it must be noted that what each method really returns are the totals over the selected dimension.

quantTotals(x, ...)

Generic function

quantTotals(x)

yearTotals(x)

See also

Author

The FLR Team

Examples

flq <- FLQuant(rlnorm(100), dim=c(10,10))
quantTotals(flq)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1      2      3      4      5      6      7      8      9      10    
#>    1  16.209 16.209 16.209 16.209 16.209 16.209 16.209 16.209 16.209 16.209
#>    2  26.465 26.465 26.465 26.465 26.465 26.465 26.465 26.465 26.465 26.465
#>    3  14.844 14.844 14.844 14.844 14.844 14.844 14.844 14.844 14.844 14.844
#>    4  16.844 16.844 16.844 16.844 16.844 16.844 16.844 16.844 16.844 16.844
#>    5  11.483 11.483 11.483 11.483 11.483 11.483 11.483 11.483 11.483 11.483
#>    6  14.587 14.587 14.587 14.587 14.587 14.587 14.587 14.587 14.587 14.587
#>    7  27.367 27.367 27.367 27.367 27.367 27.367 27.367 27.367 27.367 27.367
#>    8  13.632 13.632 13.632 13.632 13.632 13.632 13.632 13.632 13.632 13.632
#>    9  12.128 12.128 12.128 12.128 12.128 12.128 12.128 12.128 12.128 12.128
#>    10 17.925 17.925 17.925 17.925 17.925 17.925 17.925 17.925 17.925 17.925
#> 
#> units:  NA 
# See how the values obtained by yearSums are being replicated
  yearSums(flq)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1     
#>    1  16.209
#>    2  26.465
#>    3  14.844
#>    4  16.844
#>    5  11.483
#>    6  14.587
#>    7  27.367
#>    8  13.632
#>    9  12.128
#>    10 17.925
#> 
#> units:  NA 
# Get the proportions by quant
  flq / quantTotals(flq)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1         2         3         4         5         6         7        
#>    1  0.2497635 0.2058009 0.0337005 0.0155815 0.3503807 0.0461866 0.0190504
#>    2  0.0212926 0.0226310 0.0782168 0.6124777 0.0047139 0.0101764 0.1202873
#>    3  0.0335430 0.3695958 0.1133043 0.1859330 0.0216383 0.0036856 0.0257384
#>    4  0.0341373 0.0503159 0.1594226 0.1563479 0.2533902 0.0375913 0.1639330
#>    5  0.0477196 0.1297126 0.0527786 0.0202636 0.5335427 0.0030375 0.0895161
#>    6  0.0444914 0.0185366 0.3875156 0.0375306 0.0729187 0.0813906 0.0062151
#>    7  0.0783561 0.3925173 0.0680749 0.1067364 0.0386176 0.0893218 0.0497134
#>    8  0.0377336 0.0086277 0.0223027 0.0840032 0.3133743 0.1086742 0.1162591
#>    9  0.4030879 0.1041242 0.0868423 0.1940958 0.0452601 0.0645231 0.0166893
#>    10 0.0690799 0.0722811 0.1293420 0.0363324 0.1137285 0.0063995 0.0125352
#>      year
#> quant 8         9         10       
#>    1  0.0310464 0.0314328 0.0170568
#>    2  0.0307190 0.0931764 0.0063089
#>    3  0.0264539 0.1720691 0.0480386
#>    4  0.0930636 0.0187341 0.0330641
#>    5  0.0303117 0.0769902 0.0161274
#>    6  0.2595788 0.0325147 0.0593080
#>    7  0.0047661 0.1326026 0.0392937
#>    8  0.0359293 0.1731280 0.0999678
#>    9  0.0407272 0.0075644 0.0370857
#>    10 0.1436416 0.0847158 0.3319441
#> 
#> units:  NA 
# or year
  flq / yearTotals(flq)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1         2         3         4         5         6         7        
#>    1  0.2583958 0.1310895 0.0290852 0.0085017 0.2227157 0.0965839 0.0277520
#>    2  0.0359668 0.0235365 0.1102182 0.5456365 0.0048923 0.0347454 0.2861055
#>    3  0.0317805 0.2156006 0.0895541 0.0929085 0.0125961 0.0070584 0.0343379
#>    4  0.0367016 0.0333062 0.1429836 0.0886519 0.1673786 0.0816913 0.2481737
#>    5  0.0349758 0.0585352 0.0322708 0.0078330 0.2402673 0.0045001 0.0923860
#>    6  0.0414223 0.0106256 0.3009737 0.0184283 0.0417112 0.1531673 0.0081478
#>    7  0.1368704 0.4221430 0.0991981 0.0983307 0.0414453 0.3153747 0.1222767
#>    8  0.0328310 0.0046219 0.0161880 0.0385471 0.1675225 0.1911237 0.1424348
#>    9  0.3120201 0.0496248 0.0560783 0.0792390 0.0215255 0.1009556 0.0181909
#>    10 0.0790356 0.0509167 0.1234500 0.0219233 0.0799456 0.0147996 0.0201947
#>      year
#> quant 8         9         10       
#>    1  0.0453361 0.0344183 0.0238299
#>    2  0.0732414 0.1665824 0.0143911
#>    3  0.0353774 0.1725488 0.0614637
#>    4  0.1412255 0.0213176 0.0480045
#>    5  0.0313587 0.0597251 0.0159626
#>    6  0.3411186 0.0320397 0.0745661
#>    7  0.0117510 0.2451539 0.0926892
#>    8  0.0441247 0.1594311 0.1174586
#>    9  0.0444985 0.0061974 0.0387667
#>    10 0.2319682 0.1025855 0.5128676
#> 
#> units:  NA