Methods quantTotals

Methods to compute totals over selected dimensions of FLQuant objects These methods return an object of same dimensions as the input but with the sums along the first (yearTotals) or second dimension (quantTotals). Although the names might appear contradictory, it must be noted that what each method really returns are the totals over the selected dimension.

quantTotals(x, ...)

Generic function

quantTotals(x)

yearTotals(x)

See also

Author

The FLR Team

Examples

flq <- FLQuant(rlnorm(100), dim=c(10,10))
quantTotals(flq)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1       2       3       4       5       6       7       8       9      
#>    1  19.3535 19.3535 19.3535 19.3535 19.3535 19.3535 19.3535 19.3535 19.3535
#>    2  16.4669 16.4669 16.4669 16.4669 16.4669 16.4669 16.4669 16.4669 16.4669
#>    3   8.5935  8.5935  8.5935  8.5935  8.5935  8.5935  8.5935  8.5935  8.5935
#>    4  13.6053 13.6053 13.6053 13.6053 13.6053 13.6053 13.6053 13.6053 13.6053
#>    5  32.8741 32.8741 32.8741 32.8741 32.8741 32.8741 32.8741 32.8741 32.8741
#>    6   7.9165  7.9165  7.9165  7.9165  7.9165  7.9165  7.9165  7.9165  7.9165
#>    7   9.3365  9.3365  9.3365  9.3365  9.3365  9.3365  9.3365  9.3365  9.3365
#>    8  13.4078 13.4078 13.4078 13.4078 13.4078 13.4078 13.4078 13.4078 13.4078
#>    9  11.9501 11.9501 11.9501 11.9501 11.9501 11.9501 11.9501 11.9501 11.9501
#>    10 23.4792 23.4792 23.4792 23.4792 23.4792 23.4792 23.4792 23.4792 23.4792
#>      year
#> quant 10     
#>    1  19.3535
#>    2  16.4669
#>    3   8.5935
#>    4  13.6053
#>    5  32.8741
#>    6   7.9165
#>    7   9.3365
#>    8  13.4078
#>    9  11.9501
#>    10 23.4792
#> 
#> units:  NA 
# See how the values obtained by yearSums are being replicated
  yearSums(flq)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1      
#>    1  19.3535
#>    2  16.4669
#>    3   8.5935
#>    4  13.6053
#>    5  32.8741
#>    6   7.9165
#>    7   9.3365
#>    8  13.4078
#>    9  11.9501
#>    10 23.4792
#> 
#> units:  NA 
# Get the proportions by quant
  flq / quantTotals(flq)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1         2         3         4         5         6         7        
#>    1  0.0091499 0.0181072 0.2138843 0.0170791 0.1273371 0.0640048 0.3182395
#>    2  0.1226421 0.1123895 0.0400801 0.0693892 0.1615175 0.0274214 0.1421287
#>    3  0.0383656 0.0917744 0.3578025 0.0971157 0.1191148 0.0750435 0.0343739
#>    4  0.3004000 0.0290072 0.1025031 0.0431377 0.2003738 0.0645411 0.1169614
#>    5  0.0184775 0.0421930 0.0356046 0.0164842 0.5159776 0.2468424 0.0254166
#>    6  0.0134545 0.0369128 0.0215038 0.0525985 0.0796738 0.0851642 0.2483788
#>    7  0.0708023 0.0470658 0.2843517 0.0798389 0.0452504 0.1069187 0.1512132
#>    8  0.0564793 0.0203940 0.1230327 0.0552124 0.0264409 0.0330213 0.4404000
#>    9  0.1641105 0.1403384 0.0815561 0.0820469 0.0727983 0.0410232 0.0292893
#>    10 0.0770338 0.1153049 0.1971530 0.0737889 0.1455036 0.0372280 0.0829476
#>      year
#> quant 8         9         10       
#>    1  0.0261172 0.0579136 0.1481672
#>    2  0.0891119 0.1457406 0.0895791
#>    3  0.0716658 0.0396096 0.0751343
#>    4  0.0336212 0.0773577 0.0320967
#>    5  0.0473573 0.0295762 0.0220705
#>    6  0.2496635 0.1151620 0.0974881
#>    7  0.0907220 0.0698566 0.0539803
#>    8  0.0457928 0.1832678 0.0159589
#>    9  0.1684357 0.1137481 0.1066536
#>    10 0.0092800 0.1344636 0.1272967
#> 
#> units:  NA 
# or year
  flq / yearTotals(flq)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1         2         3         4         5         6         7        
#>    1  0.0141492 0.0344888 0.2017498 0.0410528 0.0781606 0.0836547 0.2701077
#>    2  0.1613636 0.1821396 0.0321675 0.1419129 0.0843540 0.0304944 0.1026403
#>    3  0.0263429 0.0776169 0.1498604 0.1036514 0.0324644 0.0435512 0.0129545
#>    4  0.3265601 0.0388403 0.0679707 0.0728928 0.0864618 0.0593014 0.0697873
#>    5  0.0485346 0.1365089 0.0570473 0.0673040 0.5379705 0.5480152 0.0366433
#>    6  0.0085105 0.0287592 0.0082970 0.0517159 0.0200043 0.0455312 0.0862327
#>    7  0.0528184 0.0432470 0.1293942 0.0925798 0.0133992 0.0674150 0.0619152
#>    8  0.0605065 0.0269110 0.0803998 0.0919420 0.0112437 0.0299001 0.2589587
#>    9  0.1566973 0.1650495 0.0475011 0.1217731 0.0275909 0.0331070 0.0153499
#>    10 0.1445170 0.2664390 0.2256121 0.2151753 0.1083506 0.0590299 0.0854105
#>      year
#> quant 8         9         10       
#>    1  0.0492116 0.0777091 0.2408974
#>    2  0.1428660 0.1663892 0.1239197
#>    3  0.0599599 0.0235994 0.0542410
#>    4  0.0445353 0.0729702 0.0366852
#>    5  0.1515728 0.0674106 0.0609520
#>    6  0.1924285 0.0632084 0.0648345
#>    7  0.0824664 0.0452192 0.0423390
#>    8  0.0597774 0.1703639 0.0179756
#>    9  0.1959686 0.0942427 0.1070701
#>    10 0.0212135 0.2188873 0.2510856
#> 
#> units:  NA