Methods quantTotals

Methods to compute totals over selected dimensions of FLQuant objects These methods return an object of same dimensions as the input but with the sums along the first (yearTotals) or second dimension (quantTotals). Although the names might appear contradictory, it must be noted that what each method really returns are the totals over the selected dimension.

quantTotals(x, ...)

Generic function

quantTotals(x)

yearTotals(x)

See also

Examples

flq <- FLQuant(rlnorm(100), dim=c(10,10))
quantTotals(flq)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1       2       3       4       5       6       7       8       9      
#>    1  21.9537 21.9537 21.9537 21.9537 21.9537 21.9537 21.9537 21.9537 21.9537
#>    2  18.0071 18.0071 18.0071 18.0071 18.0071 18.0071 18.0071 18.0071 18.0071
#>    3  13.5336 13.5336 13.5336 13.5336 13.5336 13.5336 13.5336 13.5336 13.5336
#>    4  11.4293 11.4293 11.4293 11.4293 11.4293 11.4293 11.4293 11.4293 11.4293
#>    5   5.2574  5.2574  5.2574  5.2574  5.2574  5.2574  5.2574  5.2574  5.2574
#>    6  20.4760 20.4760 20.4760 20.4760 20.4760 20.4760 20.4760 20.4760 20.4760
#>    7  39.2379 39.2379 39.2379 39.2379 39.2379 39.2379 39.2379 39.2379 39.2379
#>    8  26.6997 26.6997 26.6997 26.6997 26.6997 26.6997 26.6997 26.6997 26.6997
#>    9  21.3521 21.3521 21.3521 21.3521 21.3521 21.3521 21.3521 21.3521 21.3521
#>    10  7.3493  7.3493  7.3493  7.3493  7.3493  7.3493  7.3493  7.3493  7.3493
#>      year
#> quant 10     
#>    1  21.9537
#>    2  18.0071
#>    3  13.5336
#>    4  11.4293
#>    5   5.2574
#>    6  20.4760
#>    7  39.2379
#>    8  26.6997
#>    9  21.3521
#>    10  7.3493
#> 
#> units:  NA 

# See how the values obtained by yearSums are being replicated
  yearSums(flq)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1      
#>    1  21.9537
#>    2  18.0071
#>    3  13.5336
#>    4  11.4293
#>    5   5.2574
#>    6  20.4760
#>    7  39.2379
#>    8  26.6997
#>    9  21.3521
#>    10  7.3493
#> 
#> units:  NA 

# Get the proportions by quant
  flq / quantTotals(flq)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1         2         3         4         5         6         7        
#>    1  0.0058778 0.1645491 0.0825122 0.3922428 0.0214897 0.0478850 0.1004016
#>    2  0.1171447 0.0247539 0.0099370 0.5932302 0.0467590 0.0475659 0.0435286
#>    3  0.0731258 0.1318516 0.0318695 0.0206425 0.0738486 0.0731922 0.0675211
#>    4  0.0564680 0.0789265 0.0742360 0.0608600 0.1622307 0.0490317 0.1100730
#>    5  0.0212503 0.0369676 0.1078018 0.1021458 0.0180600 0.0752170 0.0609760
#>    6  0.0733275 0.2470707 0.0131845 0.0124773 0.0276988 0.2008971 0.0160131
#>    7  0.0290580 0.3089707 0.0297240 0.0083190 0.1877670 0.0430327 0.0752352
#>    8  0.1780919 0.3320373 0.0525292 0.0562061 0.0300890 0.0235183 0.0088622
#>    9  0.0268358 0.0313449 0.0355197 0.3460665 0.0140748 0.0413210 0.0206980
#>    10 0.0185859 0.0537724 0.2691591 0.1341805 0.0422436 0.0585867 0.0591511
#>      year
#> quant 8         9         10       
#>    1  0.0671393 0.0176567 0.1002458
#>    2  0.0430202 0.0312881 0.0427723
#>    3  0.4601053 0.0492022 0.0186412
#>    4  0.2812841 0.0484244 0.0784656
#>    5  0.1775544 0.1982331 0.2017940
#>    6  0.0316755 0.0333917 0.3442637
#>    7  0.0574190 0.2431781 0.0172964
#>    8  0.0432811 0.0898135 0.1855714
#>    9  0.0309707 0.2151199 0.2380488
#>    10 0.0133569 0.2135562 0.1374076
#> 
#> units:  NA 

# or year
  flq / yearTotals(flq)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1         2         3         4         5         6         7        
#>    1  0.0106719 0.1060889 0.1924562 0.2754385 0.0346604 0.0906483 0.2232382
#>    2  0.1744571 0.0130905 0.0190110 0.3416886 0.0618594 0.0738572 0.0793854
#>    3  0.0818471 0.0524041 0.0458241 0.0089359 0.0734260 0.0854141 0.0925491
#>    4  0.0533757 0.0264917 0.0901449 0.0222492 0.1362225 0.0483225 0.1274154
#>    5  0.0092397 0.0057077 0.0602154 0.0171774 0.0069757 0.0340991 0.0324679
#>    6  0.1241745 0.1485707 0.0286824 0.0081720 0.0416678 0.3547078 0.0332078
#>    7  0.0942960 0.3560337 0.1239138 0.0104409 0.5412790 0.1455988 0.2989840
#>    8  0.3932525 0.2603520 0.1490095 0.0480012 0.0590215 0.0541459 0.0239644
#>    9  0.0473888 0.0196551 0.0805781 0.2363538 0.0220790 0.0760788 0.0447600
#>    10 0.0112966 0.0116057 0.2101647 0.0315425 0.0228087 0.0371275 0.0440279
#>      year
#> quant 8         9         10       
#>    1  0.0845133 0.0176205 0.0918662
#>    2  0.0444179 0.0256109 0.0321507
#>    3  0.3570348 0.0302689 0.0105310
#>    4  0.1843346 0.0251585 0.0374354
#>    5  0.0535238 0.0473752 0.0442859
#>    6  0.0371886 0.0310801 0.2942511
#>    7  0.1291824 0.4337413 0.0283299
#>    8  0.0662593 0.1090054 0.2068236
#>    9  0.0379169 0.2087955 0.2121724
#>    10 0.0056285 0.0713438 0.0421539
#> 
#> units:  NA