Methods quantTotals

Methods to compute totals over selected dimensions of FLQuant objects These methods return an object of same dimensions as the input but with the sums along the first (yearTotals) or second dimension (quantTotals). Although the names might appear contradictory, it must be noted that what each method really returns are the totals over the selected dimension.

quantTotals(x, ...)

Generic function

quantTotals(x)

yearTotals(x)

See also

Examples

flq <- FLQuant(rlnorm(100), dim=c(10,10))
quantTotals(flq)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1       2       3       4       5       6       7       8       9      
#>    1  38.5776 38.5776 38.5776 38.5776 38.5776 38.5776 38.5776 38.5776 38.5776
#>    2  17.6257 17.6257 17.6257 17.6257 17.6257 17.6257 17.6257 17.6257 17.6257
#>    3  18.2585 18.2585 18.2585 18.2585 18.2585 18.2585 18.2585 18.2585 18.2585
#>    4  16.5016 16.5016 16.5016 16.5016 16.5016 16.5016 16.5016 16.5016 16.5016
#>    5  18.1636 18.1636 18.1636 18.1636 18.1636 18.1636 18.1636 18.1636 18.1636
#>    6   9.9439  9.9439  9.9439  9.9439  9.9439  9.9439  9.9439  9.9439  9.9439
#>    7  15.9619 15.9619 15.9619 15.9619 15.9619 15.9619 15.9619 15.9619 15.9619
#>    8  14.7203 14.7203 14.7203 14.7203 14.7203 14.7203 14.7203 14.7203 14.7203
#>    9   5.6614  5.6614  5.6614  5.6614  5.6614  5.6614  5.6614  5.6614  5.6614
#>    10 10.6381 10.6381 10.6381 10.6381 10.6381 10.6381 10.6381 10.6381 10.6381
#>      year
#> quant 10     
#>    1  38.5776
#>    2  17.6257
#>    3  18.2585
#>    4  16.5016
#>    5  18.1636
#>    6   9.9439
#>    7  15.9619
#>    8  14.7203
#>    9   5.6614
#>    10 10.6381
#> 
#> units:  NA 

# See how the values obtained by yearSums are being replicated
  yearSums(flq)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1      
#>    1  38.5776
#>    2  17.6257
#>    3  18.2585
#>    4  16.5016
#>    5  18.1636
#>    6   9.9439
#>    7  15.9619
#>    8  14.7203
#>    9   5.6614
#>    10 10.6381
#> 
#> units:  NA 

# Get the proportions by quant
  flq / quantTotals(flq)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1         2         3         4         5         6         7        
#>    1  0.0077551 0.0616871 0.1018878 0.0054608 0.0190057 0.1108415 0.4276760
#>    2  0.2447784 0.0521962 0.2018799 0.0322176 0.0760834 0.1926591 0.0309530
#>    3  0.0247254 0.0594526 0.1924387 0.0036761 0.3256938 0.0435257 0.0166490
#>    4  0.1478434 0.1493119 0.0443665 0.1391403 0.0825659 0.0286720 0.1044930
#>    5  0.0237746 0.0406489 0.0072697 0.1918976 0.3026471 0.0091886 0.1547708
#>    6  0.2282193 0.0617334 0.0952930 0.0727171 0.2094592 0.0654451 0.0343156
#>    7  0.0714054 0.0699881 0.0718073 0.1694787 0.0263568 0.0936493 0.1652904
#>    8  0.3530713 0.1156419 0.0917717 0.1330929 0.0434657 0.1224001 0.0457977
#>    9  0.0527130 0.0671012 0.0317914 0.3001304 0.1113755 0.1603558 0.0194972
#>    10 0.1227967 0.0156051 0.0628475 0.0539318 0.1600635 0.1675983 0.0768842
#>      year
#> quant 8         9         10       
#>    1  0.1026051 0.1005887 0.0624923
#>    2  0.0182967 0.0246327 0.1263031
#>    3  0.0161195 0.0134907 0.3042284
#>    4  0.1537036 0.0213917 0.1285118
#>    5  0.0782445 0.0644030 0.1271552
#>    6  0.0713831 0.0905883 0.0708458
#>    7  0.2107064 0.0095377 0.1117798
#>    8  0.0716684 0.0039693 0.0191209
#>    9  0.1453008 0.0463893 0.0653454
#>    10 0.0747449 0.2021087 0.0634193
#> 
#> units:  NA 

# or year
  flq / yearTotals(flq)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1         2         3         4         5         6         7        
#>    1  0.0164855 0.2057415 0.2432310 0.0147446 0.0360164 0.2715849 0.6233981
#>    2  0.2377376 0.0795386 0.2201917 0.0397449 0.0658746 0.2156770 0.0206141
#>    3  0.0248764 0.0938488 0.2174299 0.0046978 0.2921169 0.0504753 0.0114860
#>    4  0.1344332 0.2130164 0.0453046 0.1607020 0.0669282 0.0300505 0.0651522
#>    5  0.0237954 0.0638326 0.0081711 0.2439569 0.2700350 0.0106003 0.1062199
#>    6  0.1250506 0.0530723 0.0586378 0.0506097 0.1023144 0.0413333 0.0128933
#>    7  0.0628049 0.0965830 0.0709274 0.1893396 0.0206661 0.0949417 0.0996890
#>    8  0.2863892 0.1471715 0.0835962 0.1371238 0.0314300 0.1144367 0.0254727
#>    9  0.0164444 0.0328431 0.0111376 0.1189248 0.0309737 0.0576598 0.0041707
#>    10 0.0719828 0.0143523 0.0413727 0.0401560 0.0836447 0.1132404 0.0309041
#>      year
#> quant 8         9         10       
#>    1  0.2590747 0.4038838 0.1307608
#>    2  0.0211077 0.0451888 0.1207468
#>    3  0.0192636 0.0256373 0.3012871
#>    4  0.1660089 0.0367403 0.1150231
#>    5  0.0930200 0.1217530 0.1252713
#>    6  0.0464592 0.0937559 0.0382107
#>    7  0.2201318 0.0158453 0.0967750
#>    8  0.0690501 0.0060814 0.0152665
#>    9  0.0538405 0.0273345 0.0200656
#>    10 0.0520435 0.2237797 0.0365932
#> 
#> units:  NA