Methods quantTotals

Methods to compute totals over selected dimensions of FLQuant objects These methods return an object of same dimensions as the input but with the sums along the first (yearTotals) or second dimension (quantTotals). Although the names might appear contradictory, it must be noted that what each method really returns are the totals over the selected dimension.

quantTotals(x, ...)

Generic function

quantTotals(x)

yearTotals(x)

See also

FLQuant

Examples

flq <- FLQuant(rlnorm(100), dim=c(10,10))
quantTotals(flq)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1       2       3       4       5       6       7       8       9      
#>    1  35.3206 35.3206 35.3206 35.3206 35.3206 35.3206 35.3206 35.3206 35.3206
#>    2  14.7977 14.7977 14.7977 14.7977 14.7977 14.7977 14.7977 14.7977 14.7977
#>    3  25.7589 25.7589 25.7589 25.7589 25.7589 25.7589 25.7589 25.7589 25.7589
#>    4  22.4847 22.4847 22.4847 22.4847 22.4847 22.4847 22.4847 22.4847 22.4847
#>    5   7.3188  7.3188  7.3188  7.3188  7.3188  7.3188  7.3188  7.3188  7.3188
#>    6  23.3671 23.3671 23.3671 23.3671 23.3671 23.3671 23.3671 23.3671 23.3671
#>    7  23.1012 23.1012 23.1012 23.1012 23.1012 23.1012 23.1012 23.1012 23.1012
#>    8  17.1677 17.1677 17.1677 17.1677 17.1677 17.1677 17.1677 17.1677 17.1677
#>    9   8.0095  8.0095  8.0095  8.0095  8.0095  8.0095  8.0095  8.0095  8.0095
#>    10 18.1923 18.1923 18.1923 18.1923 18.1923 18.1923 18.1923 18.1923 18.1923
#>      year
#> quant 10     
#>    1  35.3206
#>    2  14.7977
#>    3  25.7589
#>    4  22.4847
#>    5   7.3188
#>    6  23.3671
#>    7  23.1012
#>    8  17.1677
#>    9   8.0095
#>    10 18.1923
#> 
#> units:  NA 

# See how the values obtained by yearSums are being replicated
  yearSums(flq)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1      
#>    1  35.3206
#>    2  14.7977
#>    3  25.7589
#>    4  22.4847
#>    5   7.3188
#>    6  23.3671
#>    7  23.1012
#>    8  17.1677
#>    9   8.0095
#>    10 18.1923
#> 
#> units:  NA 

# Get the proportions by quant
  flq / quantTotals(flq)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1         2         3         4         5         6         7        
#>    1  0.0165309 0.0261081 0.0493415 0.0328247 0.0207092 0.0135612 0.0725860
#>    2  0.0668568 0.0690730 0.1185312 0.0688462 0.2427471 0.0360923 0.2605697
#>    3  0.1589810 0.1409390 0.0080560 0.1460981 0.0104552 0.1528695 0.0106336
#>    4  0.0214140 0.0298584 0.0597591 0.0691864 0.2495590 0.2113575 0.1418519
#>    5  0.1049300 0.1206478 0.0527402 0.1393330 0.0941516 0.0816468 0.0246745
#>    6  0.0654298 0.0630311 0.2300614 0.0817010 0.1254898 0.0399770 0.0083892
#>    7  0.0077221 0.3976830 0.0280535 0.0135329 0.0302351 0.0627803 0.0633551
#>    8  0.0485459 0.0734624 0.0939537 0.0931964 0.0248424 0.0305014 0.0902100
#>    9  0.2420324 0.1036127 0.1509464 0.1341014 0.0399370 0.0754756 0.1239413
#>    10 0.0830732 0.0293425 0.2273497 0.2943019 0.0676079 0.0220646 0.0860021
#>      year
#> quant 8         9         10       
#>    1  0.0202881 0.7083158 0.0397345
#>    2  0.0197518 0.0700248 0.0475070
#>    3  0.0185813 0.0032530 0.3501333
#>    4  0.0205181 0.0496235 0.1468721
#>    5  0.1783414 0.1024844 0.1010503
#>    6  0.2055800 0.1002966 0.0800440
#>    7  0.0503124 0.0090595 0.3372662
#>    8  0.1279128 0.3846394 0.0327353
#>    9  0.0363401 0.0329688 0.0606443
#>    10 0.0523915 0.0832786 0.0545880
#> 
#> units:  NA 

# or year
  flq / yearTotals(flq)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1         2         3         4         5         6         7        
#>    1  0.0452327 0.0451733 0.0946343 0.0617790 0.0443302 0.0336969 0.1619652
#>    2  0.0766422 0.0500704 0.0952435 0.0542859 0.2176989 0.0375726 0.2435901
#>    3  0.3172489 0.1778427 0.0112682 0.2005322 0.0163218 0.2770194 0.0173041
#>    4  0.0373003 0.0328875 0.0729624 0.0828934 0.3400696 0.3343235 0.2014945
#>    5  0.0594931 0.0432550 0.0209599 0.0543382 0.0417614 0.0420379 0.0114085
#>    6  0.1184425 0.0721501 0.2919149 0.1017289 0.1777136 0.0657169 0.0123841
#>    7  0.0138196 0.4500374 0.0351907 0.0166585 0.0423305 0.1020281 0.0924606
#>    8  0.0645641 0.0617809 0.0875856 0.0852556 0.0258472 0.0368378 0.0978379
#>    9  0.1501782 0.0406533 0.0656503 0.0572337 0.0193860 0.0425280 0.0627138
#>    10 0.1170784 0.0261495 0.2245902 0.2852946 0.0745407 0.0282389 0.0988413
#>      year
#> quant 8         9         10       
#>    1  0.0566012 0.6424900 0.0522312
#>    2  0.0230864 0.0266108 0.0261630
#>    3  0.0378058 0.0021519 0.3356565
#>    4  0.0364400 0.0286540 0.1229025
#>    5  0.1030972 0.0192623 0.0275240
#>    6  0.3794377 0.0601869 0.0696094
#>    7  0.0918046 0.0053746 0.2899621
#>    8  0.1734524 0.1695806 0.0209152
#>    9  0.0229904 0.0067814 0.0180772
#>    10 0.0752842 0.0389075 0.0369590
#> 
#> units:  NA