Methods quantTotals

Methods to compute totals over selected dimensions of FLQuant objects These methods return an object of same dimensions as the input but with the sums along the first (yearTotals) or second dimension (quantTotals). Although the names might appear contradictory, it must be noted that what each method really returns are the totals over the selected dimension.

quantTotals(x, ...)

Generic function

quantTotals(x)

yearTotals(x)

See also

Author

The FLR Team

Examples

flq <- FLQuant(rlnorm(100), dim=c(10,10))
quantTotals(flq)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1      2      3      4      5      6      7      8      9      10    
#>    1  19.873 19.873 19.873 19.873 19.873 19.873 19.873 19.873 19.873 19.873
#>    2  25.220 25.220 25.220 25.220 25.220 25.220 25.220 25.220 25.220 25.220
#>    3  10.999 10.999 10.999 10.999 10.999 10.999 10.999 10.999 10.999 10.999
#>    4  15.701 15.701 15.701 15.701 15.701 15.701 15.701 15.701 15.701 15.701
#>    5  13.216 13.216 13.216 13.216 13.216 13.216 13.216 13.216 13.216 13.216
#>    6  16.369 16.369 16.369 16.369 16.369 16.369 16.369 16.369 16.369 16.369
#>    7  13.594 13.594 13.594 13.594 13.594 13.594 13.594 13.594 13.594 13.594
#>    8  19.814 19.814 19.814 19.814 19.814 19.814 19.814 19.814 19.814 19.814
#>    9  20.770 20.770 20.770 20.770 20.770 20.770 20.770 20.770 20.770 20.770
#>    10 33.210 33.210 33.210 33.210 33.210 33.210 33.210 33.210 33.210 33.210
#> 
#> units:  NA 

# See how the values obtained by yearSums are being replicated
  yearSums(flq)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1     
#>    1  19.873
#>    2  25.220
#>    3  10.999
#>    4  15.701
#>    5  13.216
#>    6  16.369
#>    7  13.594
#>    8  19.814
#>    9  20.770
#>    10 33.210
#> 
#> units:  NA 

# Get the proportions by quant
  flq / quantTotals(flq)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1         2         3         4         5         6         7        
#>    1  0.0222383 0.0332753 0.2311280 0.2557631 0.1252179 0.1445951 0.0342522
#>    2  0.0172367 0.0038922 0.0622304 0.0400407 0.0305306 0.2247676 0.0283643
#>    3  0.1340137 0.0352438 0.2000963 0.0406634 0.0766029 0.0143575 0.1222992
#>    4  0.0493400 0.0358845 0.0490557 0.1544690 0.0388780 0.1469465 0.2417634
#>    5  0.4711680 0.0503852 0.0190894 0.0895051 0.0920876 0.0305282 0.0982096
#>    6  0.1964019 0.0338115 0.0547873 0.4239039 0.0549989 0.0756498 0.0454504
#>    7  0.0686686 0.0766660 0.0780432 0.3493816 0.0771965 0.0666157 0.0737917
#>    8  0.0327337 0.0345072 0.3557645 0.0512350 0.0152455 0.1533453 0.0226226
#>    9  0.1084722 0.4593960 0.0326753 0.0816666 0.0117624 0.0301680 0.0250640
#>    10 0.0347971 0.0722080 0.1491953 0.2373982 0.0332403 0.1056377 0.1221208
#>      year
#> quant 8         9         10       
#>    1  0.0340240 0.0588936 0.0606126
#>    2  0.0460821 0.4988350 0.0480204
#>    3  0.0843162 0.0122547 0.2801522
#>    4  0.0499469 0.2109894 0.0227265
#>    5  0.0240100 0.0568825 0.0681342
#>    6  0.0235610 0.0700753 0.0213600
#>    7  0.0985525 0.0751800 0.0359041
#>    8  0.0584644 0.2156811 0.0604007
#>    9  0.0502750 0.1662248 0.0342957
#>    10 0.1499777 0.0362433 0.0591815
#> 
#> units:  NA 

# or year
  flq / yearTotals(flq)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1         2         3         4         5         6         7        
#>    1  0.0251711 0.0398473 0.1911769 0.1567244 0.2611655 0.1386331 0.0466028
#>    2  0.0247594 0.0059151 0.0653237 0.0311376 0.0808108 0.2734841 0.0489757
#>    3  0.0839491 0.0233574 0.0915984 0.0137901 0.0884220 0.0076183 0.0920902
#>    4  0.0441214 0.0339494 0.0320569 0.0747807 0.0640623 0.1113069 0.2598748
#>    5  0.3546513 0.0401240 0.0105002 0.0364730 0.1277249 0.0194643 0.0888593
#>    6  0.1831040 0.0333497 0.0373262 0.2139525 0.0944832 0.0597410 0.0509346
#>    7  0.0531665 0.0627999 0.0441567 0.1464462 0.1101352 0.0436888 0.0686770
#>    8  0.0369403 0.0411994 0.2933931 0.0313018 0.0317026 0.1465846 0.0306881
#>    9  0.1283200 0.5749616 0.0282473 0.0523020 0.0256402 0.0302298 0.0356410
#>    10 0.0658169 0.1444961 0.2062205 0.2430917 0.1158532 0.1692491 0.2776565
#>      year
#> quant 8         9         10       
#>    1  0.0529249 0.0402905 0.1050518
#>    2  0.0909686 0.4330879 0.1056213
#>    3  0.0725858 0.0046398 0.2687203
#>    4  0.0613809 0.1140366 0.0311188
#>    5  0.0248366 0.0258784 0.0785291
#>    6  0.0301870 0.0394866 0.0304925
#>    7  0.1048629 0.0351816 0.0425662
#>    8  0.0906715 0.1471129 0.1043726
#>    9  0.0817338 0.1188516 0.0621234
#>    10 0.3898481 0.0414339 0.1714040
#> 
#> units:  NA