Method show

Standard complete display of an object contents in an interactive session. Objects of class FLArray with length > 1 alonArray sixth dimension (iter) are output in a summarised form, as median(mad), where mad is the median absolute deviation. See mad.

Usage

# S4 method for class 'FLArray'
show(object)

# S4 method for class 'FLArray'
print(x)

Details

The same format is used for objects of class FLPar with length > 1 on the last dimension (iter).

Generic function

show(object)

show(object)

See also

Author

The FLR Team

Examples

# no 'iter'
  flq <- FLQuant(rnorm(80), dim=c(4,20), quant='age', units='kg')
  flq
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1       2       3       4       5       6       7       8       9      
#>   1  1.9689  0.7547  1.2813 -0.6067  0.3531 -0.0928 -0.4314 -0.7416  0.1343
#>   2  1.4246  0.7543 -1.0181 -1.2673  0.6946  0.5561 -0.5547  1.7454  0.0664
#>   3  1.6707 -1.6545 -2.0456 -0.1158 -1.0479  0.3939 -1.6194  1.5964  0.7333
#>   4  1.3983 -0.8773 -0.5659  0.8310 -0.7683 -0.8635 -1.0393  0.8814 -1.5713
#>    year
#> age 10      11      12      13      14      15      16      17      18     
#>   1  0.3898 -0.8074  1.8014 -1.1418 -0.9801 -0.4088 -0.2007  0.7351 -1.0549
#>   2 -0.4147 -1.5721 -1.1692 -0.0613  0.6202  1.1144  1.2800  1.9829 -0.6860
#>   3  1.4237 -0.6937 -0.2360 -0.2946 -0.1128 -0.6219 -1.5845 -1.0265 -1.7087
#>   4  0.1656 -0.7367  1.9859  0.1321 -1.0302 -1.9053 -0.6399  1.6206  1.4134
#>    year
#> age 19      20     
#>   1 -1.0139 -0.9474
#>   2  0.9816 -1.0705
#>   3 -0.5048  0.9192
#>   4  0.9338  2.4380
#> 
#> units:  kg 

# with 'iter'
  flq <- FLQuant(rnorm(800), dim=c(4,20,1,1,1,10), quant='age', units='kg')
  flq
#> An object of class "FLQuant"
#> iters:  10 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1              2              3              4              5             
#>   1  0.1144(0.668)  0.5059(0.597) -0.3418(0.522) -0.4550(0.442) -0.3205(0.369)
#>   2 -0.4907(1.444)  0.3590(1.670) -0.1162(0.736)  0.2656(1.129) -0.6552(0.657)
#>   3 -0.1297(1.017) -0.1584(0.765)  0.0933(0.482)  0.1975(0.792) -0.0397(0.802)
#>   4  0.0797(0.825) -0.3877(0.664) -0.5352(0.295)  0.6076(1.090)  0.3320(1.667)
#>    year
#> age 6              7              8              9              10            
#>   1 -0.2005(0.945)  0.1474(0.922) -0.3551(1.881) -0.3945(0.721) -0.7061(0.856)
#>   2  0.8885(0.933) -0.3535(1.108) -0.0316(0.728)  0.3876(0.602) -0.1582(0.530)
#>   3  0.1132(0.799)  0.2065(0.417) -0.1742(1.073) -0.0211(0.328) -0.1808(1.351)
#>   4 -0.1046(1.340) -0.1219(0.902)  0.1932(1.329) -0.3543(1.314)  0.7195(1.091)
#>    year
#> age 11             12             13             14             15            
#>   1  0.4676(0.623) -0.4195(1.366) -0.2379(1.401) -0.8183(0.647) -0.5283(1.192)
#>   2  0.0259(0.805)  0.5662(0.872)  0.2210(1.520) -0.0902(0.240)  0.5920(0.958)
#>   3  0.1184(0.873)  0.7422(0.636) -0.0244(0.788)  0.5280(0.701) -0.0468(0.822)
#>   4  0.1629(0.841)  0.2412(0.530)  0.4678(0.790)  0.0917(0.348) -0.2259(0.593)
#>    year
#> age 16             17             18             19             20            
#>   1  0.2801(1.406)  0.1562(0.999) -0.3213(0.856) -0.2233(0.644) -0.4665(0.598)
#>   2  0.2609(0.729) -0.0216(0.564) -0.1498(0.974)  0.5651(0.437)  0.7218(0.537)
#>   3 -0.1300(1.079)  0.0676(0.486) -0.1913(0.242)  0.5229(0.635)  0.0649(1.119)
#>   4 -0.3939(1.051) -0.2948(0.313) -0.4457(0.842)  0.2684(0.937)  0.4319(0.636)
#> 
#> units:  kg 

# no 'iter'
  flq <- FLQuant(rnorm(80), dim=c(4,20), quant='age', units='kg')
  print(flq)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1        2        3        4        5        6        7        8       
#>   1 -1.42750  0.10770  1.75494 -0.22130  0.15833  0.31297 -0.56692 -1.31361
#>   2 -0.28917 -0.31297  1.49195 -0.58088 -0.68052 -0.42846 -0.53705 -0.41579
#>   3 -0.23162  1.61953 -0.21610 -0.40791  0.25933 -1.20183  0.01234 -0.56465
#>   4  0.10801  0.00358  0.15862 -1.73155  0.35792  0.14314  0.13500  0.47246
#>    year
#> age 9        10       11       12       13       14       15       16      
#>   1 -1.62311  2.05337  1.07333 -1.63066 -1.61676 -0.58427  0.39066  2.23443
#>   2  1.00470 -0.12086 -0.58736 -0.99563  0.14657  0.08536 -0.27317 -0.01571
#>   3 -0.17441  0.29930  0.94404 -0.38025 -0.27040 -0.14937  0.23710 -0.89101
#>   4 -0.93493 -1.05931 -0.06643  0.16307  0.49622  0.99456 -0.81332  0.06408
#>    year
#> age 17       18       19       20      
#>   1 -0.51628  0.47571  1.04122 -0.13877
#>   2  1.02562 -0.22306  0.44452 -0.54147
#>   3 -0.42974 -0.76841 -2.27069  0.94408
#>   4  1.45493 -0.34793  0.94599  0.56966
#> 
#> units:  kg 

# with 'iter'
  flq <- FLQuant(rnorm(800), dim=c(4,20,1,1,1,10), quant='age', units='kg')
  print(flq)
#> An object of class "FLQuant"
#> iters:  10 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1               2               3               4              
#>   1  0.31808(0.571)  0.34238(0.879)  0.03490(1.340) -0.03663(0.877)
#>   2 -0.08389(0.698)  0.23320(0.596) -0.06539(1.239)  0.71328(1.065)
#>   3  0.39290(0.751)  0.87579(0.652) -0.14346(0.896) -0.01298(1.278)
#>   4  0.48086(0.533)  0.19069(1.078)  0.31578(0.613)  0.19438(0.770)
#>    year
#> age 5               6               7               8              
#>   1  0.35240(0.542)  0.20593(1.230)  0.19095(0.767)  0.26942(0.668)
#>   2  0.25504(0.869)  0.00156(0.604)  0.24380(0.642) -0.32949(0.761)
#>   3  0.03251(0.953)  0.40428(0.650) -0.27740(1.307) -0.88371(0.484)
#>   4 -0.80751(0.812) -0.05616(0.641)  0.11569(0.707)  0.46866(0.730)
#>    year
#> age 9               10              11              12             
#>   1 -0.65405(0.360)  0.21549(0.789) -0.15101(1.297) -0.30782(1.120)
#>   2 -0.26554(1.184)  0.20096(1.016) -0.31567(1.500)  0.04768(1.302)
#>   3 -0.17511(0.391)  0.75097(0.923)  0.00923(0.869) -0.23566(1.310)
#>   4 -0.81367(0.627)  0.06233(1.305) -0.54924(0.833)  0.33269(1.059)
#>    year
#> age 13              14              15              16             
#>   1 -0.05947(0.885)  0.14169(0.670) -0.15553(1.495)  0.06160(1.160)
#>   2  0.27583(1.578)  0.19309(0.528) -0.36745(0.417)  0.06643(1.009)
#>   3 -0.19389(0.683) -0.26118(1.066)  0.21743(0.936)  0.14658(1.710)
#>   4 -0.09150(0.827) -0.03163(1.342) -0.10790(0.956)  0.56864(0.870)
#>    year
#> age 17              18              19              20             
#>   1  0.46916(1.020) -0.39623(1.219)  0.35855(0.942) -0.11907(0.896)
#>   2 -0.14145(1.051) -0.12738(0.995) -0.06890(0.695) -0.14660(0.976)
#>   3  0.48043(0.951)  0.29468(1.185)  0.16376(0.781) -0.02563(0.882)
#>   4  0.25811(0.473) -0.27875(1.016)  0.38644(0.761) -0.19465(0.654)
#> 
#> units:  kg