Method show

Standard display of an object contents in an interactive session. Objects of class FLQuant with length > 1 along the sixth dimension (iter) are output in a summarised form, as median(mad), where mad is the median absolute deviation. See mad.

# S4 method for FLArray
show(object)

Details

The same format is used for objects of class FLPar with length > 1 on the last dimension (iter).

Generic function

show(object)

See also

Author

The FLR Team

Examples

# no 'iter'
  flq <- FLQuant(rnorm(80), dim=c(4,20), quant='age', units='kg')
  flq

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1         2         3         4         5         6         7        
#>   1 -0.382067  0.662967  0.815614 -0.173074  0.597373  1.294623  0.722492
#>   2  0.695212  2.052921 -0.407486  0.961231 -0.090217  1.811744  0.063760
#>   3 -1.255059  1.039265  1.539860 -0.665834 -1.803165 -1.907233  0.428448
#>   4  0.526065 -0.894941 -0.830625  1.168181  0.098290  0.691816 -0.923430
#>    year
#> age 8         9         10        11        12        13        14       
#>   1  1.020554  0.527543 -0.565474 -0.854946  1.043942  0.328088  1.665067
#>   2 -0.720291  1.473150 -1.540570 -0.053979 -1.463891  0.666904 -0.139416
#>   3  1.686743 -1.172696 -0.673382  0.402331  0.261062 -1.344601  0.597639
#>   4 -0.036980  0.546661 -1.212156  1.648685 -0.532787  0.117347  1.471131
#>    year
#> age 15        16        17        18        19        20       
#>   1  0.438538 -0.123762  1.127649  0.826410  1.311104  1.838461
#>   2  0.239954  0.368577 -0.076716  0.910800  0.665271  0.577038
#>   3 -0.247162 -1.254689  0.502071 -0.279890  1.238448 -1.364207
#>   4 -0.778187 -0.936184  0.057296  0.209825 -0.809189 -1.553709
#> 
#> units:  kg 

# with 'iter'
  flq <- FLQuant(rnorm(800), dim=c(4,20,1,1,1,10), quant='age', units='kg')
  flq

#> An object of class "FLQuant"
#> iters:  10 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1                  2                  3                  4                 
#>   1 -0.66421146(1.511) -0.07088577(0.846) -0.20968082(1.031) -0.07163740(1.472)
#>   2 -0.28242857(0.907)  0.60098725(1.226) -0.14344672(0.930) -0.29085788(0.487)
#>   3 -0.55796682(0.597) -0.08891829(0.944) -0.08880975(0.318) -0.11619816(0.911)
#>   4 -0.24386911(1.436)  0.31536879(0.745) -0.13639467(0.672)  0.26950214(0.547)
#>    year
#> age 5                  6                  7                  8                 
#>   1 -0.22813944(0.625)  0.33909138(0.490) -0.06203316(0.987)  0.60104839(0.446)
#>   2 -0.04025446(0.948)  0.31590341(0.695)  0.51255479(0.634) -0.27997772(0.573)
#>   3 -0.04012695(0.340) -0.47725285(0.423)  0.17788439(0.709) -0.01191644(1.273)
#>   4  0.34155312(0.646) -0.05390258(0.442) -0.08971690(0.586) -0.26551873(0.762)
#>    year
#> age 9                  10                 11                 12                
#>   1  0.20512412(0.589) -0.38363301(0.952) -0.10091267(0.817)  0.22432143(0.641)
#>   2 -0.05655899(0.856) -0.58961472(0.638) -0.59955486(0.865) -0.14746087(0.527)
#>   3  0.35891162(0.550)  0.39739118(0.885) -0.40835542(0.761) -0.32154346(0.724)
#>   4  0.29320601(0.996) -0.28781549(0.920) -0.19023080(0.789) -0.04162828(0.872)
#>    year
#> age 13                 14                 15                 16                
#>   1  0.00211365(0.501)  0.40427438(0.669) -0.06241383(1.427) -0.04203392(0.341)
#>   2 -0.36762397(1.158) -0.82937957(1.573) -0.30751013(0.318) -0.72665018(0.941)
#>   3  0.37576164(1.211)  0.54868588(1.591) -0.64140036(0.920)  0.78280306(0.562)
#>   4 -0.24452423(0.677) -0.08265155(0.791) -0.19126380(0.571) -0.33430009(0.892)
#>    year
#> age 17                 18                 19                 20                
#>   1 -0.86196387(1.384)  0.10182223(1.211)  0.42695801(0.706) -0.00082856(0.799)
#>   2  0.06489357(1.442)  0.34758093(0.437)  0.09740406(1.307) -0.52093009(1.645)
#>   3 -0.11802375(0.574) -0.50244185(1.038)  0.05920272(1.283)  0.50954951(1.191)
#>   4 -0.18483306(0.847)  0.40895525(0.806)  0.15287784(0.926)  0.07533767(1.318)
#> 
#> units:  kg