Method show

Standard display of an object contents in an interactive session. Objects of class FLQuant with length > 1 along the sixth dimension (iter) are output in a summarised form, as median(mad), where mad is the median absolute deviation. See mad.

# S4 method for FLArray
show(object)

Details

The same format is used for objects of class FLPar with length > 1 on the last dimension (iter).

Generic function

show(object)

See also

Examples

# no 'iter'
  flq <- FLQuant(rnorm(80), dim=c(4,20), quant='age', units='kg')
  flq

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1          2          3          4          5          6         
#>   1  1.4175776  0.3697536  0.4797195  0.7117035 -0.5173404 -0.0068206
#>   2 -1.0258831  0.2280138 -0.9448011  1.0549057 -1.0708702 -0.5415214
#>   3  0.3030562  0.7058424 -0.5078071 -2.0386944 -1.1658526 -0.3720247
#>   4  0.3516501 -1.1001776  0.4179661  0.2685623  1.5531689  2.5603311
#>    year
#> age 7          8          9          10         11         12        
#>   1 -1.1947754 -1.1976701 -0.1195907  0.0633525  0.2662276  2.2613427
#>   2  0.3249450 -0.0349465  0.9414032  0.5594499 -1.3491959  1.5988160
#>   3  0.5191995  0.8382481  0.2212313  0.4238692 -0.3210743  0.0204583
#>   4 -1.2876453 -0.0701980  2.0009655  0.5994409 -0.9441326  0.3369306
#>    year
#> age 13         14         15         16         17         18        
#>   1  0.1419021 -0.0575766 -0.5206190 -0.0824558 -1.0213645 -1.4211684
#>   2 -1.2533916 -2.2447618  0.4053855 -0.3126234  0.1441703 -0.7433550
#>   3 -2.1931488  0.7177715  0.6983281 -0.6191096 -0.2227379  0.4934890
#>   4 -0.1058296  0.4384489 -1.0322589  0.5162480 -1.6201217 -1.2371822
#>    year
#> age 19         20        
#>   1 -1.4283957  1.3054933
#>   2  0.1052284 -0.2922847
#>   3 -0.0273812  0.7596918
#>   4 -0.7283719 -1.8766007
#> 
#> units:  kg 

# with 'iter'
  flq <- FLQuant(rnorm(800), dim=c(4,20,1,1,1,10), quant='age', units='kg')
  flq

#> An object of class "FLQuant"
#> iters:  10 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1                 2                 3                 4                
#>   1  0.0683721(0.988) -0.0640551(0.529)  0.3038643(0.227)  0.1470192(0.713)
#>   2  0.4030989(0.745)  0.0171047(0.891)  0.2039202(0.667)  0.1422683(0.812)
#>   3 -0.5147766(0.852)  0.3329365(0.898) -0.2868419(0.829) -0.1936920(0.659)
#>   4  0.2795495(0.468) -0.1678714(0.887)  0.0564621(1.168)  0.1568689(1.573)
#>    year
#> age 5                 6                 7                 8                
#>   1  0.2675702(0.632)  0.4483864(0.960) -0.2150566(1.643) -0.4535624(1.466)
#>   2  0.6810675(0.682)  0.1455449(1.185) -0.3984221(1.182)  0.1306944(1.144)
#>   3  0.1158763(0.516) -0.4805621(0.586)  0.0993014(1.014)  0.0201888(0.854)
#>   4  0.2166143(1.630)  0.4316111(0.943)  0.3454788(0.651) -0.1912657(0.721)
#>    year
#> age 9                 10                11                12               
#>   1 -0.0318063(0.959)  0.1652270(0.572)  0.4414173(0.913) -0.3777181(0.607)
#>   2  0.1691378(0.366)  0.5922544(0.868) -0.3386157(0.801)  0.2106468(1.530)
#>   3 -0.0940108(0.843)  0.3648390(0.279)  0.6234538(1.079)  0.3022529(0.571)
#>   4  0.5671030(0.600)  0.0437387(1.035)  0.2475237(0.918)  0.3491853(0.790)
#>    year
#> age 13                14                15                16               
#>   1 -0.1255425(1.168) -0.7371511(0.854) -0.2789615(0.650) -0.3899848(1.014)
#>   2  0.0753220(1.142) -0.0722542(1.219) -0.4404240(0.775)  0.0745463(1.201)
#>   3 -0.3234403(0.735) -0.0332767(1.077)  0.2795983(0.539) -0.3770571(1.009)
#>   4  0.2304744(0.391) -0.1850169(0.987)  0.3832265(0.850)  0.1831642(1.376)
#>    year
#> age 17                18                19                20               
#>   1 -0.0854048(0.473)  0.0563853(0.522)  0.1742273(1.221) -0.5422025(0.863)
#>   2 -0.0797258(0.777) -0.0030458(0.705) -0.6756818(1.254)  0.4072805(0.895)
#>   3 -0.2397169(0.784) -0.1938120(0.778)  0.0159480(1.101)  0.2277230(0.848)
#>   4 -0.0509196(0.487)  0.2328993(1.162) -0.1250849(1.403) -0.3680383(0.750)
#> 
#> units:  kg