Method show

Standard display of an object contents in an interactive session. Objects of class FLQuant with length > 1 along the sixth dimension (iter) are output in a summarised form, as median(mad), where mad is the median absolute deviation. See mad.

Usage

# S4 method for class 'FLArray'
show(object)

Details

The same format is used for objects of class FLPar with length > 1 on the last dimension (iter).

Generic function

show(object)

See also

Author

The FLR Team

Examples

# no 'iter'
  flq <- FLQuant(rnorm(80), dim=c(4,20), quant='age', units='kg')
  flq
#> An x of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1       2       3       4       5       6       7       8       9      
#>   1  1.2713  0.0329  0.3757 -0.4390 -0.8383  1.1631  0.8187  0.1533 -0.7127
#>   2  0.2440  2.0947  1.5724 -0.9957 -0.5069  1.1667 -0.4809  0.4956 -0.4660
#>   3  0.8486  1.1307 -1.1362 -0.0523 -0.2039  0.8414 -0.8644 -0.3730  0.1871
#>   4  0.0712  1.7538 -0.4906  0.2683  0.3112 -0.6502 -0.9059  0.3142  0.1106
#>    year
#> age 10      11      12      13      14      15      16      17      18     
#>   1  0.5331  1.8302  0.3439  0.2268  0.3603  0.0150  1.0165 -0.7381  0.6274
#>   2 -0.8126  0.6897 -0.4465  0.6428 -0.5128  0.1901 -0.1836  0.5261  0.9250
#>   3 -0.8133 -0.2227 -0.7947  0.2197  0.1100 -0.7795 -0.9890 -0.4387  0.1602
#>   4  0.4686  0.3385 -0.7823 -0.0152  1.3128 -1.3669  0.4405  0.6712 -1.1199
#>    year
#> age 19      20     
#>   1  1.8476 -0.9952
#>   2 -0.2695  0.6021
#>   3 -0.2789 -1.8563
#>   4  0.2506 -0.6104
#> 
#> units:  kg 

# with 'iter'
  flq <- FLQuant(rnorm(800), dim=c(4,20,1,1,1,10), quant='age', units='kg')
  flq
#> An x of class "FLQuant"
#> iters:  10 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1               2               3               4              
#>   1 -0.30893(0.899)  0.06314(0.592)  0.25519(0.591)  0.08628(1.159)
#>   2  0.89428(0.616) -0.01788(0.910) -0.02904(0.976) -0.04503(0.705)
#>   3  0.01465(0.744) -0.34310(1.009) -0.17362(1.202) -0.14751(0.764)
#>   4  0.33684(0.777)  0.26620(0.958) -0.21653(1.078) -0.39970(0.634)
#>    year
#> age 5               6               7               8              
#>   1  0.01146(1.308) -0.30463(0.304) -0.13510(0.919) -0.49446(1.792)
#>   2  0.26939(1.314)  0.20288(0.523) -0.40814(0.475) -0.17465(0.869)
#>   3 -0.15600(1.105) -0.10457(0.672)  0.05793(0.899)  0.17127(1.382)
#>   4 -0.23905(0.822) -0.23016(0.916)  0.23409(0.449) -0.14446(1.356)
#>    year
#> age 9               10              11              12             
#>   1 -0.36682(1.180)  0.19110(1.253)  0.33193(0.479)  0.15309(0.997)
#>   2  0.32996(0.976)  0.53838(1.010) -0.21584(1.170)  0.70369(1.290)
#>   3  0.03548(1.055)  0.00316(0.898) -0.39115(1.132) -0.01334(1.542)
#>   4 -0.41604(1.094) -0.25958(1.280)  0.10286(0.807) -0.06193(0.887)
#>    year
#> age 13              14              15              16             
#>   1  0.31433(0.677) -0.67119(0.554) -0.53408(0.489)  0.41956(0.884)
#>   2  0.01794(0.543) -0.29302(0.862) -0.58340(0.522)  0.15344(0.955)
#>   3  0.23479(0.734) -0.13371(0.770)  0.31448(0.872)  0.29399(1.431)
#>   4  0.41039(1.485)  0.10704(0.661)  0.41497(0.945) -0.10320(0.644)
#>    year
#> age 17              18              19              20             
#>   1 -0.44091(0.625)  0.56305(0.390) -0.25187(1.103)  0.46553(0.714)
#>   2 -0.31650(1.062) -0.36082(0.396) -0.44306(0.566)  0.70606(0.436)
#>   3  0.17383(1.078)  0.19501(1.081)  0.07630(0.903) -0.53911(0.851)
#>   4 -0.68295(0.679) -0.13808(0.420) -0.04652(1.498) -0.45807(0.530)
#> 
#> units:  kg