Method show

Standard display of an object contents in an interactive session. Objects of class FLQuant with length > 1 along the sixth dimension (iter) are output in a summarised form, as median(mad), where mad is the median absolute deviation. See mad.

# S4 method for FLArray
show(object)

Details

The same format is used for objects of class FLPar with length > 1 on the last dimension (iter).

Generic function

show(object)

See also

FLComp

Examples

# no 'iter'
  flq <- FLQuant(rnorm(80), dim=c(4,20), quant='age', units='kg')
  flq

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1          2          3          4          5          6         
#>   1  1.7107347  1.6890421 -1.2320528 -1.0713005  0.1598333 -0.2788052
#>   2  1.0818325 -0.3088327  1.3668096 -0.2415736 -1.3764010  1.3239398
#>   3  0.4816399  1.3051733  1.1624438 -0.2129905  2.8034198 -0.6728176
#>   4  0.2986779 -1.8455352 -1.8784739 -2.0600328  0.5859310 -0.8757588
#>    year
#> age 7          8          9          10         11         12        
#>   1  0.8058549  0.4910919  1.4548478 -1.0687396 -0.5920325  0.6927434
#>   2  0.1818426  1.4054670 -2.4728017  0.2485942 -0.5127169 -0.5592794
#>   3  0.1906407  1.5678740  0.9494104  0.0087216  1.6006151  0.0855997
#>   4 -0.5528425 -0.4543415 -1.3699782 -0.0134129  0.0122653 -1.3190884
#>    year
#> age 13         14         15         16         17         18        
#>   1  0.0504650 -1.2297777  0.3290518 -0.5617825  0.4519208  0.4013079
#>   2 -0.4561743 -0.3361049 -1.0157586  0.8571496 -0.3331017 -0.9536048
#>   3  0.1984053 -0.4228592 -0.0574091  0.1605772 -1.7412438  0.4143226
#>   4  0.4638367 -0.5129392  1.0334341  1.4474257  0.0312717 -0.7416205
#>    year
#> age 19         20        
#>   1 -0.7645804 -0.9646186
#>   2 -0.5801996 -0.0301683
#>   3 -1.3796681  2.2492242
#>   4  1.4602868 -0.8601541
#> 
#> units:  kg 

# with 'iter'
  flq <- FLQuant(rnorm(800), dim=c(4,20,1,1,1,10), quant='age', units='kg')
  flq

#> An object of class "FLQuant"
#> iters:  10 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1                 2                 3                 4                
#>   1 -0.2813507(0.839) -0.0122614(1.293)  0.2868668(0.930)  0.5722015(0.844)
#>   2  0.2457713(0.552)  0.2766602(0.657)  0.2291091(0.337)  0.0160693(1.241)
#>   3  0.3434944(0.867)  0.4434178(0.898)  0.3636921(0.467)  0.3090718(0.598)
#>   4 -0.0943708(1.081)  0.3112263(0.546) -0.3902902(0.463)  0.3519026(0.807)
#>    year
#> age 5                 6                 7                 8                
#>   1  0.1325904(0.933)  0.2496214(0.878) -0.3236594(0.930)  0.2818203(0.845)
#>   2  0.2952405(0.975) -0.5565622(1.169)  0.4688073(0.982) -0.3502609(1.205)
#>   3 -0.0733706(0.734)  0.1377822(0.468)  0.5148941(1.379)  0.1730433(0.553)
#>   4  0.1813252(0.200) -0.3837379(0.982)  0.6253074(1.368) -0.0580490(1.459)
#>    year
#> age 9                 10                11                12               
#>   1 -0.0719935(0.660) -0.3312671(1.044)  0.0871817(1.187)  0.3434138(0.529)
#>   2  0.3988765(1.046)  0.0023172(0.345)  0.4692642(1.045)  0.2974606(1.466)
#>   3  0.6598014(1.483) -0.0670067(0.833)  0.0257732(0.724) -0.2458399(0.711)
#>   4 -0.1275144(1.208) -0.2936771(0.664)  0.0846874(0.737) -0.3007748(1.282)
#>    year
#> age 13                14                15                16               
#>   1 -0.0662944(0.861) -0.2169032(0.617) -0.0133682(1.323)  0.5247960(0.957)
#>   2 -0.0444874(0.644)  0.3072787(0.314) -0.4586382(0.544) -0.6792492(0.722)
#>   3  0.1469054(0.951)  0.3230443(1.468) -0.3824736(1.281)  0.2053726(0.564)
#>   4  0.4881452(0.875)  0.3919451(0.573)  0.1193747(1.174) -0.0565361(0.735)
#>    year
#> age 17                18                19                20               
#>   1  0.1081572(0.604)  0.5741923(0.768) -0.3649358(0.403) -0.0743102(0.600)
#>   2  0.3628086(0.881)  0.0903149(0.902) -0.1851682(1.559) -0.4023801(0.764)
#>   3 -0.0354174(0.797) -0.4428682(0.877) -0.6884036(0.636)  0.3128164(1.372)
#>   4  0.1439167(0.864) -0.3202839(1.150)  0.3535367(1.279)  0.2866680(0.740)
#> 
#> units:  kg