Method show

Standard display of an object contents in an interactive session. Objects of class FLQuant with length > 1 along the sixth dimension (iter) are output in a summarised form, as median(mad), where mad is the median absolute deviation. See mad.

# S4 method for FLArray
show(object)

Details

The same format is used for objects of class FLPar with length > 1 on the last dimension (iter).

Generic function

show(object)

See also

Examples

# no 'iter'
  flq <- FLQuant(rnorm(80), dim=c(4,20), quant='age', units='kg')
  flq

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1         2         3         4         5         6         7        
#>   1  0.360175 -1.003347  0.406650  1.117579  0.170095  0.901350  0.486099
#>   2 -0.494506  0.754435  0.804831  1.637659  1.007737 -0.769411 -1.151365
#>   3 -0.413291 -0.178355 -1.010925  0.412339 -0.794225  0.555077 -1.965689
#>   4  1.131109  0.289623 -0.240162 -2.223200  0.106403  0.152232  0.585826
#>    year
#> age 8         9         10        11        12        13        14       
#>   1  1.058928 -1.119406 -0.679898 -0.788722 -1.964372  1.118950  0.526065
#>   2  0.492364  0.072820  0.374191  0.610215  0.120318 -0.382067  0.662967
#>   3 -0.438109 -0.170812  0.392892 -0.545844  0.198190  0.695212  2.052921
#>   4  0.302637  0.907947 -0.245903  1.290938 -0.933153 -1.255059  1.039265
#>    year
#> age 15        16        17        18        19        20       
#>   1 -0.894941 -0.830625  1.168181  0.098290  0.691816 -0.923430
#>   2  0.815614 -0.173074  0.597373  1.294623  0.722492  1.020554
#>   3 -0.407486  0.961231 -0.090217  1.811744  0.063760 -0.720291
#>   4  1.539860 -0.665834 -1.803165 -1.907233  0.428448  1.686743
#> 
#> units:  kg 

# with 'iter'
  flq <- FLQuant(rnorm(800), dim=c(4,20,1,1,1,10), quant='age', units='kg')
  flq

#> An object of class "FLQuant"
#> iters:  10 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1                 2                 3                 4                
#>   1 -0.2655187(0.762)  0.3499799(0.996) -0.5841853(1.156) -0.1543706(0.802)
#>   2  0.2051241(0.522) -0.3695812(0.973) -0.2730657(0.887)  0.2243214(0.817)
#>   3  0.3054665(1.107) -0.5896147(0.899) -0.5995549(0.681) -0.2231261(0.640)
#>   4  0.3589116(0.697)  0.3723756(0.980) -0.1272131(0.715) -0.3215435(0.724)
#>    year
#> age 5                 6                 7                 8                
#>   1 -0.2524673(0.549) -0.1289676(0.708) -0.0826515(0.791) -0.2133355(0.728)
#>   2  0.0021137(0.501)  0.6322558(0.818)  0.0442438(1.326) -0.0914388(0.341)
#>   3 -0.3676240(1.174) -0.4072142(1.194) -0.3075101(0.318) -0.7266502(0.941)
#>   4 -0.1017052(1.380)  0.5829020(1.242) -0.5007081(0.810)  0.7828031(0.597)
#>    year
#> age 9                 10                11                12               
#>   1 -0.3343001(0.717) -0.1848331(0.847)  0.4089553(0.643)  0.1528778(0.926)
#>   2 -0.8300436(1.694)  0.4700141(1.164)  0.6313460(0.507)  0.0715385(1.321)
#>   3 -0.1149658(0.686)  0.5797203(0.551)  0.4419735(1.409) -0.0976332(1.729)
#>   4 -0.0619858(0.752) -0.3455555(1.017)  0.0592027(1.312)  0.2395320(1.017)
#>    year
#> age 13                14                15                16               
#>   1  0.0753377(1.318) -0.2438691(1.436)  0.3153688(0.745) -0.1363947(0.672)
#>   2 -0.6642115(1.511) -0.0708858(0.846) -0.2096808(1.031) -0.0716374(1.472)
#>   3 -0.2824286(0.907)  0.6009873(1.226) -0.1434467(0.930) -0.2908579(0.487)
#>   4 -0.5579668(0.597) -0.0889183(0.944) -0.0888097(0.318) -0.1161982(0.911)
#>    year
#> age 17                18                19                20               
#>   1  0.2695021(0.547)  0.3415531(0.646) -0.0539026(0.442) -0.0897169(0.586)
#>   2 -0.2281394(0.625)  0.3390914(0.490) -0.0620332(0.987)  0.6010484(0.446)
#>   3 -0.0402545(0.948)  0.3159034(0.695)  0.5125548(0.634) -0.2799777(0.573)
#>   4 -0.0401269(0.340) -0.4772528(0.423)  0.1778844(0.709) -0.0119164(1.273)
#> 
#> units:  kg