Method show

Standard display of an object contents in an interactive session. Objects of class FLQuant with length > 1 along the sixth dimension (iter) are output in a summarised form, as median(mad), where mad is the median absolute deviation. See mad.

# S4 method for FLArray
show(object)

Details

The same format is used for objects of class FLPar with length > 1 on the last dimension (iter).

Generic function

show(object)

See also

Author

The FLR Team

Examples

# no 'iter'
  flq <- FLQuant(rnorm(80), dim=c(4,20), quant='age', units='kg')
  flq
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1          2          3          4          5          6         
#>   1  0.9584906 -1.3305503  0.0109892 -1.5491918  1.1592182 -2.2542053
#>   2  1.5311187 -1.4330366 -1.2507800 -0.5518211  0.6345983 -0.1831026
#>   3  0.1363503 -2.2047737 -0.6481155  0.1124633 -0.9274678  1.7817377
#>   4  0.0061896 -0.3941467 -0.5427160  1.0065689  0.0338205 -0.4612802
#>    year
#> age 7          8          9          10         11         12        
#>   1  0.8050782 -0.0568726  1.6126220  0.2239686  0.4071716 -0.6280759
#>   2 -0.6795054  0.7586843 -0.0122528  0.0859024 -0.6687745 -0.1901627
#>   3 -0.8120483  0.1928770  0.9038220 -1.3797128  0.2516513 -0.6176154
#>   4 -0.9014480  0.6889826 -0.4472200 -0.6525469  0.9962662  0.7922829
#>    year
#> age 13         14         15         16         17         18        
#>   1  1.5313679  1.0609124 -1.3938686 -0.9762452 -0.1942163 -2.8209283
#>   2  0.2928952 -0.6415124 -0.3218565  0.5881691 -0.3269656  0.5578169
#>   3  1.3755545  0.7315000 -1.4107806  1.8136208  0.2454349  0.3083116
#>   4 -0.3645759  0.3645177 -1.4625778 -1.2324439  0.0141575 -0.4864723
#>    year
#> age 19         20        
#>   1  0.6116812  0.5322276
#>   2  0.5904584  0.8485907
#>   3  0.6503077 -1.2207072
#>   4 -0.0668189  0.0447145
#> 
#> units:  kg 

# with 'iter'
  flq <- FLQuant(rnorm(800), dim=c(4,20,1,1,1,10), quant='age', units='kg')
  flq
#> An object of class "FLQuant"
#> iters:  10 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1                2                3                4               
#>   1 -0.076935(0.779) -0.329529(0.428) -0.948205(0.605)  0.041409(1.008)
#>   2 -0.405515(0.750) -0.301921(1.031)  0.493424(1.100)  0.151267(1.078)
#>   3  0.145276(0.814) -0.483257(0.346) -0.565829(1.418) -0.020724(1.653)
#>   4  0.251423(1.212) -0.030706(1.229) -0.431626(0.780)  0.155858(0.622)
#>    year
#> age 5                6                7                8               
#>   1  0.083172(0.755) -0.124004(0.728)  0.165931(0.949)  0.175790(1.052)
#>   2  1.047257(1.077)  0.338048(0.558)  0.049425(0.840)  0.334123(1.183)
#>   3  0.113679(0.606)  0.350536(0.955)  0.712849(0.614) -0.366156(0.236)
#>   4  0.283582(1.203) -0.196089(0.914) -0.411942(1.060)  0.777586(0.323)
#>    year
#> age 9                10               11               12              
#>   1 -0.068413(0.758)  0.570132(0.794) -0.136054(0.892)  0.036137(0.888)
#>   2 -0.230530(1.040) -0.330827(0.965) -0.320914(0.508) -0.714974(1.302)
#>   3 -0.231025(0.870) -0.012452(0.827)  0.170652(0.949) -0.072477(0.650)
#>   4  0.484179(0.974) -0.273085(0.995) -0.598989(1.180)  0.403682(0.830)
#>    year
#> age 13               14               15               16              
#>   1 -0.841073(0.805)  0.018822(1.079) -0.015697(0.491)  0.248134(0.665)
#>   2  0.024195(0.614)  0.088407(0.902)  0.186787(0.923)  0.065725(1.070)
#>   3 -0.261560(0.450) -0.296643(0.902) -0.271846(1.264) -0.230310(0.758)
#>   4  0.574737(0.667)  0.221841(1.439)  0.112249(1.646) -0.157107(0.665)
#>    year
#> age 17               18               19               20              
#>   1  0.453849(0.641)  0.134848(0.292)  0.292328(0.893) -0.498745(0.401)
#>   2 -0.527569(0.723)  0.065695(1.193) -0.216223(0.673) -0.116541(0.973)
#>   3 -0.530664(0.777) -0.083638(0.512)  0.204280(0.598)  0.606429(1.091)
#>   4 -0.137988(0.633)  0.010394(1.533)  0.258866(0.860) -0.413255(0.562)
#> 
#> units:  kg