Method show

Standard complete display of an object contents in an interactive session. Objects of class FLArray with length > 1 alonArray sixth dimension (iter) are output in a summarised form, as median(mad), where mad is the median absolute deviation. See mad.

Usage

# S4 method for class 'FLArray'
show(object)

# S4 method for class 'FLArray'
print(x)

Details

The same format is used for objects of class FLPar with length > 1 on the last dimension (iter).

Generic function

show(object)

show(object)

See also

Author

The FLR Team

Examples

# no 'iter'
  flq <- FLQuant(rnorm(80), dim=c(4,20), quant='age', units='kg')
  flq
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1       2       3       4       5       6       7       8       9      
#>   1 -0.2039  0.8414 -0.8644 -0.3730  0.1871 -0.8133 -0.2227 -0.7947  0.2197
#>   2  0.3112 -0.6502 -0.9059  0.3142  0.1106  0.4686  0.3385 -0.7823 -0.0152
#>   3  1.1631  0.8187  0.1533 -0.7127  0.5331  1.8302  0.3439  0.2268  0.3603
#>   4  1.1667 -0.4809  0.4956 -0.4660 -0.8126  0.6897 -0.4465  0.6428 -0.5128
#>    year
#> age 10      11      12      13      14      15      16      17      18     
#>   1  0.1100 -0.7795 -0.9890 -0.4387  0.1602 -0.2789 -1.8563 -0.6652 -0.3657
#>   2  1.3128 -1.3669  0.4405  0.6712 -1.1199  0.2506 -0.6104  2.5244 -0.6878
#>   3  0.0150  1.0165 -0.7381  0.6274  1.8476 -0.9952  1.1790 -0.1814  0.8312
#>   4  0.1901 -0.1836  0.5261  0.9250 -0.2695  0.6021  0.9062 -0.1414 -0.6004
#>    year
#> age 19      20     
#>   1 -0.5943 -0.9393
#>   2  0.6810 -0.1233
#>   3 -2.0930 -0.6285
#>   4  1.5733 -0.1089
#> 
#> units:  kg 

# with 'iter'
  flq <- FLQuant(rnorm(800), dim=c(4,20,1,1,1,10), quant='age', units='kg')
  flq
#> An object of class "FLQuant"
#> iters:  10 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1               2               3               4              
#>   1 -0.15600(1.105) -0.10457(0.672)  0.05793(0.899)  0.17127(1.382)
#>   2 -0.23905(0.822) -0.23016(0.916)  0.23409(0.449) -0.14446(1.356)
#>   3 -0.30463(0.304) -0.13510(0.919) -0.49446(1.792) -0.36682(1.180)
#>   4  0.20288(0.523) -0.40814(0.475) -0.17465(0.869)  0.32996(0.976)
#>    year
#> age 5               6               7               8              
#>   1  0.03548(1.055)  0.00316(0.898) -0.39115(1.132) -0.01334(1.542)
#>   2 -0.41604(1.094) -0.25958(1.280)  0.10286(0.807) -0.06193(0.887)
#>   3  0.19110(1.253)  0.33193(0.479)  0.15309(0.997)  0.31433(0.677)
#>   4  0.53838(1.010) -0.21584(1.170)  0.70369(1.290)  0.01794(0.543)
#>    year
#> age 9               10              11              12             
#>   1  0.23479(0.734) -0.13371(0.770)  0.31448(0.872)  0.29399(1.431)
#>   2  0.41039(1.485)  0.10704(0.661)  0.41497(0.945) -0.10320(0.644)
#>   3 -0.67119(0.554) -0.53408(0.489)  0.41956(0.884) -0.44091(0.625)
#>   4 -0.29302(0.862) -0.58340(0.522)  0.15344(0.955) -0.31650(1.062)
#>    year
#> age 13              14              15              16             
#>   1  0.17383(1.078)  0.19501(1.081)  0.07630(0.903) -0.53911(0.851)
#>   2 -0.68295(0.679) -0.13808(0.420) -0.04652(1.498) -0.45807(0.530)
#>   3  0.56305(0.390) -0.25187(1.103)  0.46553(0.714) -0.30893(0.855)
#>   4 -0.36082(0.396) -0.44306(0.566)  0.70606(0.436)  0.83692(0.836)
#>    year
#> age 17              18              19              20             
#>   1  0.06023(0.812) -0.37216(1.052)  0.21283(1.775)  0.05655(0.983)
#>   2  0.12946(0.622)  0.35312(0.958) -0.39520(0.499) -0.39970(0.954)
#>   3  0.06314(0.815)  0.16533(0.586)  0.08628(0.928)  0.01146(1.364)
#>   4  0.25425(0.895) -0.02904(1.092) -0.04503(0.645)  0.26939(1.638)
#> 
#> units:  kg 

# no 'iter'
  flq <- FLQuant(rnorm(80), dim=c(4,20), quant='age', units='kg')
  print(flq)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1       2       3       4       5       6       7       8       9      
#>   1 -0.1524  2.0169  0.5767  0.7395  0.5087  1.2233 -0.5987  0.4155  0.1875
#>   2 -1.4836 -0.1239  1.6111 -2.9687 -0.1608  0.3247 -1.7113  0.5001  1.3723
#>   3  1.4834 -0.8084 -0.2372  1.3378  0.0503  0.8589  0.5840 -0.3335 -0.2508
#>   4 -0.0180  0.3716 -1.6753  0.6821  1.5753 -1.4146  0.1446  0.0153  0.1258
#>    year
#> age 10      11      12      13      14      15      16      17      18     
#>   1 -0.9756 -0.1306 -0.4771 -0.4924 -0.9056  0.0951  0.9082  1.4087 -0.3584
#>   2 -0.3144  1.1098  1.0124 -0.6715  0.9827  0.6551  0.3786  0.7864 -0.8675
#>   3 -1.0844  0.0176  0.1741 -0.8234 -0.0643  0.8305 -0.4270 -1.4501  1.7472
#>   4  0.1222  0.4005  0.7402  0.0163 -0.0560 -0.3412 -0.5320  0.5967  0.3369
#>    year
#> age 19      20     
#>   1 -0.6728  0.3537
#>   2  0.7528 -1.2512
#>   3 -1.1725  0.1859
#>   4 -1.0852  0.3264
#> 
#> units:  kg 

# with 'iter'
  flq <- FLQuant(rnorm(800), dim=c(4,20,1,1,1,10), quant='age', units='kg')
  print(flq)
#> An object of class "FLQuant"
#> iters:  10 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1               2               3               4              
#>   1 -0.26926(0.370) -0.13547(1.597) -0.32689(0.773) -0.03155(1.216)
#>   2  0.13650(0.819) -0.04211(0.303) -0.55362(1.256) -0.04585(0.700)
#>   3 -0.09564(1.195)  0.14548(0.547) -0.14876(1.386)  0.04494(0.991)
#>   4  0.10165(0.678) -0.32174(0.640)  0.81639(0.565) -0.25147(1.085)
#>    year
#> age 5               6               7               8              
#>   1 -0.35486(0.594) -0.87462(1.193)  0.34934(1.036)  0.63953(1.329)
#>   2  0.29172(0.691) -0.18992(0.595)  0.30546(1.062)  0.28884(1.021)
#>   3  0.33936(1.505) -0.43709(0.865)  0.03739(0.944)  0.42094(1.478)
#>   4  0.20021(0.674)  0.24232(0.872)  0.00581(0.452) -0.71420(0.720)
#>    year
#> age 9               10              11              12             
#>   1  0.81176(1.024) -0.14407(0.832) -0.45324(1.242) -0.02556(0.839)
#>   2 -0.08255(1.315) -0.36394(0.772) -0.38176(0.551) -0.23785(0.878)
#>   3  0.00110(1.431) -0.64659(1.460) -0.59818(0.839)  0.31861(0.936)
#>   4  0.34925(0.649)  0.14518(1.232) -0.02424(0.814)  0.14493(0.634)
#>    year
#> age 13              14              15              16             
#>   1 -0.07728(0.850) -0.26280(1.197) -0.46750(0.786)  0.14371(1.214)
#>   2  0.34344(0.764)  0.24528(0.985) -0.30053(1.458) -0.37297(0.617)
#>   3  0.24120(0.379)  0.47300(1.047)  0.67742(1.008) -0.06377(0.903)
#>   4 -0.28875(1.246)  0.17509(0.819)  0.04566(0.907)  0.60143(0.707)
#>    year
#> age 17              18              19              20             
#>   1  0.12645(1.075)  0.08210(0.834) -0.13833(1.040) -0.68354(0.928)
#>   2 -0.44668(0.690) -0.30416(0.498) -0.09110(0.629) -0.31991(0.936)
#>   3  0.15853(0.674) -0.05363(1.300)  0.18942(0.748)  0.75001(0.887)
#>   4  0.19172(1.268) -0.22934(0.530) -0.34190(0.704) -0.20849(0.815)
#> 
#> units:  kg