Method show

Standard display of an object contents in an interactive session. Objects of class FLQuant with length > 1 along the sixth dimension (iter) are output in a summarised form, as median(mad), where mad is the median absolute deviation. See mad.

Usage

# S4 method for FLArray
show(object)

Details

The same format is used for objects of class FLPar with length > 1 on the last dimension (iter).

Generic function

show(object)

See also

Author

The FLR Team

Examples

# no 'iter'
  flq <- FLQuant(rnorm(80), dim=c(4,20), quant='age', units='kg')
  flq
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1         2         3         4         5         6         7        
#>   1  1.271252  0.032911  0.375711 -0.439019 -0.838299  1.163149  0.818696
#>   2  0.243976  2.094692  1.572370 -0.995664 -0.506914  1.166703 -0.480893
#>   3  0.848595  1.130655 -1.136223 -0.052321 -0.203883  0.841376 -0.864438
#>   4  0.071205  1.753819 -0.490602  0.268262  0.311160 -0.650181 -0.905919
#>    year
#> age 8         9         10        11        12        13        14       
#>   1  0.153291 -0.712742  0.533097  1.830176  0.343902  0.226830  0.360262
#>   2  0.495628 -0.465994 -0.812553  0.689716 -0.446518  0.642837 -0.512805
#>   3 -0.373004  0.187112 -0.813350 -0.222695 -0.794747  0.219696  0.110008
#>   4  0.314186  0.110642  0.468605  0.338550 -0.782291 -0.015230  1.312848
#>    year
#> age 15        16        17        18        19        20       
#>   1  0.014958  1.016462 -0.738116  0.627374  1.847553 -0.995209
#>   2  0.190096 -0.183635  0.526129  0.925017 -0.269524  0.602072
#>   3 -0.779454 -0.988987 -0.438680  0.160177 -0.278919 -1.856290
#>   4 -1.366896  0.440471  0.671194 -1.119939  0.250624 -0.610363
#> 
#> units:  kg 

# with 'iter'
  flq <- FLQuant(rnorm(800), dim=c(4,20,1,1,1,10), quant='age', units='kg')
  flq
#> An object of class "FLQuant"
#> iters:  10 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1                 2                 3                 4                
#>   1 -0.3089296(0.899)  0.0631446(0.592)  0.2551947(0.591)  0.0862811(1.159)
#>   2  0.8942780(0.616) -0.0178774(0.910) -0.0290428(0.976) -0.0450279(0.705)
#>   3  0.0146506(0.744) -0.3430955(1.009) -0.1736159(1.202) -0.1475128(0.764)
#>   4  0.3368354(0.777)  0.2662023(0.958) -0.2165251(1.078) -0.3996970(0.634)
#>    year
#> age 5                 6                 7                 8                
#>   1  0.0114645(1.308) -0.3046330(0.304) -0.1350969(0.919) -0.4944649(1.792)
#>   2  0.2693854(1.314)  0.2028819(0.523) -0.4081432(0.475) -0.1746503(0.869)
#>   3 -0.1559973(1.105) -0.1045653(0.672)  0.0579263(0.899)  0.1712732(1.382)
#>   4 -0.2390533(0.822) -0.2301584(0.916)  0.2340851(0.449) -0.1444576(1.356)
#>    year
#> age 9                 10                11                12               
#>   1 -0.3668160(1.180)  0.1910958(1.253)  0.3319283(0.479)  0.1530920(0.997)
#>   2  0.3299629(0.976)  0.5383774(1.010) -0.2158444(1.170)  0.7036873(1.290)
#>   3  0.0354759(1.055)  0.0031582(0.898) -0.3911477(1.132) -0.0133351(1.542)
#>   4 -0.4160365(1.094) -0.2595765(1.280)  0.1028569(0.807) -0.0619296(0.887)
#>    year
#> age 13                14                15                16               
#>   1  0.3143253(0.677) -0.6711947(0.554) -0.5340848(0.489)  0.4195613(0.884)
#>   2  0.0179421(0.543) -0.2930193(0.862) -0.5833958(0.522)  0.1534410(0.955)
#>   3  0.2347924(0.734) -0.1337129(0.770)  0.3144784(0.872)  0.2939867(1.431)
#>   4  0.4103916(1.485)  0.1070403(0.661)  0.4149749(0.945) -0.1032012(0.644)
#>    year
#> age 17                18                19                20               
#>   1 -0.4409126(0.625)  0.5630499(0.390) -0.2518662(1.103)  0.4655326(0.714)
#>   2 -0.3165033(1.062) -0.3608176(0.396) -0.4430595(0.566)  0.7060585(0.436)
#>   3  0.1738286(1.078)  0.1950127(1.081)  0.0762994(0.903) -0.5391107(0.851)
#>   4 -0.6829497(0.679) -0.1380817(0.420) -0.0465230(1.498) -0.4580658(0.530)
#> 
#> units:  kg