Method show

Standard display of an object contents in an interactive session. Objects of class FLQuant with length > 1 along the sixth dimension (iter) are output in a summarised form, as median(mad), where mad is the median absolute deviation. See mad.

# S4 method for FLArray
show(object)

Details

The same format is used for objects of class FLPar with length > 1 on the last dimension (iter).

Generic function

show(object)

See also

Examples

# no 'iter'
  flq <- FLQuant(rnorm(80), dim=c(4,20), quant='age', units='kg')
  flq

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1         2         3         4         5         6         7        
#>   1 -0.804666  1.558039  0.911674  0.196398 -1.409321 -1.845722 -0.099184
#>   2  0.885941  0.015058 -0.261370 -0.105059  0.098843  0.836016  1.111328
#>   3  0.167953  0.528415 -0.171360  0.048228  1.055556 -0.907691 -0.467448
#>   4  1.937135  2.064820 -0.493625 -1.197082  1.734968  0.213743  1.255096
#>    year
#> age 8         9         10        11        12        13        14       
#>   1 -0.384632  0.260762 -0.652797 -0.075420  0.292464  0.157351 -0.285354
#>   2 -0.334968 -0.295751  1.400092 -0.243091  0.147054  2.532177  0.137199
#>   3  0.296476  0.003121 -0.391316 -1.147870  0.043278 -2.004083  0.021760
#>   4  1.333907 -0.802416  0.150325 -0.952779  1.605568  1.197756  1.452436
#>    year
#> age 15        16        17        18        19        20       
#>   1  1.239111  1.125339 -0.717273  0.508289  0.772120  0.035459
#>   2  0.185336 -1.030520  0.179635 -0.497275 -0.638661  0.844138
#>   3  0.186121 -0.104863 -0.339209  1.317702  0.052431 -0.787307
#>   4  0.191528 -1.050851  0.675691  2.359824 -1.241107  0.039526
#> 
#> units:  kg 

# with 'iter'
  flq <- FLQuant(rnorm(800), dim=c(4,20,1,1,1,10), quant='age', units='kg')
  flq

#> An object of class "FLQuant"
#> iters:  10 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1                 2                 3                 4                
#>   1  0.8622282(0.609)  0.3730216(1.167) -0.3455555(0.571) -0.2273871(1.303)
#>   2 -0.3343001(1.011) -0.1848331(0.590)  0.4089553(0.806)  0.0152951(1.146)
#>   3 -0.7650333(1.384)  0.5620389(0.727)  0.5651198(0.550)  0.1017373(1.082)
#>   4 -0.1149658(0.686)  0.3981372(0.551)  0.4419735(1.307)  0.2750508(1.506)
#>    year
#> age 5                 6                 7                 8                
#>   1 -0.2938872(0.655) -0.6702315(0.468) -0.1303827(1.351) -0.1532884(0.593)
#>   2  0.1435784(1.561)  0.0836686(1.282)  0.1297751(0.832) -0.4418952(0.453)
#>   3 -0.4319859(1.620) -0.0708858(0.994)  0.1430177(0.848)  0.4384531(1.472)
#>   4 -0.2824286(1.291)  0.6009873(1.226) -0.1434467(0.930) -0.0275120(0.736)
#>    year
#> age 9                 10                11                12               
#>   1 -0.2593717(0.613) -0.2693278(0.465) -0.4772528(0.454)  0.2087724(0.755)
#>   2  0.1477648(0.871)  0.4893043(0.497) -0.0539026(0.811) -0.1210779(0.672)
#>   3 -0.2281394(0.677)  0.3629960(0.499) -0.2435506(0.987)  0.7499791(0.569)
#>   4 -0.1723560(0.654)  0.2278663(0.751)  0.3557481(0.423) -0.2799777(0.557)
#>    year
#> age 13                14                15                16               
#>   1  0.2061077(1.408)  0.4036049(0.957)  0.3241361(0.908) -0.1153186(0.639)
#>   2 -0.2655187(0.762) -0.1262751(1.149) -0.5841853(1.156) -0.1543706(0.802)
#>   3  0.1429225(0.643) -0.3695812(1.138) -0.1009127(0.817)  0.4603387(0.972)
#>   4  0.3054665(0.939) -0.5896147(0.899) -0.5995549(0.581) -0.1474609(0.708)
#>    year
#> age 17                18                19                20               
#>   1 -0.6091201(0.803)  0.1706145(1.415)  0.5486859(0.881) -0.3372954(1.355)
#>   2 -0.2524673(0.539) -0.1289676(0.708)  0.1402595(0.461)  0.0703781(0.959)
#>   3  0.0021137(0.501)  0.6322558(0.818)  0.0442438(1.326)  0.0853533(0.360)
#>   4 -0.3676240(0.870) -0.4072142(1.381) -0.3796039(0.517) -0.7266502(0.941)
#> 
#> units:  kg