Method show

Standard display of an object contents in an interactive session. Objects of class FLQuant with length > 1 along the sixth dimension (iter) are output in a summarised form, as median(mad), where mad is the median absolute deviation. See mad.

# S4 method for FLArray
show(object)

Details

The same format is used for objects of class FLPar with length > 1 on the last dimension (iter).

Generic function

show(object)

See also

FLComp

Examples

# no 'iter'
  flq <- FLQuant(rnorm(80), dim=c(4,20), quant='age', units='kg')
  flq

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1         2         3         4         5         6         7        
#>   1 -0.608148  1.058819  0.433217 -2.284705  0.857973  0.056473  0.775255
#>   2  0.232961 -1.268336 -1.072297  0.243341 -0.167162  1.093280  0.260089
#>   3  0.598652  2.297597 -1.493171  0.580964  0.831572  2.076709 -1.382079
#>   4  0.671307 -2.247847 -2.197331  0.532523  0.421662  0.569146  0.142487
#>    year
#> age 8         9         10        11        12        13        14       
#>   1  0.933568  0.718962 -0.837299 -0.066640  0.076583  1.399147 -1.219435
#>   2  0.575018 -0.105009 -1.663812  0.993225 -2.089923  1.793255 -0.646229
#>   3  0.422619  0.267890 -1.685419  2.124517  2.599646 -0.802554 -0.920994
#>   4  0.551624  0.194397  0.520552 -0.924118 -0.014254  0.753149  0.033031
#>    year
#> age 15        16        17        18        19        20       
#>   1 -0.656180  1.487845 -0.076917 -1.096555 -0.609107 -0.419249
#>   2 -1.591679  0.406434 -0.993032 -1.026123  0.525843 -2.150908
#>   3 -0.918139 -0.659196  0.715015 -0.809150 -0.438312  0.019314
#>   4  0.368786 -1.185767  2.319667 -0.695101  0.980514  1.425292
#> 
#> units:  kg 

# with 'iter'
  flq <- FLQuant(rnorm(800), dim=c(4,20,1,1,1,10), quant='age', units='kg')
  flq

#> An object of class "FLQuant"
#> iters:  10 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1                 2                 3                 4                
#>   1 -0.6382588(0.490)  0.0556407(0.436)  0.0249078(0.873)  0.2890605(0.861)
#>   2  0.4609146(0.892) -0.9259422(1.908)  0.0265672(1.310) -0.4191337(1.385)
#>   3 -0.8121444(0.920)  0.1779395(0.335)  0.1033510(0.374)  0.3636902(0.469)
#>   4 -0.0271298(1.202)  0.1495889(1.334)  0.3481920(0.908) -0.2777281(1.155)
#>    year
#> age 5                 6                 7                 8                
#>   1 -0.5296163(0.656) -0.5034570(0.915)  0.2048886(1.078) -0.4086647(1.001)
#>   2 -0.3305000(0.764)  0.1888136(0.662)  0.0911980(0.806) -0.5001321(0.624)
#>   3  0.3417861(1.281)  0.3390723(1.079)  0.2138741(0.457) -0.2803628(1.303)
#>   4  0.0718833(0.451)  0.1995532(0.504)  0.2716007(0.915)  0.1289731(1.009)
#>    year
#> age 9                 10                11                12               
#>   1 -0.3958983(0.858)  0.0430409(0.642) -0.4206789(0.996)  0.1415978(0.266)
#>   2 -0.2550037(0.749) -0.2432334(1.183)  0.1928284(0.440) -0.1557233(0.686)
#>   3 -0.3413814(1.050)  0.0466285(0.915) -0.0857602(1.508) -0.7095823(0.932)
#>   4  0.4387365(0.734) -0.3163457(0.678) -0.0905474(0.994) -0.4484071(0.652)
#>    year
#> age 13                14                15                16               
#>   1 -0.0061205(0.743) -0.4556827(1.599) -0.3083809(0.930) -0.2237581(1.648)
#>   2  0.2045075(0.731)  0.2594592(1.025)  0.4714255(0.699) -0.6909415(1.008)
#>   3 -0.0318063(0.764)  0.1402200(0.661)  0.3493170(0.549) -0.2499423(0.990)
#>   4  0.0921632(0.975)  0.6184478(0.200) -0.7381214(0.850) -0.4201608(1.158)
#>    year
#> age 17                18                19                20               
#>   1  0.0443752(0.929) -0.3025488(1.298)  0.4127451(0.673)  0.2795983(0.926)
#>   2 -0.1331483(0.542) -0.1742012(0.915)  0.0534981(0.579) -0.2153684(1.084)
#>   3 -0.0807939(1.069) -0.4132601(0.855) -0.2789615(0.375) -0.1883049(0.995)
#>   4 -0.5813534(0.838) -0.1765819(1.381) -0.1387872(1.119) -0.0758094(0.688)
#> 
#> units:  kg