Method show

Standard display of an object contents in an interactive session. Objects of class FLQuant with length > 1 along the sixth dimension (iter) are output in a summarised form, as median(mad), where mad is the median absolute deviation. See mad.

# S4 method for FLArray
show(object)

Details

The same format is used for objects of class FLPar with length > 1 on the last dimension (iter).

Generic function

show(object)

See also

FLComp

Examples

# no 'iter'
  flq <- FLQuant(rnorm(80), dim=c(4,20), quant='age', units='kg')
  flq

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1           2           3           4           5           6          
#>   1  0.48170611 -0.69253741  1.68881022 -0.00024576 -2.15948236  0.22392581
#>   2 -0.05039241 -0.80225178  1.17649560  1.13142455 -1.68986667 -0.02961192
#>   3  1.25761991  1.92075922  0.59933185  0.85857981 -0.93562753  0.60394780
#>   4  0.40558118  0.38176296 -0.54843221  0.34116064 -0.09225381  0.02597704
#>    year
#> age 7           8           9           10          11          12         
#>   1  1.79086530  0.15811619 -0.95149618  0.12984334 -0.68652056  2.07962649
#>   2  0.91306774 -0.28834436  0.54295038  1.16485361 -0.11936252 -0.67659044
#>   3 -0.64441007  0.17520761  1.86223866  0.12929518 -1.73371444  0.08803754
#>   4 -1.07671087 -1.02795354 -0.18578523 -0.09965415  0.75809597 -0.79072971
#>    year
#> age 13          14          15          16          17          18         
#>   1  1.81848695 -2.17151399 -1.16556431  0.32974719  0.45433809  0.25108916
#>   2 -0.38481106 -0.41933657 -0.54094321  0.65781799  0.80153705 -1.77433989
#>   3 -0.87267813  0.92502924 -0.80713855 -2.34004281  0.72621545 -0.10597923
#>   4  0.99071014 -0.46895955  1.78744563 -1.72036523 -0.80621516 -0.75911180
#>    year
#> age 19          20         
#>   1 -0.59873741 -0.44225161
#>   2  1.17555077  2.06425423
#>   3  0.47398785  1.94310364
#>   4 -0.20918346 -2.56111720
#> 
#> units:  kg 

# with 'iter'
  flq <- FLQuant(rnorm(800), dim=c(4,20,1,1,1,10), quant='age', units='kg')
  flq

#> An object of class "FLQuant"
#> iters:  10 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1                2                3                4               
#>   1 -0.108469(1.106) -0.289012(0.776) -0.765722(1.203)  0.284142(1.224)
#>   2 -0.235313(0.846) -0.010909(0.206)  0.056157(0.616) -0.223369(1.263)
#>   3 -0.824780(0.471) -0.101640(0.360)  0.247927(0.655)  0.015948(0.845)
#>   4  0.183164(1.433) -0.874266(1.697)  0.067725(1.020) -0.393440(1.060)
#>    year
#> age 5                6                7                8               
#>   1 -0.198784(1.076)  0.315597(1.320) -0.262715(0.854)  0.305797(0.271)
#>   2  0.103993(1.043)  0.286609(1.202)  0.141038(0.422)  0.014992(0.730)
#>   3 -0.168647(0.804) -0.302752(1.163)  0.171817(1.074) -0.286842(0.739)
#>   4 -0.439536(0.848)  0.231599(0.588) -0.323134(1.039)  0.056462(0.825)
#>    year
#> age 9                10               11               12              
#>   1 -0.109865(0.840)  0.267570(0.682)  0.249506(1.034) -0.215057(1.731)
#>   2  0.309053(0.827)  0.374025(0.758) -0.372214(0.608) -0.202718(1.251)
#>   3 -0.193692(0.885) -0.047677(0.551) -0.420679(0.586)  0.099301(1.027)
#>   4 -0.151347(1.058)  0.137490(1.697)  0.431630(0.943)  0.345479(0.651)
#>    year
#> age 13               14               15               16              
#>   1 -0.709582(0.977) -0.174681(0.531)  0.284033(0.523)  0.441417(0.713)
#>   2  0.130694(0.658)  0.083109(0.732)  0.574108(0.614) -0.595207(0.799)
#>   3  0.426079(0.821)  0.153984(0.743)  0.308847(0.362) -0.104461(1.286)
#>   4 -0.110926(0.721)  0.467131(0.518)  0.349680(0.879)  0.230445(0.898)
#>    year
#> age 17               18               19               20              
#>   1 -0.116198(0.906) -0.125543(1.000) -0.917235(0.685)  0.022134(0.836)
#>   2  0.210647(1.658) -0.064146(0.861) -0.328039(1.605) -0.047579(0.874)
#>   3  0.254387(0.894) -0.274901(1.350)  0.078302(0.677)  0.279598(0.637)
#>   4  0.297888(0.525) -0.033512(0.405)  0.047107(0.679) -0.228712(1.649)
#> 
#> units:  kg