Method show

Standard display of an object contents in an interactive session. Objects of class FLQuant with length > 1 along the sixth dimension (iter) are output in a summarised form, as median(mad), where mad is the median absolute deviation. See mad.

# S4 method for FLArray
show(object)

Details

The same format is used for objects of class FLPar with length > 1 on the last dimension (iter).

Generic function

show(object)

See also

Author

The FLR Team

Examples

# no 'iter'
  flq <- FLQuant(rnorm(80), dim=c(4,20), quant='age', units='kg')
  flq
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1         2         3         4         5         6         7        
#>   1  1.557916 -0.347803  1.007905 -0.583724  1.356284 -0.437159 -0.115601
#>   2  0.173441  1.587372 -0.888456 -1.661516 -0.627546  1.042725 -0.052162
#>   3  1.559921  1.612016 -0.551808  0.014928  0.032217  1.232493 -1.456018
#>   4 -1.536523  0.154738  0.788305  0.593925  0.997732 -1.155536 -0.725186
#>    year
#> age 8         9         10        11        12        13        14       
#>   1 -0.559099  0.809641  1.134771 -0.726359  1.253807 -0.045341  1.343965
#>   2  0.283793  0.832207  0.716719  0.771158 -1.387840  1.745693  0.370357
#>   3 -0.361939 -1.776787 -2.437677 -1.930468  1.259452 -1.385648  0.308475
#>   4  0.778330 -1.634975 -0.629288 -0.204675  0.089652  0.490683 -0.335306
#>    year
#> age 15        16        17        18        19        20       
#>   1 -1.146767  0.034164 -0.114665  0.423231 -0.649969  0.167272
#>   2  0.217412  0.485699  1.240292 -1.252347 -1.656012 -1.698891
#>   3 -0.978325  0.018593  1.190224 -1.388843 -0.393575  1.255632
#>   4 -0.451394  0.892306 -0.367092  0.563481  1.397245 -0.254431
#> 
#> units:  kg 

# with 'iter'
  flq <- FLQuant(rnorm(800), dim=c(4,20,1,1,1,10), quant='age', units='kg')
  flq
#> An object of class "FLQuant"
#> iters:  10 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 1                2                3                4               
#>   1  0.267227(0.603)  0.261416(0.725) -0.144549(1.479)  0.243766(1.171)
#>   2  0.284052(0.944) -0.243228(1.343) -0.505009(0.786)  0.424922(1.047)
#>   3  0.328033(0.727)  0.771059(0.729)  0.364152(0.766)  0.208123(0.523)
#>   4  0.231360(0.846)  0.073671(1.082)  0.253835(1.032) -0.058034(0.758)
#>    year
#> age 5                6                7                8               
#>   1 -0.645743(1.246) -0.424834(1.341) -0.047618(0.890) -0.291875(1.084)
#>   2 -0.399492(0.760) -0.239837(1.150) -0.069452(1.759)  0.807117(1.471)
#>   3 -0.541136(1.697)  0.253887(0.753) -0.149676(1.153)  0.391100(1.069)
#>   4  0.523422(0.363)  0.257256(1.307) -0.176605(0.627) -0.279083(0.597)
#>    year
#> age 9                10               11               12              
#>   1 -0.207403(0.569) -1.137264(0.791)  0.139538(0.525) -0.126607(1.157)
#>   2  0.615906(0.561)  0.113074(1.053) -0.263241(0.745)  0.240948(1.101)
#>   3 -0.114278(1.074)  0.497172(1.088)  0.636352(0.894)  0.200290(0.622)
#>   4  0.064911(1.356) -0.191763(1.078)  0.045558(0.720)  0.373747(1.078)
#>    year
#> age 13               14               15               16              
#>   1  0.301376(0.699) -0.604990(0.927)  0.364917(0.584)  0.307159(0.753)
#>   2 -0.716432(0.733) -0.040335(1.432) -0.574993(0.562) -0.410109(0.519)
#>   3 -0.234620(1.839)  0.061442(1.126)  0.321433(0.723)  0.622441(1.319)
#>   4 -0.111933(1.663)  0.213673(0.541)  0.279615(0.682)  0.932599(0.660)
#>    year
#> age 17               18               19               20              
#>   1  0.288330(0.369)  0.118460(1.438) -0.105543(0.808)  0.030938(0.990)
#>   2  0.388567(0.986)  0.459125(0.744)  0.637165(0.794) -0.073731(1.072)
#>   3 -0.278260(0.999) -0.143665(0.747)  0.111505(1.534)  0.084528(0.367)
#>   4  0.405634(1.528) -0.278881(1.264) -0.105961(1.061) -0.127765(1.490)
#> 
#> units:  kg