splits x along the iter dimension into the groups defined by f.

Similar to base::split, but working along the 6th, iter, dimension of any singular FLR object. The object is divided into as many objects as unique values in f, and returned as an FLlst-derived object, e.g. an FLQuants object when applied to an FLQuant.

# S4 method for FLQuant,vector
split(x, f)

# S4 method for FLComp,vector
split(x, f)

Arguments

x
The object to be split.

f
The vector of group names.

Value

An object of the corresponding plural class (FLQuants from FLQuant).

Author

Iago Mosqueira (WMR).

Examples

# FROM FLQuant to FLQuants
flq <- rlnorm(20, FLQuant(seq(0.1, 0.8, length=10)), 0.2)
split(flq, c(rep(1, 5), rep(2,15)))
#> $ 1 
#> An object of class "FLQuant"
#> iters:  5 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1             2             3             4             5            
#>   all 1.1664(0.114) 1.1146(0.130) 1.1773(0.249) 1.2129(0.102) 1.5423(0.125)
#>      year
#> quant 6             7             8             9             10           
#>   all 1.5149(0.331) 1.7447(0.405) 1.9451(0.453) 1.9356(0.872) 1.7367(0.153)
#> 
#> units:  NA 
#> 
#> $ 2 
#> An object of class "FLQuant"
#> iters:  15 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1             2             3             4             5            
#>   all 1.0297(0.202) 1.2015(0.225) 1.2063(0.242) 1.4787(0.332) 1.5591(0.520)
#>      year
#> quant 6             7             8             9             10           
#>   all 1.7280(0.384) 1.6532(0.382) 1.9874(0.573) 1.9556(0.344) 2.4854(0.713)
#> 
#> units:  NA 
#>