splits x along the iter dimension into the groups defined by f.

Similar to base::split, but working along the 6th, iter, dimension of any singular FLR object. The object is divided into as many objects as unique values in f, and returned as an FLlst-derived object, e.g. an FLQuants object when applied to an FLQuant.

# S4 method for FLQuant,vector
split(x, f)

# S4 method for FLComp,vector
split(x, f)

Arguments

x

The object to be split.

f

The vector of group names.

Value

An object of the corresponding plural class (FLQuants from FLQuant).

Author

Iago Mosqueira (WMR).

Examples

# FROM FLQuant to FLQuants
flq <- rlnorm(20, FLQuant(seq(0.1, 0.8, length=10)), 0.2)
split(flq, c(rep(1, 5), rep(2,15)))
#> $ 1 
#> An object of class "FLQuant"
#> iters:  5 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1              2              3              4             
#>   all 0.96929(0.133) 1.04377(0.251) 1.30453(0.196) 1.39159(0.303)
#>      year
#> quant 5              6              7              8             
#>   all 1.27397(0.253) 1.63434(0.114) 2.13546(0.389) 1.62028(0.362)
#>      year
#> quant 9              10            
#>   all 1.97835(0.636) 2.47918(0.124)
#> 
#> units:  NA 
#> 
#> $ 2 
#> An object of class "FLQuant"
#> iters:  15 
#> 
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1             2             3             4             5            
#>   all 1.2175(0.320) 1.3289(0.408) 1.3880(0.191) 1.1967(0.187) 1.5771(0.355)
#>      year
#> quant 6             7             8             9             10           
#>   all 1.7799(0.489) 1.7215(0.243) 2.0242(0.489) 2.0705(0.380) 2.2505(0.406)
#> 
#> units:  NA 
#>