Method sweep for FLCore classes

Use R's sweep method on FLCore classes

# S4 method for FLArray
sweep(x, MARGIN, STATS, FUN = "-", check.margin = TRUE, ...)

# S4 method for FLPar
sweep(x, MARGIN, STATS, FUN = "-", check.margin = TRUE, ...)

Details

These methods call base R sweep method on **FLCore** classes and then ensure that the returned object is of same class.

Generic function

sweep(x, MARGIN, STATS, FUN = "-", check.margin = TRUE, ...)

See also

Examples

flq <- FLQuant(rlnorm(90), dim=c(3,10), units='kg')
# Get ratio of max value by year
sweep(flq, 2, apply(flq, 2, max), "/")

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1        2        3        4        5        6        7        8       
#>     1 0.557913 0.041438 1.000000 1.000000 1.000000 0.603209 0.103144 0.194408
#>     2 1.000000 1.000000 0.536867 0.247774 0.071106 1.000000 0.057120 0.205908
#>     3 0.701483 0.124447 0.072326 0.050407 0.155818 0.302047 1.000000 1.000000
#>      year
#> quant 9        10      
#>     1 1.000000 0.099767
#>     2 0.212493 0.282905
#>     3 0.040591 1.000000
#> 
#> units:  kg