Method sweep for FLCore classes

Use R's sweep method on FLCore classes

# S4 method for FLArray
sweep(x, MARGIN, STATS, FUN = "-", check.margin = TRUE, ...)

# S4 method for FLPar
sweep(x, MARGIN, STATS, FUN = "-", check.margin = TRUE, ...)

Details

These methods call base R sweep method on **FLCore** classes and then ensure that the returned object is of same class.

Generic function

sweep(x, MARGIN, STATS, FUN = "-", check.margin = TRUE, ...)

See also

Author

The FLR Team

Examples

flq <- FLQuant(rlnorm(90), dim=c(3,10), units='kg')
# Get ratio of max value by year
sweep(flq, 2, apply(flq, 2, max), "/")
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1        2        3        4        5        6        7        8       
#>     1 0.400268 0.725612 0.598806 0.667533 0.284982 0.630106 0.953210 1.000000
#>     2 0.407992 0.356385 0.148712 0.926709 1.000000 1.000000 0.259341 0.179182
#>     3 1.000000 1.000000 1.000000 1.000000 0.487271 0.880442 1.000000 0.092186
#>      year
#> quant 9        10      
#>     1 0.023906 0.188462
#>     2 1.000000 1.000000
#>     3 0.035232 0.017740
#> 
#> units:  kg