Method sweep for FLCore classes

Use R's sweep method on FLCore classes

# S4 method for FLArray
sweep(x, MARGIN, STATS, FUN = "-", check.margin = TRUE,
  ...)

# S4 method for FLPar
sweep(x, MARGIN, STATS, FUN = "-", check.margin = TRUE,
  ...)

Details

These methods call base R sweep method on **FLCore** classes and then ensure that the returned object is of same class.

Generic function

sweep(x, MARGIN, STATS, FUN = "-", check.margin = TRUE, ...)

See also

sweep

Examples

flq <- FLQuant(rlnorm(90), dim=c(3,10), units='kg')
# Get ratio of max value by year
sweep(flq, 2, apply(flq, 2, max), "/")

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> quant 1        2        3        4        5        6        7        8       
#>     1 0.162638 0.189928 1.000000 0.712634 1.000000 0.668809 0.351496 0.378866
#>     2 0.165631 1.000000 0.300181 0.455249 0.215919 0.643504 1.000000 0.264676
#>     3 1.000000 0.293084 0.015858 1.000000 0.201699 1.000000 0.428292 1.000000
#>      year
#> quant 9        10      
#>     1 0.697488 0.112345
#>     2 0.741427 1.000000
#>     3 1.000000 0.046137
#> 
#> units:  kg