Method trim

Trim FLR objects using named dimensions

trim(x, ...)

# S4 method for FLArray
trim(x, ...)

# S4 method for FLComp
trim(x, ...)

# S4 method for FLS
trim(x, ...)

# S4 method for FLBiol
trim(x, ...)

Details

Subsetting of FLR objects can be carried out with dimension names by using trim. A number of dimension names and selected dimensions are passed to the method and those are used to subset the input object.

Exceptions are made for those classes where certain slots might differ in one or more dimensions. If trim is applied to an FLQuant object of length 1 in its first dimension and with dimension name equal to 'all', values to trim specified for that dimension will be ignored. For example, FLStock objects contain slots with length=1 in their first dimension. Specifying values to trim over the first dimension will have no effect on those slots (catch, landings, discards, and stock). Calculations might need to be carried out to recalculate those slots (e.g. using computeCatch, computeLandings, computeDiscards and computeStock) if their quant-structured counterparts are modified along the first dimension.

Generic function

trim(x)

See also

Author

The FLR Team

Examples

flq <- FLQuant(rnorm(90), dimnames=list(age=1:10, year=2000:2016))

trim(flq, year=2000:2005)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  2000      2001      2002      2003      2004      2005     
#>   1   0.443618 -1.031444  0.544584 -0.072533  0.829478 -0.142301
#>   2   0.162873 -1.137342 -1.066313  1.021597 -1.479107 -0.389543
#>   3  -0.964560  1.443426 -0.252627  1.864426  0.143306 -1.989195
#>   4  -1.013760  0.694784  0.063771 -0.337851  0.165793 -0.741900
#>   5   2.708965 -0.225671  0.045644 -1.090324 -0.060586 -1.210582
#>   6  -0.380977  2.813363 -0.533420 -0.973438 -1.803545  0.465329
#>   7  -1.545012 -0.469176 -0.811372  0.295335 -0.106300 -0.940978
#>   8   1.781029 -0.501935 -1.658422  2.632927  0.157206 -2.118858
#>   9  -0.503447  1.561535 -0.380963 -0.735961  1.209309  0.108023
#>   10 -0.219656 -0.531966 -1.088712  0.815094  0.315732 -1.228406
#> 
#> units:  NA 
# which is equivalent to
window(flq, start=2000, end=2005)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  2000      2001      2002      2003      2004      2005     
#>   1   0.443618 -1.031444  0.544584 -0.072533  0.829478 -0.142301
#>   2   0.162873 -1.137342 -1.066313  1.021597 -1.479107 -0.389543
#>   3  -0.964560  1.443426 -0.252627  1.864426  0.143306 -1.989195
#>   4  -1.013760  0.694784  0.063771 -0.337851  0.165793 -0.741900
#>   5   2.708965 -0.225671  0.045644 -1.090324 -0.060586 -1.210582
#>   6  -0.380977  2.813363 -0.533420 -0.973438 -1.803545  0.465329
#>   7  -1.545012 -0.469176 -0.811372  0.295335 -0.106300 -0.940978
#>   8   1.781029 -0.501935 -1.658422  2.632927  0.157206 -2.118858
#>   9  -0.503447  1.561535 -0.380963 -0.735961  1.209309  0.108023
#>   10 -0.219656 -0.531966 -1.088712  0.815094  0.315732 -1.228406
#> 
#> units:  NA 

trim(flq, year=2000:2005, age=1:2)
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>    year
#> age 2000      2001      2002      2003      2004      2005     
#>   1  0.443618 -1.031444  0.544584 -0.072533  0.829478 -0.142301
#>   2  0.162873 -1.137342 -1.066313  1.021597 -1.479107 -0.389543
#> 
#> units:  NA 

# Now on an FLStock
data(ple4)
summary(trim(ple4, year=1990:1995))
#> An object of class "FLStock"
#> 
#> Name: PLE 
#> Description: Plaice in IV. ICES WGNSSK 2018. FLAAP 
#> Quant: age 
#> Dims:  age 	year	unit	season	area	iter
#> 	10	6	1	1	1	1	
#> 
#> Range:  min	max	pgroup	minyear	maxyear	minfbar	maxfbar 
#> 	1	10	10	1990	1995	2	6	
#> 
#> Metrics: 
#>   rec: 550376 - 1083810  (1000) 
#>   ssb: 222203 - 396458  (t) 
#>   catch: 132629 - 250604  (t) 
#>   fbar: 0.60 - 0.61  (f) 

# If 'age' is trimmed in ple4, catch, landings and discards need to be
# recalculated
  shpl4 <- trim(ple4, age=1:4)
  landings(shpl4) <- computeLandings(shpl4)
  discards(shpl4) <- computeDiscards(shpl4)
  catch(shpl4) <- computeCatch(shpl4)
  summary(shpl4)
#> An object of class "FLStock"
#> 
#> Name: PLE 
#> Description: Plaice in IV. ICES WGNSSK 2018. FLAAP 
#> Quant: age 
#> Dims:  age 	year	unit	season	area	iter
#> 	4	61	1	1	1	1	
#> 
#> Range:  min	max	pgroup	minyear	maxyear	minfbar	maxfbar 
#> 	1	4	NA	1957	2017	2	4	
#> 
#> Metrics: 
#>   rec: 367450 - 4303680  (1000) 
#>   ssb: 91016 - 316967  (t) 
#>   catch: 36219 - 253152  (t) 
#>   fbar: 0.22 - 0.71  (f)