Total mortality z

Returns the calculation of total mortality, z, usually as the sum of fishing mortality, f, and natural mortality, m.

Usage

z(object, ...)

# S4 method for class 'FLS'
z(object, ...)

Arguments

object

Object to calculate on.

...

Any extra arguments.

Value

An object of the corresponding class, usually FLQuant.

See also

Author

The FLR Team

Examples

data(ple4)

z(ple4)
#> An x of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957  1958  1959  1960  1961  1962  1963  1964  1965  1966  1967  1968 
#>   1  0.199 0.214 0.228 0.237 0.236 0.219 0.194 0.175 0.168 0.175 0.198 0.242
#>   2  0.271 0.312 0.349 0.362 0.355 0.354 0.370 0.379 0.359 0.333 0.338 0.382
#>   3  0.362 0.410 0.450 0.459 0.452 0.469 0.531 0.589 0.571 0.511 0.477 0.482
#>   4  0.409 0.429 0.445 0.450 0.451 0.460 0.483 0.507 0.514 0.502 0.480 0.457
#>   5  0.356 0.392 0.424 0.441 0.446 0.453 0.473 0.493 0.500 0.493 0.481 0.470
#>   6  0.311 0.350 0.387 0.410 0.423 0.442 0.476 0.502 0.495 0.471 0.464 0.480
#>   7  0.325 0.326 0.330 0.342 0.362 0.387 0.414 0.426 0.413 0.392 0.392 0.417
#>   8  0.328 0.323 0.319 0.316 0.319 0.332 0.357 0.375 0.366 0.346 0.340 0.357
#>   9  0.299 0.333 0.351 0.331 0.297 0.289 0.320 0.361 0.363 0.335 0.317 0.320
#>   10 0.299 0.333 0.351 0.331 0.297 0.289 0.320 0.361 0.363 0.335 0.317 0.320
#>     year
#> age  1969  1970  1971  1972  1973  1974  1975  1976  1977  1978  1979  1980 
#>   1  0.289 0.304 0.295 0.292 0.305 0.337 0.387 0.443 0.479 0.471 0.417 0.351
#>   2  0.428 0.430 0.413 0.417 0.448 0.479 0.491 0.499 0.522 0.553 0.575 0.578
#>   3  0.495 0.492 0.493 0.525 0.579 0.608 0.583 0.551 0.555 0.599 0.668 0.741
#>   4  0.445 0.453 0.484 0.542 0.612 0.645 0.616 0.576 0.570 0.598 0.642 0.688
#>   5  0.461 0.460 0.475 0.522 0.586 0.617 0.586 0.552 0.564 0.607 0.633 0.621
#>   6  0.492 0.481 0.470 0.490 0.533 0.553 0.528 0.506 0.532 0.586 0.611 0.582
#>   7  0.443 0.449 0.445 0.454 0.475 0.487 0.479 0.471 0.482 0.507 0.532 0.540
#>   8  0.383 0.404 0.419 0.435 0.451 0.461 0.461 0.455 0.446 0.443 0.454 0.477
#>   9  0.338 0.367 0.400 0.432 0.456 0.469 0.468 0.453 0.427 0.403 0.395 0.406
#>   10 0.338 0.367 0.400 0.432 0.456 0.469 0.468 0.453 0.427 0.403 0.395 0.406
#>     year
#> age  1981  1982  1983  1984  1985  1986  1987  1988  1989  1990  1991  1992 
#>   1  0.316 0.330 0.372 0.409 0.420 0.406 0.378 0.350 0.330 0.322 0.324 0.331
#>   2  0.564 0.544 0.528 0.527 0.542 0.563 0.582 0.583 0.557 0.527 0.523 0.543
#>   3  0.775 0.739 0.680 0.647 0.652 0.686 0.737 0.768 0.743 0.696 0.680 0.703
#>   4  0.721 0.728 0.719 0.708 0.705 0.716 0.741 0.758 0.744 0.715 0.697 0.699
#>   5  0.604 0.607 0.630 0.667 0.709 0.743 0.761 0.771 0.783 0.792 0.786 0.765
#>   6  0.539 0.514 0.519 0.565 0.645 0.722 0.759 0.770 0.781 0.797 0.815 0.828
#>   7  0.522 0.483 0.456 0.473 0.535 0.612 0.667 0.679 0.646 0.614 0.642 0.737
#>   8  0.486 0.465 0.438 0.433 0.456 0.506 0.569 0.602 0.560 0.509 0.531 0.645
#>   9  0.426 0.443 0.447 0.430 0.411 0.427 0.493 0.558 0.540 0.490 0.501 0.593
#>   10 0.426 0.443 0.447 0.430 0.411 0.427 0.493 0.558 0.540 0.490 0.501 0.593
#>     year
#> age  1993  1994  1995  1996  1997  1998  1999  2000  2001  2002  2003  2004 
#>   1  0.322 0.290 0.254 0.232 0.225 0.232 0.251 0.271 0.276 0.274 0.290 0.323
#>   2  0.547 0.511 0.482 0.510 0.575 0.576 0.482 0.419 0.458 0.568 0.625 0.559
#>   3  0.724 0.715 0.725 0.824 0.974 0.969 0.761 0.617 0.655 0.800 0.836 0.691
#>   4  0.711 0.726 0.753 0.805 0.868 0.891 0.850 0.785 0.736 0.696 0.646 0.581
#>   5  0.748 0.747 0.761 0.785 0.816 0.849 0.872 0.855 0.781 0.683 0.599 0.534
#>   6  0.821 0.791 0.769 0.787 0.829 0.837 0.786 0.735 0.727 0.722 0.647 0.519
#>   7  0.813 0.780 0.715 0.713 0.752 0.721 0.599 0.520 0.554 0.634 0.597 0.443
#>   8  0.751 0.716 0.632 0.616 0.648 0.619 0.506 0.423 0.419 0.446 0.416 0.327
#>   9  0.674 0.631 0.550 0.529 0.555 0.558 0.494 0.408 0.336 0.286 0.254 0.230
#>   10 0.674 0.631 0.550 0.529 0.555 0.558 0.494 0.408 0.336 0.286 0.254 0.230
#>     year
#> age  2005  2006  2007  2008  2009  2010  2011  2012  2013  2014  2015  2016 
#>   1  0.335 0.296 0.257 0.255 0.276 0.264 0.213 0.184 0.198 0.243 0.262 0.218
#>   2  0.482 0.452 0.436 0.394 0.338 0.298 0.282 0.282 0.292 0.304 0.304 0.289
#>   3  0.562 0.527 0.515 0.455 0.376 0.335 0.341 0.362 0.364 0.349 0.340 0.344
#>   4  0.512 0.446 0.395 0.364 0.350 0.344 0.342 0.342 0.343 0.343 0.341 0.336
#>   5  0.466 0.390 0.329 0.293 0.279 0.281 0.293 0.304 0.303 0.294 0.288 0.289
#>   6  0.413 0.350 0.306 0.266 0.235 0.224 0.233 0.246 0.246 0.237 0.231 0.236
#>   7  0.331 0.288 0.271 0.249 0.221 0.198 0.186 0.181 0.184 0.190 0.192 0.189
#>   8  0.256 0.221 0.203 0.189 0.176 0.164 0.153 0.148 0.149 0.154 0.155 0.150
#>   9  0.202 0.170 0.147 0.136 0.133 0.132 0.132 0.132 0.132 0.131 0.128 0.125
#>   10 0.202 0.170 0.147 0.136 0.133 0.132 0.132 0.132 0.132 0.131 0.128 0.125
#>     year
#> age  2017 
#>   1  0.169
#>   2  0.268
#>   3  0.355
#>   4  0.330
#>   5  0.295
#>   6  0.245
#>   7  0.182
#>   8  0.142
#>   9  0.121
#>   10 0.121
#> 
#> units:  z