Total mortality z

Returns the calculation of total mortality, z, usually as the sum of fishing mortality, f, and natural mortality, m.

z(object, ...)

# S4 method for FLS
z(object, ...)

Arguments

object
Object to calculate on.

...
Any extra arguments.

Value

An object of the corresponding class, usually FLQuant.

See also

Examples

data(ple4)

z(ple4)

#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>     year
#> age  1957    1958    1959    1960    1961    1962    1963    1964    1965   
#>   1  0.19870 0.21365 0.22783 0.23708 0.23562 0.21903 0.19365 0.17462 0.16825
#>   2  0.27086 0.31223 0.34887 0.36159 0.35486 0.35402 0.36995 0.37938 0.35853
#>   3  0.36151 0.41008 0.44974 0.45911 0.45168 0.46891 0.53097 0.58899 0.57128
#>   4  0.40882 0.42889 0.44462 0.45025 0.45073 0.45967 0.48326 0.50741 0.51404
#>   5  0.35628 0.39233 0.42428 0.44117 0.44569 0.45338 0.47277 0.49289 0.49961
#>   6  0.31094 0.34975 0.38669 0.41030 0.42299 0.44197 0.47568 0.50218 0.49506
#>   7  0.32540 0.32609 0.33037 0.34195 0.36164 0.38743 0.41368 0.42610 0.41259
#>   8  0.32773 0.32274 0.31858 0.31606 0.31856 0.33191 0.35659 0.37475 0.36611
#>   9  0.29867 0.33277 0.35097 0.33147 0.29744 0.28907 0.32035 0.36149 0.36326
#>   10 0.29867 0.33277 0.35097 0.33147 0.29744 0.28907 0.32035 0.36149 0.36326
#>     year
#> age  1966    1967    1968    1969    1970    1971    1972    1973    1974   
#>   1  0.17465 0.19774 0.24156 0.28898 0.30371 0.29494 0.29163 0.30513 0.33698
#>   2  0.33270 0.33829 0.38235 0.42813 0.43004 0.41262 0.41717 0.44766 0.47866
#>   3  0.51109 0.47683 0.48170 0.49480 0.49244 0.49309 0.52456 0.57933 0.60832
#>   4  0.50195 0.47953 0.45703 0.44543 0.45313 0.48365 0.54170 0.61160 0.64506
#>   5  0.49314 0.48121 0.46954 0.46101 0.45962 0.47548 0.52200 0.58629 0.61667
#>   6  0.47146 0.46447 0.47999 0.49206 0.48098 0.47026 0.49000 0.53293 0.55323
#>   7  0.39201 0.39173 0.41671 0.44342 0.44864 0.44500 0.45410 0.47491 0.48674
#>   8  0.34577 0.34040 0.35726 0.38316 0.40385 0.41948 0.43526 0.45092 0.46085
#>   9  0.33546 0.31658 0.31965 0.33819 0.36662 0.40016 0.43209 0.45645 0.46932
#>   10 0.33546 0.31658 0.31965 0.33819 0.36662 0.40016 0.43209 0.45645 0.46932
#>     year
#> age  1975    1976    1977    1978    1979    1980    1981    1982    1983   
#>   1  0.38748 0.44345 0.47857 0.47127 0.41715 0.35065 0.31639 0.32981 0.37186
#>   2  0.49117 0.49939 0.52161 0.55308 0.57531 0.57755 0.56445 0.54402 0.52806
#>   3  0.58315 0.55094 0.55543 0.59864 0.66808 0.74099 0.77471 0.73907 0.67963
#>   4  0.61613 0.57586 0.56993 0.59760 0.64224 0.68819 0.72080 0.72827 0.71929
#>   5  0.58613 0.55211 0.56409 0.60687 0.63252 0.62135 0.60423 0.60711 0.62975
#>   6  0.52797 0.50634 0.53187 0.58614 0.61096 0.58192 0.53865 0.51374 0.51874
#>   7  0.47907 0.47110 0.48188 0.50738 0.53163 0.53960 0.52207 0.48265 0.45565
#>   8  0.46145 0.45523 0.44631 0.44276 0.45407 0.47667 0.48613 0.46500 0.43790
#>   9  0.46813 0.45293 0.42688 0.40305 0.39522 0.40592 0.42559 0.44312 0.44735
#>   10 0.46813 0.45293 0.42688 0.40305 0.39522 0.40592 0.42559 0.44312 0.44735
#>     year
#> age  1984    1985    1986    1987    1988    1989    1990    1991    1992   
#>   1  0.40912 0.41985 0.40628 0.37835 0.34982 0.33022 0.32160 0.32426 0.33089
#>   2  0.52728 0.54164 0.56324 0.58183 0.58309 0.55748 0.52744 0.52285 0.54274
#>   3  0.64699 0.65196 0.68640 0.73690 0.76756 0.74311 0.69619 0.67994 0.70285
#>   4  0.70823 0.70501 0.71645 0.74140 0.75802 0.74450 0.71494 0.69670 0.69887
#>   5  0.66654 0.70877 0.74302 0.76071 0.77054 0.78272 0.79197 0.78589 0.76547
#>   6  0.56480 0.64520 0.72235 0.75939 0.77044 0.78093 0.79696 0.81509 0.82752
#>   7  0.47272 0.53470 0.61240 0.66720 0.67918 0.64582 0.61430 0.64230 0.73714
#>   8  0.43263 0.45624 0.50564 0.56911 0.60161 0.56049 0.50933 0.53139 0.64500
#>   9  0.42997 0.41099 0.42665 0.49347 0.55838 0.53998 0.48961 0.50078 0.59291
#>   10 0.42997 0.41099 0.42665 0.49347 0.55838 0.53998 0.48961 0.50078 0.59291
#>     year
#> age  1993    1994    1995    1996    1997    1998    1999    2000    2001   
#>   1  0.32241 0.28970 0.25411 0.23198 0.22502 0.23193 0.25120 0.27111 0.27573
#>   2  0.54722 0.51124 0.48238 0.51046 0.57523 0.57555 0.48164 0.41899 0.45783
#>   3  0.72399 0.71472 0.72487 0.82405 0.97402 0.96885 0.76099 0.61747 0.65524
#>   4  0.71108 0.72592 0.75264 0.80504 0.86753 0.89096 0.84959 0.78545 0.73628
#>   5  0.74811 0.74706 0.76089 0.78523 0.81601 0.84889 0.87166 0.85542 0.78122
#>   6  0.82085 0.79123 0.76910 0.78687 0.82920 0.83735 0.78630 0.73473 0.72746
#>   7  0.81309 0.77978 0.71487 0.71340 0.75239 0.72067 0.59888 0.52005 0.55361
#>   8  0.75055 0.71565 0.63227 0.61637 0.64798 0.61938 0.50630 0.42281 0.41883
#>   9  0.67366 0.63134 0.55041 0.52852 0.55526 0.55764 0.49418 0.40842 0.33641
#>   10 0.67366 0.63134 0.55041 0.52852 0.55526 0.55764 0.49418 0.40842 0.33641
#>     year
#> age  2002    2003    2004    2005    2006    2007    2008    2009    2010   
#>   1  0.27438 0.28953 0.32301 0.33492 0.29592 0.25698 0.25526 0.27566 0.26394
#>   2  0.56848 0.62481 0.55932 0.48172 0.45244 0.43646 0.39440 0.33803 0.29788
#>   3  0.80043 0.83642 0.69080 0.56215 0.52670 0.51465 0.45521 0.37627 0.33543
#>   4  0.69588 0.64566 0.58133 0.51163 0.44569 0.39461 0.36404 0.34975 0.34410
#>   5  0.68268 0.59948 0.53408 0.46556 0.39001 0.32891 0.29325 0.27912 0.28056
#>   6  0.72212 0.64750 0.51933 0.41323 0.35008 0.30648 0.26635 0.23549 0.22415
#>   7  0.63370 0.59666 0.44316 0.33101 0.28847 0.27119 0.24868 0.22057 0.19839
#>   8  0.44573 0.41566 0.32710 0.25617 0.22085 0.20254 0.18889 0.17602 0.16363
#>   9  0.28562 0.25370 0.23044 0.20232 0.16971 0.14662 0.13591 0.13259 0.13211
#>   10 0.28562 0.25370 0.23044 0.20232 0.16971 0.14662 0.13591 0.13259 0.13211
#>     year
#> age  2011    2012    2013    2014    2015    2016    2017   
#>   1  0.21287 0.18408 0.19808 0.24334 0.26224 0.21847 0.16858
#>   2  0.28164 0.28157 0.29207 0.30395 0.30423 0.28918 0.26823
#>   3  0.34138 0.36238 0.36396 0.34943 0.34013 0.34397 0.35514
#>   4  0.34229 0.34246 0.34317 0.34302 0.34065 0.33588 0.32999
#>   5  0.29270 0.30435 0.30325 0.29388 0.28778 0.28921 0.29475
#>   6  0.23275 0.24616 0.24612 0.23655 0.23142 0.23568 0.24544
#>   7  0.18571 0.18116 0.18393 0.18994 0.19249 0.18865 0.18186
#>   8  0.15313 0.14778 0.14940 0.15437 0.15546 0.14976 0.14171
#>   9  0.13223 0.13234 0.13205 0.13086 0.12835 0.12485 0.12127
#>   10 0.13223 0.13234 0.13205 0.13086 0.12835 0.12485 0.12127
#> 
#> units:  z